+++ CookieMonster.js
@@ -1,2 +1,2 @@
 /*! For license information please see CookieMonster.js.LICENSE.txt */
-(()=>{var e={877:function(e){!function(t,o){"use strict";"object"!=typeof e.exports?o(t):e.exports=t.document?o(t):function(e){if(!e.document)throw new Error("jscolor needs a window with document");return o(e)}}("undefined"!=typeof window?window:this,(function(e){"use strict";var t,o,n,i,a=((i={initialized:!1,instances:[],readyQueue:[],register:function(){void 0!==e&&e.document&&e.document.addEventListener("DOMContentLoaded",i.pub.init,!1)},installBySelector:function(t,o){if(!(o=o?i.node(o):e.document))throw new Error("Missing root node");for(var n=o.querySelectorAll(t),a=new RegExp("(^|\\s)("+i.pub.lookupClass+")(\\s*(\\{[^}]*\\})|\\s|$)","i"),r=0;r<n.length;r+=1){var l,s;if(!(n[r].jscolor&&n[r].jscolor instanceof i.pub)&&(void 0===n[r].type||"color"!=n[r].type.toLowerCase()||!i.isColorAttrSupported)&&(null!==(l=i.getDataAttr(n[r],"jscolor"))||n[r].className&&(s=n[r].className.match(a)))){var c=n[r],d="";null!==l?d=l:s&&(console.warn('Installation using class name is DEPRECATED. Use data-jscolor="" attribute instead.'+i.docsRef),s[4]&&(d=s[4]));var p=null;if(d.trim())try{p=i.parseOptionsStr(d)}catch(e){console.warn(e+"\n"+d)}try{new i.pub(c,p)}catch(e){console.warn(e)}}}},parseOptionsStr:function(e){var t=null;try{t=JSON.parse(e)}catch(o){if(!i.pub.looseJSON)throw new Error("Could not parse jscolor options as JSON: "+o);try{t=new Function("var opts = ("+e+'); return typeof opts === "object" ? opts : {};')()}catch(e){throw new Error("Could not evaluate jscolor options: "+e)}}return t},getInstances:function(){for(var e=[],t=0;t<i.instances.length;t+=1)i.instances[t]&&i.instances[t].targetElement&&e.push(i.instances[t]);return e},createEl:function(t){var o=e.document.createElement(t);return i.setData(o,"gui",!0),o},node:function(t){if(!t)return null;if("string"==typeof t){var o=t,n=null;try{n=e.document.querySelector(o)}catch(e){return console.warn(e),null}return n||console.warn("No element matches the selector: %s",o),n}return i.isNode(t)?t:(console.warn("Invalid node of type %s: %s",typeof t,t),null)},isNode:function(e){return"object"==typeof Node?e instanceof Node:e&&"object"==typeof e&&"number"==typeof e.nodeType&&"string"==typeof e.nodeName},nodeName:function(e){return!(!e||!e.nodeName)&&e.nodeName.toLowerCase()},removeChildren:function(e){for(;e.firstChild;)e.removeChild(e.firstChild)},isTextInput:function(e){return e&&"input"===i.nodeName(e)&&"text"===e.type.toLowerCase()},isButton:function(e){if(!e)return!1;var t=i.nodeName(e);return"button"===t||"input"===t&&["button","submit","reset"].indexOf(e.type.toLowerCase())>-1},isButtonEmpty:function(e){switch(i.nodeName(e)){case"input":return!e.value||""===e.value.trim();case"button":return""===e.textContent.trim()}return null},isPassiveEventSupported:function(){var t=!1;try{var o=Object.defineProperty({},"passive",{get:function(){t=!0}});e.addEventListener("testPassive",null,o),e.removeEventListener("testPassive",null,o)}catch(e){}return t}(),isColorAttrSupported:(n=e.document.createElement("input"),!(!n.setAttribute||(n.setAttribute("type","color"),"color"!=n.type.toLowerCase()))),dataProp:"_data_jscolor",setData:function(){var e=arguments[0];if(3===arguments.length){var t=e.hasOwnProperty(i.dataProp)?e[i.dataProp]:e[i.dataProp]={},o=arguments[2];return t[a=arguments[1]]=o,!0}if(2===arguments.length&&"object"==typeof arguments[1]){t=e.hasOwnProperty(i.dataProp)?e[i.dataProp]:e[i.dataProp]={};var n=arguments[1];for(var a in n)n.hasOwnProperty(a)&&(t[a]=n[a]);return!0}throw new Error("Invalid arguments")},removeData:function(){var e=arguments[0];if(!e.hasOwnProperty(i.dataProp))return!0;for(var t=1;t<arguments.length;t+=1){var o=arguments[t];delete e[i.dataProp][o]}return!0},getData:function(e,t,o){if(!e.hasOwnProperty(i.dataProp)){if(void 0===o)return;e[i.dataProp]={}}var n=e[i.dataProp];return n.hasOwnProperty(t)||void 0===o||(n[t]=o),n[t]},getDataAttr:function(e,t){var o="data-"+t;return e.getAttribute(o)},setDataAttr:function(e,t,o){var n="data-"+t;e.setAttribute(n,o)},_attachedGroupEvents:{},attachGroupEvent:function(e,t,o,n){i._attachedGroupEvents.hasOwnProperty(e)||(i._attachedGroupEvents[e]=[]),i._attachedGroupEvents[e].push([t,o,n]),t.addEventListener(o,n,!1)},detachGroupEvents:function(e){if(i._attachedGroupEvents.hasOwnProperty(e)){for(var t=0;t<i._attachedGroupEvents[e].length;t+=1){var o=i._attachedGroupEvents[e][t];o[0].removeEventListener(o[1],o[2],!1)}delete i._attachedGroupEvents[e]}},preventDefault:function(e){e.preventDefault&&e.preventDefault(),e.returnValue=!1},captureTarget:function(e){e.setCapture&&(i._capturedTarget=e,i._capturedTarget.setCapture())},releaseTarget:function(){i._capturedTarget&&(i._capturedTarget.releaseCapture(),i._capturedTarget=null)},triggerEvent:function(t,o,n,a){if(t){var r=null;return"function"==typeof Event?r=new Event(o,{bubbles:n,cancelable:a}):(r=e.document.createEvent("Event")).initEvent(o,n,a),!!r&&(i.setData(r,"internal",!0),t.dispatchEvent(r),!0)}},triggerInputEvent:function(e,t,o,n){e&&i.isTextInput(e)&&i.triggerEvent(e,t,o,n)},eventKey:function(e){var t={9:"Tab",13:"Enter",27:"Escape"};return"string"==typeof e.code?e.code:void 0!==e.keyCode&&t.hasOwnProperty(e.keyCode)?t[e.keyCode]:null},strList:function(e){return e?e.replace(/^\s+|\s+$/g,"").split(/\s+/):[]},hasClass:function(e,t){return!!t&&(void 0!==e.classList?e.classList.contains(t):-1!=(" "+e.className.replace(/\s+/g," ")+" ").indexOf(" "+t+" "))},addClass:function(e,t){var o=i.strList(t);if(void 0===e.classList)for(n=0;n<o.length;n+=1)i.hasClass(e,o[n])||(e.className+=(e.className?" ":"")+o[n]);else for(var n=0;n<o.length;n+=1)e.classList.add(o[n])},removeClass:function(e,t){var o=i.strList(t);if(void 0===e.classList)for(a=0;a<o.length;a+=1){var n=new RegExp("^\\s*"+o[a]+"\\s*|\\s*"+o[a]+"\\s*$|\\s+"+o[a]+"(\\s+)","g");e.className=e.className.replace(n,"$1")}else for(var a=0;a<o.length;a+=1)e.classList.remove(o[a])},getCompStyle:function(t){return(e.getComputedStyle?e.getComputedStyle(t):t.currentStyle)||{}},setStyle:function(e,t,o,n){var a=o?"important":"",r=null;for(var l in t)if(t.hasOwnProperty(l)){var s=null;null===t[l]?(r||(r=i.getData(e,"origStyle")),r&&r.hasOwnProperty(l)&&(s=r[l])):(n&&(r||(r=i.getData(e,"origStyle",{})),r.hasOwnProperty(l)||(r[l]=e.style[l])),s=t[l]),null!==s&&e.style.setProperty(l,s,a)}},hexColor:function(e,t,o){return"#"+(("0"+Math.round(e).toString(16)).substr(-2)+("0"+Math.round(t).toString(16)).substr(-2)+("0"+Math.round(o).toString(16)).substr(-2)).toUpperCase()},hexaColor:function(e,t,o,n){return"#"+(("0"+Math.round(e).toString(16)).substr(-2)+("0"+Math.round(t).toString(16)).substr(-2)+("0"+Math.round(o).toString(16)).substr(-2)+("0"+Math.round(255*n).toString(16)).substr(-2)).toUpperCase()},rgbColor:function(e,t,o){return"rgb("+Math.round(e)+","+Math.round(t)+","+Math.round(o)+")"},rgbaColor:function(e,t,o,n){return"rgba("+Math.round(e)+","+Math.round(t)+","+Math.round(o)+","+Math.round(100*(null==n?1:n))/100+")"},linearGradient:(o=function(){for(var t="linear-gradient",o=["","-webkit-","-moz-","-o-","-ms-"],n=e.document.createElement("div"),i=0;i<o.length;i+=1){var a=o[i]+t,r=a+"(to right, rgba(0,0,0,0), rgba(0,0,0,0))";if(n.style.background=r,n.style.background)return a}return t}(),function(){return o+"("+Array.prototype.join.call(arguments,", ")+")"}),setBorderRadius:function(e,t){i.setStyle(e,{"border-radius":t||"0"})},setBoxShadow:function(e,t){i.setStyle(e,{"box-shadow":t||"none"})},getElementPos:function(e,t){var o=0,n=0,a=e.getBoundingClientRect();if(o=a.left,n=a.top,!t){var r=i.getViewPos();o+=r[0],n+=r[1]}return[o,n]},getElementSize:function(e){return[e.offsetWidth,e.offsetHeight]},getAbsPointerPos:function(e){var t=0,o=0;return void 0!==e.changedTouches&&e.changedTouches.length?(t=e.changedTouches[0].clientX,o=e.changedTouches[0].clientY):"number"==typeof e.clientX&&(t=e.clientX,o=e.clientY),{x:t,y:o}},getRelPointerPos:function(e){var t=(e.target||e.srcElement).getBoundingClientRect(),o=0,n=0;return void 0!==e.changedTouches&&e.changedTouches.length?(o=e.changedTouches[0].clientX,n=e.changedTouches[0].clientY):"number"==typeof e.clientX&&(o=e.clientX,n=e.clientY),{x:o-t.left,y:n-t.top}},getViewPos:function(){var t=e.document.documentElement;return[(e.pageXOffset||t.scrollLeft)-(t.clientLeft||0),(e.pageYOffset||t.scrollTop)-(t.clientTop||0)]},getViewSize:function(){var t=e.document.documentElement;return[e.innerWidth||t.clientWidth,e.innerHeight||t.clientHeight]},RGB_HSV:function(e,t,o){e/=255,t/=255,o/=255;var n=Math.min(Math.min(e,t),o),i=Math.max(Math.max(e,t),o),a=i-n;if(0===a)return[null,0,100*i];var r=e===n?3+(o-t)/a:t===n?5+(e-o)/a:1+(t-e)/a;return[60*(6===r?0:r),a/i*100,100*i]},HSV_RGB:function(e,t,o){var n=o/100*255;if(null===e)return[n,n,n];e/=60,t/=100;var i=Math.floor(e),a=n*(1-t),r=n*(1-t*(i%2?e-i:1-(e-i)));switch(i){case 6:case 0:return[n,r,a];case 1:return[r,n,a];case 2:return[a,n,r];case 3:return[a,r,n];case 4:return[r,a,n];case 5:return[n,a,r]}},parseColorString:function(e){var t,o={rgba:null,format:null};if(t=e.match(/^\W*([0-9A-F]{3,8})\W*$/i)){if(8===t[1].length)o.format="hexa",o.rgba=[parseInt(t[1].substr(0,2),16),parseInt(t[1].substr(2,2),16),parseInt(t[1].substr(4,2),16),parseInt(t[1].substr(6,2),16)/255];else if(6===t[1].length)o.format="hex",o.rgba=[parseInt(t[1].substr(0,2),16),parseInt(t[1].substr(2,2),16),parseInt(t[1].substr(4,2),16),null];else{if(3!==t[1].length)return!1;o.format="hex",o.rgba=[parseInt(t[1].charAt(0)+t[1].charAt(0),16),parseInt(t[1].charAt(1)+t[1].charAt(1),16),parseInt(t[1].charAt(2)+t[1].charAt(2),16),null]}return o}if(t=e.match(/^\W*rgba?\(([^)]*)\)\W*$/i)){var n,i,a,r,l=t[1].split(","),s=/^\s*(\d+|\d*\.\d+|\d+\.\d*)\s*$/;if(l.length>=3&&(n=l[0].match(s))&&(i=l[1].match(s))&&(a=l[2].match(s)))return o.format="rgb",o.rgba=[parseFloat(n[1])||0,parseFloat(i[1])||0,parseFloat(a[1])||0,null],l.length>=4&&(r=l[3].match(s))&&(o.format="rgba",o.rgba[3]=parseFloat(r[1])||0),o}return!1},parsePaletteValue:function(e){var t=[];"string"==typeof e?e.replace(/#[0-9A-F]{3}([0-9A-F]{3})?|rgba?\(([^)]*)\)/gi,(function(e){t.push(e)})):Array.isArray(e)&&(t=e);for(var o=[],n=0;n<t.length;n++){var a=i.parseColorString(t[n]);a&&o.push(a)}return o},containsTranparentColor:function(e){for(var t=0;t<e.length;t++){var o=e[t].rgba[3];if(null!==o&&o<1)return!0}return!1},isAlphaFormat:function(e){switch(e.toLowerCase()){case"hexa":case"rgba":return!0}return!1},scaleCanvasForHighDPR:function(t){var o=e.devicePixelRatio||1;t.width*=o,t.height*=o,t.getContext("2d").scale(o,o)},genColorPreviewCanvas:function(e,t,o,n){var a=Math.round(i.pub.previewSeparator.length),r=i.pub.chessboardSize,l=i.pub.chessboardColor1,s=i.pub.chessboardColor2,c=o||2*r,d=2*r,p=i.createEl("canvas"),u=p.getContext("2d");p.width=c,p.height=d,n&&i.scaleCanvasForHighDPR(p),u.fillStyle=l,u.fillRect(0,0,c,d),u.fillStyle=s;for(var m=0;m<c;m+=2*r)u.fillRect(m,0,r,r),u.fillRect(m+r,r,r,r);e&&(u.fillStyle=e,u.fillRect(0,0,c,d));var h=null;switch(t){case"left":h=0,u.clearRect(0,0,a/2,d);break;case"right":h=c-a,u.clearRect(c-a/2,0,a/2,d)}if(null!==h){u.lineWidth=1;for(var g=0;g<i.pub.previewSeparator.length;g+=1)u.beginPath(),u.strokeStyle=i.pub.previewSeparator[g],u.moveTo(.5+h+g,0),u.lineTo(.5+h+g,d),u.stroke()}return{canvas:p,width:c,height:d}},genColorPreviewGradient:function(e,t,o){var n;return n=t&&o?["to "+{left:"right",right:"left"}[t],e+" 0%",e+" "+o+"px","rgba(0,0,0,0) "+(o+1)+"px","rgba(0,0,0,0) 100%"]:["to right",e+" 0%",e+" 100%"],i.linearGradient.apply(this,n)},redrawPosition:function(){if(i.picker&&i.picker.owner){var e,t,o=i.picker.owner;o.fixed?(e=i.getElementPos(o.targetElement,!0),t=[0,0]):(e=i.getElementPos(o.targetElement),t=i.getViewPos());var n,a,r,l=i.getElementSize(o.targetElement),s=i.getViewSize(),c=i.getPickerDims(o),d=[c.borderW,c.borderH];switch(o.position.toLowerCase()){case"left":n=1,a=0,r=-1;break;case"right":n=1,a=0,r=1;break;case"top":n=0,a=1,r=-1;break;default:n=0,a=1,r=1}var p=(l[a]+d[a])/2;if(o.smartPosition)u=[-t[n]+e[n]+d[n]>s[n]&&-t[n]+e[n]+l[n]/2>s[n]/2&&e[n]+l[n]-d[n]>=0?e[n]+l[n]-d[n]:e[n],-t[a]+e[a]+l[a]+d[a]-p+p*r>s[a]?-t[a]+e[a]+l[a]/2>s[a]/2&&e[a]+l[a]-p-p*r>=0?e[a]+l[a]-p-p*r:e[a]+l[a]-p+p*r:e[a]+l[a]-p+p*r>=0?e[a]+l[a]-p+p*r:e[a]+l[a]-p-p*r];else var u=[e[n],e[a]+l[a]-p+p*r];var m=u[n],h=u[a],g=o.fixed?"fixed":"absolute",f=(u[0]+d[0]>e[0]||u[0]<e[0]+l[0])&&u[1]+d[1]<e[1]+l[1];i._drawPosition(o,m,h,g,f)}},_drawPosition:function(e,t,o,n,a){var r=a?0:e.shadowBlur;i.picker.wrap.style.position=n,i.picker.wrap.style.left=t+"px",i.picker.wrap.style.top=o+"px",i.setBoxShadow(i.picker.boxS,e.shadow?new i.BoxShadow(0,r,e.shadowBlur,0,e.shadowColor):null)},getPickerDims:function(e){var t=2*e.controlBorderWidth+e.width,o=2*e.controlBorderWidth+e.height,n=2*e.controlBorderWidth+2*i.getControlPadding(e)+e.sliderSize;i.getSliderChannel(e)&&(t+=n),e.hasAlphaChannel()&&(t+=n);var a=i.getPaletteDims(e,t);a.height&&(o+=a.height+e.padding),e.closeButton&&(o+=2*e.controlBorderWidth+e.padding+e.buttonHeight);var r=t+2*e.padding,l=o+2*e.padding;return{contentW:t,contentH:o,paddedW:r,paddedH:l,borderW:r+2*e.borderWidth,borderH:l+2*e.borderWidth,palette:a}},getPaletteDims:function(e,t){var o=0,n=0,i=0,a=0,r=0,l=e._palette?e._palette.length:0;return l&&(n=(o=e.paletteCols)>0?Math.ceil(l/o):0,i=Math.max(1,Math.floor((t-(o-1)*e.paletteSpacing)/o)),a=e.paletteHeight?Math.min(e.paletteHeight,i):i),n&&(r=n*a+(n-1)*e.paletteSpacing),{cols:o,rows:n,cellW:i,cellH:a,width:t,height:r}},getControlPadding:function(e){return Math.max(e.padding/2,2*e.pointerBorderWidth+e.pointerThickness-e.controlBorderWidth)},getPadYChannel:function(e){switch(e.mode.charAt(1).toLowerCase()){case"v":return"v"}return"s"},getSliderChannel:function(e){if(e.mode.length>2)switch(e.mode.charAt(2).toLowerCase()){case"s":return"s";case"v":return"v"}return null},triggerCallback:function(e,t){if(e[t]){var o=null;if("string"==typeof e[t])try{o=new Function(e[t])}catch(e){console.error(e)}else o=e[t];o&&o.call(e)}},triggerGlobal:function(e){for(var t=i.getInstances(),o=0;o<t.length;o+=1)t[o].trigger(e)},_pointerMoveEvent:{mouse:"mousemove",touch:"touchmove"},_pointerEndEvent:{mouse:"mouseup",touch:"touchend"},_pointerOrigin:null,_capturedTarget:null,onDocumentKeyUp:function(e){-1!==["Tab","Escape"].indexOf(i.eventKey(e))&&i.picker&&i.picker.owner&&i.picker.owner.tryHide()},onWindowResize:function(e){i.redrawPosition()},onWindowScroll:function(e){i.redrawPosition()},onParentScroll:function(e){i.picker&&i.picker.owner&&i.picker.owner.tryHide()},onDocumentMouseDown:function(e){var t=e.target||e.srcElement;t.jscolor&&t.jscolor instanceof i.pub?t.jscolor.showOnClick&&!t.disabled&&t.jscolor.show():i.getData(t,"gui")?i.getData(t,"control")&&i.onControlPointerStart(e,t,i.getData(t,"control"),"mouse"):i.picker&&i.picker.owner&&i.picker.owner.tryHide()},onPickerTouchStart:function(e){var t=e.target||e.srcElement;i.getData(t,"control")&&i.onControlPointerStart(e,t,i.getData(t,"control"),"touch")},onControlPointerStart:function(t,o,n,a){var r=i.getData(o,"instance");i.preventDefault(t),i.captureTarget(o);var l=function(e,r){i.attachGroupEvent("drag",e,i._pointerMoveEvent[a],i.onDocumentPointerMove(t,o,n,a,r)),i.attachGroupEvent("drag",e,i._pointerEndEvent[a],i.onDocumentPointerEnd(t,o,n,a))};if(l(e.document,[0,0]),e.parent&&e.frameElement){var s=e.frameElement.getBoundingClientRect(),c=[-s.left,-s.top];l(e.parent.window.document,c)}var d=i.getAbsPointerPos(t),p=i.getRelPointerPos(t);switch(i._pointerOrigin={x:d.x-p.x,y:d.y-p.y},n){case"pad":"v"===i.getSliderChannel(r)&&0===r.channels.v&&r.fromHSVA(null,null,100,null),i.setPad(r,t,0,0);break;case"sld":i.setSld(r,t,0);break;case"asld":i.setASld(r,t,0)}r.trigger("input")},onDocumentPointerMove:function(e,t,o,n,a){return function(e){var n=i.getData(t,"instance");switch(o){case"pad":i.setPad(n,e,a[0],a[1]);break;case"sld":i.setSld(n,e,a[1]);break;case"asld":i.setASld(n,e,a[1])}n.trigger("input")}},onDocumentPointerEnd:function(e,t,o,n){return function(e){var o=i.getData(t,"instance");i.detachGroupEvents("drag"),i.releaseTarget(),o.trigger("input"),o.trigger("change")}},onPaletteSampleClick:function(e){var t=e.currentTarget,o=i.getData(t,"instance"),n=i.getData(t,"color");"any"===o.format.toLowerCase()&&(o._setFormat(n.format),i.isAlphaFormat(o.getFormat())||(n.rgba[3]=1)),null===n.rgba[3]&&(!0===o.paletteSetsAlpha||"auto"===o.paletteSetsAlpha&&o._paletteHasTransparency)&&(n.rgba[3]=1),o.fromRGBA.apply(o,n.rgba),o.trigger("input"),o.trigger("change"),o.hideOnPaletteClick&&o.hide()},setPad:function(e,t,o,n){var a=i.getAbsPointerPos(t),r=o+a.x-i._pointerOrigin.x-e.padding-e.controlBorderWidth,l=n+a.y-i._pointerOrigin.y-e.padding-e.controlBorderWidth,s=r*(360/(e.width-1)),c=100-l*(100/(e.height-1));switch(i.getPadYChannel(e)){case"s":e.fromHSVA(s,c,null,null);break;case"v":e.fromHSVA(s,null,c,null)}},setSld:function(e,t,o){var n=100-(o+i.getAbsPointerPos(t).y-i._pointerOrigin.y-e.padding-e.controlBorderWidth)*(100/(e.height-1));switch(i.getSliderChannel(e)){case"s":e.fromHSVA(null,n,null,null);break;case"v":e.fromHSVA(null,null,n,null)}},setASld:function(e,t,o){var n=1-(o+i.getAbsPointerPos(t).y-i._pointerOrigin.y-e.padding-e.controlBorderWidth)*(1/(e.height-1));if(n<1){var a=e.getFormat();"any"!==e.format.toLowerCase()||i.isAlphaFormat(a)||e._setFormat("hex"===a?"hexa":"rgba")}e.fromHSVA(null,null,null,n)},createPadCanvas:function(){var e={elm:null,draw:null},t=i.createEl("canvas"),o=t.getContext("2d");return e.elm=t,e.draw=function(e,n,i){t.width=e,t.height=n,o.clearRect(0,0,t.width,t.height);var a=o.createLinearGradient(0,0,t.width,0);a.addColorStop(0,"#F00"),a.addColorStop(1/6,"#FF0"),a.addColorStop(2/6,"#0F0"),a.addColorStop(.5,"#0FF"),a.addColorStop(4/6,"#00F"),a.addColorStop(5/6,"#F0F"),a.addColorStop(1,"#F00"),o.fillStyle=a,o.fillRect(0,0,t.width,t.height);var r=o.createLinearGradient(0,0,0,t.height);switch(i.toLowerCase()){case"s":r.addColorStop(0,"rgba(255,255,255,0)"),r.addColorStop(1,"rgba(255,255,255,1)");break;case"v":r.addColorStop(0,"rgba(0,0,0,0)"),r.addColorStop(1,"rgba(0,0,0,1)")}o.fillStyle=r,o.fillRect(0,0,t.width,t.height)},e},createSliderGradient:function(){var e={elm:null,draw:null},t=i.createEl("canvas"),o=t.getContext("2d");return e.elm=t,e.draw=function(e,n,i,a){t.width=e,t.height=n,o.clearRect(0,0,t.width,t.height);var r=o.createLinearGradient(0,0,0,t.height);r.addColorStop(0,i),r.addColorStop(1,a),o.fillStyle=r,o.fillRect(0,0,t.width,t.height)},e},createASliderGradient:function(){var e={elm:null,draw:null},t=i.createEl("canvas"),o=t.getContext("2d");return e.elm=t,e.draw=function(e,n,a){t.width=e,t.height=n,o.clearRect(0,0,t.width,t.height);var r=t.width/2,l=i.pub.chessboardColor1,s=i.pub.chessboardColor2;if(o.fillStyle=l,o.fillRect(0,0,t.width,t.height),r>0)for(var c=0;c<t.height;c+=2*r)o.fillStyle=s,o.fillRect(0,c,r,r),o.fillRect(r,c+r,r,r);var d=o.createLinearGradient(0,0,0,t.height);d.addColorStop(0,a),d.addColorStop(1,"rgba(0,0,0,0)"),o.fillStyle=d,o.fillRect(0,0,t.width,t.height)},e},BoxShadow:(t=function(e,t,o,n,i,a){this.hShadow=e,this.vShadow=t,this.blur=o,this.spread=n,this.color=i,this.inset=!!a},t.prototype.toString=function(){var e=[Math.round(this.hShadow)+"px",Math.round(this.vShadow)+"px",Math.round(this.blur)+"px",Math.round(this.spread)+"px",this.color];return this.inset&&e.push("inset"),e.join(" ")},t),flags:{leaveValue:1,leaveAlpha:2,leavePreview:4},enumOpts:{format:["auto","any","hex","hexa","rgb","rgba"],previewPosition:["left","right"],mode:["hsv","hvs","hs","hv"],position:["left","right","top","bottom"],alphaChannel:["auto",!0,!1],paletteSetsAlpha:["auto",!0,!1]},deprecatedOpts:{styleElement:"previewElement",onFineChange:"onInput",overwriteImportant:"forceStyle",closable:"closeButton",insetWidth:"controlBorderWidth",insetColor:"controlBorderColor",refine:null},docsRef:" See https://jscolor.com/docs/",pub:function(t,o){var n=this;function a(e,t){if("string"!=typeof e)throw new Error("Invalid value for option name: "+e);if(i.enumOpts.hasOwnProperty(e)&&("string"==typeof t&&(t=t.toLowerCase()),-1===i.enumOpts[e].indexOf(t)))throw new Error("Option '"+e+"' has invalid value: "+t);if(i.deprecatedOpts.hasOwnProperty(e)){var o=e,a=i.deprecatedOpts[e];if(!a)throw new Error("Option '"+e+"' is DEPRECATED");console.warn("Option '%s' is DEPRECATED, using '%s' instead."+i.docsRef,o,a),e=a}var r="set__"+e;if("function"==typeof n[r])return n[r](t),!0;if(e in n)return n[e]=t,!0;throw new Error("Unrecognized configuration option: "+e)}function r(e){if("string"!=typeof e)throw new Error("Invalid value for option name: "+e);if(i.deprecatedOpts.hasOwnProperty(e)){var t=e,o=i.deprecatedOpts[e];if(!o)throw new Error("Option '"+e+"' is DEPRECATED");console.warn("Option '%s' is DEPRECATED, using '%s' instead."+i.docsRef,t,o),e=o}var a="get__"+e;if("function"==typeof n[a])return n[a](value);if(e in n)return n[e];throw new Error("Unrecognized configuration option: "+e)}function l(){n._processParentElementsInDOM(),i.picker||(i.picker={owner:null,wrap:i.createEl("div"),box:i.createEl("div"),boxS:i.createEl("div"),boxB:i.createEl("div"),pad:i.createEl("div"),padB:i.createEl("div"),padM:i.createEl("div"),padCanvas:i.createPadCanvas(),cross:i.createEl("div"),crossBY:i.createEl("div"),crossBX:i.createEl("div"),crossLY:i.createEl("div"),crossLX:i.createEl("div"),sld:i.createEl("div"),sldB:i.createEl("div"),sldM:i.createEl("div"),sldGrad:i.createSliderGradient(),sldPtrS:i.createEl("div"),sldPtrIB:i.createEl("div"),sldPtrMB:i.createEl("div"),sldPtrOB:i.createEl("div"),asld:i.createEl("div"),asldB:i.createEl("div"),asldM:i.createEl("div"),asldGrad:i.createASliderGradient(),asldPtrS:i.createEl("div"),asldPtrIB:i.createEl("div"),asldPtrMB:i.createEl("div"),asldPtrOB:i.createEl("div"),pal:i.createEl("div"),btn:i.createEl("div"),btnT:i.createEl("span")},i.picker.pad.appendChild(i.picker.padCanvas.elm),i.picker.padB.appendChild(i.picker.pad),i.picker.cross.appendChild(i.picker.crossBY),i.picker.cross.appendChild(i.picker.crossBX),i.picker.cross.appendChild(i.picker.crossLY),i.picker.cross.appendChild(i.picker.crossLX),i.picker.padB.appendChild(i.picker.cross),i.picker.box.appendChild(i.picker.padB),i.picker.box.appendChild(i.picker.padM),i.picker.sld.appendChild(i.picker.sldGrad.elm),i.picker.sldB.appendChild(i.picker.sld),i.picker.sldB.appendChild(i.picker.sldPtrOB),i.picker.sldPtrOB.appendChild(i.picker.sldPtrMB),i.picker.sldPtrMB.appendChild(i.picker.sldPtrIB),i.picker.sldPtrIB.appendChild(i.picker.sldPtrS),i.picker.box.appendChild(i.picker.sldB),i.picker.box.appendChild(i.picker.sldM),i.picker.asld.appendChild(i.picker.asldGrad.elm),i.picker.asldB.appendChild(i.picker.asld),i.picker.asldB.appendChild(i.picker.asldPtrOB),i.picker.asldPtrOB.appendChild(i.picker.asldPtrMB),i.picker.asldPtrMB.appendChild(i.picker.asldPtrIB),i.picker.asldPtrIB.appendChild(i.picker.asldPtrS),i.picker.box.appendChild(i.picker.asldB),i.picker.box.appendChild(i.picker.asldM),i.picker.box.appendChild(i.picker.pal),i.picker.btn.appendChild(i.picker.btnT),i.picker.box.appendChild(i.picker.btn),i.picker.boxB.appendChild(i.picker.box),i.picker.wrap.appendChild(i.picker.boxS),i.picker.wrap.appendChild(i.picker.boxB),i.picker.wrap.addEventListener("touchstart",i.onPickerTouchStart,!!i.isPassiveEventSupported&&{passive:!1}));var t,o,a=i.picker,r=!!i.getSliderChannel(n),l=n.hasAlphaChannel(),p=i.getPickerDims(n),u=2*n.pointerBorderWidth+n.pointerThickness+2*n.crossSize,m=i.getControlPadding(n),h=Math.min(n.borderRadius,Math.round(n.padding*Math.PI));a.wrap.className="jscolor-picker-wrap",a.wrap.style.clear="both",a.wrap.style.width=p.borderW+"px",a.wrap.style.height=p.borderH+"px",a.wrap.style.zIndex=n.zIndex,a.box.className="jscolor-picker",a.box.style.width=p.paddedW+"px",a.box.style.height=p.paddedH+"px",a.box.style.position="relative",a.boxS.className="jscolor-picker-shadow",a.boxS.style.position="absolute",a.boxS.style.left="0",a.boxS.style.top="0",a.boxS.style.width="100%",a.boxS.style.height="100%",i.setBorderRadius(a.boxS,h+"px"),a.boxB.className="jscolor-picker-border",a.boxB.style.position="relative",a.boxB.style.border=n.borderWidth+"px solid",a.boxB.style.borderColor=n.borderColor,a.boxB.style.background=n.backgroundColor,i.setBorderRadius(a.boxB,h+"px"),a.padM.style.background="rgba(255,0,0,.2)",a.sldM.style.background="rgba(0,255,0,.2)",a.asldM.style.background="rgba(0,0,255,.2)",a.padM.style.opacity=a.sldM.style.opacity=a.asldM.style.opacity="0",a.pad.style.position="relative",a.pad.style.width=n.width+"px",a.pad.style.height=n.height+"px",a.padCanvas.draw(n.width,n.height,i.getPadYChannel(n)),a.padB.style.position="absolute",a.padB.style.left=n.padding+"px",a.padB.style.top=n.padding+"px",a.padB.style.border=n.controlBorderWidth+"px solid",a.padB.style.borderColor=n.controlBorderColor,a.padM.style.position="absolute",a.padM.style.left="0px",a.padM.style.top="0px",a.padM.style.width=n.padding+2*n.controlBorderWidth+n.width+m+"px",a.padM.style.height=2*n.controlBorderWidth+2*n.padding+n.height+"px",a.padM.style.cursor="crosshair",i.setData(a.padM,{instance:n,control:"pad"}),a.cross.style.position="absolute",a.cross.style.left=a.cross.style.top="0",a.cross.style.width=a.cross.style.height=u+"px",a.crossBY.style.position=a.crossBX.style.position="absolute",a.crossBY.style.background=a.crossBX.style.background=n.pointerBorderColor,a.crossBY.style.width=a.crossBX.style.height=2*n.pointerBorderWidth+n.pointerThickness+"px",a.crossBY.style.height=a.crossBX.style.width=u+"px",a.crossBY.style.left=a.crossBX.style.top=Math.floor(u/2)-Math.floor(n.pointerThickness/2)-n.pointerBorderWidth+"px",a.crossBY.style.top=a.crossBX.style.left="0",a.crossLY.style.position=a.crossLX.style.position="absolute",a.crossLY.style.background=a.crossLX.style.background=n.pointerColor,a.crossLY.style.height=a.crossLX.style.width=u-2*n.pointerBorderWidth+"px",a.crossLY.style.width=a.crossLX.style.height=n.pointerThickness+"px",a.crossLY.style.left=a.crossLX.style.top=Math.floor(u/2)-Math.floor(n.pointerThickness/2)+"px",a.crossLY.style.top=a.crossLX.style.left=n.pointerBorderWidth+"px",a.sld.style.overflow="hidden",a.sld.style.width=n.sliderSize+"px",a.sld.style.height=n.height+"px",a.sldGrad.draw(n.sliderSize,n.height,"#000","#000"),a.sldB.style.display=r?"block":"none",a.sldB.style.position="absolute",a.sldB.style.left=n.padding+n.width+2*n.controlBorderWidth+2*m+"px",a.sldB.style.top=n.padding+"px",a.sldB.style.border=n.controlBorderWidth+"px solid",a.sldB.style.borderColor=n.controlBorderColor,a.sldM.style.display=r?"block":"none",a.sldM.style.position="absolute",a.sldM.style.left=n.padding+n.width+2*n.controlBorderWidth+m+"px",a.sldM.style.top="0px",a.sldM.style.width=n.sliderSize+2*m+2*n.controlBorderWidth+(l?0:Math.max(0,n.padding-m))+"px",a.sldM.style.height=2*n.controlBorderWidth+2*n.padding+n.height+"px",a.sldM.style.cursor="default",i.setData(a.sldM,{instance:n,control:"sld"}),a.sldPtrIB.style.border=a.sldPtrOB.style.border=n.pointerBorderWidth+"px solid "+n.pointerBorderColor,a.sldPtrOB.style.position="absolute",a.sldPtrOB.style.left=-(2*n.pointerBorderWidth+n.pointerThickness)+"px",a.sldPtrOB.style.top="0",a.sldPtrMB.style.border=n.pointerThickness+"px solid "+n.pointerColor,a.sldPtrS.style.width=n.sliderSize+"px",a.sldPtrS.style.height=i.pub.sliderInnerSpace+"px",a.asld.style.overflow="hidden",a.asld.style.width=n.sliderSize+"px",a.asld.style.height=n.height+"px",a.asldGrad.draw(n.sliderSize,n.height,"#000"),a.asldB.style.display=l?"block":"none",a.asldB.style.position="absolute",a.asldB.style.left=n.padding+n.width+2*n.controlBorderWidth+m+(r?n.sliderSize+3*m+2*n.controlBorderWidth:0)+"px",a.asldB.style.top=n.padding+"px",a.asldB.style.border=n.controlBorderWidth+"px solid",a.asldB.style.borderColor=n.controlBorderColor,a.asldM.style.display=l?"block":"none",a.asldM.style.position="absolute",a.asldM.style.left=n.padding+n.width+2*n.controlBorderWidth+m+(r?n.sliderSize+2*m+2*n.controlBorderWidth:0)+"px",a.asldM.style.top="0px",a.asldM.style.width=n.sliderSize+2*m+2*n.controlBorderWidth+Math.max(0,n.padding-m)+"px",a.asldM.style.height=2*n.controlBorderWidth+2*n.padding+n.height+"px",a.asldM.style.cursor="default",i.setData(a.asldM,{instance:n,control:"asld"}),a.asldPtrIB.style.border=a.asldPtrOB.style.border=n.pointerBorderWidth+"px solid "+n.pointerBorderColor,a.asldPtrOB.style.position="absolute",a.asldPtrOB.style.left=-(2*n.pointerBorderWidth+n.pointerThickness)+"px",a.asldPtrOB.style.top="0",a.asldPtrMB.style.border=n.pointerThickness+"px solid "+n.pointerColor,a.asldPtrS.style.width=n.sliderSize+"px",a.asldPtrS.style.height=i.pub.sliderInnerSpace+"px",a.pal.className="jscolor-palette",a.pal.style.display=p.palette.rows?"block":"none",a.pal.style.position="absolute",a.pal.style.left=n.padding+"px",a.pal.style.top=2*n.controlBorderWidth+2*n.padding+n.height+"px",a.pal.innerHTML="";for(var g=i.genColorPreviewCanvas("rgba(0,0,0,0)"),f=0,C=0;C<p.palette.rows;C++)for(var y=0;y<p.palette.cols&&f<n._palette.length;y++,f++){var b=n._palette[f],k=i.rgbaColor.apply(null,b.rgba),G=i.createEl("div");G.style.width=p.palette.cellW-2*n.controlBorderWidth+"px",G.style.height=p.palette.cellH-2*n.controlBorderWidth+"px",G.style.backgroundColor=k;var v=i.createEl("div");v.className="jscolor-palette-sample",v.style.display="block",v.style.position="absolute",v.style.left=(p.palette.cols<=1?0:Math.round(y*((p.contentW-p.palette.cellW)/(p.palette.cols-1))*10)/10)+"px",v.style.top=C*(p.palette.cellH+n.paletteSpacing)+"px",v.style.border=n.controlBorderWidth+"px solid",v.style.borderColor=n.controlBorderColor,v.style.cursor="pointer",null!==b.rgba[3]&&b.rgba[3]<1&&(v.style.backgroundImage="url('"+g.canvas.toDataURL()+"')",v.style.backgroundRepeat="repeat",v.style.backgroundPosition="center center"),i.setData(v,{instance:n,control:"palette-sample",color:b}),v.addEventListener("click",i.onPaletteSampleClick,!1),v.appendChild(G),a.pal.appendChild(v)}a.btn.className="jscolor-btn-close",a.btn.style.display=n.closeButton?"block":"none",a.btn.style.position="absolute",a.btn.style.left=n.padding+"px",a.btn.style.bottom=n.padding+"px",a.btn.style.padding="0 15px",a.btn.style.maxWidth=p.contentW-2*n.controlBorderWidth-30+"px",a.btn.style.overflow="hidden",a.btn.style.height=n.buttonHeight+"px",a.btn.style.whiteSpace="nowrap",a.btn.style.border=n.controlBorderWidth+"px solid",o=(t=n.controlBorderColor.split(/\s+/)).length<2?t[0]:t[1]+" "+t[0]+" "+t[0]+" "+t[1],a.btn.style.borderColor=o,a.btn.style.color=n.buttonColor,a.btn.style.font="12px sans-serif",a.btn.style.textAlign="center",a.btn.style.cursor="pointer",a.btn.onmousedown=function(){n.hide()},a.btnT.style.lineHeight=n.buttonHeight+"px",a.btnT.innerHTML="",a.btnT.appendChild(e.document.createTextNode(n.closeText)),s(),c(),d(),i.picker.owner&&i.picker.owner!==n&&i.removeClass(i.picker.owner.targetElement,i.pub.activeClassName),i.picker.owner=n,n.container===e.document.body?i.redrawPosition():i._drawPosition(n,0,0,"relative",!1),a.wrap.parentNode!==n.container&&n.container.appendChild(a.wrap),i.addClass(n.targetElement,i.pub.activeClassName)}function s(){var e=i.getPadYChannel(n),t=Math.round(n.channels.h/360*(n.width-1)),o=Math.round((1-n.channels[e]/100)*(n.height-1)),a=2*n.pointerBorderWidth+n.pointerThickness+2*n.crossSize,r=-Math.floor(a/2);switch(i.picker.cross.style.left=t+r+"px",i.picker.cross.style.top=o+r+"px",i.getSliderChannel(n)){case"s":var l=i.HSV_RGB(n.channels.h,100,n.channels.v),s=i.HSV_RGB(n.channels.h,0,n.channels.v),c="rgb("+Math.round(l[0])+","+Math.round(l[1])+","+Math.round(l[2])+")",d="rgb("+Math.round(s[0])+","+Math.round(s[1])+","+Math.round(s[2])+")";i.picker.sldGrad.draw(n.sliderSize,n.height,c,d);break;case"v":var p=i.HSV_RGB(n.channels.h,n.channels.s,100);c="rgb("+Math.round(p[0])+","+Math.round(p[1])+","+Math.round(p[2])+")",d="#000",i.picker.sldGrad.draw(n.sliderSize,n.height,c,d)}i.picker.asldGrad.draw(n.sliderSize,n.height,n.toHEXString())}function c(){var e=i.getSliderChannel(n);if(e){var t=Math.round((1-n.channels[e]/100)*(n.height-1));i.picker.sldPtrOB.style.top=t-(2*n.pointerBorderWidth+n.pointerThickness)-Math.floor(i.pub.sliderInnerSpace/2)+"px"}i.picker.asldGrad.draw(n.sliderSize,n.height,n.toHEXString())}function d(){var e=Math.round((1-n.channels.a)*(n.height-1));i.picker.asldPtrOB.style.top=e-(2*n.pointerBorderWidth+n.pointerThickness)-Math.floor(i.pub.sliderInnerSpace/2)+"px"}function p(){return i.picker&&i.picker.owner===n}if(o||(o={}),this.channels={r:255,g:255,b:255,h:0,s:0,v:100,a:1},this.format="auto",this.value=void 0,this.alpha=void 0,this.onChange=void 0,this.onInput=void 0,this.valueElement=void 0,this.alphaElement=void 0,this.previewElement=void 0,this.previewPosition="left",this.previewSize=32,this.previewPadding=8,this.required=!0,this.hash=!0,this.uppercase=!0,this.forceStyle=!0,this.width=181,this.height=101,this.mode="HSV",this.alphaChannel="auto",this.position="bottom",this.smartPosition=!0,this.showOnClick=!0,this.hideOnLeave=!0,this.palette=[],this.paletteCols=10,this.paletteSetsAlpha="auto",this.paletteHeight=16,this.paletteSpacing=4,this.hideOnPaletteClick=!1,this.sliderSize=16,this.crossSize=8,this.closeButton=!1,this.closeText="Close",this.buttonColor="rgba(0,0,0,1)",this.buttonHeight=18,this.padding=12,this.backgroundColor="rgba(255,255,255,1)",this.borderWidth=1,this.borderColor="rgba(187,187,187,1)",this.borderRadius=8,this.controlBorderWidth=1,this.controlBorderColor="rgba(187,187,187,1)",this.shadow=!0,this.shadowBlur=15,this.shadowColor="rgba(0,0,0,0.2)",this.pointerColor="rgba(76,76,76,1)",this.pointerBorderWidth=1,this.pointerBorderColor="rgba(255,255,255,1)",this.pointerThickness=2,this.zIndex=5e3,this.container=void 0,this.minS=0,this.maxS=100,this.minV=0,this.maxV=100,this.minA=0,this.maxA=1,this.option=function(){if(!arguments.length)throw new Error("No option specified");if(1===arguments.length&&"string"==typeof arguments[0]){try{return r(arguments[0])}catch(e){console.warn(e)}return!1}if(arguments.length>=2&&"string"==typeof arguments[0]){try{if(!a(arguments[0],arguments[1]))return!1}catch(e){return console.warn(e),!1}return this.redraw(),this.exposeColor(),!0}if(1===arguments.length&&"object"==typeof arguments[0]){var e=arguments[0],t=!0;for(var o in e)if(e.hasOwnProperty(o))try{a(o,e[o])||(t=!1)}catch(e){console.warn(e),t=!1}return this.redraw(),this.exposeColor(),t}throw new Error("Invalid arguments")},this.channel=function(e,t){if("string"!=typeof e)throw new Error("Invalid value for channel name: "+e);if(void 0===t)return this.channels.hasOwnProperty(e.toLowerCase())?this.channels[e.toLowerCase()]:(console.warn("Getting unknown channel: "+e),!1);var o=!1;switch(e.toLowerCase()){case"r":o=this.fromRGBA(t,null,null,null);break;case"g":o=this.fromRGBA(null,t,null,null);break;case"b":o=this.fromRGBA(null,null,t,null);break;case"h":o=this.fromHSVA(t,null,null,null);break;case"s":o=this.fromHSVA(null,t,null,null);break;case"v":o=this.fromHSVA(null,null,t,null);break;case"a":o=this.fromHSVA(null,null,null,t);break;default:return console.warn("Setting unknown channel: "+e),!1}return!!o&&(this.redraw(),!0)},this.trigger=function(e){for(var t=i.strList(e),o=0;o<t.length;o+=1){var n=t[o].toLowerCase(),a=null;switch(n){case"input":a="onInput";break;case"change":a="onChange"}a&&i.triggerCallback(this,a),i.triggerInputEvent(this.valueElement,n,!0,!0)}},this.fromHSVA=function(e,t,o,n,a){if(void 0===e&&(e=null),void 0===t&&(t=null),void 0===o&&(o=null),void 0===n&&(n=null),null!==e){if(isNaN(e))return!1;this.channels.h=Math.max(0,Math.min(360,e))}if(null!==t){if(isNaN(t))return!1;this.channels.s=Math.max(0,Math.min(100,this.maxS,t),this.minS)}if(null!==o){if(isNaN(o))return!1;this.channels.v=Math.max(0,Math.min(100,this.maxV,o),this.minV)}if(null!==n){if(isNaN(n))return!1;this.channels.a=this.hasAlphaChannel()?Math.max(0,Math.min(1,this.maxA,n),this.minA):1}var r=i.HSV_RGB(this.channels.h,this.channels.s,this.channels.v);return this.channels.r=r[0],this.channels.g=r[1],this.channels.b=r[2],this.exposeColor(a),!0},this.fromRGBA=function(e,t,o,n,a){if(void 0===e&&(e=null),void 0===t&&(t=null),void 0===o&&(o=null),void 0===n&&(n=null),null!==e){if(isNaN(e))return!1;e=Math.max(0,Math.min(255,e))}if(null!==t){if(isNaN(t))return!1;t=Math.max(0,Math.min(255,t))}if(null!==o){if(isNaN(o))return!1;o=Math.max(0,Math.min(255,o))}if(null!==n){if(isNaN(n))return!1;this.channels.a=this.hasAlphaChannel()?Math.max(0,Math.min(1,this.maxA,n),this.minA):1}var r=i.RGB_HSV(null===e?this.channels.r:e,null===t?this.channels.g:t,null===o?this.channels.b:o);null!==r[0]&&(this.channels.h=Math.max(0,Math.min(360,r[0]))),0!==r[2]&&(this.channels.s=Math.max(0,this.minS,Math.min(100,this.maxS,r[1]))),this.channels.v=Math.max(0,this.minV,Math.min(100,this.maxV,r[2]));var l=i.HSV_RGB(this.channels.h,this.channels.s,this.channels.v);return this.channels.r=l[0],this.channels.g=l[1],this.channels.b=l[2],this.exposeColor(a),!0},this.fromHSV=function(e,t,o,n){return console.warn("fromHSV() method is DEPRECATED. Using fromHSVA() instead."+i.docsRef),this.fromHSVA(e,t,o,null,n)},this.fromRGB=function(e,t,o,n){return console.warn("fromRGB() method is DEPRECATED. Using fromRGBA() instead."+i.docsRef),this.fromRGBA(e,t,o,null,n)},this.fromString=function(e,t){if(!this.required&&""===e.trim())return this.setPreviewElementBg(null),this.setValueElementValue(""),!0;var o=i.parseColorString(e);return!!o&&("any"===this.format.toLowerCase()&&(this._setFormat(o.format),i.isAlphaFormat(this.getFormat())||(o.rgba[3]=1)),this.fromRGBA(o.rgba[0],o.rgba[1],o.rgba[2],o.rgba[3],t),!0)},this.toString=function(e){switch(void 0===e&&(e=this.getFormat()),e.toLowerCase()){case"hex":return this.toHEXString();case"hexa":return this.toHEXAString();case"rgb":return this.toRGBString();case"rgba":return this.toRGBAString()}return!1},this.toHEXString=function(){return i.hexColor(this.channels.r,this.channels.g,this.channels.b)},this.toHEXAString=function(){return i.hexaColor(this.channels.r,this.channels.g,this.channels.b,this.channels.a)},this.toRGBString=function(){return i.rgbColor(this.channels.r,this.channels.g,this.channels.b)},this.toRGBAString=function(){return i.rgbaColor(this.channels.r,this.channels.g,this.channels.b,this.channels.a)},this.toGrayscale=function(){return.213*this.channels.r+.715*this.channels.g+.072*this.channels.b},this.toCanvas=function(){return i.genColorPreviewCanvas(this.toRGBAString()).canvas},this.toDataURL=function(){return this.toCanvas().toDataURL()},this.toBackground=function(){return i.pub.background(this.toRGBAString())},this.isLight=function(){return this.toGrayscale()>127.5},this.hide=function(){p()&&(i.removeClass(n.targetElement,i.pub.activeClassName),i.picker.wrap.parentNode.removeChild(i.picker.wrap),delete i.picker.owner)},this.show=function(){l()},this.redraw=function(){p()&&l()},this.getFormat=function(){return this._currentFormat},this._setFormat=function(e){this._currentFormat=e.toLowerCase()},this.hasAlphaChannel=function(){return"auto"===this.alphaChannel?"any"===this.format.toLowerCase()||i.isAlphaFormat(this.getFormat())||void 0!==this.alpha||void 0!==this.alphaElement:this.alphaChannel},this.processValueInput=function(e){this.fromString(e)||this.exposeColor()},this.processAlphaInput=function(e){this.fromHSVA(null,null,null,parseFloat(e))||this.exposeColor()},this.exposeColor=function(e){var t=this.toString(),o=this.getFormat();if(i.setDataAttr(this.targetElement,"current-color",t),e&i.flags.leaveValue||!this.valueElement||("hex"!==o&&"hexa"!==o||(this.uppercase||(t=t.toLowerCase()),this.hash||(t=t.replace(/^#/,""))),this.setValueElementValue(t)),!(e&i.flags.leaveAlpha)&&this.alphaElement){var n=Math.round(100*this.channels.a)/100;this.setAlphaElementValue(n)}e&i.flags.leavePreview||!this.previewElement||((i.isTextInput(this.previewElement)||i.isButton(this.previewElement)&&!i.isButtonEmpty(this.previewElement))&&this.previewPosition,this.setPreviewElementBg(this.toRGBAString())),p()&&(s(),c(),d())},this.setPreviewElementBg=function(e){if(this.previewElement){var t=null,o=null;(i.isTextInput(this.previewElement)||i.isButton(this.previewElement)&&!i.isButtonEmpty(this.previewElement))&&(t=this.previewPosition,o=this.previewSize);var n=[];if(e){n.push({image:i.genColorPreviewGradient(e,t,o?o-i.pub.previewSeparator.length:null),position:"left top",size:"auto",repeat:t?"repeat-y":"repeat",origin:"padding-box"});var a=i.genColorPreviewCanvas("rgba(0,0,0,0)",t?{left:"right",right:"left"}[t]:null,o,!0);n.push({image:"url('"+a.canvas.toDataURL()+"')",position:(t||"left")+" top",size:a.width+"px "+a.height+"px",repeat:t?"repeat-y":"repeat",origin:"padding-box"})}else n.push({image:"none",position:"left top",size:"auto",repeat:"no-repeat",origin:"padding-box"});for(var r={image:[],position:[],size:[],repeat:[],origin:[]},l=0;l<n.length;l+=1)r.image.push(n[l].image),r.position.push(n[l].position),r.size.push(n[l].size),r.repeat.push(n[l].repeat),r.origin.push(n[l].origin);var s={"background-image":r.image.join(", "),"background-position":r.position.join(", "),"background-size":r.size.join(", "),"background-repeat":r.repeat.join(", "),"background-origin":r.origin.join(", ")};i.setStyle(this.previewElement,s,this.forceStyle);var c={left:null,right:null};t&&(c[t]=this.previewSize+this.previewPadding+"px"),s={"padding-left":c.left,"padding-right":c.right},i.setStyle(this.previewElement,s,this.forceStyle,!0)}},this.setValueElementValue=function(e){this.valueElement&&("input"===i.nodeName(this.valueElement)?this.valueElement.value=e:this.valueElement.innerHTML=e)},this.setAlphaElementValue=function(e){this.alphaElement&&("input"===i.nodeName(this.alphaElement)?this.alphaElement.value=e:this.alphaElement.innerHTML=e)},this._processParentElementsInDOM=function(){if(!this._parentElementsProcessed){this._parentElementsProcessed=!0;var e=this.targetElement;do{var t=i.getCompStyle(e);t.position&&"fixed"===t.position.toLowerCase()&&(this.fixed=!0),e!==this.targetElement&&(i.getData(e,"hasScrollListener")||(e.addEventListener("scroll",i.onParentScroll,!1),i.setData(e,"hasScrollListener",!0)))}while((e=e.parentNode)&&"body"!==i.nodeName(e))}},this.tryHide=function(){this.hideOnLeave&&this.hide()},this.set__palette=function(e){this.palette=e,this._palette=i.parsePaletteValue(e),this._paletteHasTransparency=i.containsTranparentColor(this._palette)},i.pub.options)for(var u in i.pub.options)if(i.pub.options.hasOwnProperty(u))try{a(u,i.pub.options[u])}catch(e){console.warn(e)}var m=[];o.preset&&("string"==typeof o.preset?m=o.preset.split(/\s+/):Array.isArray(o.preset)?m=o.preset.slice():console.warn("Unrecognized preset value")),-1===m.indexOf("default")&&m.push("default");for(var h=m.length-1;h>=0;h-=1){var g=m[h];if(g)if(i.pub.presets.hasOwnProperty(g)){for(var u in i.pub.presets[g])if(i.pub.presets[g].hasOwnProperty(u))try{a(u,i.pub.presets[g][u])}catch(e){console.warn(e)}}else console.warn("Unknown preset: %s",g)}var f=["preset"];for(var u in o)if(o.hasOwnProperty(u)&&-1===f.indexOf(u))try{a(u,o[u])}catch(e){console.warn(e)}if(void 0===this.container?this.container=e.document.body:this.container=i.node(this.container),!this.container)throw new Error("Cannot instantiate color picker without a container element");if(this.targetElement=i.node(t),!this.targetElement){if("string"==typeof t&&/^[a-zA-Z][\w:.-]*$/.test(t))throw new Error("If '"+t+"' is supposed to be an ID, please use '#"+t+"' or any valid CSS selector.");throw new Error("Cannot instantiate color picker without a target element")}if(this.targetElement.jscolor&&this.targetElement.jscolor instanceof i.pub)throw new Error("Color picker already installed on this element");if(this.targetElement.jscolor=this,i.addClass(this.targetElement,i.pub.className),i.instances.push(this),i.isButton(this.targetElement)&&("button"!==this.targetElement.type.toLowerCase()&&(this.targetElement.type="button"),i.isButtonEmpty(this.targetElement))){i.removeChildren(this.targetElement),this.targetElement.appendChild(e.document.createTextNode(" "));var C=i.getCompStyle(this.targetElement);(parseFloat(C["min-width"])||0)<this.previewSize&&i.setStyle(this.targetElement,{"min-width":this.previewSize+"px"},this.forceStyle)}if(void 0===this.valueElement?i.isTextInput(this.targetElement)&&(this.valueElement=this.targetElement):null===this.valueElement||(this.valueElement=i.node(this.valueElement)),this.alphaElement&&(this.alphaElement=i.node(this.alphaElement)),void 0===this.previewElement?this.previewElement=this.targetElement:null===this.previewElement||(this.previewElement=i.node(this.previewElement)),this.valueElement&&i.isTextInput(this.valueElement)){var y={onInput:this.valueElement.oninput};this.valueElement.oninput=null,this.valueElement.addEventListener("keydown",(function(e){"Enter"===i.eventKey(e)&&(n.valueElement&&n.processValueInput(n.valueElement.value),n.tryHide())}),!1),this.valueElement.addEventListener("change",(function(e){if(!i.getData(e,"internal")){var t=n.valueElement.value;n.processValueInput(n.valueElement.value),i.triggerCallback(n,"onChange"),n.valueElement.value!==t&&i.triggerInputEvent(n.valueElement,"change",!0,!0)}}),!1),this.valueElement.addEventListener("input",(function(e){i.getData(e,"internal")||(n.valueElement&&n.fromString(n.valueElement.value,i.flags.leaveValue),i.triggerCallback(n,"onInput"))}),!1),y.onInput&&this.valueElement.addEventListener("input",y.onInput,!1),this.valueElement.setAttribute("autocomplete","off"),this.valueElement.setAttribute("autocorrect","off"),this.valueElement.setAttribute("autocapitalize","off"),this.valueElement.setAttribute("spellcheck",!1)}this.alphaElement&&i.isTextInput(this.alphaElement)&&(this.alphaElement.addEventListener("keydown",(function(e){"Enter"===i.eventKey(e)&&(n.alphaElement&&n.processAlphaInput(n.alphaElement.value),n.tryHide())}),!1),this.alphaElement.addEventListener("change",(function(e){if(!i.getData(e,"internal")){var t=n.alphaElement.value;n.processAlphaInput(n.alphaElement.value),i.triggerCallback(n,"onChange"),i.triggerInputEvent(n.valueElement,"change",!0,!0),n.alphaElement.value!==t&&i.triggerInputEvent(n.alphaElement,"change",!0,!0)}}),!1),this.alphaElement.addEventListener("input",(function(e){i.getData(e,"internal")||(n.alphaElement&&n.fromHSVA(null,null,null,parseFloat(n.alphaElement.value),i.flags.leaveAlpha),i.triggerCallback(n,"onInput"),i.triggerInputEvent(n.valueElement,"input",!0,!0))}),!1),this.alphaElement.setAttribute("autocomplete","off"),this.alphaElement.setAttribute("autocorrect","off"),this.alphaElement.setAttribute("autocapitalize","off"),this.alphaElement.setAttribute("spellcheck",!1));var b="FFFFFF";void 0!==this.value?b=this.value:this.valueElement&&void 0!==this.valueElement.value&&(b=this.valueElement.value);var k=void 0;if(void 0!==this.alpha?k=""+this.alpha:this.alphaElement&&void 0!==this.alphaElement.value&&(k=this.alphaElement.value),this._currentFormat=null,["auto","any"].indexOf(this.format.toLowerCase())>-1){var G=i.parseColorString(b);this._currentFormat=G?G.format:"hex"}else this._currentFormat=this.format.toLowerCase();this.processValueInput(b),void 0!==k&&this.processAlphaInput(k)}}).pub.className="jscolor",i.pub.activeClassName="jscolor-active",i.pub.looseJSON=!0,i.pub.presets={},i.pub.presets.default={},i.pub.presets.light={backgroundColor:"rgba(255,255,255,1)",controlBorderColor:"rgba(187,187,187,1)",buttonColor:"rgba(0,0,0,1)"},i.pub.presets.dark={backgroundColor:"rgba(51,51,51,1)",controlBorderColor:"rgba(153,153,153,1)",buttonColor:"rgba(240,240,240,1)"},i.pub.presets.small={width:101,height:101,padding:10,sliderSize:14,paletteCols:8},i.pub.presets.medium={width:181,height:101,padding:12,sliderSize:16,paletteCols:10},i.pub.presets.large={width:271,height:151,padding:12,sliderSize:24,paletteCols:15},i.pub.presets.thin={borderWidth:1,controlBorderWidth:1,pointerBorderWidth:1},i.pub.presets.thick={borderWidth:2,controlBorderWidth:2,pointerBorderWidth:2},i.pub.sliderInnerSpace=3,i.pub.chessboardSize=8,i.pub.chessboardColor1="#666666",i.pub.chessboardColor2="#999999",i.pub.previewSeparator=["rgba(255,255,255,.65)","rgba(128,128,128,.65)"],i.pub.init=function(){if(!i.initialized)for(e.document.addEventListener("mousedown",i.onDocumentMouseDown,!1),e.document.addEventListener("keyup",i.onDocumentKeyUp,!1),e.addEventListener("resize",i.onWindowResize,!1),e.addEventListener("scroll",i.onWindowScroll,!1),i.pub.install(),i.initialized=!0;i.readyQueue.length;)i.readyQueue.shift()()},i.pub.install=function(e){var t=!0;try{i.installBySelector("[data-jscolor]",e)}catch(e){t=!1,console.warn(e)}if(i.pub.lookupClass)try{i.installBySelector("input."+i.pub.lookupClass+", button."+i.pub.lookupClass,e)}catch(e){}return t},i.pub.ready=function(e){return"function"!=typeof e?(console.warn("Passed value is not a function"),!1):(i.initialized?e():i.readyQueue.push(e),!0)},i.pub.trigger=function(e){var t=function(){i.triggerGlobal(e)};i.initialized?t():i.pub.ready(t)},i.pub.hide=function(){i.picker&&i.picker.owner&&i.picker.owner.hide()},i.pub.chessboard=function(e){return e||(e="rgba(0,0,0,0)"),i.genColorPreviewCanvas(e).canvas.toDataURL()},i.pub.background=function(e){var t=[];t.push(i.genColorPreviewGradient(e));var o=i.genColorPreviewCanvas();return t.push(["url('"+o.canvas.toDataURL()+"')","left top","repeat"].join(" ")),t.join(", ")},i.pub.options={},i.pub.lookupClass="jscolor",i.pub.installByClassName=function(){return console.error('jscolor.installByClassName() is DEPRECATED. Use data-jscolor="" attribute instead of a class name.'+i.docsRef),!1},i.register(),i.pub);return void 0===e.jscolor&&(e.jscolor=e.JSColor=a),a}))}},t={};function o(n){var i=t[n];if(void 0!==i)return i.exports;var a=t[n]={exports:{}};return e[n].call(a.exports,a,a.exports,o),a.exports}o.n=e=>{var t=e&&e.__esModule?()=>e.default:()=>e;return o.d(t,{a:t}),t},o.d=(e,t)=>{for(var n in t)o.o(t,n)&&!o.o(e,n)&&Object.defineProperty(e,n,{enumerable:!0,get:t[n]})},o.o=(e,t)=>Object.prototype.hasOwnProperty.call(e,t),(()=>{"use strict";let e={};function t(){e.WrinklerButtons&&Game.elderWrath?(l("PopAllNormalWrinklerButton").style.display="",l("PopFattestWrinklerButton").style.display=""):(l("PopAllNormalWrinklerButton").style.display="none",l("PopFattestWrinklerButton").style.display="none")}const n=["","","M","G","T","P","E","Z","Y"],i=["","","M","B","Tr","Quadr","Quint","Sext","Sept","Oct","Non","Dec","Undec","Duodec","Tredec","Quattuordec","Quindec","Sexdec","Septendec","Octodec","Novemdec","Vigint","Unvigint","Duovigint","Trevigint","Quattuorvigint"],a=["","K","M","B","T","Qa","Qi","Sx","Sp","Oc","No","De","UDe","DDe","TDe","QaDe","QiDe","SxDe","SpDe","ODe","NDe","Vi","UVi","DVi","TVi","QaVi","QiVi","SxVi","SpVi","OVi","NVi","Tr","UTr","DTr","TTr","QaTr","QiTr","SxTr","SpTr","OTr","NTr","Qaa","UQa","DQa","TQa","QaQa","QiQa","SxQa","SpQa","OQa","NQa","Qia","UQi","DQi","TQi","QaQi","QiQi","SxQi","SpQi","OQi","NQi","Sxa","USx","DSx","TSx","QaSx","QiSx","SxSx","SpSx","OSx","NSx","Spa","USp","DSp","TSp","QaSp","QiSp","SxSp","SpSp","OSp","NSp","Oco","UOc","DOc","TOc","QaOc","QiOc","SxOc","SpOc","OOc","NOc","Noa","UNo","DNo","TNo","QaNo","QiNo","SxNo","SpNo","ONo","NNo","Ct","UCt"];let r,s,c,d,p,u,m,h,g,f,C,y,b,k,G,v,w,x=[],T=[],M=[],S={},P=Date.now(),B=Date.now();const E="CMText",N="CMBack",O="Blue",F="Green",L="Yellow",W="Orange",A="Red",j="Purple",D="Gray",H="Pink",U="Brown",$=[D,O,F,L,W,A,j,H,U],R=[["GoldCookTooltipPlaceholder","Calculated with Golden Switch off","200px"],["GoldCookDragonsFortuneTooltipPlaceholder","Calculated with Golden Switch off and at least one golden cookie on-screen","240px"],["PrestMaxTooltipPlaceholder","The MAX prestige is calculated with the cookies gained from popping all wrinklers with Skruuia god in Diamond slot, selling all stock market goods, selling all buildings with Earth Shatterer and Reality Bending auras, and buying Chocolate egg","320px"],["NextPrestTooltipPlaceholder","Calculated with cookies gained from wrinklers and Chocolate egg","200px"],["HeavenChipMaxTooltipPlaceholder","The MAX heavenly chips is calculated with the cookies gained from popping all wrinklers with Skruuia god in Diamond slot, selling all stock market goods, selling all buildings with Earth Shatterer and Reality Bending auras, and buying Chocolate egg","330px"],["ResetTooltipPlaceholder","The bonus income you would get from new prestige levels unlocked at 100% of its potential and from ascension achievements if you have the same buildings/upgrades after reset","370px"],["ChoEggTooltipPlaceholder","The amount of cookies you would get from popping all wrinklers with Skruuia god in Diamond slot, selling all stock market goods, selling all buildings with Earth Shatterer and Reality Bending auras, and then buying Chocolate egg","300px"],["ChainNextLevelPlaceholder","Cheated cookies might break this formula","250px"],["FavouriteSettingPlaceholder","Click to set this setting as favourite and show it in 'favourite' settings at the top of the Cookie Monster Settings","250px"]],I={};let z,V,_=0,Q=-1,X=[],q=1,Y=1,K=1;const J=[10,15,30,60,300,600,900,1800],Z=[1,5,10,15,30],ee={Frenzy:L,"Dragon Harvest":U,"Elder frenzy":F,Clot:A,"Click frenzy":O,Dragonflight:H};let te,oe,ne,ie,ae,re,le,se={},ce="",de=[];function pe(t,o,r){const l=e.ScaleDecimals+1;if(t===1/0)return"Infinity";if(void 0===t)return"0";if(0===e.Scale)return S.Beautify(t,o);if(Number.isFinite(t)){if(t<0)return`-${pe(Math.abs(t))}`;let s="";if(0===t)return t.toString();if(t>.001&&t<e.ScaleCutoff)return s=e.ScaleSeparator?t.toLocaleString("nl"):t.toLocaleString("en"),s;if(4===e.Scale&&!r||4===r)s=t.toExponential(l).toString().replace("e","E");else{const o=t.toExponential().toString(),c=Math.floor(o.slice(o.indexOf("e")+1)/3);s=(t/Number("1e"+3*c)).toFixed(l),1===e.Scale&&!r||1===r?t>=.01&&t<Number("1e"+3*n.length)?s+=` ${n[c]}`:s=pe(t,0,4):2===e.Scale&&!r||2===r?t>=.01&&t<Number("1e"+3*i.length)?s+=` ${i[c]}`:s=pe(t,0,4):3===e.Scale&&!r||3===r?t>=.01&&t<Number("1e"+3*a.length)?s+=` ${a[c]}`:s=pe(t,0,4):(5===e.Scale&&!r||5===r)&&(s+="E"+3*c)}return""===s&&(console.log(`Could not beautify number with Cookie Monster Beautify: ${t}`),s=S.Beautify(t,o)),e.ScaleSeparator&&(s=s.replace(".",",")),s}return console.log(`Could not beautify number with Cookie Monster Beautify: ${t}`),S.Beautify(t,o)}function ue(t,o){let n=t;if(t===1/0)return t;if(t<0)return"Negative time period";n=Math.ceil(t);const i=Math.floor(n/31557600),a=Math.floor(n%31557600/86400),r=Math.floor(n%86400/3600),l=Math.floor(n%3600/60),s=Math.floor(n%60);let c="";if(e.TimeFormat){if(n>315576e4)return"XX:XX:XX:XX:XX";c+=(i<10?"0":"")+i+":",c+=(a<10?"0":"")+a+":",c+=(r<10?"0":"")+r+":",c+=(l<10?"0":"")+l+":",c+=(s<10?"0":"")+s}else{if(n>7776e5)return o?"Over 9000 days!":">9000d";c+=i>0?i+(o?1===i?" year":" years":"y")+", ":"",c+=a>0?a+(o?1===a?" day":" days":"d")+", ":"",(c.length>0||r>0)&&(c+=r+(o?1===r?" hour":" hours":"h")+", "),(c.length>0||l>0)&&(c+=l+(o?1===l?" minute":" minutes":"m")+", "),c+=s+(o?1===s?" second":" seconds":"s")}return c}function me(t){let o,n;return t<=0?(n=e.TimeFormat?"00:00:00:00:00":"Done!",o=F):(n=ue(t),o=t>300?A:t>60?W:L),{text:n,color:o}}let he,ge,fe,Ce,ye,be,ke,Ge,ve,we,xe,Te,Me,Se,Pe,Be,Ee,Ne,Oe,Fe,Le,We,Ae,je,De,He,Ue,$e,Re,Ie,ze,Ve,_e,Qe,Xe,qe,Ye=0,Ke=0,Je=0,Ze=0,et=0,tt=0,ot=0,nt=0,it=0,at=0,rt=0,lt=0,st=0,ct=0,dt=0,pt=1,ut=1,mt=1,ht=0,gt=0,ft=0,Ct=0,yt={},bt={},kt={},Gt={},vt=0,wt=0,xt=[0,null],Tt=0,Mt=0,St={},Pt={},Bt={},Et={},Nt={},Ot=0,Ft=0,Lt=[],Wt=0,At=0,jt=[],Dt=0,Ht=0,Ut=[],$t=0,Rt=0,It=[],zt=0,Vt=0,_t={},Qt={},Xt=0,qt=0,Yt=[],Kt={},Jt=0,Zt={0:[0,0,0],1:[0,0,0],2:[0,0,0],3:[0,0,0],4:[0,0,0],5:[0,0,0],6:[0,0,0],7:[0,0,0],8:[0,0,0],9:[0,0,0],10:[0,0,0]},eo=[],to=[],oo=[];function no(e){let t=0;return Game.dragonAuras[$e].name!==e&&Game.dragonAuras[Re].name!==e||(t=1),"Reality Bending"!==Game.dragonAuras[$e].name&&"Reality Bending"!==Game.dragonAuras[Re].name||(t+=.1),t}function io(){let e=.25;return e*=1+no("Earth Shatterer"),e}function ao(e,t){return void 0===Qe[e]?void 0===t?1:t:Qe[e]}function ro(e){const t=to[e];return(1!==Game.ascensionMode||"prestige"!==t.pool&&"fortune"!==t.tier)&&t?t.bought:0}function lo(e){if(Game.hasGod){void 0===eo.Temple.minigame&&(eo.Temple.minigame=Game.Objects.Temple.minigame);const t=eo.Temple.minigame.gods[e];if(Ie===t.id)return 1;if(ze===t.id)return 2;if(Ve===t.id)return 3}return!1}function so(e,t){let o=t;if(ro("Season savings")&&(o*=.99),ro("Santa's dominion")&&(o*=.99),ro("Faberge egg")&&(o*=.99),ro("Divine discount")&&(o*=.99),ro("Fortune #100")&&(o*=.99),o*=1-.02*no("Fierce Hoarder"),Game.hasBuff("Everything must go")&&(o*=.95),Game.hasBuff("Crafty pixies")&&(o*=.98),Game.hasBuff("Nasty goblins")&&(o*=1.02),e.fortune&&ro(e.fortune.name)&&(o*=.93),o*=ao("buildingCost"),eo.Temple.minigameLoaded){const e=lo("creation");1===e?o*=.93:2===e?o*=.95:3===e&&(o*=.98)}return o}function co(e,t,o,n,i,a){const r=void 0===a?0:a;let l=i,s=o,c=0;-1===i&&(l=s),i||(l=Game.buyBulk);for(let o=0;o<l;o++){let o=t*Game.priceIncrease**Math.max(0,s-n);o=r?Game.modifyBuildingPrice(e,o):so(e,o),o=Math.ceil(o);const i=r?e.getSellMultiplier():io();o=Math.floor(o*i),s>0&&(c+=o,s-=1)}return c}function po(){let t=Game.buyBulk;if(1===Game.buyMode?Y=t:t=Y,1===t?t=yt:10===t?t=bt:100===t&&(t=kt),1===Game.buyMode?1===e.BuildColour?Object.keys(t).forEach((o=>{l(`productPrice${Game.Objects[o].id}`).style.color=e[`Colour${t[o].color}`]})):Object.keys(Game.Objects).forEach((e=>{l(`productPrice${Game.Objects[e].id}`).style.removeProperty("color")})):-1===Game.buyMode&&Object.keys(yt).forEach((e=>{const t=Game.Objects[e];l(`productPrice${t.id}`).style.color="",l(`productPrice${t.id}`).innerHTML=pe(co(t,t.basePrice,t.amount,t.free,Game.buyBulk,1))})),1===Game.buyMode&&e.SortBuildings){let o;1===e.SortBuildings?(o=Object.keys(yt).map((e=>{const t=yt[e];return t.name=e,t.id=Game.Objects[e].id,t})),o.sort((function(e,t){return $.indexOf(e.color)>$.indexOf(t.color)?1:$.indexOf(e.color)<$.indexOf(t.color)||e.pp<t.pp?-1:0}))):2===e.SortBuildings&&(o=Object.keys(t).map((e=>{const o=t[e];return o.name=e,o.id=Game.Objects[e].id,o})),o.sort((function(e,t){return $.indexOf(e.color)>$.indexOf(t.color)?1:$.indexOf(e.color)<$.indexOf(t.color)||e.pp<t.pp?-1:0})));for(let e=0;e<o.length;e++)Game.Objects[o[e].name].l.style.gridRow=`${e+2}/${e+2}`}else{const e=Object.keys(yt).map((e=>{const t=yt[e];return t.name=e,t.id=Game.Objects[e].id,t}));e.sort(((e,t)=>e.id-t.id));for(let t=0;t<e.length;t++)Game.Objects[e[t].name].l.style.gridRow=`${t+2}/${t+2}`}}function uo(){if(e.UpBarColour>0){let e=0,t=0,o=0,n=0,i=0,a=0,r=0;Object.keys(Game.UpgradesInStore).forEach((s=>{const c=Game.UpgradesInStore[s];let d=!1;for(let e=0;e<l(`upgrade${s}`).childNodes.length;e+=1)if(-1!==l(`upgrade${s}`).childNodes[e].className.indexOf(N)){l(`upgrade${s}`).childNodes[e].className=N+St[c.name].color,d=!0;break}if(!d){const e=document.createElement("div");e.style.width="10px",e.style.height="10px",e.className=N+St[c.name].color,l(`upgrade${s}`).appendChild(e)}St[c.name].color===O?e+=1:St[c.name].color===F?t+=1:St[c.name].color===L?o+=1:St[c.name].color===W?n+=1:St[c.name].color===A?i+=1:St[c.name].color===j?a+=1:St[c.name].color===D&&(r+=1)})),l("CMUpgradeBarBlue").textContent=e,l("CMUpgradeBarGreen").textContent=t,l("CMUpgradeBarYellow").textContent=o,l("CMUpgradeBarOrange").textContent=n,l("CMUpgradeBarRed").textContent=i,l("CMUpgradeBarPurple").textContent=a,l("CMUpgradeBarGray").textContent=r}const t=[];for(let e=0;e<Game.UpgradesInStore.length;e+=1){const o={};o.name=Game.UpgradesInStore[e].name,o.price=Game.UpgradesInStore[e].basePrice,o.pp=St[o.name].pp,t.push(o)}e.SortUpgrades?t.sort((function(e,t){return $.indexOf(e.color)>$.indexOf(t.color)?1:$.indexOf(e.color)<$.indexOf(t.color)||e.pp<t.pp?-1:0})):t.sort(((e,t)=>e.price-t.price));const o=function(e,t){return e.findIndex((e=>e.name===t.name))};for(let e=0;e<Game.UpgradesInStore.length;e+=1)l(`upgrade${e}`).style.order=o(t,Game.UpgradesInStore[e])+1}function mo(){return e.CPSMode?Ee:0===e.CalcWrink?Game.cookiesPs*(1-Game.cpsSucked):1===e.CalcWrink?Game.cookiesPs*(Tt+(1-.05*Mt)):null!==xt[1]&&2===e.CalcWrink&&1===Game.wrinklers[xt[1]].type?Game.cookiesPs*(3*Tt/Mt+(1-.05*Mt)):Game.cookiesPs*(Tt/Mt+(1-.05*Mt))}function ho(){return 1===e.CalcWrink?vt:2===e.CalcWrink?xt[0]:0}function go(e,t,o){const n=document.createElement("div");n.id=e,n.style.height="12px",n.style.margin="0px 10px",n.style.position="relative";const i=document.createElement("div");i.style.width="100%",i.style.height="10px",i.style.margin="auto",i.style.position="absolute",i.style.left="0px",i.style.top="0px",i.style.right="0px",i.style.bottom="0px";const a=document.createElement("span");a.style.display="inline-block",a.style.textAlign="right",a.style.fontSize="10px",a.style.width="108px",a.style.marginRight="5px",a.style.verticalAlign="text-top",a.textContent=t,i.appendChild(a);for(let e=0;e<o.length;e++){const t=document.createElement("span");t.id=o[e].id,t.style.display="inline-block",t.style.height="10px",t.style.verticalAlign="text-top",t.style.textAlign="center",o.length-1===e&&(t.style.borderTopRightRadius="10px",t.style.borderBottomRightRadius="10px"),void 0!==o[e].color&&(t.className=N+o[e].color),i.appendChild(t)}const r=document.createElement("span");return r.id=`${e}Time`,r.style.marginLeft="5px",r.style.verticalAlign="text-top",i.appendChild(r),n.appendChild(i),n}function fo(e){if(null!==l("CMBotBar")){const t=l("CMBotBar").firstChild.firstChild.childNodes[0],o=l("CMBotBar").firstChild.firstChild.childNodes[1],n=l("CMBotBar").firstChild.firstChild.childNodes[2],i=l("CMBotBar").firstChild.firstChild.childNodes[3],a=e,r=t.appendChild(document.createElement("td"));r.appendChild(document.createTextNode(`${-1!==a.indexOf(" ")?a.substring(0,a.indexOf(" ")):a} (`)),r.appendChild(document.createElement("span")).className="CMTextBlue",r.appendChild(document.createTextNode(")")),t.lastChild.style.paddingLeft="8px",o.appendChild(document.createElement("td")),o.lastChild.style.paddingLeft="8px",n.appendChild(document.createElement("td")),n.lastChild.style.paddingLeft="8px",i.appendChild(document.createElement("td")),i.lastChild.style.paddingLeft="2px"}}function Co(){if(1===e.BotBar&&yt&&1===Game.buyMode){let t=0;Object.keys(yt).forEach((o=>{let n,i=Game.buyBulk;1===Game.buyMode?q=i:i=q,1===i&&(i=yt),10===i&&(i=bt),100===i&&(i=kt),t+=1,l("CMBotBar").firstChild.firstChild.childNodes[0].childNodes[t].childNodes[1].textContent=Game.Objects[o].amount,l("CMBotBar").firstChild.firstChild.childNodes[1].childNodes[t].textContent=pe(i[o].bonus,2),l("CMBotBar").firstChild.firstChild.childNodes[2].childNodes[t].className=E+i[o].color,n=e.PPDisplayTime?ue(Math.round(i[o].pp)):pe(Math.round(i[o].pp),2),l("CMBotBar").firstChild.firstChild.childNodes[2].childNodes[t].textContent=n;const a=me((Game.Objects[o].bulkPrice-(Game.cookies+ho()))/mo());l("CMBotBar").firstChild.firstChild.childNodes[3].childNodes[t].className=E+a.color,"Done!"===a.text&&Game.cookies<Game.Objects[o].bulkPrice?l("CMBotBar").firstChild.firstChild.childNodes[3].childNodes[t].textContent=`${a.text} (with Wrink)`:l("CMBotBar").firstChild.firstChild.childNodes[3].childNodes[t].textContent=a.text}))}}function yo(){Game.Background.canvas.width=Game.Background.canvas.parentNode.offsetWidth,Game.Background.canvas.height=Game.Background.canvas.parentNode.offsetHeight,Game.LeftBackground.canvas.width=Game.LeftBackground.canvas.parentNode.offsetWidth,Game.LeftBackground.canvas.height=Game.LeftBackground.canvas.parentNode.offsetHeight,Game.DrawBackground()}function bo(){1===e.BotBar&&1===e.TimerBar&&1===e.TimerBarPos?(l("CMBotBar").style.bottom=l("CMTimerBar").style.height,l("game").style.bottom=`${Number(l("CMTimerBar").style.height.replace("px",""))+70}px`):1===e.BotBar?(l("CMBotBar").style.bottom="0px",l("game").style.bottom="70px"):1===e.TimerBar&&1===e.TimerBarPos?l("game").style.bottom=l("CMTimerBar").style.height:l("game").style.bottom="0px",1===e.TimerBar&&0===e.TimerBarPos?l("sectionLeft").style.top=l("CMTimerBar").style.height:l("sectionLeft").style.top="",yo()}function ko(){1===e.TimerBar?l("CMTimerBar").style.display="":l("CMTimerBar").style.display="none",bo()}function Go(){Ye=Game.dragonAura,Ke=Game.dragonAura2}function vo(e){const t=Game.Achievements[e],o={};return o.name=t.name,o}function wo(e){let t=1;if(Object.keys(e.tieredUpgrades).forEach((o=>{!Game.Tiers[e.tieredUpgrades[o].tier].special&&ro(e.tieredUpgrades[o].name)&&(t*=2)})),Object.keys(e.synergies).forEach((o=>{if(ro(e.synergies[o].name)){const n=e.synergies[o];n.buildingTie1.name===e.name?t*=1+.05*n.buildingTie2.amount:n.buildingTie2.name===e.name&&(t*=1+.001*n.buildingTie1.amount)}})),e.fortune&&ro(e.fortune.name)&&(t*=1.07),e.grandma&&ro(e.grandma.name)&&(t*=1+.01*eo.Grandma.amount*(1/(e.id-1))),"object"==typeof e.tieredUpgrades.misfortune&&1===e.vanilla&&ro(e.tieredUpgrades.misfortune.name))switch(Game.elderWrath){default:t*=1;break;case 1:t*=1.02;break;case 2:t*=1.04;break;case 3:t*=1.06}return t}function xo(e){const t=Game.Objects[e],o={};return"Cursor"===t.name?o.cps=function(e){let t=0;ro("Thousand fingers")&&(t+=.1),ro("Million fingers")&&(t*=5),ro("Billion fingers")&&(t*=10),ro("Trillion fingers")&&(t*=20),ro("Quadrillion fingers")&&(t*=20),ro("Quintillion fingers")&&(t*=20),ro("Sextillion fingers")&&(t*=20),ro("Septillion fingers")&&(t*=20),ro("Octillion fingers")&&(t*=20),ro("Nonillion fingers")&&(t*=20);let o=1,n=0;return Object.keys(eo).forEach((e=>{"Cursor"!==eo[e].name&&(n+=eo[e].amount)})),t*=n,o*=wo(e),o*=Game.magicCpS("Cursor"),o*=ao("cursorCps"),Game.ComputeCps(.1,ro("Reinforced index finger")+ro("Carpal tunnel prevention cream")+ro("Ambidextrous"),t)*o}:"Grandma"===t.name?o.cps=function(e){let o=1;if(Object.keys(Game.GrandmaSynergies).forEach((e=>{ro(Game.GrandmaSynergies[e])&&(o*=2)})),ro("Bingo center/Research facility")&&(o*=4),ro("Ritual rolling pins")&&(o*=2),ro("Naughty list")&&(o*=2),ro("Elderwort biscuits")&&(o*=1.02),o*=ao("grandmaCps"),ro("Cat ladies"))for(let e=0;e<Game.UpgradesByPool.kitten.length;e++)ro(Game.UpgradesByPool.kitten[e].name)&&(o*=1.29);o*=wo(e);let n=0;ro("One mind")&&(n+=.02*eo.Grandma.amount),ro("Communal brainsweep")&&(n+=.02*eo.Grandma.amount),ro("Elder Pact")&&(n+=.05*eo.Portal.amount);let i=0;return Object.keys(eo).forEach((e=>{"Grandma"!==eo[e].name&&(i+=eo[e].amount)})),o*=1+.01*no("Elder Battalion")*i,o*=Game.magicCpS(t.name),(t.baseCps+n)*o}:o.cps=function(e){let t=1;return t*=wo(e),t*=Game.magicCpS(e.name),e.baseCPS*t},o.baseCps=t.baseCps,o.name=t.name,o.tieredUpgrades=t.tieredUpgrades,o.synergies=t.synergies,o.fortune=t.fortune,o.grandma=t.grandma,o.baseCPS=t.baseCps,o.id=t.id,o.vanilla=t.vanilla,o}function To(e){const t=Game.Upgrades[e],o={};return o.power=t.power,"function"==typeof o.power&&("Sugar crystal cookies"===t.name?o.power=function(){let e=5;return Object.keys(eo).forEach((t=>{eo[t].level>=10&&(e+=1)})),e}:o.power=function(){let e=2;if(ro("Starlove")&&(e=3),Game.hasGod){const t=lo("seasons");1===t?e*=1.3:2===t?e*=1.2:3===t&&(e*=1.1)}return e}),o.pool=t.pool,o.name=t.name,o}function Mo(){Ae=Game.UpgradesOwned,je=Game.pledges,De=Game.AchievementsOwned,He=Game.heavenlyPower,Ue=Game.prestige,Object.keys(Game.Objects).forEach((e=>{const t=Game.Objects[e];let o=eo[e];void 0===o&&(eo[e]=xo(e),o=eo[e],fo(e)),o.amount=t.amount,o.level=t.level,o.totalCookies=t.totalCookies,o.basePrice=t.basePrice,o.free=t.free,t.minigameLoaded&&("Temple"===t.name&&(Ie=t.minigame.slot[0],ze=t.minigame.slot[1],Ve=t.minigame.slot[2]),o.minigameLoaded=t.minigameLoaded,o.minigame=t.minigame),eo[e]=o})),Object.keys(Game.Upgrades).forEach((e=>{const t=Game.Upgrades[e];let o=to[e];void 0===o&&(to[e]=To(e),o=to[e]),o.bought=t.bought,to[e]=o})),Object.keys(Game.Achievements).forEach((e=>{const t=Game.Achievements[e];let o=oo[e];void 0===o&&(oo[e]=vo(e),o=oo[e]),o.won=t.won,oo[e]=o})),Go(),$e=Ye,Re=Ke}function So(e){const t=document.createElement("div");return t.style.fontWeight="bold",t.id=`${e}Title`,t.className="CMTextBlue",t.textContent=e,t}function Po(e){e.appendChild(So("Bonus Income"));const t=document.createElement("div");t.style.marginBottom="4px",t.style.color="white",t.id="CMTooltipIncome",e.appendChild(t),e.appendChild(So("Bonus Cookies per Click")),e.lastChild.style.display="none";const o=document.createElement("div");o.style.marginBottom="4px",o.style.color="white",o.style.display="none",o.id="CMTooltipCookiePerClick",e.appendChild(o),e.appendChild(So("Payback Period"));const n=document.createElement("div");n.style.marginBottom="4px",n.id="CMTooltipPP",e.appendChild(n),e.appendChild(So("Time Left"));const i=document.createElement("div");if(i.id="CMTooltipTime",e.appendChild(i),"b"===ne){e.appendChild(So("Production left till next achievement")),e.lastChild.id="CMTooltipProductionLeftHeader";const t=document.createElement("div");t.id="CMTooltipProductionLeft",e.appendChild(t)}if("b"===ne){e.appendChild(So("Buildings (price / PP) left till next achievement")),e.lastChild.id="CMTooltipNextAchievementHeader";const t=document.createElement("div");t.id="CMTooltipNextAchievement",e.appendChild(t)}}function Bo(t,o){let n="";return n=t.pp<=0||t.pp===1/0?D:t.pp<qt?O:t.pp===qt?F:t.pp<Yt[10][0]?L:t.pp<Yt[20][0]?W:t.pp<Yt[30][0]?A:j,0!==Number(e.PPSecondsLowerLimit)&&o/mo()<Number(e.PPSecondsLowerLimit)&&(n=O),e.PPOnlyConsiderBuyable&&o-Game.cookies>0&&(n=A),n}function Eo(){let e=1;return Object.keys(Game.buffs).forEach((t=>{void 0!==Game.buffs[t].multCpS&&(e*=Game.buffs[t].multCpS)})),e}function No(e){oo[e]&&0===oo[e].won&&(oo[e].won=1,"shadow"!==Game.Achievements[e].pool&&(De+=1))}function Oo(){Xe=0;let e=1;const t={};Object.keys(Game.Objects).forEach((e=>{if(Game.Objects[e].minigameLoaded&&Game.Objects[e].minigame.effs){const o=Game.Objects[e].minigame.effs;Object.keys(o).forEach((e=>{t[e]?t[e]*=o[e]:t[e]=o[e]}))}})),Qe=t,1!==Game.ascensionMode&&(e+=.01*parseFloat(Ue)*He*function(){let e=0;if(ro("Heavenly chip secret")&&(e+=.05),ro("Heavenly cookie stand")&&(e+=.2),ro("Heavenly bakery")&&(e+=.25),ro("Heavenly confectionery")&&(e+=.25),ro("Heavenly key")&&(e+=.25),e*=1+.05*no("Dragon God"),ro("Lucky digit")&&(e*=1.01),ro("Lucky number")&&(e*=1.01),ro("Lucky payout")&&(e*=1.01),Game.hasGod){const t=lo("creation");1===t?e*=.7:2===t?e*=.8:3===t&&(e*=.9)}return e}()),e*=ao("cps"),ro("Heralds")&&1!==Game.ascensionMode&&(e*=1+.01*Game.heralds),Object.keys(Game.cookieUpgrades).forEach((t=>{const o=Game.cookieUpgrades[t];ro(o.name)&&("function"==typeof o.power?e*=1+.01*to[o.name].power(to[o.name]):e*=1+.01*o.power)})),ro("Specialized chocolate chips")&&(e*=1.01),ro("Designer cocoa beans")&&(e*=1.02),ro("Underworld ovens")&&(e*=1.03),ro("Exotic nuts")&&(e*=1.04),ro("Arcane sugar")&&(e*=1.05),ro("Increased merriness")&&(e*=1.15),ro("Improved jolliness")&&(e*=1.15),ro("A lump of coal")&&(e*=1.01),ro("An itchy sweater")&&(e*=1.01),ro("Santa's dominion")&&(e*=1.2),ro("Fortune #100")&&(e*=1.01),ro("Fortune #101")&&(e*=1.07),ro("Dragon scale")&&(e*=1.03);let o=1;if(lo){let t=lo("asceticism");1===t?e*=1.15:2===t?e*=1.1:3===t&&(e*=1.05),t=lo("ages"),1===t?e*=1+.15*Math.sin(P/1e3/10800*Math.PI*2):2===t?e*=1+.15*Math.sin(P/1e3/43200*Math.PI*2):3===t&&(e*=1+.15*Math.sin(P/1e3/86400*Math.PI*2)),t=lo("decadence"),1===t?o*=.93:2===t?o*=.95:3===t&&(o*=.98),t=lo("industry"),1===t?o*=1.1:2===t?o*=1.06:3===t&&(o*=1.03),t=lo("labor"),1===t?o*=.97:2===t?o*=.98:3===t&&(o*=.99)}ro("Santa's legacy")&&(e*=1+.03*(Game.santaLevel+1));const n=De/25;let i=1;if(ro("Santa's milk and cookies")&&(i*=1.05),i*=1+.05*no("Breath of Milk"),lo){const e=lo("mother");1===e?i*=1.1:2===e?i*=1.05:3===e&&(i*=1.03)}i*=ao("milk");let a=1;ro("Kitten helpers")&&(a*=1+.1*n*i),ro("Kitten workers")&&(a*=1+.125*n*i),ro("Kitten engineers")&&(a*=1+.15*n*i),ro("Kitten overseers")&&(a*=1+.175*n*i),ro("Kitten managers")&&(a*=1+.2*n*i),ro("Kitten accountants")&&(a*=1+.2*n*i),ro("Kitten specialists")&&(a*=1+.2*n*i),ro("Kitten experts")&&(a*=1+.2*n*i),ro("Kitten consultants")&&(a*=1+.2*n*i),ro("Kitten assistants to the regional manager")&&(a*=1+.175*n*i),ro("Kitten marketeers")&&(a*=1+.15*n*i),ro("Kitten analysts")&&(a*=1+.125*n*i),ro("Kitten executives")&&(a*=1+.115*n*i),ro("Kitten angels")&&(a*=1+.1*n*i),ro("Fortune #103")&&(a*=1+.05*n*i),Object.keys(eo).forEach((e=>{const t=eo[e];let a=t.cps(t);1!==Game.ascensionMode&&(a*=(1+.01*t.level)*o),"Grandma"===t.name&&ro("Milkhelp&reg; lactose intolerance relief tablets")&&(a*=1+.05*n*i),Xe+=t.amount*a})),ro('"egg"')&&(Xe+=9),e*=a;let r=1;if(ro("Chicken egg")&&(r*=1.01),ro("Duck egg")&&(r*=1.01),ro("Turkey egg")&&(r*=1.01),ro("Quail egg")&&(r*=1.01),ro("Robin egg")&&(r*=1.01),ro("Ostrich egg")&&(r*=1.01),ro("Cassowary egg")&&(r*=1.01),ro("Salmon roe")&&(r*=1.01),ro("Frogspawn")&&(r*=1.01),ro("Shark egg")&&(r*=1.01),ro("Turtle egg")&&(r*=1.01),ro("Ant larva")&&(r*=1.01),ro("Century egg")){let e=10*Math.floor((B-Game.startDate)/1e3/10)/60/60/24;e=Math.min(e,100),gt=1+.1*(1-(1-e/100)**3),r*=gt}e*=r,ro("Sugar baking")&&(e*=1+.01*Math.min(100,Game.lumps)),e*=1+no("Radiant Appetite");const l=Xe*e;Object.keys(Game.CpsAchievements).forEach((e=>{l>=Game.CpsAchievements[e].threshold&&No(Game.CpsAchievements[e].name)})),qe=l;const{n:s}=Game.shimmerTypes.golden,c=no("Dragon's Fortune");for(let t=0;t<s;t++)e*=1+1.23*c;const d=Game.bakeryName.toLowerCase();if("orteil"===d?e*=.99:"ortiel"===d&&(e*=.98),ro("Elder Covenant")&&(e*=.95),ro("Golden switch [off]")){let t=1.5;if(ro("Residual luck")){const e=Game.goldenCookieUpgrades;Object.keys(e).forEach((o=>{ro(e[o])&&(t+=.1)}))}e*=t}if(ro("Shimmering veil [off]")){let t=.5;ro("Reinforced membrane")&&(t+=.1),e*=1+t}ro("Magic shenanigans")&&(e*=1e3),ro("Occult obstruction")&&(e*=0),Xe=Game.runModHookOnValue("cps",Xe),e*=Eo(),Xe*=e}const Fo=["Fortune #001","Fortune #002","Fortune #003","Fortune #004","Fortune #005","Fortune #006","Fortune #007","Fortune #008","Fortune #009","Fortune #010","Fortune #011","Fortune #012","Fortune #013","Fortune #014","Fortune #015","Fortune #016","Fortune #017","Fortune #018","Fortune #100","Fortune #101","Fortune #102","Fortune #103","Fortune #104"],Lo=["Skull cookies","Ghost cookies","Bat cookies","Slime cookies","Pumpkin cookies","Eyeball cookies","Spider cookies"],Wo=["Christmas tree biscuits","Snowflake biscuits","Snowman biscuits","Holly biscuits","Candy cane biscuits","Bell biscuits","Present biscuits"],Ao=["Pure heart biscuits","Ardent heart biscuits","Sour heart biscuits","Weeping heart biscuits","Golden heart biscuits","Eternal heart biscuits","Prism heart biscuits"],jo=["Elderwort biscuits","Bakeberry cookies","Duketater cookies","Green yeast digestives","Wheat slims","Fern tea","Ichor syrup"];function Do(e){return oo[e]?oo[e].won:0}function Ho(){let e=0;Object.keys(Game.GrandmaSynergies).forEach((t=>{ro(Game.GrandmaSynergies[t])&&(e+=1)})),!Do("Elder")&&e>=7&&No("Elder"),!Do("Veteran")&&e>=14&&No("Veteran");let t=0,o=1,n=1,i=1e5;Object.keys(eo).forEach((e=>{t+=eo[e].amount,i=Math.min(eo[e].amount,i),Do("Mathematician")||eo[e].amount<Math.min(128,2**(Game.ObjectsById.length-Game.Objects[e].id-1))&&(o=0),Do("Base 10")||eo[e].amount<10*(Game.ObjectsById.length-Game.Objects[e].id)&&(n=0)})),i>=1&&No("One with everything"),1===o&&No("Mathematician"),1===n&&No("Base 10"),i>=100&&No("Centennial"),i>=150&&No("Centennial and a half"),i>=200&&No("Bicentennial"),i>=250&&No("Bicentennial and a half"),i>=300&&No("Tricentennial"),i>=350&&No("Tricentennial and a half"),i>=400&&No("Quadricentennial"),i>=450&&No("Quadricentennial and a half"),i>=500&&No("Quincentennial"),i>=550&&No("Quincentennial and a half"),i>=600&&No("Sexcentennial"),t>=100&&No("Builder"),t>=500&&No("Architect"),t>=1e3&&No("Engineer"),t>=2e3&&No("Lord of Constructs"),t>=4e3&&No("Grand design"),t>=8e3&&No("Ecumenopolis"),Ae>=20&&No("Enhancer"),Ae>=50&&No("Augmenter"),Ae>=100&&No("Upgrader"),Ae>=200&&No("Lord of Progress"),Ae>=300&&No("The full picture"),Ae>=400&&No("When there's nothing left to add"),t>=4e3&&Ae>=300&&No("Polymath"),t>=8e3&&Ae>=400&&No("Renaissance baker"),eo.Cursor.amount+eo.Grandma.amount>=777&&No("The elder scrolls");let a=!0;Object.keys(Lo).forEach((e=>{ro(Lo[e])||(a=!1)})),a&&No("Spooky cookies");let r=!0;if(Object.keys(Wo).forEach((e=>{ro(Wo[e])||(r=!1)})),r&&No("Let it snow"),ro("Fortune cookies")){const e=Game.Tiers.fortune.upgrades;let t=0;Object.keys(e).forEach((o=>{ro(e[o].name)&&(t+=1)})),t>=e.length&&No("O Fortuna")}}function Uo(e,t){Mo(),eo[e].amount+=t;const o=eo[e];"Cursor"===e?(o.amount>=1&&No("Click"),o.amount>=2&&No("Double-click"),o.amount>=50&&No("Mouse wheel"),o.amount>=100&&No("Of Mice and Men"),o.amount>=200&&No("The Digital"),o.amount>=300&&No("Extreme polydactyly"),o.amount>=400&&No("Dr. T"),o.amount>=500&&No("Thumbs, phalanges, metacarpals"),o.amount>=600&&No("With her finger and her thumb"),o.amount>=700&&No("Gotta hand it to you"),o.amount>=800&&No("The devil's workshop")):Object.keys(Game.Objects[o.name].tieredAchievs).forEach((e=>{o.amount>=Game.Tiers[Game.Objects[o.name].tieredAchievs[e].tier].achievUnlock&&No(Game.Objects[o.name].tieredAchievs[e].name)}));const n=De;return Oo(),Ho(),n!==De&&Oo(),Xe-Game.cookiesPs}function $o(e,t,o){let n=e,i=0;for(;n<o;)n+=.002*Math.max(.002,(n/Math.max(t,100))**.5),i+=1;return i/Game.fps}function Ro(){null!==l("CMDispTooltipWarningParent")&&(0===e.ToolWarnPos?(l("CMDispTooltipWarningParent").style.top="auto",l("CMDispTooltipWarningParent").style.margin="4px -4px",l("CMDispTooltipWarningParent").style.padding="3px 4px"):(l("CMDispTooltipWarningParent").style.right="auto",l("CMDispTooltipWarningParent").style.margin="4px",l("CMDispTooltipWarningParent").style.padding="4px 3px"))}function Io(){if(Mo(),"none"!==l("tooltipAnchor").style.display&&l("CMTooltipArea")){l("CMTooltipArea").innerHTML="";const t=function(){l("tooltip").firstChild.style.paddingBottom="4px";const e=document.createElement("div");return e.style.border="1px solid",e.style.padding="4px",e.style.margin="0px -4px",e.id="CMTooltipBorder",e.className="CMTextGray",e}();l("CMTooltipArea").appendChild(t),"b"===ne?function(){if(1===e.TooltipBuildUpgrade&&1===Game.buyMode){let t;if(Po(l("CMTooltipBorder")),1===Game.buyMode?K=t:t=K,1===Game.buyBulk?t=yt:10===Game.buyBulk?t=bt:100===Game.buyBulk&&(t=kt),te=Game.Objects[ie].bulkPrice,oe=t[ie].bonus,1===e.TooltipBuildUpgrade&&1===Game.buyMode){l("CMTooltipIncome").textContent=pe(oe,2);const o=Math.round(oe/Game.cookiesPs*1e4);Number.isFinite(o)&&0!==o?l("CMTooltipIncome").textContent+=` (${o/100}% of income)`:l("CMTooltipIncome").textContent+=` (<0${e.ScaleSeparator?",":"."}01% of income)`,l("CMTooltipBorder").className=E+t[ie].color,e.PPDisplayTime?l("CMTooltipPP").textContent=ue(t[ie].pp):l("CMTooltipPP").textContent=pe(t[ie].pp,2),l("CMTooltipPP").className=E+t[ie].color;const n=me((te-(Game.cookies+ho()))/mo());l("CMTooltipTime").textContent=n.text,"Done!"===n.text&&Game.cookies<t[ie].price?l("CMTooltipTime").textContent=`${n.text} (with Wrink)`:l("CMTooltipTime").textContent=n.text,l("CMTooltipTime").className=E+n.color}l("CMTooltipProductionLeftHeader").style.display="none",l("CMTooltipTime").style.marginBottom="0px";for(const e of Object.keys(Game.Objects[ie].productionAchievs))if(!Game.HasAchiev(Game.Objects[ie].productionAchievs[e].achiev.name)){const t=Game.Objects[ie].productionAchievs[e];l("CMTooltipTime").style.marginBottom="4px",l("CMTooltipProductionLeftHeader").style.display="",l("CMTooltipProductionLeft").className=`ProdAchievement${ie}`,l("CMTooltipProductionLeft").textContent=pe(t.pow-eo[ie].totalCookies,15),l("CMTooltipProductionLeft").style.color="white";break}const o=Gt[ie];if(o.AmountNeeded<101){let t;l("CMTooltipProductionLeft").style.marginBottom="4px",l("CMTooltipNextAchievementHeader").style.display="",t=Game.cookiesPs?Math.max(o.price-(Game.cookies+ho()),0)/Game.cookiesPs+o.price/Uo(ie,o.AmountNeeded):o.price/Uo(ie,o.AmountNeeded),l("CMTooltipNextAchievement").textContent=`${pe(o.AmountNeeded)} / ${pe(o.price)} / `,l("CMTooltipNextAchievement").style.color="white";const n=document.createElement("span");e.PPDisplayTime?n.textContent=ue(t):n.textContent=pe(t),n.className=E+Bo({pp:t},o.price),l("CMTooltipNextAchievement").appendChild(n)}else l("CMTooltipNextAchievementHeader").style.display="none",l("CMTooltipProductionLeft").style.marginBottom="0px"}else l("CMTooltipArea").style.display="none"}():"u"===ne?function(){if(Po(l("CMTooltipBorder")),oe=St[Game.UpgradesInStore[ie].name].bonus,te=Game.Upgrades[Game.UpgradesInStore[ie].name].getPrice(),ae=St[Game.UpgradesInStore[ie].name].bonusMouse,1===e.TooltipBuildUpgrade){l("CMTooltipIncome").textContent=pe(oe,2);const t=Math.round(oe/Game.cookiesPs*1e4);"0"!==l("CMTooltipIncome").textContent||"b"!==ne&&"u"!==ne?(Number.isFinite(t)&&0!==t?l("CMTooltipIncome").textContent+=` (${t/100}% of income)`:l("CMTooltipIncome").textContent+=` (<0${e.ScaleSeparator?",":"."}01% of income)`,l("CMTooltipBorder").className=E+St[Game.UpgradesInStore[ie].name].color,ae&&(l("CMTooltipCookiePerClick").textContent=pe(ae),l("CMTooltipCookiePerClick").style.display="block",l("CMTooltipCookiePerClick").previousSibling.style.display="block"),0===oe&&ae?(l("CMTooltipPP").textContent=`${pe(te/ae)} Clicks`,l("CMTooltipPP").style.color="white"):(e.PPDisplayTime?l("CMTooltipPP").textContent=ue(St[Game.UpgradesInStore[ie].name].pp):l("CMTooltipPP").textContent=pe(St[Game.UpgradesInStore[ie].name].pp,2),l("CMTooltipPP").className=E+St[Game.UpgradesInStore[ie].name].color)):(l("Bonus IncomeTitle").style.display="none",l("CMTooltipIncome").style.display="none",l("Payback PeriodTitle").style.display="none",l("CMTooltipPP").style.display="none");const o=me((te-(Game.cookies+ho()))/mo());if(l("CMTooltipTime").textContent=o.text,"Done!"===o.text&&Game.cookies<Game.UpgradesInStore[ie].getPrice()?l("CMTooltipTime").textContent=`${o.text} (with Wrink)`:l("CMTooltipTime").textContent=o.text,l("CMTooltipTime").className=E+o.color,"Chocolate egg"===Game.UpgradesInStore[ie].name){l("CMTooltipBorder").lastChild.style.marginBottom="4px",l("CMTooltipBorder").appendChild(So("Cookies to be gained (Currently/Max)"));const e=document.createElement("div");e.style.color="white",e.textContent=`${pe(.05*Game.cookies)} / ${pe(Ct)}`,l("CMTooltipBorder").appendChild(e)}}else l("CMTooltipArea").style.display="none"}():"s"===ne?function(){if(1===e.TooltipLump){const e=l("CMTooltipBorder");e.appendChild(So("Current Sugar Lump"));const o=document.createElement("div");o.id="CMTooltipTime",e.appendChild(o);const n=0===(t=Game.lumpCurrentType)?{text:"Normal",color:D}:1===t?{text:"Bifurcated",color:F}:2===t?{text:"Golden",color:L}:3===t?{text:"Meaty",color:W}:4===t?{text:"Caramelized",color:j}:{text:"Unknown Sugar Lump",color:A};o.textContent=n.text,o.className=E+n.color}else l("CMTooltipArea").style.display="none";var t}():"g"===ne?function(){const{minigame:t}=Game.Objects["Wizard tower"],o=t.getSpellCost(t.spellsById[ie]);if(1===e.TooltipGrim&&o<=t.magicM){const e=l("CMTooltipBorder");e.appendChild(So("Time Left"));const n=document.createElement("div");n.id="CMTooltipTime",e.appendChild(n);const i=me($o(t.magic,t.magicM,o));if(n.textContent=i.text,n.className=E+i.color,o<=t.magic){e.appendChild(So("Recover Time"));const n=document.createElement("div");n.id="CMTooltipRecover",e.appendChild(n);const i=me($o(Math.max(0,t.magic-o),t.magicM,t.magic));n.textContent=i.text,n.className=E+i.color}if("0"===ie){e.appendChild(So("Cookies to be gained/lost"));const t=document.createElement("div");t.id="x",e.appendChild(t);const o=document.createElement("span");o.style.color="#33FF00",o.textContent=pe(Math.min(.15*(Game.cookies+ho()),60*dt*30),2),t.appendChild(o);const n=document.createElement("span");n.textContent=" / ",t.appendChild(n);const i=document.createElement("span");i.style.color="red",i.textContent=pe(60*dt*15,2),t.appendChild(i)}l("CMTooltipArea").appendChild(e)}else l("CMTooltipArea").style.display="none"}():"p"===ne?function(){const{minigame:t}=Game.Objects.Farm;if(e.TooltipPlots&&0!==t.plot[ie[1]][ie[0]][0]){const e=t.plot[ie[1]][ie[0]][1]>t.plantsById[t.plot[ie[1]][ie[0]][0]-1].mature,o=t.plantsById[t.plot[ie[1]][ie[0]][0]-1].name;l("CMTooltipBorder").appendChild(So("Reward (Current / Maximum)"));const n=document.createElement("div");n.id="CMTooltipPlantReward",l("CMTooltipBorder").appendChild(n),"Bakeberry"===o?l("CMTooltipPlantReward").textContent=`${e?pe(Math.min(.03*Game.cookies,60*Game.cookiesPs*30)):"0"} / ${pe(60*Game.cookiesPs*30)}`:"Chocoroot"===o||"White chocoroot"===o?l("CMTooltipPlantReward").textContent=`${e?pe(Math.min(.03*Game.cookies,60*Game.cookiesPs*3)):"0"} / ${pe(60*Game.cookiesPs*3)}`:"Queenbeet"===o?l("CMTooltipPlantReward").textContent=`${e?pe(Math.min(.04*Game.cookies,60*Game.cookiesPs*60)):"0"} / ${pe(60*Game.cookiesPs*60)}`:"Duketater"===o?l("CMTooltipPlantReward").textContent=`${e?pe(Math.min(.08*Game.cookies,60*Game.cookiesPs*120)):"0"} / ${pe(60*Game.cookiesPs*120)}`:l("CMTooltipArea").style.display="none"}else l("CMTooltipArea").style.display="none"}():"ha"===ne?function(){const{minigame:t}=Game.Objects.Farm;if(e.TooltipLump){l("CMTooltipBorder").appendChild(So("Cookies gained from harvesting:"));let e=0,o=0;Game.keys[16]&&Game.keys[17]&&(o=1);for(let n=0;n<6;n++)for(let i=0;i<6;i++)if(t.plot[n][i][0]>=1){const a=t.plot[n][i],r=t.plantsById[a[0]-1],l=r.name;let s=!0;o&&r.immortal&&(s=!1),a[1]<r.matureBase&&(s=!1),s&&"Bakeberry"===l?e+=Math.min(.03*Game.cookies,60*Game.cookiesPs*30):s&&"Chocoroot"===l||"White chocoroot"===l?e+=Math.min(.03*Game.cookies,60*Game.cookiesPs*3):s&&"Queenbeet"===l?e+=Math.min(.04*Game.cookies,60*Game.cookiesPs*60):s&&"Duketater"===l&&(e+=Math.min(.08*Game.cookies,60*Game.cookiesPs*120))}l("CMTooltipBorder").appendChild(document.createTextNode(pe(e)))}else l("CMTooltipArea").style.display="none"}():"wb"===ne?function(){l("tooltip").innerHTML="",l("tooltip").appendChild(So("Reward:"));const e=document.createElement("div");e.id="CMWrinklerReward","PopAll"===ie?e.textContent=pe(vt):"PopFattest"===ie&&(e.textContent=pe(xt[0])),l("tooltip").appendChild(e)}():("pag"===ne||"pas"===ne&&-1!==ie[1])&&function(){if(1===e.TooltipPantheon){const t=l("CMTooltipBorder");let o;o="pas"===ne?ie[1]:ie,t.appendChild(So("Effect in position 1:"));const n=document.createElement("div");if(n.id="CMPantheonTooltipPosition1",0!==Zt[o][0]){n.textContent=pe(Zt[o][0]);const t=Math.round(Zt[o][0]/Game.cookiesPs*1e4);Number.isFinite(t)&&0!==t?n.textContent+=` (${t/100}% of income)`:n.textContent+=` (<0${e.ScaleSeparator?",":"."}01% of income)`}else n.textContent="No effect to CPS";t.appendChild(n),t.appendChild(So("Effect in position 2:"));const i=document.createElement("div");if(i.id="CMPantheonTooltipPosition2",0!==Zt[o][1]){i.textContent=pe(Zt[o][1]);const t=Math.round(Zt[o][1]/Game.cookiesPs*1e4);Number.isFinite(t)&&0!==t?i.textContent+=` (${t/100}% of income)`:i.textContent+=` (<0${e.ScaleSeparator?",":"."}01% of income)`}else i.textContent="No effect to CPS";t.appendChild(i),t.appendChild(So("Effect in position 3:"));const a=document.createElement("div");if(a.id="CMPantheonTooltipPosition2",0!==Zt[o][2]){a.textContent=pe(Zt[o][2]);const t=Math.round(Zt[o][2]/Game.cookiesPs*1e4);Number.isFinite(t)&&0!==t?a.textContent+=` (${t/100}% of income)`:a.textContent+=` (<0${e.ScaleSeparator?",":"."}01% of income)`}else a.textContent="No effect to CPS";t.appendChild(a),l("CMTooltipArea").appendChild(t)}else l("CMTooltipArea").style.display="none"}(),function(){if("b"===ne||"u"===ne){null===document.getElementById("CMDispTooltipWarningParent")&&(l("tooltipAnchor").appendChild(function(){const t=document.createElement("div");t.style.position="absolute",t.style.display="block",t.style.left="auto",t.style.bottom="auto",t.id="CMDispTooltipWarningParent";const o=function(e,t,o,n,i){const a=document.createElement("div");a.id=e,a.style.display="none",a.style.transition="opacity 0.1s ease-out",a.className="CMBorder"+t,a.style.padding="2px",a.style.background="#000 url(img/darkNoise.png)";const r=document.createElement("div");a.appendChild(r);const l=document.createElement("span");l.className=E+t,l.style.fontWeight="bold",l.textContent=o,r.appendChild(l),r.appendChild(document.createTextNode(n));const s=document.createElement("div");a.appendChild(s);const c=document.createElement("span");return c.id=i,s.appendChild(document.createTextNode("Deficit: ")),s.appendChild(c),a};return t.appendChild(o("CMDispTooltipWarnLucky",A,"Warning: ",'Purchase of this item will put you under the number of Cookies required for "Lucky!"',"CMDispTooltipWarnLuckyText")),t.firstChild.style.marginBottom="4px",t.appendChild(o("CMDispTooltipWarnLuckyFrenzy",L,"Warning: ",'Purchase of this item will put you under the number of Cookies required for "Lucky!" (Frenzy)',"CMDispTooltipWarnLuckyFrenzyText")),t.lastChild.style.marginBottom="4px",t.appendChild(o("CMDispTooltipWarnConjure",j,"Warning: ",'Purchase of this item will put you under the number of Cookies required for "Conjure Baked Goods"',"CMDispTooltipWarnConjureText")),t.lastChild.style.marginBottom="4px",t.appendChild(o("CMDispTooltipWarnConjureFrenzy",j,"Warning: ",'Purchase of this item will put you under the number of Cookies required for "Conjure Baked Goods" (Frenzy)',"CMDispTooltipWarnConjureFrenzyText")),t.lastChild.style.marginBottom="4px",t.appendChild(o("CMDispTooltipWarnEdifice",j,"Warning: ",'Purchase of this item will put you under the number of Cookies needed for "Spontaneous Edifice" to possibly give you your most expensive building"',"CMDispTooltipWarnEdificeText")),t.lastChild.style.marginBottom="4px",t.appendChild(o("CMDispTooltipWarnUser",A,"Warning: ",`Purchase of this item will put you under the number of Cookies equal to ${e.ToolWarnUser} seconds of CPS`,"CMDispTooltipWarnUserText")),t}()),Ro()),0===e.ToolWarnPos?l("CMDispTooltipWarningParent").style.right="0px":l("CMDispTooltipWarningParent").style.top=`${l("tooltip").offsetHeight}px`,l("CMDispTooltipWarningParent").style.width=l("tooltip").offsetWidth-6+"px";const t=Game.cookies+ho()-te,o=e.ToolWarnBon?oe:0;let n=et;if(1===e.ToolWarnBon){let e=oe;e/=Eo(),n+=60*e*15/.15}if(1===e.ToolWarnLucky&&t<n&&("b"!==ne||1===Game.buyMode)?(l("CMDispTooltipWarnLucky").style.display="",l("CMDispTooltipWarnLuckyText").textContent=`${pe(n-t)} (${ue((n-t)/(mo()+o))})`):l("CMDispTooltipWarnLucky").style.display="none",1===e.ToolWarnLuckyFrenzy){const e=7*n;t<e&&("b"!==ne||1===Game.buyMode)?(l("CMDispTooltipWarnLuckyFrenzy").style.display="",l("CMDispTooltipWarnLuckyFrenzyText").textContent=`${pe(e-t)} (${ue((e-t)/(mo()+o))})`):l("CMDispTooltipWarnLuckyFrenzy").style.display="none"}else l("CMDispTooltipWarnLuckyFrenzy").style.display="none";if(1===e.ToolWarnConjure){const e=2*n;t<e&&("b"!==ne||1===Game.buyMode)?(l("CMDispTooltipWarnConjure").style.display="",l("CMDispTooltipWarnConjureText").textContent=`${pe(e-t)} (${ue((e-t)/(mo()+o))})`):l("CMDispTooltipWarnConjure").style.display="none"}else l("CMDispTooltipWarnConjure").style.display="none";if(1===e.ToolWarnConjureFrenzy){const e=2*n*7;t<e&&("b"!==ne||1===Game.buyMode)?(l("CMDispTooltipWarnConjureFrenzy").style.display="",l("CMDispTooltipWarnConjureFrenzyText").textContent=`${pe(e-t)} (${ue((e-t)/(mo()+o))})`):l("CMDispTooltipWarnConjureFrenzy").style.display="none"}else l("CMDispTooltipWarnConjureFrenzy").style.display="none";1===e.ToolWarnEdifice&&Game.Objects["Wizard tower"].minigameLoaded&&st&&t<st&&("b"!==ne||1===Game.buyMode)?(l("CMDispTooltipWarnEdifice").style.display="",l("CMDispTooltipWarnEdificeText").textContent=`${pe(st-t)} (${ue((st-t)/(mo()+o))})`):l("CMDispTooltipWarnEdifice").style.display="none",e.ToolWarnUser>0&&t<e.ToolWarnUser*mo()&&("b"!==ne||1===Game.buyMode)?(l("CMDispTooltipWarnUser").style.display="",l("CMDispTooltipWarnUser").children[0].textContent=`Purchase of this item will put you under the number of Cookies equal to ${e.ToolWarnUser} seconds of CPS`,l("CMDispTooltipWarnUserText").textContent=`${pe(e.ToolWarnUser*mo()-t)} (${ue((e.ToolWarnUser*mo()-t)/(mo()+o))})`):l("CMDispTooltipWarnUser").style.display="none"}else null!==l("CMDispTooltipWarningParent")&&l("CMDispTooltipWarningParent").remove()}()}else null===l("CMTooltipArea")&&null!==l("CMDispTooltipWarningParent")&&l("CMDispTooltipWarningParent").remove()}function zo(){if(Game.prefs.autosave&&Game.drawT%10==0&&"stats"===Game.onMenu&&e.Stats){const e=document.getElementById("CMStatsAutosaveTimer");e&&(e.innerText=Game.sayTime(60*Game.fps-Game.T%(60*Game.fps),4))}if(po(),uo(),function(){if(1===e.TimerBar){const t=l("CMTimerBar").offsetWidth-163,o=l("CMTimerBar").offsetWidth-133;let n=0;0!==Game.shimmerTypes.golden.spawned||Game.Has("Golden switch [off]")?l("CMTimerBarGC").style.display="none":(l("CMTimerBarGC").style.display="",l("CMTimerBarGCMinBar").style.width=`${Math.round(Math.max(0,Game.shimmerTypes.golden.minTime-Game.shimmerTypes.golden.time)*t/Game.shimmerTypes.golden.maxTime)}px`,e.TimerBarOverlay>=1?l("CMTimerBarGCMinBar").textContent=Math.ceil((Game.shimmerTypes.golden.minTime-Game.shimmerTypes.golden.time)/Game.fps):l("CMTimerBarGCMinBar").textContent="",Game.shimmerTypes.golden.minTime===Game.shimmerTypes.golden.maxTime?(l("CMTimerBarGCMinBar").style.borderTopRightRadius="10px",l("CMTimerBarGCMinBar").style.borderBottomRightRadius="10px"):(l("CMTimerBarGCMinBar").style.borderTopRightRadius="",l("CMTimerBarGCMinBar").style.borderBottomRightRadius=""),l("CMTimerBarGCBar").style.width=`${Math.round(Math.min(Game.shimmerTypes.golden.maxTime-Game.shimmerTypes.golden.minTime,Game.shimmerTypes.golden.maxTime-Game.shimmerTypes.golden.time)*t/Game.shimmerTypes.golden.maxTime)}px`,e.TimerBarOverlay>=1?l("CMTimerBarGCBar").textContent=Math.ceil(Math.min(Game.shimmerTypes.golden.maxTime-Game.shimmerTypes.golden.minTime,Game.shimmerTypes.golden.maxTime-Game.shimmerTypes.golden.time)/Game.fps):l("CMTimerBarGCBar").textContent="",l("CMTimerBarGCTime").textContent=Math.ceil((Game.shimmerTypes.golden.maxTime-Game.shimmerTypes.golden.time)/Game.fps),n+=1),"christmas"===Game.season&&0===Game.shimmerTypes.reindeer.spawned?(l("CMTimerBarRen").style.display="",l("CMTimerBarRenMinBar").style.width=`${Math.round(Math.max(0,Game.shimmerTypes.reindeer.minTime-Game.shimmerTypes.reindeer.time)*t/Game.shimmerTypes.reindeer.maxTime)}px`,e.TimerBarOverlay>=1?l("CMTimerBarRenMinBar").textContent=Math.ceil((Game.shimmerTypes.reindeer.minTime-Game.shimmerTypes.reindeer.time)/Game.fps):l("CMTimerBarRenMinBar").textContent="",l("CMTimerBarRenBar").style.width=`${Math.round(Math.min(Game.shimmerTypes.reindeer.maxTime-Game.shimmerTypes.reindeer.minTime,Game.shimmerTypes.reindeer.maxTime-Game.shimmerTypes.reindeer.time)*t/Game.shimmerTypes.reindeer.maxTime)}px`,e.TimerBarOverlay>=1?l("CMTimerBarRenBar").textContent=Math.ceil(Math.min(Game.shimmerTypes.reindeer.maxTime-Game.shimmerTypes.reindeer.minTime,Game.shimmerTypes.reindeer.maxTime-Game.shimmerTypes.reindeer.time)/Game.fps):l("CMTimerBarRenBar").textContent="",l("CMTimerBarRenTime").textContent=Math.ceil((Game.shimmerTypes.reindeer.maxTime-Game.shimmerTypes.reindeer.time)/Game.fps),n+=1):l("CMTimerBarRen").style.display="none";const i={};l("CMTimerBarBuffTimers").innerHTML="",Object.keys(Game.buffs).forEach((t=>{if(Game.buffs[t]){const a=go(Game.buffs[t].name,Game.buffs[t].name,[{id:`${Game.buffs[t].name}Bar`}]);a.style.display="";let r="";r=void 0!==ee[Game.buffs[t].name]?ee[Game.buffs[t].name]:j,a.lastChild.children[1].className=N+r,a.lastChild.children[1].style.color="black",2===e.TimerBarOverlay?a.lastChild.children[1].textContent=`${Math.round(Game.buffs[t].time/Game.buffs[t].maxTime*100)}%`:a.lastChild.children[1].textContent="",a.lastChild.children[1].style.width=`${Math.round(Game.buffs[t].time*(o-8*Math.ceil(Game.buffs[t].time/Game.fps).toString().length)/Game.buffs[t].maxTime)}px`,a.lastChild.children[2].textContent=Math.ceil(Game.buffs[t].time/Game.fps),n+=1,i[Game.buffs[t].name]=a}})),Object.keys(i).forEach((e=>{l("CMTimerBarBuffTimers").appendChild(i[e])})),0!==n&&(l("CMTimerBar").style.height=12*n+2+"px"),le!==n&&(le=n,bo())}}(),Co(),Io(),function(){if(1===e.TooltipWrink&&1===_){let e=!1;Object.keys(Game.wrinklers).forEach((t=>{const o=Game.wrinklers[t];if(o.phase>0&&o.selected){if(e=!0,0===X[t]||void 0===X[t]){const e=document.createElement("div"),o=document.createElement("div");o.style.minWidth="120px",o.style.marginBottom="4px";const n=document.createElement("div");n.style.textAlign="center",n.id="CMTooltipWrinkler",o.appendChild(n),e.appendChild(o),Game.tooltip.draw(this,escape(e.innerHTML)),Q=t,X[t]=1}}else X[t]=0})),e||Game.tooltip.hide()}}(),function(){if(1===e.TooltipWrink&&null!==l("CMTooltipWrinkler")){let{sucked:e}=Game.wrinklers[Q],t=1.1;if(Game.Has("Sacrilegious corruption")&&(t*=1.05),1===Game.wrinklers[Q].type&&(t*=3),e*=t,Game.Has("Wrinklerspawn")&&(e*=1.05),eo.Temple.minigameLoaded){const t=Game.hasGod("scorn");1===t?e*=1.15:2===t?e*=1.1:3===t&&(e*=1.05)}l("CMTooltipWrinkler").textContent=pe(e)}}(),e.UpStats&&"stats"===Game.onMenu&&(Game.drawT-1)%(5*Game.fps)!=0&&(Game.drawT-1)%Game.fps==0&&Game.UpdateMenu(),t(),e.Scale){let e=l("cookies").innerHTML.replace(/.*(?=<br>)/i,pe(Game.cookies));Game.prefs.monospace&&(e=`<span class="monospace">${e}</span>`),l("cookies").innerHTML=e}}function Vo(){he.addLatest(Game.computedMouseCps)}class _o{constructor(e){this.maxLength=e,this.queue=[]}addLatest(e){this.queue.push(e)>this.maxLength&&this.queue.shift()}calcAverage(e){let t=e;t>this.maxLength&&(t=this.maxLength),t>this.queue.length&&(t=this.queue.length);let o=0;for(let e=this.queue.length-1;e>=0&&e>this.queue.length-1-t;e--)o+=this.queue[e];return 0===o?0:o/t}calcSum(e){let t=e;return t>this.maxLength&&(t=this.maxLength),t>this.queue.length&&(t=this.queue.length),0===t?0:this.queue.slice(-t).reduce(((e,t)=>e+t),0)}}function Qo(){const t=Math.floor(Date.now()/1e3);if(Game.T/Game.fps%1==0){let o=Game.cookies+Jt;Game.cpsSucked>0&&(o+=vt),zt=Math.max(Game.cookiesEarned,o),o*=.05;const n=t-Ge,i=Math.max(0,Game.cookies-ve)/n,a=Math.max(0,vt-we)/n,r=Math.max(0,xt[0]-xe)/n,l=Math.max(0,o-Ct)/n,s=(Game.cookieClicks-Te)/n;for(let e=0;e<n;e++)ge.addLatest(i),fe.addLatest(a),Ce.addLatest(r),ye.addLatest(l),be.addLatest(s);Ge=t,ve=Game.cookies,we=vt,xe=xt[0],Ct=o,Te=Game.cookieClicks;const c=J[e.AvgCPSHist];Me=ge.calcAverage(c),Se=fe.calcAverage(c),Pe=Ce.calcAverage(c),Be=ye.calcAverage(c),Ee=Me,1===e.CalcWrink&&(Ee+=Se),2===e.CalcWrink&&(Ee+=Pe);const d=Game.HasUnlocked("Chocolate egg")&&!Game.Has("Chocolate egg");Vt=d||0===e.CalcWrink?Me+Se+(d?Be:0):Ee,Pt=be.calcAverage(Z[e.AvgClicksHist])}}function Xo(t,o){Object.keys(t).forEach((n=>{if(e.PPRigidelMode&&1===o)t[n].color=D;else{t[n].color=Bo(t[n],Game.Objects[n].getSumPrice(o));for(let o=0;o<e.PPExcludeTop;o++)t[n].pp===Yt[o][0]&&(t[n].color=D)}}))}function qo(t,o){Object.keys(t).forEach((n=>{const i=Game.Objects[n].getSumPrice(o);Game.cookiesPs?t[n].pp=Math.max(i-(Game.cookies+ho()),0)/Game.cookiesPs+i/t[n].bonus:t[n].pp=i/t[n].bonus,e.PPRigidelMode&&1===o||Yt.push([t[n].pp,o])}))}function Yo(){!function(){qt=1/0,Yt=[],void 0===e.PPExcludeTop&&(e.PPExcludeTop=0),qo(yt,1),qo(bt,10),qo(kt,100),Yt.sort(((e,t)=>e[0]-t[0]));let t=0;if(e.PPOnlyConsiderBuyable)for(;Yt[t][1]>Game.cookies&&(t+=1,Yt.length!==t+1););qt=Yt[e.PPExcludeTop][t],Xo(yt,1),Xo(bt,10),Xo(kt,100)}(),Object.keys(St).forEach((e=>{Game.cookiesPs?St[e].pp=Math.max(Game.Upgrades[e].getPrice()-(Game.cookies+ho()),0)/Game.cookiesPs+Game.Upgrades[e].getPrice()/St[e].bonus:St[e].pp=Game.Upgrades[e].getPrice()/St[e].bonus,Number.isNaN(St[e].pp)&&(St[e].pp=1/0),St[e].color=Bo(St[e],Game.Upgrades[e].getPrice())}))}function Ko(e,t,o,n,i){let a=o,r=0;for(let o=0;o<i;o+=1){let o=t*Game.priceIncrease**Math.max(0,a-n);o=Game.modifyBuildingPrice(e,o),o=Math.ceil(o),r+=o,a+=1}return r}function Jo(e){const t={};return Object.keys(Game.Objects).forEach((o=>{t[o]={},t[o].bonus=Uo(o,e),1!==e&&(Fe=1)})),t}function Zo(){Object.keys(Game.Objects).forEach((e=>{yt[e].price=Ko(Game.Objects[e],Game.Objects[e].basePrice,Game.Objects[e].amount,Game.Objects[e].free,1),bt[e].price=Ko(Game.Objects[e],Game.Objects[e].basePrice,Game.Objects[e].amount,Game.Objects[e].free,10),kt[e].price=Ko(Game.Objects[e],Game.Objects[e].basePrice,Game.Objects[e].amount,Game.Objects[e].free,100),Gt[e].price=Ko(Game.Objects[e],Game.Objects[e].basePrice,Game.Objects[e].amount,Game.Objects[e].free,Gt[e].AmountNeeded)}))}function en(){yt=Jo(1),bt=Jo(10),kt=Jo(100),St=[],Object.keys(Game.Upgrades).forEach((t=>{const o=function(e){if("toggle"===Game.Upgrades[e].pool||0===Game.Upgrades[e].bought&&Game.Upgrades[e].unlocked&&"prestige"!==Game.Upgrades[e].pool){Mo(),"Shimmering veil [on]"===to[e].name?to["Shimmering veil [off]"].bought=0:"Golden switch [on]"===to[e].name?to["Golden switch [off]"].bought=0:to[e].bought=(to[e].bought+1)%2,Game.CountsAsUpgradeOwned(Game.Upgrades[e].pool)&&(Ae+=1),"Elder Pledge"===e?(je+=1,je>0&&No("Elder nap"),je>=5&&No("Elder slumber")):"Elder Covenant"===e?No("Elder calm"):"Prism heart biscuits"===e?No("Lovely cookies"):"Heavenly key"===e&&No("Wholesome");const t=De;Oo(),Ho(),t!==De&&Oo();const o=function(){let e=0;ro("Thousand fingers")&&(e+=.1),ro("Million fingers")&&(e*=5),ro("Billion fingers")&&(e*=10),ro("Trillion fingers")&&(e*=20),ro("Quadrillion fingers")&&(e*=20),ro("Quintillion fingers")&&(e*=20),ro("Sextillion fingers")&&(e*=20),ro("Septillion fingers")&&(e*=20),ro("Octillion fingers")&&(e*=20),ro("Nonillion fingers")&&(e*=20);let t=0;Object.keys(eo).forEach((e=>{t+=eo[e].amount})),t-=eo.Cursor.amount,e*=t,ro("Plastic mouse")&&(e+=.01*Xe),ro("Iron mouse")&&(e+=.01*Xe),ro("Titanium mouse")&&(e+=.01*Xe),ro("Adamantium mouse")&&(e+=.01*Xe),ro("Unobtainium mouse")&&(e+=.01*Xe),ro("Eludium mouse")&&(e+=.01*Xe),ro("Wishalloy mouse")&&(e+=.01*Xe),ro("Fantasteel mouse")&&(e+=.01*Xe),ro("Nevercrack mouse")&&(e+=.01*Xe),ro("Armythril mouse")&&(e+=.01*Xe),ro("Technobsidian mouse")&&(e+=.01*Xe),ro("Plasmarble mouse")&&(e+=.01*Xe),ro("Miraculite mouse")&&(e+=.01*Xe),ro("Fortune #104")&&(e+=.01*Xe);let o=1;if(ro("Santa's helpers")&&(o*=1.1),ro("Cookie egg")&&(o*=1.1),ro("Halo gloves")&&(o*=1.1),ro("Dragon claw")&&(o*=1.03),ro("Aura gloves")&&(o*=1+.05*Math.min(Game.Objects.Cursor.level,ro("Luminous gloves")?20:10)),o*=ao("click"),eo.Temple.minigameLoaded&&lo){const e=lo("labor");1===e?o*=1.15:2===e?o*=1.1:3===e&&(o*=1.05)}Object.keys(Game.buffs).forEach((e=>{void 0!==Game.buffs[e].multClick&&(o*=Game.buffs[e].multClick)})),o*=1+.05*no("Dragon Cursor");let n=o*Game.ComputeCps(1,ro("Reinforced index finger")+ro("Carpal tunnel prevention cream")+ro("Ambidextrous"),e);return n=Game.runModHookOnValue("cookiesPerClick",n),Game.hasBuff("Cursed finger")&&(n=Game.buffs["Cursed finger"].power),n}()-Game.computedMouseCps;return o?[Xe-Game.cookiesPs,o]:[Xe-Game.cookiesPs]}return[]}(t);"Elder Pledge"===t?(St[t]={bonus:Game.cookiesPs-Me},1===e.CalcWrink?St[t].bonus-=Se:2===e.CalcWrink&&(St[t].bonus-=Pe),Number.isFinite(St[t].bonus)||(St[t].bonus=0)):(St[t]={},o[0]&&(St[t].bonus=o[0]),o[1]&&(St[t].bonusMouse=o[1]))}))}function tn(e,t,o){let n=0,i=0,a=0,r=0,l=1+Math.max(0,Math.ceil(Math.log(Game.cookies)/Math.LN10)-10);for(;a<t*o;)i=Math.max(e,Math.min(Math.floor(1/9*10**l*e*o),t*o)),a=Math.max(e,Math.min(Math.floor(1/9*10**(l+1)*e*o),t*o)),r=Math.floor(1/9*10**(l+1)*e*o),n+=i,l+=1;return[n,i,r]}function on(){let e=60*dt*60*6*pt;const t=Eo();t>0?e/=t:e=0,Lt=tn(7,e,ut),Ot=2*Lt[1]/ut,Ft=Lt[2]/60/60/6/pt,jt=tn(6,e,mt),Wt=2*jt[1]/mt,At=jt[2]/60/60/6/pt,Ut=tn(7,7*e,ut),Dt=2*Ut[1]/ut,Ht=Ut[2]/60/60/6/pt,It=tn(6,7*e,mt),$t=2*It[1]/mt,Rt=It[2]/60/60/6/pt}function nn(){const e=Math.floor(Date.now()/1e3);if(Game.T/Game.fps%1==0){const t=Game.HowMuchPrestige(Game.cookiesReset),o=Math.floor(Game.HowMuchPrestige(Game.cookiesReset+Game.cookiesEarned))-Math.floor(t),n=e-Ne,i=Math.max(0,o-Oe)/n;for(let e=0;e<n;e++)ke.addLatest(i);Ne=e,Oe=o,ht=ke.calcAverage(5)}}function an(){Bt="",Nt="",Et="";const e=[];Object.keys(Game.Upgrades).forEach((t=>{e.push(Game.Upgrades[t])})),e.sort((function(e,t){return e.order>t.order?1:e.order<t.order?-1:0})),Object.keys(e).forEach((t=>{const o=e[t];if(0===o.bought){let e="";e+=function(e){let t="crate upgrade missing";"prestige"===e.pool&&(t+=" heavenly");let o=0;Game.prefs.crates||(o=1),o&&(t+=" noFrame");let{icon:n}=e;e.iconFunction&&(n=e.iconFunction());const i=`function() {return Game.crateTooltip(Game.UpgradesById[${e.id}], 'stats');}`;return`<div class="${t}"\n\t${Game.getDynamicTooltip(i,"top",!0)}\n\tstyle = "${n[2]?`background-image: url(${n[2]});`:""}background-position:${48*-n[0]}px ${48*-n[1]}px;">\n\t</div>`}(o),"prestige"===o.pool?Et+=e:"cookie"===o.pool?Nt+=e:"toggle"!==o.pool&&"unused"!==o.pool&&"debug"!==o.pool&&(Bt+=e)}}))}function rn(){if("christmas"===Game.season){let e=60*Game.cookiesPs;Game.hasBuff("Elder frenzy")&&(e*=.5),Game.hasBuff("Frenzy")&&(e*=.75),ft=Math.max(25,e),Game.Has("Ho ho ho-flavored frosting")&&(ft*=2)}}function ln(){et=900*dt/.15,et*=pt;const e=Eo();e>0?et/=e:et=0,tt=ut*(.15*et)+13,ot=mt*(.15*et)+13,nt=7*et,it=ut*(.15*nt)+13,at=mt*(.15*nt)+13,rt=2*et,lt=.15*rt,st=0;let t=0,o=0;Object.keys(Game.Objects).forEach((e=>{Game.Objects[e].amount>t&&(t=Game.Objects[e].amount),Game.Objects[e].amount>0&&(o+=1)})),Object.keys(Game.Objects).forEach((e=>{(Game.Objects[e].amount<t||1===o)&&Game.Objects[e].amount<400&&2*Game.Objects[e].price>st&&(st=2*Game.Objects[e].price,ct=e)}))}function sn(){let e=1,t=1,o=1;ro("Green yeast digestives")&&(o*=1.01),ro("Dragon fang")&&(o*=1.03),e*=1+.1*Game.auraMult("Ancestral Metamorphosis"),e*=Game.eff("goldenCookieGain"),t*=1+.1*Game.auraMult("Unholy Dominion"),t*=Game.eff("wrathCookieGain"),ut=o*e,mt=o*t,pt=1,0===Game.shimmerTypes.golden.n&&(pt*=1+1.23*Game.auraMult("Dragon's Fortune"))}function cn(){const e={};Object.keys(Game.Objects).forEach((t=>{if(0!==Object.keys(Gt).length&&Gt[t].TotalNeeded>Game.Objects[t].amount)e[t]={AmountNeeded:Gt[t].TotalNeeded-Game.Objects[t].amount,TotalNeeded:Gt[t].TotalNeeded};else{const o=function(e){const t=Game.AchievementsOwned;for(let o=0;o<101;o++)if(Uo(e,o),De>t)return o;return 101}(t);e[t]={AmountNeeded:o,TotalNeeded:Game.Objects[t].amount+o}}})),Gt=e}function dn(){vt=0,wt=0,xt=[0,null];for(let e=0;e<Game.wrinklers.length;e++){let{sucked:t}=Game.wrinklers[e],o=1.1;if(Game.Has("Sacrilegious corruption")&&(o*=1.05),1===Game.wrinklers[e].type&&(o*=3),t*=o,Game.Has("Wrinklerspawn")&&(t*=1.05),eo.Temple.minigameLoaded){const e=Game.hasGod("scorn");1===e?t*=1.15:2===e?t*=1.1:3===e&&(t*=1.05)}vt+=t,0===Game.wrinklers[e].type&&(wt+=t,t>xt[0]&&(xt=[t,e]))}}function pn(){Go(),dn(),ln(),sn(),on(),an(),rn(),ge=new _o(J[J.length-1]),fe=new _o(J[J.length-1]),Ce=new _o(J[J.length-1]),ye=new _o(J[J.length-1]),be=new _o(Z[Z.length-1]),ke=new _o(5),he=new _o(20*Z[Z.length-1]),nn(),cn(),Qo(),en(),Zo(),Yo()}const un={CPSMode:1,AvgCPSHist:3,AvgClicksHist:0,CalcWrink:0,Scale:2,ScaleDecimals:2,ScaleSeparator:0,ScaleCutoff:999999,TimeFormat:0,DetailedTime:1,PPDisplayTime:0,BuildColour:1,PPOnlyConsiderBuyable:0,PPExcludeTop:0,PPRigidelMode:0,PPSecondsLowerLimit:0,ColourBlue:"#4bb8f0",ColourGreen:"#00ff00",ColourYellow:"#ffff00",ColourOrange:"#ff7f00",ColourRed:"#ff0000",ColourPurple:"#ff00ff",ColourGray:"#b3b3b3",ColourPink:"#ff1493",ColourBrown:"#8b4513",BotBar:1,TimerBar:1,TimerBarPos:0,TimerBarOverlay:2,UpBarColour:1,UpgradeBarFixedPos:1,SortBuildings:0,SortUpgrades:0,DragonAuraInfo:1,GrimoireBar:1,GCTimer:1,Favicon:1,WrinklerButtons:1,HideSectionsButtons:0,TooltipBuildUpgrade:1,TooltipAmor:0,ToolWarnLucky:1,ToolWarnLuckyFrenzy:1,ToolWarnConjure:1,ToolWarnConjureFrenzy:1,ToolWarnEdifice:1,ToolWarnUser:0,ToolWarnBon:1,ToolWarnPos:1,TooltipGrim:1,TooltipWrink:1,TooltipLump:1,TooltipPlots:1,TooltipPantheon:1,TooltipAscendButton:1,Stats:1,MissingUpgrades:1,MissingAchievements:0,UpStats:1,HeavenlyChipsTarget:1,ShowMissedGC:1,Title:1,GeneralSound:1,GCNotification:0,GCFlash:1,ColourGCFlash:"#ffffff",GCSound:1,GCVolume:100,GCSoundURL:"https://freesound.org/data/previews/66/66717_931655-lq.mp3",FortuneNotification:0,FortuneFlash:1,ColourFortuneFlash:"#ffffff",FortuneSound:1,FortuneVolume:100,FortuneSoundURL:"https://freesound.org/data/previews/174/174027_3242494-lq.mp3",SeaNotification:0,SeaFlash:1,ColourSeaFlash:"#ffffff",SeaSound:1,SeaVolume:100,SeaSoundURL:"https://www.freesound.org/data/previews/121/121099_2193266-lq.mp3",GardFlash:1,ColourGardFlash:"#ffffff",GardSound:1,GardVolume:100,GardSoundURL:"https://freesound.org/data/previews/103/103046_861714-lq.mp3",MagicNotification:0,MagicFlash:1,ColourMagicFlash:"#ffffff",MagicSound:1,MagicVolume:100,MagicSoundURL:"https://freesound.org/data/previews/221/221683_1015240-lq.mp3",WrinklerNotification:0,WrinklerFlash:1,ColourWrinklerFlash:"#ffffff",WrinklerSound:1,WrinklerVolume:100,WrinklerSoundURL:"https://freesound.org/data/previews/124/124186_8043-lq.mp3",WrinklerMaxNotification:0,WrinklerMaxFlash:1,ColourWrinklerMaxFlash:"#ffffff",WrinklerMaxSound:1,WrinklerMaxVolume:100,WrinklerMaxSoundURL:"https://freesound.org/data/previews/152/152743_15663-lq.mp3",BulkBuyBlock:0,FavouriteSettings:1,Header:{Favourite:1,Calculation:1,Notation:1,Colours:1,BarsDisplay:1,Tooltip:1,Statistics:1,Notification:1,NotificationGeneral:1,NotificationGC:1,NotificationFC:1,NotificationSea:1,NotificationGard:1,NotificationMagi:1,NotificationWrink:1,NotificationWrinkMax:1,Miscellaneous:1,Lucky:1,Chain:1,Spells:1,Garden:1,Prestige:1,Wrink:1,Sea:1,Misc:1,InfoTab:1}},mn=function(e){if(1===e){const e=function(){try{Notification.requestPermission().then()}catch(e){return!1}return!0};"Notification"in window?e()?Notification.requestPermission().then():Notification.requestPermission():console.log("This browser does not support notifications.")}};function hn(){1===e.BotBar?(l("CMBotBar").style.display="",Co()):l("CMBotBar").style.display="none",bo()}function gn(){BeautifyAll(),Game.RefreshStore(),Game.RebuildUpgrades(),Co(),po(),uo()}function fn(){1===e.Favicon&&y>0?_t.wrath?l("CMFavicon").href="https://aktanusa.github.io/CookieMonster/favicon/wrathCookie.ico":l("CMFavicon").href="https://aktanusa.github.io/CookieMonster/favicon/goldenCookie.ico":l("CMFavicon").href="https://orteil.dashnet.org/cookieclicker/favicon.ico"}class Cn{constructor(e,t){this.type=e,this.group=t}}class yn extends Cn{constructor(e,t,o){super(e,t),this.desc=o}}class bn extends Cn{constructor(e,t,o,n,i,a){super(e,t),this.label=o,this.desc=n,this.min=i,this.max=a}}class kn extends Cn{constructor(e,t,o,n,i,a){super(e,t),this.label=o,this.desc=n,this.toggle=i,void 0!==a&&(this.func=a)}}class Gn extends Cn{constructor(e,t,o,n){super(e,t),this.label=o,this.desc=n;for(let e=0;e<101;e++)this.label[e]=`${e}%`}}const vn={CPSMode:new kn("bool","Calculation",["Current cookies per second","Average cookies per second"],"Calculate times using current cookies per second or average cookies per second",!1),AvgCPSHist:new kn("bool","Calculation",["Average CPS in past 10s","Average CPS in past 15s","Average CPS in past 30s","Average CPS in past 1m","Average CPS in past 5m","Average CPS in past 10m","Average CPS in past 15m","Average CPS in past 30m"],"How much time average Cookies Per Second should consider",!1),AvgClicksHist:new kn("bool","Calculation",["Average clicks in past 1s","Average clicks in past 5s","Average clicks in past 10s","Average clicks in past 15s","Average clicks in past 30s"],"How much time average Cookie Clicks should consider",!1),CalcWrink:new kn("bool","Calculation",["Calculate with wrinklers OFF","Calculate with wrinklers ON","Calculate with single fattest wrinkler ON"],"Calculate times and average Cookies Per Second with (only the single non-shiny fattest) wrinklers",!0,(function(){_e=!0})),Scale:new kn("bool","Notation",["Game's setting scale","Metric","Short scale","Short scale (Abbreviated)","Scientific notation","Engineering notation"],"Change how long numbers are formatted",!1,(function(){gn()})),ScaleDecimals:new kn("bool","Notation",["1 decimals","2 decimals","3 decimals"],'Set the number of decimals used when applicable. This only works with Cookie Monster scales and not with "Game\'s Setting Scale"',!1,(function(){gn()})),ScaleSeparator:new kn("bool","Notation",[". for decimals (standard)",". for thousands"],"Set the separator used for decimals and thousands",!1,(function(){gn()})),ScaleCutoff:new bn("numscale","Notation","Notation cut-off point: ","The number from which Cookie Monster will start formatting numbers based on chosen scale. Standard is 999,999. Setting this above 999,999,999 might break certain notations",1,999999999),TimeFormat:new kn("bool","Notation",["Time XXd, XXh, XXm, XXs","Time XX:XX:XX:XX:XX"],"Change the time format",!1),DetailedTime:new kn("bool","Notation",["Detailed time OFF","Detailed time ON"],"Change how time is displayed in certain statistics and tooltips",!0,(function(){1===e.DetailedTime?Game.sayTime=V:Game.sayTime=S.sayTime})),PPDisplayTime:new kn("bool","Notation",["PP as value (standard)","PP as time unit"],"Display PP as calculated value or as approximate time unit. Note that PP does not translate directly into a time unit and this is therefore only an approximation.",!1),BuildColour:new kn("bool","Colours",["Building colours OFF","Building colours ON"],"Colour code buildings",!0,(function(){po()})),PPOnlyConsiderBuyable:new kn("bool","Colours",["Don't ignore non-buyable","Ignore non-buyable"],"Makes Cookie Monster label buildings and upgrades you can't buy right now red, useful in those situations where you just want to spend your full bank 'most optimally'",!0),PPExcludeTop:new kn("bool","Colours",["Don't ignore any","Ignore 1st best","Ignore 1st and 2nd best","Ignore 1st, 2nd and 3rd best"],"Makes Cookie Monster ignore the 1st, 2nd or 3rd best buildings in labeling and colouring PP values",!0),PPRigidelMode:new kn("bool","Colours",["Rigidel mode OFF","Rigidel mode ON"],'Makes Cookie Monster ignore all "buy 1" options when colouring PP in order to stay at a total building count ending in 10 for pantheon god Rigidel',!0),PPSecondsLowerLimit:new bn("numscale","Colours","Lower limit for PP (in seconds): ",'If a building or upgrade costs less than the specified seconds of CPS it will also be considered optimal and label it as such ("PP is less than xx seconds of CPS"); setting to 0 ignores this option',0,1/0),ColourBlue:new yn("colour","Colours","Standard colour is blue. Used to show upgrades better than best PP building, for Click Frenzy bar, and for various labels"),ColourGreen:new yn("colour","Colours","Standard colour is green. Used to show best PP building, for Blood Frenzy bar, and for various labels"),ColourYellow:new yn("colour","Colours","Standard colour is yellow. Used to show buildings within the top 10 of PP, for Frenzy bar, and for various labels"),ColourOrange:new yn("colour","Colours","Standard colour is orange. Used to show buildings within the top 20 of PP, for Next Reindeer bar, and for various labels"),ColourRed:new yn("colour","Colours","Standard colour is Red. Used to show buildings within the top 30 of PP, for Clot bar, and for various labels"),ColourPurple:new yn("colour","Colours","Standard colour is purple. Used to show buildings outside of the top 30 of PP, for Next Cookie bar, and for various labels"),ColourGray:new yn("colour","Colours","Standard colour is gray. Used to show negative or infinity PP, and for Next Cookie/Next Reindeer bar"),ColourPink:new yn("colour","Colours","Standard colour is pink. Used for Dragonflight bar"),ColourBrown:new yn("colour","Colours","Standard colour is brown. Used for Dragon Harvest bar"),BotBar:new kn("bool","BarsDisplay",["Bottom bar OFF","Bottom bar ON"],"Building information",!0,(function(){hn()})),TimerBar:new kn("bool","BarsDisplay",["Timer bar OFF","Timer bar ON"],"Bar with timers for golden cookie, season popup, Frenzy (Normal, Clot, Elder), Click Frenzy",!0,(function(){ko()})),TimerBarPos:new kn("bool","BarsDisplay",["Timer bar position (top left)","Timer bar position (bottom)"],"Placement of the timer bar",!1,(function(){0===e.TimerBarPos?(l("CMTimerBar").style.width="30%",l("CMTimerBar").style.bottom="",l("game").insertBefore(l("CMTimerBar"),l("sectionLeft"))):(l("CMTimerBar").style.width="100%",l("CMTimerBar").style.bottom="0px",l("wrapper").appendChild(l("CMTimerBar"))),bo()})),TimerBarOverlay:new kn("bool","BarsDisplay",["Timer bar overlay OFF","Timer bar overlay only seconds","Timer bar overlay full"],"Overlay on timers displaying seconds and/or percentage left",!0),UpBarColour:new kn("bool","BarsDisplay",["Upgrade colours/bar OFF","Upgrade colours with bar ON","Upgrade colours without bar ON"],"Colour code upgrades and optionally add a counter bar",!1,(function(){1===e.UpBarColor?(l("CMUpgradeBar").style.display="",uo()):2===e.UpBarColor?(l("CMUpgradeBar").style.display="none",uo()):(l("CMUpgradeBar").style.display="none",Game.RebuildUpgrades())})),UpgradeBarFixedPos:new kn("bool","BarsDisplay",["Upgrade bar fixed position OFF","Upgrade bar fixed position ON"],"Lock the upgrade bar at top of the screen to prevent it from moving ofscreen when scrolling",!0,(function(){1===e.UpgradeBarFixedPos?(l("CMUpgradeBar").style.position="sticky",l("CMUpgradeBar").style.top="0px"):l("CMUpgradeBar").style.position=""})),SortBuildings:new kn("bool","BarsDisplay",["Sort buildings: default","Sort buildings: PP of x1 purchase","Sort buildings: PP of selected bulk mode"],"Sort the display of buildings in either default order or by PP",!1,(function(){po()})),SortUpgrades:new kn("bool","BarsDisplay",["Sort upgrades: default","Sort upgrades: PP"],"Sort the display of upgrades in either default order or by PP",!1,(function(){uo()})),DragonAuraInfo:new kn("bool","BarsDisplay",["Extra dragon aura info OFF","Extra dragon aura info ON"],"Shows information about changes in CPS and costs in the dragon aura interface.",!0),GrimoireBar:new kn("bool","BarsDisplay",["Grimoire magic meter timer OFF","Grimoire magic meter timer ON"],"A timer on how long before the Grimoire magic meter is full",!0),GCTimer:new kn("bool","BarsDisplay",["Golden cookie timer OFF","Golden cookie timer ON"],"A timer on the golden cookie when it has been spawned",!0,(function(){1===e.GCTimer?Object.keys(se).forEach((e=>{se[e].style.display="block",se[e].style.left=Kt[e].l.style.left,se[e].style.top=Kt[e].l.style.top})):Object.keys(se).forEach((e=>se[e].style.display="none"))})),Favicon:new kn("bool","BarsDisplay",["Favicon OFF","Favicon ON"],"Update favicon with golden/wrath cookie",!0,(function(){fn()})),WrinklerButtons:new kn("bool","BarsDisplay",["Extra wrinkler buttons OFF","Extra wrinkler buttons ON"],"Show buttons for popping wrinklers at bottom of cookie section",!0,(function(){t()})),HideSectionsButtons:new kn("bool","BarsDisplay",["Hide buildings/upgrades button OFF","Hide buildings/upgrades button ON"],"Show buttons for hiding and showing the buildings and upgrades sections in the right column",!0,(function(){e.HideSectionsButtons?l("CMSectionHidButtons").style.display="":l("CMSectionHidButtons").style.display="none"})),TooltipBuildUpgrade:new kn("bool","Tooltip",["Building/upgrade tooltip information OFF","Building/upgrade tooltip information ON"],"Extra information in building/upgrade tooltips",!0),TooltipAmor:new kn("bool","Tooltip",["Buildings tooltip amortization information OFF","Buildings tooltip amortization information ON"],"Add amortization information to buildings tooltip",!0),ToolWarnLucky:new kn("bool","Tooltip",["Tooltip lucky warning OFF","Tooltip lucky warning ON"],'A warning when buying if it will put the bank under the amount needed for max "Lucky!" rewards',!0),ToolWarnLuckyFrenzy:new kn("bool","Tooltip",["Tooltip lucky frenzy warning OFF","Tooltip lucky frenzy warning ON"],'A warning when buying if it will put the bank under the amount needed for max "Lucky!" (Frenzy) rewards',!0),ToolWarnConjure:new kn("bool","Tooltip",["Tooltip conjure warning OFF","Tooltip conjure warning ON"],'A warning when buying if it will put the bank under the amount needed for max "Conjure Baked Goods" rewards',!0),ToolWarnConjureFrenzy:new kn("bool","Tooltip",["Tooltip conjure frenzy warning OFF","Tooltip conjure frenzy warning ON"],'A warning when buying if it will put the bank under the amount needed for max "Conjure Baked Goods" rewards with Frenzy active',!0),ToolWarnEdifice:new kn("bool","Tooltip",["Tooltip edifice warning OFF","Tooltip edifice warning ON"],'A warning when buying if it will put the bank under the amount needed for "Spontaneous Edifice" to possibly give you your most expensive building',!0),ToolWarnUser:new bn("numscale","Tooltip","Tooltip warning at x times CPS: ","Use this to show a customized warning if buying it will put the bank under the amount equal to value times cps; setting to 0 disables the function altogether",0,1/0),ToolWarnBon:new kn("bool","Tooltip",["Calculate tooltip warning with bonus CPS OFF","Calculate tooltip warning with bonus CPS ON"],"Calculate the warning with or without the bonus CPS you get from buying",!0),ToolWarnPos:new kn("bool","Tooltip",["Tooltip warning position (left)","Tooltip warning position (bottom)"],"Placement of the warning boxes",!1,(function(){Ro()})),TooltipGrim:new kn("bool","Tooltip",["Grimoire tooltip information OFF","Grimoire tooltip information ON"],"Extra information in tooltip for grimoire",!0),TooltipWrink:new kn("bool","Tooltip",["Wrinkler tooltip OFF","Wrinkler tooltip ON"],"Shows the amount of cookies a wrinkler will give when popping it",!0),TooltipLump:new kn("bool","Tooltip",["Sugar lump tooltip OFF","Sugar lump tooltip ON"],"Shows the current Sugar Lump type in Sugar lump tooltip.",!0),TooltipPlots:new kn("bool","Tooltip",["Garden plots tooltip OFF","Garden plots tooltip ON"],"Shows a tooltip for plants that have a cookie reward.",!0),TooltipPantheon:new kn("bool","Tooltip",["Pantheon tooltip OFF","Pantheon tooltip ON"],"Shows additional info in the pantheon tooltip",!0),TooltipAscendButton:new kn("bool","Tooltip",["Show Extra Info Ascend Tooltip OFF","Show Extra Info Ascend Tooltip ON"],"Shows additional info in the ascend tooltip",!0),Stats:new kn("bool","Statistics",["Statistics OFF","Statistics ON"],"Extra Cookie Monster statistics!",!0),MissingUpgrades:new kn("bool","Statistics",["Missing upgrades OFF","Missing upgrades ON"],"Shows missing upgrades in statistics menu",!0),MissingAchievements:new kn("bool","Statistics",["Missing Achievements OFF","Missing Normal Achievements ON"],"Shows missing normal achievements in statistics menu.",!0),UpStats:new kn("bool","Statistics",["Statistics update rate (default)","Statistics update rate (1s)"],"Default rate is once every 5 seconds",!1),HeavenlyChipsTarget:new bn("numscale","Statistics","Heavenly chips target: ",'Use this to set a heavenly chips target that will be counted towards in the "prestige" statsistics sections',1,1/0),ShowMissedGC:new kn("bool","Statistics",["Missed GC OFF","Missed GC ON"],"Show a stat in the statistics screen that counts how many golden cookies you have missed",!0),Title:new kn("bool","NotificationGeneral",["Title OFF","Title ON","Title pinned tab highlight"],'Update title with colden cookie/season popup timers; pinned tab highlight only changes the title when a golden cookie/season popup spawns; "!" means that golden cookie/reindeer can spawn',!0),GeneralSound:new kn("bool","NotificationGeneral",["Consider game volume setting OFF","Consider game volume setting ON"],'Turning this toggle to "off" makes Cookie Monster no longer consider the volume setting of the base game, allowing mod notifications to play with base game volume turned down',!0),GCNotification:new kn("bool","NotificationGC",["Notification OFF","Notification ON"],"Create a notification when golden cookie spawns",!0,(function(){mn(e.GCNotification)})),GCFlash:new kn("bool","NotificationGC",["Flash OFF","Flash ON"],"Flash screen on golden cookie",!0),ColourGCFlash:new yn("colour","NotificationGC","The colour of the GC flash, standard colour is white"),GCSound:new kn("bool","NotificationGC",["Sound OFF","Sound ON"],"Play a sound on golden cookie",!0),GCVolume:new Gn("vol","NotificationGC",[],"Volume"),GCSoundURL:new kn("url","NotificationGC","Sound URL:","URL of the sound to be played when a golden cookie spawns"),FortuneNotification:new kn("bool","NotificationFC",["Notification OFF","Notification ON"],"Create a notification when fortune cookie is on the ticker",!0,(function(){mn(e.FortuneNotification)})),FortuneFlash:new kn("bool","NotificationFC",["Flash OFF","Flash ON"],"Flash screen on fortune cookie spawn",!0),ColourFortuneFlash:new yn("colour","NotificationFC","The colour of the fortune flash, standard colour is white"),FortuneSound:new kn("bool","NotificationFC",["Sound OFF","Sound ON"],"Play a sound on fortune cookie spawn",!0),FortuneVolume:new Gn("vol","NotificationFC",[],"Volume"),FortuneSoundURL:new kn("url","NotificationFC","Sound URL:","URL of the sound to be played when the ticker has a fortune cookie"),SeaNotification:new kn("bool","NotificationSea",["Notification OFF","Notification ON"],"Create a notification on season popup",!0,(function(){mn(e.SeaNotification)})),SeaFlash:new kn("bool","NotificationSea",["Flash OFF","Flash ON"],"Flash screen on season popup",!0),ColourSeaFlash:new yn("colour","NotificationSea","The colour of the season popup flash, standard colour is white"),SeaSound:new kn("bool","NotificationSea",["Sound OFF","Sound ON"],"Play a sound on season popup",!0),SeaVolume:new Gn("vol","NotificationSea",[],"Volume"),SeaSoundURL:new kn("url","NotificationSea","Sound URL:","URL of the sound to be played when on season popup spawns"),GardFlash:new kn("bool","NotificationGard",["Garden Tick Flash OFF","Flash ON"],"Flash screen on garden tick",!0),ColourGardFlash:new yn("colour","NotificationGard","The colour of the garden flash, standard colour is white"),GardSound:new kn("bool","NotificationGard",["Sound OFF","Sound ON"],"Play a sound on garden tick",!0),GardVolume:new Gn("vol","NotificationGard",[],"Volume"),GardSoundURL:new kn("url","NotificationGard","Garden Tick Sound URL:","URL of the sound to be played when the garden ticks"),MagicNotification:new kn("bool","NotificationMagi",["Notification OFF","Notification ON"],"Create a notification when magic reaches maximum",!0,(function(){mn(e.MagicNotification)})),MagicFlash:new kn("bool","NotificationMagi",["Flash OFF","Flash ON"],"Flash screen when magic reaches maximum",!0),ColourMagicFlash:new yn("colour","NotificationMagi","The colour of the magic flash, standard colour is white"),MagicSound:new kn("bool","NotificationMagi",["Sound OFF","Sound ON"],"Play a sound when magic reaches maximum",!0),MagicVolume:new Gn("vol","NotificationMagi",[],"Volume"),MagicSoundURL:new kn("url","NotificationMagi","Sound URL:","URL of the sound to be played when magic reaches maxium"),WrinklerNotification:new kn("bool","NotificationWrink",["Notification OFF","Notification ON"],"Create a notification when a wrinkler appears",!0,(function(){mn(e.WrinklerNotification)})),WrinklerFlash:new kn("bool","NotificationWrink",["Flash OFF","Flash ON"],"Flash screen when a wrinkler appears",!0),ColourWrinklerFlash:new yn("colour","NotificationWrink","The colour of the wrinkler flash, standard colour is white"),WrinklerSound:new kn("bool","NotificationWrink",["Sound OFF","Sound ON"],"Play a sound when a wrinkler appears",!0),WrinklerVolume:new Gn("vol","NotificationWrink",[],"Volume"),WrinklerSoundURL:new kn("url","NotificationWrink","Sound URL:","URL of the sound to be played when a wrinkler appears"),WrinklerMaxNotification:new kn("bool","NotificationWrinkMax",["Notification OFF","Notification ON"],"Create a notification when the maximum amount of wrinklers has appeared",!0,(function(){mn(e.WrinklerMaxNotification)})),WrinklerMaxFlash:new kn("bool","NotificationWrinkMax",["Flash OFF","Flash ON"],"Flash screen when the maximum amount of Wrinklers has appeared",!0),ColourWrinklerMaxFlash:new yn("colour","NotificationWrinkMax","The colour of the maximum wrinkler flash, standard colour is white"),WrinklerMaxSound:new kn("bool","NotificationWrinkMax",["Sound OFF","Sound ON"],"Play a sound when the maximum amount of wrinklers has appeared",!0),WrinklerMaxVolume:new Gn("vol","NotificationWrinkMax",[],"Volume"),WrinklerMaxSoundURL:new kn("url","NotificationWrinkMax","Sound URL:","URL of the sound to be played when the maximum amount of wrinklers has appeared"),BulkBuyBlock:new kn("bool","Miscellaneous",["Block bulk buying OFF","Block bulk buying ON"],"Block clicking bulk buying when you can't buy all. This prevents buying 7 of a building when you are in buy-10 or buy-100 mode.",!0),FavouriteSettings:new kn("bool","Miscellaneous",["Favourite settings section OFF","Favourite settings section ON","Favourite settings section ON (Locked)"],"Show stars before each setting which allows selecting it for a 'favourites' section at the top of the Cookie Monster settings. Setting this to Locked removes the stars but shows the 'favourites' section",!0,(function(){Game.UpdateMenu()}))};function wn(){return JSON.stringify({favouriteSettings:de,settings:e,version:"2.031.7"})}function xn(){if(Je!==Game.dragonLevel||_e){if(Game.dragonLevel<25&&Game.dragonLevels[Game.dragonLevel].buy.toString().includes("sacrifice")){let e=Game.dragonLevels[Game.dragonLevel].buy.toString().match(/Objects\[(.*)\]/)[1];const t=Game.dragonLevels[Game.dragonLevel].buy.toString().match(/sacrifice\((.*?)\)/)[1];if("i"!==e)if(e=e.replaceAll("'",""),Game.Objects[e].amount<t)Ze="Not enough buildings to sell";else{let o=0;Mo();for(let n=0;n<t;n++){let t=eo[e].basePrice*Game.priceIncrease**Math.max(0,eo[e].amount-1-eo[e].free);t=Game.modifyBuildingPrice(eo[e],t),t=Math.ceil(t),o+=t,eo[e].amount-=1}Ze=`Cost to rebuy: ${pe(o)}`}else{let o=0;Mo(),Object.keys(Game.Objects).forEach((n=>{if(e=n,Game.Objects[e].amount<t)Ze="Not enough buildings to sell";else{for(let n=0;n<t;n++){let t=eo[e].basePrice*Game.priceIncrease**Math.max(0,eo[e].amount-1-eo[e].free);t=Game.modifyBuildingPrice(eo[e],t),t=Math.ceil(t),o+=t,eo[e].amount-=1}Ze=`Cost to rebuy: ${pe(o)}`}}))}}Je=Game.dragonLevel}}function Tn(e,t){if(!Game.Objects.Temple.minigameLoaded)return 0;Mo();const{minigame:o}=Game.Objects.Temple,n=o.godsById[e].slot;"0"===n?Ie=o.slot[t]:"1"===n?ze=o.slot[t]:"2"===n&&(Ve=o.slot[t]),0===t?Ie=e:1===t?ze=e:2===t&&(Ve=e);const i=De;return Oo(),Ho(),i!==De&&Oo(),Xe-Game.cookiesPs}function Mn(){eo=[],Object.keys(Game.Objects).forEach((e=>{eo[e]=xo(e)})),to=[],Object.keys(Game.Upgrades).forEach((e=>{to[e]=To(e)})),oo=[],Object.keys(Game.Achievements).forEach((e=>{oo[e]=vo(e)})),Mo()}let Sn=!1;function Pn(t,o,n){(1===e[o]||n)&&3===t&&!1===Sn||1===t?(l("CMFlashScreen").style.backgroundColor=e[`Colour${o}`],l("CMFlashScreen").style.opacity="0.5",3===t?(l("CMFlashScreen").style.display="inline",setTimeout((function(){Pn(2,o,!0)}),1e3/Game.fps)):setTimeout((function(){Pn(0,o,!0)}),1e3/Game.fps)):2===t?(l("CMFlashScreen").style.opacity="1",setTimeout((function(){Pn(1,o,!0)}),1e3/Game.fps)):0===t&&(l("CMFlashScreen").style.display="none")}function Bn(t,o,n,i){if((1===e[o]||i)&&!1===Sn){const o=new Audio(t);e.GeneralSound?o.volume=e[n]/100*(Game.volume/100):o.volume=e[n]/100,o.play()}}function En(t,o,n){1===e[t]&&"hidden"===document.visibilityState&&!1===Sn&&new Notification(o,{body:n,badge:"https://orteil.dashnet.org/cookieclicker/favicon.ico"})}function Nn(){if(re!==Game.OnAscend&&(re=Game.OnAscend,Game.OnAscend?(l("game").style.bottom="0px",1===e.BotBar&&(l("CMBotBar").style.display="none"),1===e.TimerBar&&(l("CMTimerBar").style.display="none")):(hn(),ko()),yo()),!Game.OnAscend&&0===Game.AscendTimer){r!==Object.keys(Game.mods).length&&(Mn(),pn(),r=Object.keys(Game.mods).length),_e&&(cn(),en(),Game.Has("Golden switch [off]")?(Mo(),to["Golden switch [off]"].bought=0,Oo(),dt=Xe):dt=Game.cookiesPs,sn(),ln(),an(),on(),xn(),function(){for(let e=0;e<11;e+=1)for(let t=0;t<3;t+=1)Zt[e][t]=Tn(e,t)}(),rn(),function(){let e=0;if(Game.Objects.Bank.minigameLoaded){const t=Game.Objects.Bank.minigame.goods;let o=0;Object.keys(t).forEach((e=>{const n=t[e];o+=n.stock*n.val})),e+=o*Game.cookiesPsRawHighest}e+=function(){let e=0;Mo();let t=2;5!==$e&&18!==$e||(t-=1),5!==Re&&18!==Re||(t-=1),$e=5,Re=18;for(let e=0;e<t;++e){let e="Cursor";Object.keys(eo).forEach((t=>{eo[t].amount>0&&(e=t)})),eo[e].amount-=1,We-=1}return Object.keys(eo).forEach((t=>{const o=eo[t];e+=co(Game.Objects[o.name],Game.Objects[t].basePrice,o.amount,Game.Objects[t].free,o.amount)})),e}(),Jt=e}(),_e=0);const t=Game.auraMult("Fierce Hoarder")>0;!Le&&t?(Le=!0,Fe=1):Le&&!t&&(Le=!1,Fe=1),Fe&&(Zo(),Fe=0),function(){dn(),Yo(),function(){Tt=0;let e=0;Object.keys(Game.wrinklers).forEach((t=>{2===Game.wrinklers[t].phase&&(e+=1)}));let t=1;if(eo.Temple.minigameLoaded){const e=Game.hasGod("scorn");1===e?t*=1.15:2===e?t*=1.1:3===e&&(t*=1.05)}Mt=e,Tt=e*(.05*e*1.1)*(.05*Game.Has("Sacrilegious corruption")+1)*(.05*Game.Has("Wrinklerspawn")+1)*t}(),Qo(),nn();const e=Game.HowManyCookiesReset(Math.floor(Game.HowMuchPrestige(Game.cookiesReset+Game.cookiesEarned))+1)-(Game.cookiesEarned+Game.cookiesReset);Xt=ue(e/mo())}(),G=0,Kt={},Object.keys(Game.shimmers).forEach((e=>{Kt[Game.shimmers[e].id]=Game.shimmers[e],Game.shimmers[e].spawnLead&&"golden"===Game.shimmers[e].type&&(_t=Game.shimmers[e],G+=1)})),Object.keys(se).forEach((e=>{void 0===Kt[e]&&(se[e].parentNode.removeChild(se[e]),delete se[e])})),y!==Game.shimmerTypes.golden.n?(y=Game.shimmerTypes.golden.n,y&&(b<G&&(Pn(3,"GCFlash",!1),Bn(e.GCSoundURL,"GCSound","GCVolume",!1),En("GCNotification","Golden Cookie Spawned","A Golden Cookie has spawned. Click it now!")),Object.keys(Game.shimmers).forEach((t=>{void 0===se[Game.shimmers[t].id]&&function(t){const o=document.createElement("div");o.id=`GCTimer${t.id}`,o.style.width="96px",o.style.height="96px",o.style.position="absolute",o.style.zIndex="10000000001",o.style.textAlign="center",o.style.lineHeight="96px",o.style.fontFamily='"Kavoon", Georgia, serif',o.style.fontSize="35px",o.style.cursor="pointer",o.style.display="block",0===e.GCTimer&&(o.style.display="none"),o.style.left=t.l.style.left,o.style.top=t.l.style.top,o.onclick=function(){t.pop()},o.onmouseover=function(){t.l.style.filter="brightness(125%) drop-shadow(0px 0px 3px rgba(255,255,255,1))",t.l.style.webkitFilter="brightness(125%) drop-shadow(0px 0px 3px rgba(255,255,255,1))"},o.onmouseout=function(){t.l.style.filter="",t.l.style.webkitFilter=""},se[t.id]=o,l("shimmers").appendChild(o)}(Game.shimmers[t])}))),fn(),b=G,0===G&&(_t=0)):1===e.GCTimer&&y&&Object.keys(se).forEach((e=>{se[e].style.opacity=Kt[e].l.style.opacity,se[e].style.transform=Kt[e].l.style.transform,se[e].textContent=Math.ceil(Kt[e].life/Game.fps)})),f!==(Game.TickerEffect&&"fortune"===Game.TickerEffect.type)&&(f=Game.TickerEffect&&"fortune"===Game.TickerEffect.type,f&&(Pn(3,"FortuneFlash",!1),Bn(e.FortuneSoundURL,"FortuneSound","FortuneVolume",!1),En("FortuneNotification","Fortune Cookie found","A Fortune Cookie has appeared on the Ticker."))),g!==Game.shimmerTypes.reindeer.spawned&&(g=Game.shimmerTypes.reindeer.spawned,Object.keys(Game.shimmers).forEach((e=>{Game.shimmers[e].spawnLead&&"reindeer"===Game.shimmers[e].type&&(Qt=Game.shimmers[e])})),Pn(3,"SeaFlash",!1),Bn(e.SeaSoundURL,"SeaSound","SeaVolume",!1),En("SeaNotification","Reindeer sighted!","A Reindeer has spawned. Click it now!")),Game.Objects.Farm.minigameLoaded&&C!==Game.Objects.Farm.minigame.nextStep&&(0!==C&&C<Date.now()&&(Pn(3,"GardFlash",!1),Bn(e.GardSoundURL,"GardSound","GardVolume",!1)),C=Game.Objects.Farm.minigame.nextStep),function(){if(Game.Objects["Wizard tower"].minigameLoaded&&1===e.GrimoireBar){const{minigame:t}=Game.Objects["Wizard tower"];t.magic<t.magicM?k=!1:k||(k=!0,Pn(3,"MagicFlash",!1),Bn(e.MagicSoundURL,"MagicSound","MagicVolume",!1),En("MagicNotification","Magic Meter full","Your Magic Meter is full. Cast a spell!"))}}(),function(){if(Game.elderWrath>0){let t=0;Object.keys(Game.wrinklers).forEach((e=>{2===Game.wrinklers[e].phase&&(t+=1)})),t>v?(v=t,t===Game.getWrinklersMax()&&e.WrinklerMaxFlash?Pn(3,"WrinklerMaxFlash",!1):Pn(3,"WrinklerFlash",!1),t===Game.getWrinklersMax()&&e.WrinklerMaxSound?Bn(e.WrinklerMaxSoundURL,"WrinklerMaxSound","WrinklerMaxVolume",!1):Bn(e.WrinklerSoundURL,"WrinklerSound","WrinklerVolume",!1),t===Game.getWrinklersMax()&&e.WrinklerMaxNotification?En("WrinklerMaxNotification","Maximum Wrinklers Reached","You have reached your maximum ammount of wrinklers"):En("WrinklerNotification","A Wrinkler appeared","A new wrinkler has appeared")):v=t}}()}}function On(){let t="";for(let o=0;o<$.length;o++)t+=`.CMText${$[o]} { color: ${e[`Colour${$[o]}`]}; }\n`;for(let o=0;o<$.length;o++)t+=`.CMBack${$[o]} { background-color: ${e[`Colour${$[o]}`]}; }\n`;for(let o=0;o<$.length;o++)t+=`.CMBorder${$[o]} { border: 1px solid ${e[`Colour${$[o]}`]}; }\n`;l("CMCSS").textContent=t,po()}function Fn(){const e=b64_to_utf8(unescape(localStorage.getItem("CookieClickerGame")).split("!END!")[0]),t=e.match(/CookieMonster.*(;|$)/);if(null!==t){const o=e.replace(t[0],`CookieMonster:${wn()}`);localStorage.setItem("CookieClickerGame",escape(`${utf8_to_b64(o)}!END!`))}}function Ln(t){if(void 0!==localStorage.CMConfig&&delete localStorage.CMConfig,void 0!==t){e=t,void 0!==e.Colors&&delete e.Colors,void 0!==e.Colours&&delete e.Colours;let o=!1;Object.keys(un).forEach((t=>{void 0===e[t]?(o=!0,e[t]=un[t]):"Header"===t&&Object.keys(un.Header).forEach((n=>{void 0!==e[t][n]&&e[t][n]>-1&&e[t][n]<2||(o=!0,e[t][n]=un[t][n])}))})),o&&Fn(),Nn(),Object.keys(un).forEach((e=>{"Header"!==e&&void 0!==vn[e].func&&vn[e].func()}))}else Ln(un);Game.UpdateMenu(),On()}function Wn(){Object.keys(Game.wrinklers).forEach((e=>{Game.wrinklers[e].sucked>0&&0===Game.wrinklers[e].type&&(Game.wrinklers[e].hp=0)}))}function An(t,o){if("b"===t){if(l("tooltip").innerHTML=Game.Objects[o].tooltip(),1===e.TooltipAmor){const e=Ko(Game.Objects[o],Game.Objects[o].basePrice,0,Game.Objects[o].free,Game.Objects[o].amount),t=e-Game.Objects[o].totalCookies;t>0&&(l("tooltip").innerHTML=l("tooltip").innerHTML.split("so far</div>").join(`so far<br/>&bull; <b>${pe(t)}</b> ${1===Math.floor(t)?"cookie":"cookies"} left to amortize (${me((e-Game.Objects[o].totalCookies)/(Game.Objects[o].storedTotalCps*Game.globalCpsMult)).text})</div>`))}-1===Game.buyMode&&(l("tooltip").innerHTML=l("tooltip").innerHTML.split(pe(Game.Objects[o].bulkPrice)).join(pe((Game.Objects[o],Game.Objects[o].basePrice,Game.Objects[o].amount,Game.Objects[o].free,Game.buyBulk,1))))}else if("u"===t){if(!Game.UpgradesInStore[o])return"";l("tooltip").innerHTML=Game.crateTooltip(Game.UpgradesInStore[o],"store")}else"s"===t?l("tooltip").innerHTML=Game.lumpTooltip():"g"===t?l("tooltip").innerHTML=Game.Objects["Wizard tower"].minigame.spellTooltip(o)():"p"===t?l("tooltip").innerHTML=Game.ObjectsById[2].minigame.tileTooltip(o[0],o[1])():"ha"===t?l("tooltip").innerHTML=Game.ObjectsById[2].minigame.toolTooltip(1)():"wb"===t?l("tooltip").innerHTML="":"pag"===t?l("tooltip").innerHTML=Game.Objects.Temple.minigame.godTooltip(o)():"pas"===t&&(l("tooltip").innerHTML=Game.Objects.Temple.minigame.slotTooltip(o[0])());if("b"===t&&1===Game.buyMode||"u"===t||"s"===t||"g"===t||"p"===t&&!Game.keys[16]||"ha"===t||"wb"===t||"pag"===t||"pas"===t&&-1!==o[1]){const e=document.createElement("div");e.id="CMTooltipArea",l("tooltip").appendChild(e)}return ne=t,ie=o,Io(),l("tooltip").innerHTML}function jn(){Game.Objects["Wizard tower"].minigameLoaded&&Object.keys(Game.Objects["Wizard tower"].minigame.spellsById).forEach((e=>{null!==l(`grimoireSpell${e}`).onmouseover&&(T[e]=l(`grimoireSpell${e}`).onmouseover,l(`grimoireSpell${e}`).onmouseover=function(){Game.tooltip.dynamic=1,Game.tooltip.draw(this,(function(){return An("g",`${e}`)}),"this"),Game.tooltip.wobble()})}))}function Dn(){if(!m&&Game.Objects["Wizard tower"].minigameLoaded){const{minigame:t}=Game.Objects["Wizard tower"];p=t.draw,Game.Objects["Wizard tower"].minigame.draw=function(){p(),1===e.GrimoireBar&&t.magic<t.magicM&&(t.magicBarTextL.innerHTML+=` (${ue($o(t.magic,t.magicM,t.magicM))})`)},m=!0}}function Hn(){!function(){if(!u&&Game.Objects["Wizard tower"].minigameLoaded){const{minigame:e}=Game.Objects["Wizard tower"];c=e.launch,d=new Function(`return ${e.launch.toString().split("=this").join("= Game.Objects['Wizard tower'].minigame")}`),Game.Objects["Wizard tower"].minigame.launch=function(){d(),jn(),m=!1,Dn(),u=!0}}}(),Dn()}var Un=o(877),$n=o.n(Un);const Rn="CMConfig";function In(t){null!==l(`slider${t}`)&&(l(`slider${t}right`).innerHTML=`${l(`slider${t}`).value}%`,e[t]=Math.round(l(`slider${t}`).value)),Fn()}function zn(t){e.Header[t]+=1,e.Header[t]>1&&(e.Header[t]=0),Fn()}function Vn(t,o){const n=document.createElement("div");n.className="title",n.style.padding="0px 16px",n.style.opacity="0.7",n.style.fontSize="17px",n.style.fontFamily='"Kavoon", Georgia, serif',n.appendChild(document.createTextNode(`${t} `));const i=document.createElement("span");return i.style.cursor="pointer",i.style.display="inline-block",i.style.height="14px",i.style.width="14px",i.style.borderRadius="7px",i.style.textAlign="center",i.style.backgroundColor="#C0C0C0",i.style.color="black",i.style.fontSize="13px",i.style.verticalAlign="middle",i.textContent=e.Header[o]?"-":"+",i.onclick=function(){zn(o),Game.UpdateMenu()},n.appendChild(i),n}function _n(e,t,o,n){const i=document.createElement("div");i.className="listing";const a=document.createElement("b");if(a.textContent=t,i.appendChild(a),"withTooltip"===e){i.className="listing",i.appendChild(document.createTextNode(" "));const e=document.createElement("span");e.onmouseout=function(){Game.tooltip.hide()},e.onmouseover=function(){Game.tooltip.draw(this,escape(I[n].innerHTML))},e.style.cursor="default",e.style.display="inline-block",e.style.height="10px",e.style.width="10px",e.style.borderRadius="5px",e.style.textAlign="center",e.style.backgroundColor="#C0C0C0",e.style.color="black",e.style.fontSize="9px",e.style.verticalAlign="bottom",e.textContent="?",i.appendChild(e)}return i.appendChild(document.createTextNode(": ")),i.appendChild(o),i}function Qn(e){const t=document.createDocumentFragment();t.appendChild(document.createTextNode(`${e.length} `));const o=document.createElement("span");o.onmouseout=function(){Game.tooltip.hide()};const n=document.createElement("div"),i=document.createElement("div");i.style.minWidth="140px",i.style.marginBottom="4px";const a=document.createElement("div");return a.className="name",a.style.marginBottom="4px",a.style.textAlign="center",a.textContent="Missing",i.appendChild(a),Object.keys(e).forEach((t=>{const o=document.createElement("div");o.style.textAlign="center",o.appendChild(document.createTextNode(e[t])),i.appendChild(o)})),n.appendChild(i),o.onmouseover=function(){Game.tooltip.draw(this,escape(n.innerHTML))},o.style.cursor="default",o.style.display="inline-block",o.style.height="10px",o.style.width="10px",o.style.borderRadius="5px",o.style.textAlign="center",o.style.backgroundColor="#C0C0C0",o.style.color="black",o.style.fontSize="9px",o.style.verticalAlign="bottom",o.textContent="?",t.appendChild(o),t}function Xn(t){const o=document.createElement("div");if(o.className="subsection",o.appendChild(t),o.appendChild(Vn("Lucky Cookies","Lucky")),e.Header.Lucky&&o.appendChild(function(){const e=Game.auraMult("Dragon's Fortune")?"GoldCookDragonsFortuneTooltipPlaceholder":"GoldCookTooltipPlaceholder",t=document.createElement("div");t.className="CMStatsLuckySection";const o=Game.cookies+ho()<et?A:F,n=Game.cookies+ho()<et?ue((et-(Game.cookies+ho()))/mo()):"",i=document.createDocumentFragment(),a=document.createElement("span");if(a.style.fontWeight="bold",a.className=E+o,a.textContent=pe(et),i.appendChild(a),""!==n){const e=document.createElement("small");e.textContent=` (${n})`,i.appendChild(e)}t.appendChild(_n("withTooltip",'"Lucky!" cookies required',i,e));const r=Game.cookies+ho()<nt?A:F,l=Game.cookies+ho()<nt?ue((nt-(Game.cookies+ho()))/mo()):"",s=document.createDocumentFragment(),c=document.createElement("span");if(c.style.fontWeight="bold",c.className=E+r,c.textContent=pe(nt),s.appendChild(c),""!==l){const e=document.createElement("small");e.textContent=` (${l})`,s.appendChild(e)}t.appendChild(_n("withTooltip",'"Lucky!" cookies required (frenzy)',s,e));const d=tt!==ot,p=document.createElement("span");p.style.fontWeight="bold",p.className=E+tt,p.textContent=pe(tt)+(d?` / ${pe(ot)}`:""),t.appendChild(_n("withTooltip",'"Lucky!" reward (max)'+(d?" (golden / wrath)":""),p,e));const u=document.createElement("span");u.style.fontWeight="bold",u.className=E+u,u.textContent=pe(it)+(d?` / ${pe(at)}`:""),t.appendChild(_n("withTooltip",'"Lucky!" reward (max) (frenzy)'+(d?" (golden / wrath)":""),u,e));const m=Math.min(.15*(Game.cookies+ho()),dt*pt*60*15)+13,h=document.createElement("span");return h.style.fontWeight="bold",h.className=E+h,h.textContent=pe(ut*m)+(d?` / ${pe(mt*m)}`:""),t.appendChild(_n("withTooltip",'"Lucky!" reward (cur)'+(d?" (golden / wrath)":""),h,e)),t}()),o.appendChild(Vn("Chain Cookies","Chain")),e.Header.Chain&&o.appendChild(function(){const e=Game.auraMult("Dragon's Fortune")?"GoldCookDragonsFortuneTooltipPlaceholder":"GoldCookTooltipPlaceholder",t=document.createElement("div");t.className="CMStatsChainSection";const o=Game.cookies+ho()<Ot?A:F,n=Game.cookies+ho()<Ot?ue((Ot-(Game.cookies+ho()))/mo()):"",i=document.createDocumentFragment(),a=document.createElement("span");if(a.style.fontWeight="bold",a.className=E+o,a.textContent=pe(Ot),i.appendChild(a),""!==n){const e=document.createElement("small");e.textContent=` (${n})`,i.appendChild(e)}t.appendChild(_n("withTooltip",'"Chain" cookies required',i,e));const r=Game.cookies+ho()<Wt?A:F,l=Game.cookies+ho()<Wt?ue((Wt-(Game.cookies+ho()))/mo()):"",s=document.createDocumentFragment(),c=document.createElement("span");if(c.style.fontWeight="bold",c.className=E+r,c.textContent=pe(Wt),s.appendChild(c),""!==l){const e=document.createElement("small");e.textContent=` (${l})`,s.appendChild(e)}t.appendChild(_n("withTooltip",'"Chain" cookies required (Wrath)',s,e));const d=Game.cookies+ho()<Dt?A:F,p=Game.cookies+ho()<Dt?ue((Dt-(Game.cookies+ho()))/mo()):"",u=document.createDocumentFragment(),m=document.createElement("span");if(m.style.fontWeight="bold",m.className=E+d,m.textContent=pe(Dt),u.appendChild(m),""!==p){const e=document.createElement("small");e.textContent=` (${p})`,u.appendChild(e)}t.appendChild(_n("withTooltip",'"Chain" cookies required (Frenzy)',u,e));const h=Game.cookies+ho()<$t?A:F,g=Game.cookies+ho()<$t?ue(($t-(Game.cookies+ho()))/mo()):"",f=document.createDocumentFragment(),C=document.createElement("span");if(C.style.fontWeight="bold",C.className=E+h,C.textContent=pe($t),f.appendChild(C),""!==g){const e=document.createElement("small");e.textContent=` (${g})`,f.appendChild(e)}t.appendChild(_n("withTooltip",'"Chain" cookies required (frenzy) (Wrath)',f,e)),t.appendChild(_n("withTooltip",'"Chain" reward (max) (golden / wrath)',document.createTextNode(`${pe(Lt[0])} / ${pe(jt[0])}`),e)),t.appendChild(_n("withTooltip",'"Chain" reward (max) (frenzy) (golden / wrath)',document.createTextNode(`${pe(Ut[0])} / ${pe(Ut[0])}`),e));const y=Math.min(60*Game.cookiesPs*60*6*pt,.5*Game.cookies),b=tn(7,y,ut)[0],k=tn(6,y,mt)[0];return t.appendChild(_n("withTooltip",'"Chain" reward (cur) (golden / wrath)',document.createTextNode(`${pe(b)} / ${pe(k)}`),e)),t.appendChild(_n("withTooltip","CPS needed for next level (g / w)",document.createTextNode(`${pe(Ft)} / ${pe(At)}`),"ChainNextLevelPlaceholder")),t.appendChild(_n("withTooltip","CPS needed for next level (frenzy) (g / w)",document.createTextNode(`${pe(Ht)} / ${pe(Rt)}`),"ChainNextLevelPlaceholder")),t}()),Game.Objects["Wizard tower"].minigameLoaded&&(o.appendChild(Vn("Spells","Spells")),e.Header.Spells&&o.appendChild(function(){const e=document.createElement("div");e.className="CMStatsSpellsSection";const t=Game.cookies+ho()<rt?A:F,o=Game.cookies+ho()<rt?ue((rt-(Game.cookies+ho()))/mo()):"",n=document.createDocumentFragment(),i=document.createElement("span");if(i.style.fontWeight="bold",i.className=E+t,i.textContent=pe(rt),n.appendChild(i),""!==o){const e=document.createElement("small");e.textContent=` (${o})`,n.appendChild(e)}e.appendChild(_n("withTooltip",'"Conjure Baked Goods" cookies required',n,"GoldCookTooltipPlaceholder")),e.appendChild(_n("withTooltip",'"Conjure Baked Goods" reward (max)',document.createTextNode(pe(lt)),"GoldCookTooltipPlaceholder"));const a=Game.cookies+ho()<7*rt?A:F,r=Math.min(.15*(Game.cookies+ho()),60*dt*30),l=Game.cookies+ho()<7*rt?ue((7*rt-(Game.cookies+ho()))/mo()):"",s=document.createDocumentFragment(),c=document.createElement("span");if(c.style.fontWeight="bold",c.className=E+a,c.textContent=pe(7*rt),s.appendChild(c),""!==l){const e=document.createElement("small");e.textContent=` (${l})`,s.appendChild(e)}return e.appendChild(_n("withTooltip",'"Conjure Baked Goods" cookies required (frenzy)',s,"GoldCookTooltipPlaceholder")),e.appendChild(_n("withTooltip",'"Conjure Baked Goods" reward (max) (frenzy)',document.createTextNode(pe(7*lt)),"GoldCookTooltipPlaceholder")),e.appendChild(_n("withTooltip",'"Conjure Baked Goods" reward (cur)',document.createTextNode(pe(r)),"GoldCookTooltipPlaceholder")),st&&e.appendChild(_n("withTooltip",'"Spontaneous Edifice" cookies required (most expensive building)',document.createTextNode(`${pe(st)} (${ct})`),"GoldCookTooltipPlaceholder")),e}())),Game.Objects.Farm.minigameLoaded&&(o.appendChild(Vn("Garden","Garden")),e.Header.Garden&&o.appendChild(function(){const e=document.createElement("div");e.className="CMStatsGardenSection";const t=Game.cookies<60*Game.cookiesPs*10*100?A:F,o=document.createElement("span");o.style.fontWeight="bold",o.className=E+t,o.textContent=pe(60*Game.cookiesPs*10*100),e.appendChild(_n("basic","Cookies required for max reward of Bakeberry: ",o));const n=Game.cookies<60*Game.cookiesPs*100?A:F,i=document.createElement("span");i.style.fontWeight="bold",i.className=E+n,i.textContent=pe(60*Game.cookiesPs*100),e.appendChild(_n("basic","Cookies required for max reward of Chocoroot: ",i));const a=Game.cookies<60*Game.cookiesPs*60*25?A:F,r=document.createElement("span");r.style.fontWeight="bold",r.className=E+a,r.textContent=pe(60*Game.cookiesPs*60*25),e.appendChild(_n("basic","Cookies required for max reward of Queenbeet: ",r));const l=Game.cookies<60*Game.cookiesPs*15*100?A:F,s=document.createElement("span");s.style.fontWeight="bold",s.className=E+l,s.textContent=pe(60*Game.cookiesPs*15*100),e.appendChild(_n("basic","Cookies required for max reward of Duketater: ",s));const c=[];return Object.keys(jo).forEach((e=>{Game.HasUnlocked(jo[e])||c.push(jo[e])})),0!==c.length&&e.appendChild(_n("basic","Rare plant drops left to unlock",Qn(c))),e}())),o.appendChild(Vn("Prestige","Prestige")),e.Header.Prestige&&o.appendChild(function(){const t=document.createElement("div");t.className="CMStatsPrestigeSection";const o=Math.floor(Game.HowMuchPrestige(zt+Game.cookiesReset+vt+(Game.HasUnlocked("Chocolate egg")&&!Game.Has("Chocolate egg")?Ct:0)));t.appendChild(_n("withTooltip","Prestige level (cur / max)",document.createTextNode(`${pe(Game.prestige)} / ${pe(o)}`),"PrestMaxTooltipPlaceholder"));const n=Math.max(0,Game.HowManyCookiesReset(o+1)-(zt+Game.cookiesReset+vt+(Game.HasUnlocked("Chocolate egg")&&!Game.Has("Chocolate egg")&&Ct?Ct:0))),i=document.createDocumentFragment();i.appendChild(document.createTextNode(pe(n)));const a=document.createElement("small");a.textContent=` (${ue(n/Vt,1)})`,i.appendChild(a),t.appendChild(_n("withTooltip","Cookies to next level",i,"NextPrestTooltipPlaceholder")),t.appendChild(_n("withTooltip","Heavenly chips (cur / max)",document.createTextNode(`${pe(Game.heavenlyChips)} / ${pe(o-Game.prestige+Game.heavenlyChips)}`),"HeavenChipMaxTooltipPlaceholder")),t.appendChild(_n("basic","Heavenly chips per second (last 5 seconds)",document.createTextNode(pe(ht,2))));const r=Number(e.HeavenlyChipsTarget);if(!Number.isNaN(r)){const e=r-Math.floor(Game.HowMuchPrestige(Game.cookiesReset+Game.cookiesEarned));e>0&&(t.appendChild(_n("basic","Heavenly chips to target set in settings (cur)",document.createTextNode(pe(e)))),t.appendChild(_n("basic","Time till target (cur, current 5 second average)",document.createTextNode(ue(e/ht)))))}const l=function(e){let t=Game.cookiesPs;Mo(),0===to["Heavenly key"].bought&&(to["Heavenly chip secret"].bought=1,to["Heavenly cookie stand"].bought=1,to["Heavenly bakery"].bought=1,to["Heavenly confectionery"].bought=1,to["Heavenly key"].bought=1,Oo(),t=Xe,Mo()),zt>=1e6&&No("Sacrifice"),zt>=1e9&&No("Oblivion"),zt>=1e12&&No("From scratch"),zt>=1e15&&No("Nihilism"),zt>=1e18&&No("Dematerialize"),zt>=1e21&&No("Nil zero zilch"),zt>=1e24&&No("Transcendence"),zt>=1e27&&No("Obliterate"),zt>=1e30&&No("Negative void"),zt>=1e33&&No("To crumbs, you say?"),zt>=1e36&&No("You get nothing"),zt>=1e39&&No("Humble rebeginnings"),zt>=1e42&&No("The end of the world"),zt>=1e45&&No("Oh, you're back"),zt>=1e48&&No("Lazarus"),zt>=1e51&&No("Smurf account"),zt>=1e54&&No("If at first you don't succeed"),to["Heavenly chip secret"].bought=1,to["Heavenly cookie stand"].bought=1,to["Heavenly bakery"].bought=1,to["Heavenly confectionery"].bought=1,to["Heavenly key"].bought=1,Ue=e;const o=De;Oo(),Ho(),o!==De&&Oo();const n=Xe-t;return Ue=Game.prestige,n}(o),s=document.createDocumentFragment();s.appendChild(document.createTextNode(pe(l)));const c=Math.round(l/Game.cookiesPs*1e4);if(Number.isFinite(c)&&0!==c){const e=document.createElement("small");e.textContent=` (${c/100}% of income)`,s.appendChild(e)}t.appendChild(_n("withTooltip","Reset bonus income",s,"ResetTooltipPlaceholder"));const d=Math.floor(Game.HowMuchPrestige(Game.cookiesReset)),p=Math.floor(Game.HowMuchPrestige(Game.cookiesReset+Game.cookiesEarned)),u=p-d;if(!Game.Has("Lucky digit")){let e=7-p%10;e<0&&(e+=10);const o=u+e,n=p+e,i=document.createDocumentFragment();i.appendChild(document.createTextNode(`${n.toLocaleString()} / ${o.toLocaleString()} (+${e})`)),t.appendChild(_n("basic",'Next "Lucky Digit" (total / reset)',i))}if(!Game.Has("Lucky number")){let e=777-p%1e3;e<0&&(e+=1e3);const o=u+e,n=p+e,i=document.createDocumentFragment();i.appendChild(document.createTextNode(`${n.toLocaleString()} / ${o.toLocaleString()} (+${e})`)),t.appendChild(_n("basic",'Next "Lucky Number" (total / reset)',i))}if(!Game.Has("Lucky payout")){let e=777777-p%1e6;e<0&&(e+=1e6);const o=u+e,n=p+e,i=document.createDocumentFragment();i.appendChild(document.createTextNode(`${n.toLocaleString()} / ${o.toLocaleString()} (+${e})`)),t.appendChild(_n("basic",'Next "Lucky Payout" (total / reset)',i))}return t}()),Game.cpsSucked>0&&(o.appendChild(Vn("Wrinklers","Wrink")),e.Header.Wrink)){const e=document.createDocumentFragment();e.appendChild(document.createTextNode(`${pe(vt)} / ${pe(wt)} `));const t=document.createElement("a");t.textContent="Pop All Normal",t.className="option",t.onclick=function(){Wn()},e.appendChild(t),o.appendChild(_n("basic","Rewards of Popping (All/Normal)",e));const n=document.createDocumentFragment();n.appendChild(document.createTextNode(`${pe(xt[0])} `));const i=document.createElement("a");i.textContent="Pop Single Fattest",i.className="option",i.onclick=function(){null!==xt[1]&&(Game.wrinklers[xt[1]].hp=0)},n.appendChild(i),o.appendChild(_n("basic",`Rewards of Popping Single Fattest Non-Shiny Wrinkler (id: ${null!==xt[1]?xt[1]:"None"})`,n))}if(o.appendChild(function(){const t=document.createElement("div");t.className="CMStatsSeasonSection";let o=!1;const n=[];Object.keys(Lo).forEach((e=>{Game.Has(Lo[e])||(n.push(Lo[e]),o=!0)}));const i=[];Object.keys(Wo).forEach((e=>{Game.Has(Wo[e])||(i.push(Wo[e]),o=!0)}));const a=[];Object.keys(Ao).forEach((e=>{Game.Has(Ao[e])||(a.push(Ao[e]),o=!0)}));const r=[];Object.keys(Game.eggDrops).forEach((e=>{Game.HasUnlocked(Game.eggDrops[e])||(r.push(Game.eggDrops[e]),o=!0)}));const l=[];Object.keys(Game.rareEggDrops).forEach((e=>{Game.HasUnlocked(Game.rareEggDrops[e])||(l.push(Game.rareEggDrops[e]),o=!0)}));const s=Game.HasUnlocked("Chocolate egg")&&!Game.Has("Chocolate egg"),c=Game.Has("Century egg");if(("christmas"===Game.season||o||s||c)&&(t.appendChild(Vn("Season Specials","Sea")),e.Header.Sea)){if(0!==n.length){t.appendChild(_n("basic","Halloween cookies left to buy",Qn(n)));let e=.95;if(Game.HasAchiev("Spooky cookies")&&(e=.8),Game.Has("Starterror")&&(e*=.9),e*=1/Game.dropRateMult(),Game.hasGod){const t=Game.hasGod("seasons");1===t?e*=.9:2===t?e*=.95:3===t&&(e*=.97)}const o=n.length/7;t.appendChild(_n("basic","Chance of receiving a cookie from wrinkler/shiny wrinkler",document.createTextNode(`${pe((1-e)*o*100)}% / ${pe((1-.9*e)*o*100)}%`)))}if(0!==i.length){t.appendChild(_n("basic","Christmas cookies left to buy",Qn(i)));let e=.8;if(Game.HasAchiev("Let it snow")&&(e=.6),e*=1/Game.dropRateMult(),Game.Has("Starsnow")&&(e*=.95),Game.hasGod){const t=Game.hasGod("seasons");1===t?e*=.9:2===t?e*=.95:3===t&&(e*=.97)}const o=i.length/7;t.appendChild(_n("basic","Chance of receiving a cookie from reindeer",document.createTextNode(`${pe((1-e)*o*100)}%`)))}0!==a.length&&t.appendChild(_n("basic","Valentine cookies left to buy",Qn(a)));const e=function(e){let t=e*(1/Game.dropRateMult());if(Game.HasAchiev("Hide & seek champion")&&(t*=.7),Game.Has("Omelette")&&(t*=.9),Game.Has("Starspawn")&&(t*=.9),Game.hasGod){const e=Game.hasGod("seasons");1===e?t*=.9:2===e?t*=.95:3===e&&(t*=.97)}const o=1-t,n=Game.eggDrops.length-r.length,i=Game.rareEggDrops.length-l.length,a=.9*o*(1-n/Game.eggDrops.length),s=.1*o*(1-i/Game.rareEggDrops.length),c=.9*o*(n/Game.eggDrops.length)+.1*o*(i/Game.rareEggDrops.length);return[a+.9*c*(1-n/Game.eggDrops.length),s+.1*c*(1-i/Game.rareEggDrops.length)]};0!==r.length&&(t.appendChild(_n("basic","Normal easter eggs left to unlock",Qn(r))),t.appendChild(_n("basic","Chance of receiving an egg from wrinkler/golden cookie",document.createTextNode(`${pe(100*e(.98)[0])}% / ${pe(100*e(.9)[0])}%`)))),0!==l.length&&(t.appendChild(_n("basic","Rare easter eggs left to unlock",Qn(l))),t.appendChild(_n("basic","Chance of receiving a rare egg from wrinkler/golden cookie",document.createTextNode(`${pe(100*e(.98)[1])}% / ${pe(100*e(.9)[1])}%`)))),"christmas"===Game.season&&t.appendChild(_n("basic","Reindeer reward",document.createTextNode(pe(ft)))),s&&t.appendChild(_n("withTooltip","Chocolate egg cookies",document.createTextNode(pe(Ct)),"ChoEggTooltipPlaceholder")),c&&t.appendChild(_n("basic","Century egg multiplier",document.createTextNode(Math.round(1e4*(gt-1))/100+"%")))}return t}()),o.appendChild(Vn("Miscellaneous","Misc")),e.Header.Misc){if(o.appendChild(_n("basic",`Average cookies per second (past ${J[e.AvgCPSHist]<60?`${J[e.AvgCPSHist]} seconds`:J[e.AvgCPSHist]/60+(3===e.AvgCPSHist?" minute":" minutes")})`,document.createTextNode(pe(mo(),3)))),o.appendChild(_n("basic",`Average cookie clicks per second (past ${Z[e.AvgClicksHist]}${0===e.AvgClicksHist?" second":" seconds"})`,document.createTextNode(pe(Pt,1)))),o.appendChild(_n("basic",`Cookies from clicking (past ${Z[e.AvgClicksHist]}${0===e.AvgClicksHist?" second":" seconds"})`,document.createTextNode(pe(he.calcSum(Pt*Z[e.AvgClicksHist]))))),Game.Has("Fortune cookies")){const e=[];Object.keys(Fo).forEach((t=>{Game.Has(Fo[t])||e.push(Fo[t])})),0!==e.length&&o.appendChild(_n("basic","Fortune Upgrades Left to Buy",Qn(e)))}if(e.ShowMissedGC&&o.appendChild(_n("basic","Missed golden cookies",document.createTextNode(pe(Game.missedGoldenClicks)))),Game.prefs.autosave){const e=document.createElement("span");e.id="CMStatsAutosaveTimer",e.innerText=Game.sayTime(60*Game.fps-(Game.OnAscend?0:Game.T%(60*Game.fps)),4),o.appendChild(_n("basic","Time till autosave",e))}}l("menu").insertBefore(o,l("menu").childNodes[2]),e.MissingUpgrades&&l("menu").childNodes.forEach((e=>{if(e.children[0])if("Prestige"===e.children[0].innerHTML&&Et){const t=Et.match(new RegExp("div","g")||0).length/2,o=document.createElement("div");o.id="CMMissingUpgradesPrestigeTitle",o.className="listing";const n=document.createElement("div");n.innerHTML=`<b>Missing Prestige upgrades:</b> ${t}/${Game.PrestigeUpgrades.length} (${Math.floor(t/Game.PrestigeUpgrades.length*100)}%)`,o.appendChild(n),e.appendChild(o);const i=document.createElement("div");i.className="listing crateBox",i.innerHTML=Et,e.appendChild(i)}else if("Upgrades"===e.children[0].innerHTML){if(Bt){const t=Bt.match(new RegExp("div","g")||0).length/2,o=document.createElement("div");o.id="CMMissingUpgradesTitle",o.className="listing";const n=document.createElement("div");n.innerHTML=`<b>Missing normal upgrades:</b> ${t}/${Game.UpgradesByPool[""].length+Game.UpgradesByPool.tech.length} (${Math.floor(t/(Game.UpgradesByPool[""].length+Game.UpgradesByPool.tech.length)*100)}%)`,o.appendChild(n),e.insertBefore(o,e.childNodes[3]);const i=document.createElement("div");i.className="listing crateBox",i.innerHTML=Bt,e.insertBefore(i,document.getElementById("CMMissingUpgradesTitle").nextSibling)}if(Nt){const t=Nt.match(new RegExp("div","g")||0).length/2,o=document.createElement("div");o.id="CMMissingUpgradesCookiesTitle",o.className="listing";const n=document.createElement("div");n.innerHTML=`<b>Missing Cookie upgrades:</b> ${t}/${Game.UpgradesByPool.cookie.length} (${Math.floor(t/Game.UpgradesByPool.cookie.length*100)}%)`,o.appendChild(n),e.appendChild(o);const i=document.createElement("div");i.className="listing crateBox",i.innerHTML=Nt,e.appendChild(i)}}})),e.MissingAchievements&&function(){let t;Object.values(document.querySelectorAll("div.title")).forEach((e=>{e.textContent.includes("Achievements")&&(t=e.parentElement.querySelectorAll("div.listing.crateBox")[0])})),e.MissingAchievements&&Object.values(t.children).forEach((e=>{if(!e.className.includes("enabled")){const t=e.onclick.toString().split(/\[(.*)\]/gi)[1],{icon:o}=Game.AchievementsById[t];e.style.backgroundPosition=`${48*-o[0]}px ${48*-o[1]}px`,e.onmouseover=function(){Game.mouseDown||(Game.setOnCrate(this),Game.tooltip.dynamic=1,Game.tooltip.draw(this,(function(){return function(e){const t=[];"shadow"===e.pool?t.push("Shadow Achievement","#9700cf"):t.push("Achievement",0),t.push("Locked",0);let o=0;(Game.Has("Neuromancy")||Game.sesame&&"debug"===e.pool)&&(o=1),o&&0===e.won?t.push("Click to win!","#00c462"):o&&e.won>0&&t.push("Click to lose!","#00c462");let{icon:n}=e;e.iconFunction&&(n=e.iconFunction());let{desc:i}=e;e.descFunc&&(i=e.descFunc("stats"));let a="";for(let e=0;e<t.length;e+=2)e%2==0&&(a+=` <div class="tag" style="color:${0===t[e+1]?"#fff":t[e+1]};">[${t[e]}]</div>`);return a=a.substring(1),`<div style="padding:8px 4px;min-width:350px;opacity:0.5">\n  <div class="icon" style="float:left;margin-left:-8px;margin-top:-8px;background-position:${48*-n[0]}px ${48*-n[1]}px;"></div>\n  <div class="name">${e.name}</div>\n  ${a}<div class="line"></div><div class="description">${i}</div></div>\n  ${Game.sesame?`<div style="font-size:9px;">Id : ${e.id} | Order : ${Math.floor(e.order)}${e.tier?` | Tier : ${e.tier}`:""}</div>`:""}`}(Game.AchievementsById[t])}),"top"),Game.tooltip.wobble())}}}))}()}const qn={Favourite:"Favourite Settings",Calculation:"Calculation",Notation:"Notation",Colours:"Colours and colour coding",BarsDisplay:"Infobars and visual settings",Tooltip:"Tooltips",Statistics:"Statistics",Notification:"Notifications",Miscellaneous:"Miscellaneous"},Yn={NotificationGeneral:"General Notifications",NotificationGC:"Golden Cookie",NotificationFC:"Fortune Cookie",NotificationSea:"Season Special",NotificationGard:"Garden Tick",NotificationMagi:"Full Magic Bar",NotificationWrink:"Wrinkler",NotificationWrinkMax:"Maximum Wrinklers"};function Kn(t,o){const n=document.createElement("div");n.className="title",n.style.opacity="0.7",n.style.fontSize="17px",n.appendChild(document.createTextNode(`${o} `));const i=document.createElement("span");return i.style.cursor="pointer",i.style.display="inline-block",i.style.height="14px",i.style.width="14px",i.style.borderRadius="7px",i.style.textAlign="center",i.style.backgroundColor="#C0C0C0",i.style.color="black",i.style.fontSize="13px",i.style.verticalAlign="middle",i.textContent=e.Header[t]?"-":"+",i.onclick=function(){zn(t),Game.UpdateMenu()},n.appendChild(i),n}function Jn(t){const o=document.createElement("div");if(o.className="listing",1===e.FavouriteSettings&&o.appendChild(function(e){const t=document.createElement("a");return de.includes(e)?(t.innerText="★",t.style.color="yellow"):t.innerText="☆",t.className="option",t.onclick=function(){!function(e){de.includes(e)?de=de.filter((function(t){return t!==e})):de.push(e)}(e),Fn(),Game.UpdateMenu()},t.onmouseover=function(){Game.tooltip.draw(this,escape(I.FavouriteSettingPlaceholder.innerHTML))},t.onmouseout=function(){Game.tooltip.hide()},t.appendChild(document.createTextNode(" ")),t}(t)),"bool"===vn[t].type){const n=document.createElement("a");vn[t].toggle&&0===e[t]?n.className="option off":n.className="option",n.id=Rn+t,n.onclick=function(){!function(t){e[t]+=1,e[t]===vn[t].label.length?(e[t]=0,vn[t].toggle&&(l(Rn+t).className="option off")):l(Rn+t).className="option",void 0!==vn[t].func&&vn[t].func(),Fn()}(t),Game.UpdateMenu()},n.textContent=vn[t].label[e[t]],o.appendChild(n);const i=document.createElement("label");return i.textContent=vn[t].desc,i.style.lineHeight="1.6",o.appendChild(i),o}if("vol"===vn[t].type){const n=document.createElement("div");n.className="sliderBox";const i=document.createElement("div");i.style.float="left",i.innerHTML=vn[t].desc,n.appendChild(i);const a=document.createElement("div");a.id=`slider${t}right`,a.style.float="right",a.innerHTML=`${e[t]}%`,n.appendChild(a);const r=document.createElement("input");r.className="slider",r.id=`slider${t}`,r.style.clear="both",r.type="range",r.min="0",r.max="100",r.step="1",r.value=e[t],r.oninput=function(){In(t),Game.UpdateMenu()},r.onchange=function(){In(t),Game.UpdateMenu()},n.appendChild(r),o.appendChild(n);const l=document.createElement("a");return l.className="option",l.onclick=function(){Bn(e[t.replace("Volume","SoundURL")],t.replace("Volume","Sound"),t,!0)},l.textContent="Test sound",o.appendChild(l),o}if("url"===vn[t].type){const n=document.createElement("span");n.className="option",n.textContent=`${vn[t].label} `,n.style.lineHeight="1.6",o.appendChild(n);const i=document.createElement("input");i.id=Rn+t,i.className="option",i.type="text",i.readOnly=!0,i.setAttribute("value",e[t]),i.style.width="300px",o.appendChild(i),o.appendChild(document.createTextNode(" "));const a=document.createElement("input");a.id=`${Rn+t}Prompt`,a.className="option",a.type="text",a.setAttribute("value",e[t]);const r=document.createElement("a");r.className="option",r.onclick=function(){!function(e,t){Game.promptWrapL.className="framed";const o=e;Game.promptL.innerHTML=`${o}<div class="optionBox"></div>`,Object.keys(t).forEach((e=>{const o=document.createElement("a");o.id=`promptOption${e}`,o.className="option",o.onclick=function(){PlaySound("snd/tick.mp3"),t[e][1]()},o.textContent=t[e][0],Game.promptL.children[1].appendChild(o)})),Game.promptAnchorL.style.display="block",Game.darkenL.style.display="block",Game.promptL.focus(),Game.promptOn=1,Game.UpdatePrompt()}(a.outerHTML,[["Save",function(){e[t]=l(`CMConfig${t}Prompt`).value,Fn(),Game.ClosePrompt(),Game.UpdateMenu()}],["Cancel",function(){Game.ClosePrompt()}]])},r.textContent="Edit",o.appendChild(r);const s=document.createElement("label");return s.textContent=vn[t].desc,s.style.lineHeight="1.6",o.appendChild(s),o}if("colour"===vn[t].type){const n=document.createElement("span");n.className="option";const i=document.createElement("input");i.id=t,i.style.width="65px",i.setAttribute("value",e[t]),n.appendChild(i),new Un(i,{hash:!0,position:"right",onInput:function(){e[this.targetElement.id]=this.toHEXString(),On(),Fn(),Game.UpdateMenu()}});const a=document.createElement("label");if(a.textContent=vn[t].desc,a.style.lineHeight="1.6",n.appendChild(a),t.includes("Flash")){const e=document.createElement("a");e.className="option",e.onclick=function(){Pn(3,t.replace("Colour",""),!0)},e.textContent="Test flash",n.appendChild(e)}return o.appendChild(n),$n().init(),o}if("numscale"===vn[t].type){const n=document.createElement("span");n.className="option",n.textContent=`${vn[t].label} `,n.style.lineHeight="1.6",o.appendChild(n);const i=document.createElement("input");i.id=Rn+t,i.className="option",i.type="number",i.value=e[t],i.min=vn[t].min,i.max=vn[t].max,i.oninput=function(){e[t]=this.value,Fn(),gn(),Game.UpdateMenu()},o.appendChild(i),o.appendChild(document.createTextNode(" "));const a=document.createElement("label");return a.textContent=vn[t].desc,a.style.lineHeight="1.6",o.appendChild(a),o}return o}function Zn(t){if(1===e.TimerBar&&0===e.TimerBarPos){const e=parseInt(l("CMTimerBar").style.height,10);Game.mouseY-=e,t(),Game.mouseY+=e}else t()}function ei(){S.Beautify=Beautify,Beautify=pe,S.CalculateGains=Game.CalculateGains,Game.CalculateGains=function(){S.CalculateGains(),_e=1,P=Date.now(),B=Date.now()},S.tooltip={},S.tooltip.draw=Game.tooltip.draw,S.tooltip.drawMod=new Function(`return ${Game.tooltip.draw.toString().split("this").join("Game.tooltip")}`)(),Game.tooltip.draw=function(e,t,o){S.tooltip.drawMod(e,t,o)},S.tooltip.update=Game.tooltip.update,S.tooltip.updateMod=new Function(`return ${Game.tooltip.update.toString().split("this.").join("Game.tooltip.")}`)(),Game.tooltip.update=function(){S.tooltip.updateMod(),function(){if("store"===Game.tooltip.origin){let t=0;1===e.ToolWarnLucky&&1===e.ToolWarnPos&&null!==l("CMDispTooltipWarningParent")&&(t=l("CMDispTooltipWarningParent").clientHeight-4),Game.tooltip.tta.style.top=`${Math.min(parseInt(Game.tooltip.tta.style.top,10),l("game").clientHeight+l("topBar").clientHeight-Game.tooltip.tt.clientHeight-t-46)}px`}}()},S.UpdateWrinklers=Game.UpdateWrinklers,Game.UpdateWrinklers=function(){Zn(S.UpdateWrinklers)},S.UpdateSpecial=Game.UpdateSpecial,Game.UpdateSpecial=function(){Zn(S.UpdateSpecial)},l("bigCookie").removeEventListener("click",Game.ClickCookie,!1),l("bigCookie").addEventListener("click",(function(){Zn(Game.ClickCookie)}),!1),S.RebuildUpgrades=Game.RebuildUpgrades,Game.RebuildUpgrades=function(){S.RebuildUpgrades(),M=[],Object.keys(Game.UpgradesInStore).forEach((e=>{null!==l(`upgrade${e}`).onmouseover&&(M[e]=l(`upgrade${e}`).onmouseover,l(`upgrade${e}`).onmouseover=function(){Game.mouseDown||(Game.setOnCrate(this),Game.tooltip.dynamic=1,Game.tooltip.draw(this,(function(){return An("u",`${e}`)}),"store"),Game.tooltip.wobble())})})),Game.CalculateGains()},S.ClickProduct=Game.ClickProduct,Game.ClickProduct=function(t){(!e.BulkBuyBlock||Game.ObjectsById[t].bulkPrice<Game.cookies||-1===Game.buyMode)&&S.ClickProduct(t)},S.DescribeDragonAura=Game.DescribeDragonAura,Game.DescribeDragonAura=function(t){S.DescribeDragonAura(t),function(t){if(1===e.DragonAuraInfo){const[e,o]=function(e){Mo(),l("promptContent").children[0].innerHTML.includes("secondary")?Re=e:$e=e;let t=0;if($e!==Ye||Re!==Ke)for(let e=Game.ObjectsById.length-1;e>-1;--e)if(Game.ObjectsById[e].amount>0){const o=eo[Game.ObjectsById[e].name].name;eo[o].amount-=1,We-=1,t=eo[o].basePrice*Game.priceIncrease**Math.max(0,eo[o].amount-1-eo[o].free),t=Game.modifyBuildingPrice(eo[o],t),t=Math.ceil(t);break}const o=De;return Oo(),Ho(),o!==De&&Oo(),[Xe-Game.cookiesPs,t]}(t),n=ue(o/(e+Game.cookiesPs));let i;i=0===Game.cookiesPs?pe(1/0):pe(e/Game.cookiesPs*100),l("dragonAuraInfo").style.minHeight="60px",l("dragonAuraInfo").style.margin="8px",l("dragonAuraInfo").appendChild(document.createElement("div")).className="line";const a=document.createElement("div");a.style.minWidth="200px",a.style.textAlign="center",a.textContent=`Picking this aura will change CPS by ${pe(e)} (${i}% of current CPS).`,l("dragonAuraInfo").appendChild(a);const r=document.createElement("div");r.style.minWidth="200px",r.style.textAlign="center",r.textContent=`It will take ${n} to recover the cost.`,l("dragonAuraInfo").appendChild(r)}}(t)},S.ToggleSpecialMenu=Game.ToggleSpecialMenu,Game.ToggleSpecialMenu=function(e){S.ToggleSpecialMenu(e),function(){if(null!==(l("specialPopup").className.match(/onScreen/)&&l("specialPopup").children[0].style.background.match(/dragon/)))for(let e=0;e<l("specialPopup").childNodes.length;e++)"optionBox"===l("specialPopup").childNodes[e].className&&(l("specialPopup").children[e].onmouseover=function(){xn(),Game.tooltip.dynamic=1,Game.tooltip.draw(l("specialPopup"),`<div style="min-width:200px;text-align:center;">${Ze}</div>`,"this"),Game.tooltip.wobble()},l("specialPopup").children[e].onmouseout=function(){Game.tooltip.shouldHide=1})}()},S.UpdateMenu=Game.UpdateMenu,Game.UpdateMenu=function(){void 0!==$n().picker&&void 0!==$n().picker.owner||(S.UpdateMenu(),function(){const t=document.createElement("div");t.className="title","prefs"===Game.onMenu?(t.textContent="Cookie Monster Settings",function(t){const o=document.createDocumentFragment();o.appendChild(t),Object.keys(qn).forEach((t=>{if("Favourite"===t){if(0!==de.length&&e.FavouriteSettings>0&&(o.appendChild(Kn(t,qn[t])),e.Header[t]))for(let e=0;e<de.length;e++)o.appendChild(Jn(de[e]))}else o.appendChild(Kn(t,qn[t])),e.Header[t]&&("Notification"===t?Object.keys(Yn).forEach((t=>{const n=Kn(t,Yn[t]);n.style.fontSize="15px",n.style.opacity="0.5",o.appendChild(n),e.Header[t]&&Object.keys(vn).forEach((e=>{vn[e].group===t&&o.appendChild(Jn(e))}))})):Object.keys(vn).forEach((e=>{vn[e].group===t&&o.appendChild(Jn(e))})))}));const n=document.createElement("div");n.className="listing";const i=document.createElement("a");i.className="option",i.onclick=function(){Ln(un)},i.textContent="Restore Default",n.appendChild(i),o.appendChild(n),l("menu").childNodes[2].insertBefore(o,l("menu").childNodes[2].childNodes[l("menu").childNodes[2].childNodes.length-1])}(t)):"stats"===Game.onMenu?e.Stats&&(t.textContent="Cookie Monster Statistics",Xn(t)):"log"===Game.onMenu&&(t.textContent="Cookie Monster ",function(t){const o=document.createElement("div");o.className="subsection";const n=document.createElement("span");if(n.style.cursor="pointer",n.style.display="inline-block",n.style.height="14px",n.style.width="14px",n.style.borderRadius="7px",n.style.textAlign="center",n.style.backgroundColor="#C0C0C0",n.style.color="black",n.style.fontSize="13px",n.style.verticalAlign="middle",n.textContent=e.Header.InfoTab?"-":"+",n.onclick=function(){zn("InfoTab"),Game.UpdateMenu()},t.appendChild(n),o.appendChild(t),e.Header.InfoTab){const e=document.createElement("div");e.innerHTML='<div class="listing">\n <a href="https://github.com/Aktanusa/CookieMonster" target="blank">Cookie Monster</a>\n offers a wide range of tools and statistics to enhance your game experience.\n It is not a cheat interface – although it does offer helpers for golden cookies and such, everything can be toggled off at will to only leave how much information you want.</br>\n Progess on new updates and all previous release notes can be found on the GitHub page linked above!</br>\n Please also report any bugs you may find over there!</br>\n </div>\n ',o.appendChild(e);const t=document.createElement("div");t.innerHTML="<div class=\"listing\">\nThese are the release notes for the latest update (v 2.031.7).</br>\n</br>\nThis update implements the following functions:</br>\n- Updated the ordering and description of all settings of Cookie Monster</br>\n- Colours of Flash notifications can now be set by the user</br>\n- You can now select a number of settings to be your 'favourite'. They will be displayed at the top of all other settings in a new section. This functionality is toggleable and you can find it in the miscellaneous section</br>\n- The statistics page now shows cookies earned by clicking for a selected time period</br>\n- The statistics page can now show missing normal achievements, this can be toggled in the settings</br>\n</br>\nThis update fixes the following bugs:</br>\n- Fixed some typo's</br>\n- No longer show the wrinkler pop buttons when wrinklers can't spawn</br>\n- Fixed incorrect calculation of plant maturity for relevant tooltips</br>\n- Fixed edit prompts not working in settings</br>\n- Fixed red and purple being switched around for PP colour coding</br>\n- Fixed incorrect notation of numbers in Dragon tooltip and interface</br>\n- Fixed incorrect calculation of changing Dragon aura cost</br>\n- Fixed incorrect calculation of changing Pantheon gods effect</br>\n- Fixed fluctuating PP values when god Cyclius was selected</br>\n- Fixed black flickering when Golden Cookies or buffs start</br>\n</div>\n",o.appendChild(t)}const i=l("menu").children[1];i.insertBefore(o,i.children[1])}(t))}())},S.sayTime=Game.sayTime,V=function(e,t){return Number.isNaN(e)||e<=0?S.sayTime(e,t):ue(e/Game.fps,1)},S.Logic=Game.Logic,Game.Logic=function(){S.Logic();let t="Cookie Clicker";"fools"===Game.season&&(t="Cookie Baker"),ce=`${Game.OnAscend?"Ascending! ":""}${pe(Game.cookies)} ${1===Game.cookies?"cookie":"cookies"} - ${t}`,function(){if(Game.OnAscend||0===e.Title)document.title=ce;else if(1===e.Title){let e,t,o,n=!1,i=!1;e=_t?_t.wrath?`[W${Math.ceil(_t.life/Game.fps)}]`:`[G${Math.ceil(_t.life/Game.fps)}]`:Game.Has("Golden switch [off]")?"[GS]":`[${Number(l("CMTimerBarGCMinBar").textContent)<0?"!":""}${Math.ceil((Game.shimmerTypes.golden.maxTime-Game.shimmerTypes.golden.time)/Game.fps)}]`,f&&(n=!0,t="[F]"),"christmas"===Game.season&&(i=!0,o=g?`[R${Math.ceil(Qt.life/Game.fps)}]`:`[${Number(l("CMTimerBarRenMinBar").textContent)<0?"!":""}${Math.ceil((Game.shimmerTypes.reindeer.maxTime-Game.shimmerTypes.reindeer.time)/Game.fps)}]`);let a=ce;"["===a.charAt(0)&&(a=a.substring(a.lastIndexOf("]")+1)),document.title=`${e+(n?t:"")+(i?o:"")} ${a}`}else if(2===e.Title){let e="",t=!1;_t&&(t=!0,_t.wrath?e+=`[W${Math.ceil(_t.life/Game.fps)}]`:e+=`[G${Math.ceil(_t.life/Game.fps)}]`),f&&(t=!0,e+="[F]"),"christmas"===Game.season&&g&&(e+=`[R${Math.ceil(Qt.life/Game.fps)}]`,t=!0),t&&(e+=" - ");let o="Cookie Clicker";"fools"===Game.season&&(o="Cookie Baker"),e+=o,document.title=e}}(),function(){const t=Math.max(0,Game.HowManyCookiesReset(Math.floor(Game.HowMuchPrestige(Game.cookiesReset+Game.cookiesEarned))+1)-(Game.cookiesEarned+Game.cookiesReset)),o=Game.sayTime((Date.now()-Game.startDate)/1e3*Game.fps,-1);let n="";n+=`You've been on this run for <b>${""===o?"not very long":o}</b>.<br>`,n+='<div class="line"></div>',Game.prestige>0&&(n+=`Your prestige level is currently <b>${pe(Game.prestige)}</b>.<br>(CpS +${pe(Game.prestige)}%)`,n+='<div class="line"></div>'),n+=Oe<1?"Ascending now would grant you no prestige.":Oe<2?"Ascending now would grant you<br><b>1 prestige level</b> (+1% CpS)<br>and <b>1 heavenly chip</b> to spend.":`Ascending now would grant you<br><b>${pe(Oe)} prestige levels</b> (+${pe(Oe)}% CpS)<br>and <b>${pe(Oe)} heavenly chips</b> to spend.`,n+='<div class="line"></div>',n+=`You need <b>${pe(t)} more cookies</b> for the next level.<br>`,n+=e.TooltipAscendButton?`<div class='line'></div>It takes ${Xt} to reach the next level and you were making ${pe(ht,2)} chips on average in the last 5 seconds.<br>`:"",l("ascendTooltip").innerHTML=n}()}}function ti(){Mn(),ln(),pn(),r=Object.keys(Game.mods).length,w=document.createElement("style"),w.type="text/css",w.id="CMCSS",document.head.appendChild(w),function(){const e=document.createElement("div");e.id="CMBotBar",e.style.height="69px",e.style.width="100%",e.style.position="absolute",e.style.display="none",e.style.backgroundColor="#262224",e.style.backgroundImage="linear-gradient(to bottom, #4d4548, #000000)",e.style.borderTop="1px solid black",e.style.overflow="auto",e.style.textShadow="-1px 0 black, 0 1px black, 1px 0 black, 0 -1px black";const t=e.appendChild(document.createElement("table"));t.style.width="100%",t.style.textAlign="center",t.style.whiteSpace="nowrap";const o=t.appendChild(document.createElement("tbody")),n=function(e,t){const o=document.createElement("td");return o.style.textAlign="right",o.className=E+t,o.textContent=e,o},i=o.appendChild(document.createElement("tr"));i.style.fontWeight="bold",i.appendChild(n("CM 2.031.7",L)),o.appendChild(document.createElement("tr")).appendChild(n("Bonus Income",O)),o.appendChild(document.createElement("tr")).appendChild(n("Payback Period",O)),o.appendChild(document.createElement("tr")).appendChild(n("Time Left",O)),l("wrapper").appendChild(e),Object.keys(Game.Objects).forEach((e=>{fo(e)}))}(),function(){const e=document.createElement("div");e.id="CMTimerBar",e.style.position="absolute",e.style.display="none",e.style.height="0px",e.style.fontSize="10px",e.style.fontWeight="bold",e.style.backgroundColor="black";const t=go("CMTimerBarGC","Next Cookie",[{id:"CMTimerBarGCMinBar",color:D},{id:"CMTimerBarGCBar",color:j}]);e.appendChild(t);const o=go("CMTimerBarRen","Next Reindeer",[{id:"CMTimerBarRenMinBar",color:D},{id:"CMTimerBarRenBar",color:W}]);e.appendChild(o);const n=document.createElement("div");n.id="CMTimerBarBuffTimers",e.appendChild(n),l("wrapper").appendChild(e)}(),function(){const e=document.createElement("div");e.id="CMUpgradeBar",e.style.width="100%",e.style.backgroundColor="black",e.style.textAlign="center",e.style.fontWeight="bold",e.style.display="none",e.style.zIndex="21",e.onmouseout=function(){Game.tooltip.hide()};const t=document.createElement("div");t.appendChild(function(){const e=document.createElement("div");e.style.minWidth="330px",e.style.marginBottom="4px";const t=document.createElement("div");t.className="name",t.style.marginBottom="4px",t.textContent="Legend",e.appendChild(t);const o=function(e,t){const o=document.createElement("div");o.style.verticalAlign="middle";const n=document.createElement("span");return n.className=N+e,n.style.display="inline-block",n.style.height="10px",n.style.width="10px",n.style.marginRight="4px",o.appendChild(n),o.appendChild(document.createTextNode(t)),o};return e.appendChild(o(O,"Better than the best PP of a building option")),e.appendChild(o(F,"Same as the best PP building option")),e.appendChild(o(L,"Within the top 10 of PP for buildings")),e.appendChild(o(W,"Within the top 20 of PP for buildings")),e.appendChild(o(A,"Within the top 30 of PP for buildings")),e.appendChild(o(j,"Outside of the top 30 of PP for buildings")),e.appendChild(o(D,"Negative or infinity PP")),e}()),e.onmouseover=function(){Game.tooltip.draw(this,escape(t.innerHTML),"store")};const o=function(e,t){const o=document.createElement("span");return o.id=e,o.className=E+t,o.style.width="14.28571428571429%",o.style.display="inline-block",o.textContent="0",o};e.appendChild(o("CMUpgradeBarBlue",O)),e.appendChild(o("CMUpgradeBarGreen",F)),e.appendChild(o("CMUpgradeBarYellow",L)),e.appendChild(o("CMUpgradeBarOrange",W)),e.appendChild(o("CMUpgradeBarRed",A)),e.appendChild(o("CMUpgradeBarPurple",j)),e.appendChild(o("CMUpgradeBarGray",D)),l("upgrades").parentNode.insertBefore(e,l("upgrades").parentNode.childNodes[3])}(),function(){const e=document.createElement("div");e.id="CMFlashScreen",e.style.width="100%",e.style.height="100%",e.style.backgroundColor="white",e.style.display="none",e.style.zIndex="9999999999",e.style.position="absolute",l("wrapper").appendChild(e)}(),function(){const e=document.createElement("div");e.id="CMSectionHidButtons",e.style.textAlign="center";const t=document.createElement("a");t.className="option",t.onclick=function(){"flex"===l("upgrades").style.display?(l("upgrades").style.display="none",l("toggleUpgrades").style.display="none",l("techUpgrades").style.display="none",l("vaultUpgrades").style.display="none"):(l("upgrades").style.display="flex",0!==l("toggleUpgrades").children.length&&(l("toggleUpgrades").style.display="block"),0!==l("techUpgrades").children.length&&(l("techUpgrades").style.display="block"),0!==l("vaultUpgrades").children.length&&(l("vaultUpgrades").style.display="block"))},t.textContent="Hide/Show Upgrades",e.appendChild(t);const o=document.createElement("a");o.className="option",o.onclick=function(){"grid"===l("products").style.display?l("products").style.display="none":l("products").style.display="grid"},o.textContent="Hide/Show Buildings",e.appendChild(o),l("store").insertBefore(e,l("store").childNodes[2])}(),function(){const e=document.createElement("link");e.id="CMFavicon",e.rel="shortcut icon",e.href="https://orteil.dashnet.org/cookieclicker/favicon.ico",document.getElementsByTagName("head")[0].appendChild(e)}(),Object.keys(R).forEach((e=>{!function(e,t,o){const n=document.createElement("div");n.id=e;const i=document.createElement("div");i.style.minWidth=o,i.style.marginBottom="4px";const a=document.createElement("div");a.style.textAlign="left",a.textContent=t,i.appendChild(a),n.appendChild(i),I[e]=n}(R[e][0],R[e][1],R[e][2])})),function(){const e=document.createElement("a");e.id="PopAllNormalWrinklerButton",e.textContent="Pop All Normal",e.className="option",e.onclick=function(){Wn()},e.onmouseout=function(){Game.tooltip.shouldHide=1},e.onmouseover=function(){Game.tooltip.dynamic=1,Game.tooltip.draw(this,(function(){return An("wb","PopAll")}),"this"),Game.tooltip.wobble()},l("sectionLeftExtra").children[0].append(e);const t=document.createElement("a");t.id="PopFattestWrinklerButton",t.textContent="Pop Single Fattest",t.className="option",t.onclick=function(){null!==xt[1]&&(Game.wrinklers[xt[1]].hp=0)},t.onmouseout=function(){Game.tooltip.shouldHide=1},t.onmouseover=function(){Game.tooltip.dynamic=1,Game.tooltip.draw(this,(function(){return An("wb","PopFattest")}),"this"),Game.tooltip.wobble()},l("sectionLeftExtra").children[0].append(t)}(),l("products").style.display="grid",l("storeBulk").style.gridRow="1/1",l("upgrades").style.display="flex",l("upgrades").style["flex-wrap"]="wrap",Object.keys(Game.Objects).forEach((e=>{const t=Game.Objects[e];null!==l(`product${t.id}`).onmouseover&&(x[e]=l(`product${t.id}`).onmouseover,l(`product${t.id}`).onmouseover=function(){Game.tooltip.dynamic=1,Game.tooltip.draw(this,(function(){return An("b",`${e}`)}),"store"),Game.tooltip.wobble()})})),Game.canLumps()&&(s=l("lumps").onmouseover,l("lumps").onmouseover=function(){Game.tooltip.dynamic=1,Game.tooltip.draw(this,(function(){return An("s","Lump")}),"this"),Game.tooltip.wobble()}),h=Game.LoadMinigames,Game.LoadMinigames=function(){h(),Game.Objects.Farm.minigameLoaded&&(l("gardenTool-1").onmouseover=function(){Game.tooltip.dynamic=1,Game.tooltip.draw(this,(function(){return An("ha","HarvestAllButton")}),"this"),Game.tooltip.wobble()},Array.from(l("gardenPlot").children).forEach((e=>{const t=e.id.slice(-3);e.onmouseover=function(){Game.tooltip.dynamic=1,Game.tooltip.draw(this,(function(){return An("p",[`${t[0]}`,`${t[2]}`])}),"this"),Game.tooltip.wobble()}}))),jn(),function(){if(Game.Objects.Temple.minigameLoaded){for(let e=0;e<11;e+=1)l(`templeGod${e}`).onmouseover=function(){Game.tooltip.dynamic=1,Game.tooltip.draw(this,(function(){return An("pag",e)}),"this"),Game.tooltip.wobble()};for(let e=0;e<3;e+=1)l(`templeSlot${e}`).onmouseover=function(){Game.tooltip.dynamic=1,Game.tooltip.draw(this,(function(){return An("pas",[e,Game.Objects.Temple.minigame.slot[e]])}),"this"),Game.tooltip.wobble()}}}(),Hn()},Game.LoadMinigames(),l("backgroundLeftCanvas").onmouseover=function(){_=1},l("backgroundLeftCanvas").onmouseout=function(){_=0,Game.tooltip.hide(),Object.keys(Game.wrinklers).forEach((e=>{X[e]=0}))},ei(),Hn(),Game.CalculateGains(),Ln(),z=Game.OnAscend,Game.prefs.popups?Game.Popup("Cookie Monster version 2.031.7 loaded!"):Game.Notify("Cookie Monster version 2.031.7 loaded!","","",1,1),Game.Win("Third-party")}const oi={init:function(){Sn=!0;let e=!0;Game.version!==Number("2.031")&&(e=confirm("Cookie Monster version 2.031.7 is meant for Game version 2.031. Loading a different version may cause errors. Do you still want to load Cookie Monster?")),e&&(ti(),Game.registerHook("draw",zo),Game.registerHook("logic",Nn),Game.registerHook("click",Vo),Sn=!1)},load:function(e){const t=JSON.parse(e);Mn(),void 0!==t.favouriteSettings&&(de=t.favouriteSettings),Ln(t.settings),"2.031.7"!==t.version&&(Game.prefs.popups?Game.Popup("A new version of Cookie Monster has been loaded, check out the release notes in the info tab!"):Game.Notify("A new version of Cookie Monster has been loaded, check out the release notes in the info tab!","","",0,1))},save:wn};Game.registerMod("CookieMonster",oi)})()})();
+(()=>{var e={877:function(e){!function(t,o){"use strict";"object"!=typeof e.exports?o(t):e.exports=t.document?o(t):function(e){if(!e.document)throw new Error("jscolor needs a window with document");return o(e)}}("undefined"!=typeof window?window:this,(function(e){"use strict";var t,o,n,i,a=((i={initialized:!1,instances:[],readyQueue:[],register:function(){void 0!==e&&e.document&&e.document.addEventListener("DOMContentLoaded",i.pub.init,!1)},installBySelector:function(t,o){if(!(o=o?i.node(o):e.document))throw new Error("Missing root node");for(var n=o.querySelectorAll(t),a=new RegExp("(^|\\s)("+i.pub.lookupClass+")(\\s*(\\{[^}]*\\})|\\s|$)","i"),r=0;r<n.length;r+=1){var l,s;if(!(n[r].jscolor&&n[r].jscolor instanceof i.pub)&&(void 0===n[r].type||"color"!=n[r].type.toLowerCase()||!i.isColorAttrSupported)&&(null!==(l=i.getDataAttr(n[r],"jscolor"))||n[r].className&&(s=n[r].className.match(a)))){var c=n[r],d="";null!==l?d=l:s&&(console.warn('Installation using class name is DEPRECATED. Use data-jscolor="" attribute instead.'+i.docsRef),s[4]&&(d=s[4]));var p=null;if(d.trim())try{p=i.parseOptionsStr(d)}catch(e){console.warn(e+"\n"+d)}try{new i.pub(c,p)}catch(e){console.warn(e)}}}},parseOptionsStr:function(e){var t=null;try{t=JSON.parse(e)}catch(o){if(!i.pub.looseJSON)throw new Error("Could not parse jscolor options as JSON: "+o);try{t=new Function("var opts = ("+e+'); return typeof opts === "object" ? opts : {};')()}catch(e){throw new Error("Could not evaluate jscolor options: "+e)}}return t},getInstances:function(){for(var e=[],t=0;t<i.instances.length;t+=1)i.instances[t]&&i.instances[t].targetElement&&e.push(i.instances[t]);return e},createEl:function(t){var o=e.document.createElement(t);return i.setData(o,"gui",!0),o},node:function(t){if(!t)return null;if("string"==typeof t){var o=t,n=null;try{n=e.document.querySelector(o)}catch(e){return console.warn(e),null}return n||console.warn("No element matches the selector: %s",o),n}return i.isNode(t)?t:(console.warn("Invalid node of type %s: %s",typeof t,t),null)},isNode:function(e){return"object"==typeof Node?e instanceof Node:e&&"object"==typeof e&&"number"==typeof e.nodeType&&"string"==typeof e.nodeName},nodeName:function(e){return!(!e||!e.nodeName)&&e.nodeName.toLowerCase()},removeChildren:function(e){for(;e.firstChild;)e.removeChild(e.firstChild)},isTextInput:function(e){return e&&"input"===i.nodeName(e)&&"text"===e.type.toLowerCase()},isButton:function(e){if(!e)return!1;var t=i.nodeName(e);return"button"===t||"input"===t&&["button","submit","reset"].indexOf(e.type.toLowerCase())>-1},isButtonEmpty:function(e){switch(i.nodeName(e)){case"input":return!e.value||""===e.value.trim();case"button":return""===e.textContent.trim()}return null},isPassiveEventSupported:function(){var t=!1;try{var o=Object.defineProperty({},"passive",{get:function(){t=!0}});e.addEventListener("testPassive",null,o),e.removeEventListener("testPassive",null,o)}catch(e){}return t}(),isColorAttrSupported:(n=e.document.createElement("input"),!(!n.setAttribute||(n.setAttribute("type","color"),"color"!=n.type.toLowerCase()))),dataProp:"_data_jscolor",setData:function(){var e=arguments[0];if(3===arguments.length){var t=e.hasOwnProperty(i.dataProp)?e[i.dataProp]:e[i.dataProp]={},o=arguments[2];return t[a=arguments[1]]=o,!0}if(2===arguments.length&&"object"==typeof arguments[1]){t=e.hasOwnProperty(i.dataProp)?e[i.dataProp]:e[i.dataProp]={};var n=arguments[1];for(var a in n)n.hasOwnProperty(a)&&(t[a]=n[a]);return!0}throw new Error("Invalid arguments")},removeData:function(){var e=arguments[0];if(!e.hasOwnProperty(i.dataProp))return!0;for(var t=1;t<arguments.length;t+=1){var o=arguments[t];delete e[i.dataProp][o]}return!0},getData:function(e,t,o){if(!e.hasOwnProperty(i.dataProp)){if(void 0===o)return;e[i.dataProp]={}}var n=e[i.dataProp];return n.hasOwnProperty(t)||void 0===o||(n[t]=o),n[t]},getDataAttr:function(e,t){var o="data-"+t;return e.getAttribute(o)},setDataAttr:function(e,t,o){var n="data-"+t;e.setAttribute(n,o)},_attachedGroupEvents:{},attachGroupEvent:function(e,t,o,n){i._attachedGroupEvents.hasOwnProperty(e)||(i._attachedGroupEvents[e]=[]),i._attachedGroupEvents[e].push([t,o,n]),t.addEventListener(o,n,!1)},detachGroupEvents:function(e){if(i._attachedGroupEvents.hasOwnProperty(e)){for(var t=0;t<i._attachedGroupEvents[e].length;t+=1){var o=i._attachedGroupEvents[e][t];o[0].removeEventListener(o[1],o[2],!1)}delete i._attachedGroupEvents[e]}},preventDefault:function(e){e.preventDefault&&e.preventDefault(),e.returnValue=!1},captureTarget:function(e){e.setCapture&&(i._capturedTarget=e,i._capturedTarget.setCapture())},releaseTarget:function(){i._capturedTarget&&(i._capturedTarget.releaseCapture(),i._capturedTarget=null)},triggerEvent:function(t,o,n,a){if(t){var r=null;return"function"==typeof Event?r=new Event(o,{bubbles:n,cancelable:a}):(r=e.document.createEvent("Event")).initEvent(o,n,a),!!r&&(i.setData(r,"internal",!0),t.dispatchEvent(r),!0)}},triggerInputEvent:function(e,t,o,n){e&&i.isTextInput(e)&&i.triggerEvent(e,t,o,n)},eventKey:function(e){var t={9:"Tab",13:"Enter",27:"Escape"};return"string"==typeof e.code?e.code:void 0!==e.keyCode&&t.hasOwnProperty(e.keyCode)?t[e.keyCode]:null},strList:function(e){return e?e.replace(/^\s+|\s+$/g,"").split(/\s+/):[]},hasClass:function(e,t){return!!t&&(void 0!==e.classList?e.classList.contains(t):-1!=(" "+e.className.replace(/\s+/g," ")+" ").indexOf(" "+t+" "))},addClass:function(e,t){var o=i.strList(t);if(void 0===e.classList)for(n=0;n<o.length;n+=1)i.hasClass(e,o[n])||(e.className+=(e.className?" ":"")+o[n]);else for(var n=0;n<o.length;n+=1)e.classList.add(o[n])},removeClass:function(e,t){var o=i.strList(t);if(void 0===e.classList)for(a=0;a<o.length;a+=1){var n=new RegExp("^\\s*"+o[a]+"\\s*|\\s*"+o[a]+"\\s*$|\\s+"+o[a]+"(\\s+)","g");e.className=e.className.replace(n,"$1")}else for(var a=0;a<o.length;a+=1)e.classList.remove(o[a])},getCompStyle:function(t){return(e.getComputedStyle?e.getComputedStyle(t):t.currentStyle)||{}},setStyle:function(e,t,o,n){var a=o?"important":"",r=null;for(var l in t)if(t.hasOwnProperty(l)){var s=null;null===t[l]?(r||(r=i.getData(e,"origStyle")),r&&r.hasOwnProperty(l)&&(s=r[l])):(n&&(r||(r=i.getData(e,"origStyle",{})),r.hasOwnProperty(l)||(r[l]=e.style[l])),s=t[l]),null!==s&&e.style.setProperty(l,s,a)}},hexColor:function(e,t,o){return"#"+(("0"+Math.round(e).toString(16)).substr(-2)+("0"+Math.round(t).toString(16)).substr(-2)+("0"+Math.round(o).toString(16)).substr(-2)).toUpperCase()},hexaColor:function(e,t,o,n){return"#"+(("0"+Math.round(e).toString(16)).substr(-2)+("0"+Math.round(t).toString(16)).substr(-2)+("0"+Math.round(o).toString(16)).substr(-2)+("0"+Math.round(255*n).toString(16)).substr(-2)).toUpperCase()},rgbColor:function(e,t,o){return"rgb("+Math.round(e)+","+Math.round(t)+","+Math.round(o)+")"},rgbaColor:function(e,t,o,n){return"rgba("+Math.round(e)+","+Math.round(t)+","+Math.round(o)+","+Math.round(100*(null==n?1:n))/100+")"},linearGradient:(o=function(){for(var t="linear-gradient",o=["","-webkit-","-moz-","-o-","-ms-"],n=e.document.createElement("div"),i=0;i<o.length;i+=1){var a=o[i]+t,r=a+"(to right, rgba(0,0,0,0), rgba(0,0,0,0))";if(n.style.background=r,n.style.background)return a}return t}(),function(){return o+"("+Array.prototype.join.call(arguments,", ")+")"}),setBorderRadius:function(e,t){i.setStyle(e,{"border-radius":t||"0"})},setBoxShadow:function(e,t){i.setStyle(e,{"box-shadow":t||"none"})},getElementPos:function(e,t){var o=0,n=0,a=e.getBoundingClientRect();if(o=a.left,n=a.top,!t){var r=i.getViewPos();o+=r[0],n+=r[1]}return[o,n]},getElementSize:function(e){return[e.offsetWidth,e.offsetHeight]},getAbsPointerPos:function(e){var t=0,o=0;return void 0!==e.changedTouches&&e.changedTouches.length?(t=e.changedTouches[0].clientX,o=e.changedTouches[0].clientY):"number"==typeof e.clientX&&(t=e.clientX,o=e.clientY),{x:t,y:o}},getRelPointerPos:function(e){var t=(e.target||e.srcElement).getBoundingClientRect(),o=0,n=0;return void 0!==e.changedTouches&&e.changedTouches.length?(o=e.changedTouches[0].clientX,n=e.changedTouches[0].clientY):"number"==typeof e.clientX&&(o=e.clientX,n=e.clientY),{x:o-t.left,y:n-t.top}},getViewPos:function(){var t=e.document.documentElement;return[(e.pageXOffset||t.scrollLeft)-(t.clientLeft||0),(e.pageYOffset||t.scrollTop)-(t.clientTop||0)]},getViewSize:function(){var t=e.document.documentElement;return[e.innerWidth||t.clientWidth,e.innerHeight||t.clientHeight]},RGB_HSV:function(e,t,o){e/=255,t/=255,o/=255;var n=Math.min(Math.min(e,t),o),i=Math.max(Math.max(e,t),o),a=i-n;if(0===a)return[null,0,100*i];var r=e===n?3+(o-t)/a:t===n?5+(e-o)/a:1+(t-e)/a;return[60*(6===r?0:r),a/i*100,100*i]},HSV_RGB:function(e,t,o){var n=o/100*255;if(null===e)return[n,n,n];e/=60,t/=100;var i=Math.floor(e),a=n*(1-t),r=n*(1-t*(i%2?e-i:1-(e-i)));switch(i){case 6:case 0:return[n,r,a];case 1:return[r,n,a];case 2:return[a,n,r];case 3:return[a,r,n];case 4:return[r,a,n];case 5:return[n,a,r]}},parseColorString:function(e){var t,o={rgba:null,format:null};if(t=e.match(/^\W*([0-9A-F]{3,8})\W*$/i)){if(8===t[1].length)o.format="hexa",o.rgba=[parseInt(t[1].substr(0,2),16),parseInt(t[1].substr(2,2),16),parseInt(t[1].substr(4,2),16),parseInt(t[1].substr(6,2),16)/255];else if(6===t[1].length)o.format="hex",o.rgba=[parseInt(t[1].substr(0,2),16),parseInt(t[1].substr(2,2),16),parseInt(t[1].substr(4,2),16),null];else{if(3!==t[1].length)return!1;o.format="hex",o.rgba=[parseInt(t[1].charAt(0)+t[1].charAt(0),16),parseInt(t[1].charAt(1)+t[1].charAt(1),16),parseInt(t[1].charAt(2)+t[1].charAt(2),16),null]}return o}if(t=e.match(/^\W*rgba?\(([^)]*)\)\W*$/i)){var n,i,a,r,l=t[1].split(","),s=/^\s*(\d+|\d*\.\d+|\d+\.\d*)\s*$/;if(l.length>=3&&(n=l[0].match(s))&&(i=l[1].match(s))&&(a=l[2].match(s)))return o.format="rgb",o.rgba=[parseFloat(n[1])||0,parseFloat(i[1])||0,parseFloat(a[1])||0,null],l.length>=4&&(r=l[3].match(s))&&(o.format="rgba",o.rgba[3]=parseFloat(r[1])||0),o}return!1},parsePaletteValue:function(e){var t=[];"string"==typeof e?e.replace(/#[0-9A-F]{3}([0-9A-F]{3})?|rgba?\(([^)]*)\)/gi,(function(e){t.push(e)})):Array.isArray(e)&&(t=e);for(var o=[],n=0;n<t.length;n++){var a=i.parseColorString(t[n]);a&&o.push(a)}return o},containsTranparentColor:function(e){for(var t=0;t<e.length;t++){var o=e[t].rgba[3];if(null!==o&&o<1)return!0}return!1},isAlphaFormat:function(e){switch(e.toLowerCase()){case"hexa":case"rgba":return!0}return!1},scaleCanvasForHighDPR:function(t){var o=e.devicePixelRatio||1;t.width*=o,t.height*=o,t.getContext("2d").scale(o,o)},genColorPreviewCanvas:function(e,t,o,n){var a=Math.round(i.pub.previewSeparator.length),r=i.pub.chessboardSize,l=i.pub.chessboardColor1,s=i.pub.chessboardColor2,c=o||2*r,d=2*r,p=i.createEl("canvas"),u=p.getContext("2d");p.width=c,p.height=d,n&&i.scaleCanvasForHighDPR(p),u.fillStyle=l,u.fillRect(0,0,c,d),u.fillStyle=s;for(var m=0;m<c;m+=2*r)u.fillRect(m,0,r,r),u.fillRect(m+r,r,r,r);e&&(u.fillStyle=e,u.fillRect(0,0,c,d));var h=null;switch(t){case"left":h=0,u.clearRect(0,0,a/2,d);break;case"right":h=c-a,u.clearRect(c-a/2,0,a/2,d)}if(null!==h){u.lineWidth=1;for(var g=0;g<i.pub.previewSeparator.length;g+=1)u.beginPath(),u.strokeStyle=i.pub.previewSeparator[g],u.moveTo(.5+h+g,0),u.lineTo(.5+h+g,d),u.stroke()}return{canvas:p,width:c,height:d}},genColorPreviewGradient:function(e,t,o){var n;return n=t&&o?["to "+{left:"right",right:"left"}[t],e+" 0%",e+" "+o+"px","rgba(0,0,0,0) "+(o+1)+"px","rgba(0,0,0,0) 100%"]:["to right",e+" 0%",e+" 100%"],i.linearGradient.apply(this,n)},redrawPosition:function(){if(i.picker&&i.picker.owner){var e,t,o=i.picker.owner;o.fixed?(e=i.getElementPos(o.targetElement,!0),t=[0,0]):(e=i.getElementPos(o.targetElement),t=i.getViewPos());var n,a,r,l=i.getElementSize(o.targetElement),s=i.getViewSize(),c=i.getPickerDims(o),d=[c.borderW,c.borderH];switch(o.position.toLowerCase()){case"left":n=1,a=0,r=-1;break;case"right":n=1,a=0,r=1;break;case"top":n=0,a=1,r=-1;break;default:n=0,a=1,r=1}var p=(l[a]+d[a])/2;if(o.smartPosition)u=[-t[n]+e[n]+d[n]>s[n]&&-t[n]+e[n]+l[n]/2>s[n]/2&&e[n]+l[n]-d[n]>=0?e[n]+l[n]-d[n]:e[n],-t[a]+e[a]+l[a]+d[a]-p+p*r>s[a]?-t[a]+e[a]+l[a]/2>s[a]/2&&e[a]+l[a]-p-p*r>=0?e[a]+l[a]-p-p*r:e[a]+l[a]-p+p*r:e[a]+l[a]-p+p*r>=0?e[a]+l[a]-p+p*r:e[a]+l[a]-p-p*r];else var u=[e[n],e[a]+l[a]-p+p*r];var m=u[n],h=u[a],g=o.fixed?"fixed":"absolute",f=(u[0]+d[0]>e[0]||u[0]<e[0]+l[0])&&u[1]+d[1]<e[1]+l[1];i._drawPosition(o,m,h,g,f)}},_drawPosition:function(e,t,o,n,a){var r=a?0:e.shadowBlur;i.picker.wrap.style.position=n,i.picker.wrap.style.left=t+"px",i.picker.wrap.style.top=o+"px",i.setBoxShadow(i.picker.boxS,e.shadow?new i.BoxShadow(0,r,e.shadowBlur,0,e.shadowColor):null)},getPickerDims:function(e){var t=2*e.controlBorderWidth+e.width,o=2*e.controlBorderWidth+e.height,n=2*e.controlBorderWidth+2*i.getControlPadding(e)+e.sliderSize;i.getSliderChannel(e)&&(t+=n),e.hasAlphaChannel()&&(t+=n);var a=i.getPaletteDims(e,t);a.height&&(o+=a.height+e.padding),e.closeButton&&(o+=2*e.controlBorderWidth+e.padding+e.buttonHeight);var r=t+2*e.padding,l=o+2*e.padding;return{contentW:t,contentH:o,paddedW:r,paddedH:l,borderW:r+2*e.borderWidth,borderH:l+2*e.borderWidth,palette:a}},getPaletteDims:function(e,t){var o=0,n=0,i=0,a=0,r=0,l=e._palette?e._palette.length:0;return l&&(n=(o=e.paletteCols)>0?Math.ceil(l/o):0,i=Math.max(1,Math.floor((t-(o-1)*e.paletteSpacing)/o)),a=e.paletteHeight?Math.min(e.paletteHeight,i):i),n&&(r=n*a+(n-1)*e.paletteSpacing),{cols:o,rows:n,cellW:i,cellH:a,width:t,height:r}},getControlPadding:function(e){return Math.max(e.padding/2,2*e.pointerBorderWidth+e.pointerThickness-e.controlBorderWidth)},getPadYChannel:function(e){switch(e.mode.charAt(1).toLowerCase()){case"v":return"v"}return"s"},getSliderChannel:function(e){if(e.mode.length>2)switch(e.mode.charAt(2).toLowerCase()){case"s":return"s";case"v":return"v"}return null},triggerCallback:function(e,t){if(e[t]){var o=null;if("string"==typeof e[t])try{o=new Function(e[t])}catch(e){console.error(e)}else o=e[t];o&&o.call(e)}},triggerGlobal:function(e){for(var t=i.getInstances(),o=0;o<t.length;o+=1)t[o].trigger(e)},_pointerMoveEvent:{mouse:"mousemove",touch:"touchmove"},_pointerEndEvent:{mouse:"mouseup",touch:"touchend"},_pointerOrigin:null,_capturedTarget:null,onDocumentKeyUp:function(e){-1!==["Tab","Escape"].indexOf(i.eventKey(e))&&i.picker&&i.picker.owner&&i.picker.owner.tryHide()},onWindowResize:function(e){i.redrawPosition()},onWindowScroll:function(e){i.redrawPosition()},onParentScroll:function(e){i.picker&&i.picker.owner&&i.picker.owner.tryHide()},onDocumentMouseDown:function(e){var t=e.target||e.srcElement;t.jscolor&&t.jscolor instanceof i.pub?t.jscolor.showOnClick&&!t.disabled&&t.jscolor.show():i.getData(t,"gui")?i.getData(t,"control")&&i.onControlPointerStart(e,t,i.getData(t,"control"),"mouse"):i.picker&&i.picker.owner&&i.picker.owner.tryHide()},onPickerTouchStart:function(e){var t=e.target||e.srcElement;i.getData(t,"control")&&i.onControlPointerStart(e,t,i.getData(t,"control"),"touch")},onControlPointerStart:function(t,o,n,a){var r=i.getData(o,"instance");i.preventDefault(t),i.captureTarget(o);var l=function(e,r){i.attachGroupEvent("drag",e,i._pointerMoveEvent[a],i.onDocumentPointerMove(t,o,n,a,r)),i.attachGroupEvent("drag",e,i._pointerEndEvent[a],i.onDocumentPointerEnd(t,o,n,a))};if(l(e.document,[0,0]),e.parent&&e.frameElement){var s=e.frameElement.getBoundingClientRect(),c=[-s.left,-s.top];l(e.parent.window.document,c)}var d=i.getAbsPointerPos(t),p=i.getRelPointerPos(t);switch(i._pointerOrigin={x:d.x-p.x,y:d.y-p.y},n){case"pad":"v"===i.getSliderChannel(r)&&0===r.channels.v&&r.fromHSVA(null,null,100,null),i.setPad(r,t,0,0);break;case"sld":i.setSld(r,t,0);break;case"asld":i.setASld(r,t,0)}r.trigger("input")},onDocumentPointerMove:function(e,t,o,n,a){return function(e){var n=i.getData(t,"instance");switch(o){case"pad":i.setPad(n,e,a[0],a[1]);break;case"sld":i.setSld(n,e,a[1]);break;case"asld":i.setASld(n,e,a[1])}n.trigger("input")}},onDocumentPointerEnd:function(e,t,o,n){return function(e){var o=i.getData(t,"instance");i.detachGroupEvents("drag"),i.releaseTarget(),o.trigger("input"),o.trigger("change")}},onPaletteSampleClick:function(e){var t=e.currentTarget,o=i.getData(t,"instance"),n=i.getData(t,"color");"any"===o.format.toLowerCase()&&(o._setFormat(n.format),i.isAlphaFormat(o.getFormat())||(n.rgba[3]=1)),null===n.rgba[3]&&(!0===o.paletteSetsAlpha||"auto"===o.paletteSetsAlpha&&o._paletteHasTransparency)&&(n.rgba[3]=1),o.fromRGBA.apply(o,n.rgba),o.trigger("input"),o.trigger("change"),o.hideOnPaletteClick&&o.hide()},setPad:function(e,t,o,n){var a=i.getAbsPointerPos(t),r=o+a.x-i._pointerOrigin.x-e.padding-e.controlBorderWidth,l=n+a.y-i._pointerOrigin.y-e.padding-e.controlBorderWidth,s=r*(360/(e.width-1)),c=100-l*(100/(e.height-1));switch(i.getPadYChannel(e)){case"s":e.fromHSVA(s,c,null,null);break;case"v":e.fromHSVA(s,null,c,null)}},setSld:function(e,t,o){var n=100-(o+i.getAbsPointerPos(t).y-i._pointerOrigin.y-e.padding-e.controlBorderWidth)*(100/(e.height-1));switch(i.getSliderChannel(e)){case"s":e.fromHSVA(null,n,null,null);break;case"v":e.fromHSVA(null,null,n,null)}},setASld:function(e,t,o){var n=1-(o+i.getAbsPointerPos(t).y-i._pointerOrigin.y-e.padding-e.controlBorderWidth)*(1/(e.height-1));if(n<1){var a=e.getFormat();"any"!==e.format.toLowerCase()||i.isAlphaFormat(a)||e._setFormat("hex"===a?"hexa":"rgba")}e.fromHSVA(null,null,null,n)},createPadCanvas:function(){var e={elm:null,draw:null},t=i.createEl("canvas"),o=t.getContext("2d");return e.elm=t,e.draw=function(e,n,i){t.width=e,t.height=n,o.clearRect(0,0,t.width,t.height);var a=o.createLinearGradient(0,0,t.width,0);a.addColorStop(0,"#F00"),a.addColorStop(1/6,"#FF0"),a.addColorStop(2/6,"#0F0"),a.addColorStop(.5,"#0FF"),a.addColorStop(4/6,"#00F"),a.addColorStop(5/6,"#F0F"),a.addColorStop(1,"#F00"),o.fillStyle=a,o.fillRect(0,0,t.width,t.height);var r=o.createLinearGradient(0,0,0,t.height);switch(i.toLowerCase()){case"s":r.addColorStop(0,"rgba(255,255,255,0)"),r.addColorStop(1,"rgba(255,255,255,1)");break;case"v":r.addColorStop(0,"rgba(0,0,0,0)"),r.addColorStop(1,"rgba(0,0,0,1)")}o.fillStyle=r,o.fillRect(0,0,t.width,t.height)},e},createSliderGradient:function(){var e={elm:null,draw:null},t=i.createEl("canvas"),o=t.getContext("2d");return e.elm=t,e.draw=function(e,n,i,a){t.width=e,t.height=n,o.clearRect(0,0,t.width,t.height);var r=o.createLinearGradient(0,0,0,t.height);r.addColorStop(0,i),r.addColorStop(1,a),o.fillStyle=r,o.fillRect(0,0,t.width,t.height)},e},createASliderGradient:function(){var e={elm:null,draw:null},t=i.createEl("canvas"),o=t.getContext("2d");return e.elm=t,e.draw=function(e,n,a){t.width=e,t.height=n,o.clearRect(0,0,t.width,t.height);var r=t.width/2,l=i.pub.chessboardColor1,s=i.pub.chessboardColor2;if(o.fillStyle=l,o.fillRect(0,0,t.width,t.height),r>0)for(var c=0;c<t.height;c+=2*r)o.fillStyle=s,o.fillRect(0,c,r,r),o.fillRect(r,c+r,r,r);var d=o.createLinearGradient(0,0,0,t.height);d.addColorStop(0,a),d.addColorStop(1,"rgba(0,0,0,0)"),o.fillStyle=d,o.fillRect(0,0,t.width,t.height)},e},BoxShadow:(t=function(e,t,o,n,i,a){this.hShadow=e,this.vShadow=t,this.blur=o,this.spread=n,this.color=i,this.inset=!!a},t.prototype.toString=function(){var e=[Math.round(this.hShadow)+"px",Math.round(this.vShadow)+"px",Math.round(this.blur)+"px",Math.round(this.spread)+"px",this.color];return this.inset&&e.push("inset"),e.join(" ")},t),flags:{leaveValue:1,leaveAlpha:2,leavePreview:4},enumOpts:{format:["auto","any","hex","hexa","rgb","rgba"],previewPosition:["left","right"],mode:["hsv","hvs","hs","hv"],position:["left","right","top","bottom"],alphaChannel:["auto",!0,!1],paletteSetsAlpha:["auto",!0,!1]},deprecatedOpts:{styleElement:"previewElement",onFineChange:"onInput",overwriteImportant:"forceStyle",closable:"closeButton",insetWidth:"controlBorderWidth",insetColor:"controlBorderColor",refine:null},docsRef:" See https://jscolor.com/docs/",pub:function(t,o){var n=this;function a(e,t){if("string"!=typeof e)throw new Error("Invalid value for option name: "+e);if(i.enumOpts.hasOwnProperty(e)&&("string"==typeof t&&(t=t.toLowerCase()),-1===i.enumOpts[e].indexOf(t)))throw new Error("Option '"+e+"' has invalid value: "+t);if(i.deprecatedOpts.hasOwnProperty(e)){var o=e,a=i.deprecatedOpts[e];if(!a)throw new Error("Option '"+e+"' is DEPRECATED");console.warn("Option '%s' is DEPRECATED, using '%s' instead."+i.docsRef,o,a),e=a}var r="set__"+e;if("function"==typeof n[r])return n[r](t),!0;if(e in n)return n[e]=t,!0;throw new Error("Unrecognized configuration option: "+e)}function r(e){if("string"!=typeof e)throw new Error("Invalid value for option name: "+e);if(i.deprecatedOpts.hasOwnProperty(e)){var t=e,o=i.deprecatedOpts[e];if(!o)throw new Error("Option '"+e+"' is DEPRECATED");console.warn("Option '%s' is DEPRECATED, using '%s' instead."+i.docsRef,t,o),e=o}var a="get__"+e;if("function"==typeof n[a])return n[a](value);if(e in n)return n[e];throw new Error("Unrecognized configuration option: "+e)}function l(){n._processParentElementsInDOM(),i.picker||(i.picker={owner:null,wrap:i.createEl("div"),box:i.createEl("div"),boxS:i.createEl("div"),boxB:i.createEl("div"),pad:i.createEl("div"),padB:i.createEl("div"),padM:i.createEl("div"),padCanvas:i.createPadCanvas(),cross:i.createEl("div"),crossBY:i.createEl("div"),crossBX:i.createEl("div"),crossLY:i.createEl("div"),crossLX:i.createEl("div"),sld:i.createEl("div"),sldB:i.createEl("div"),sldM:i.createEl("div"),sldGrad:i.createSliderGradient(),sldPtrS:i.createEl("div"),sldPtrIB:i.createEl("div"),sldPtrMB:i.createEl("div"),sldPtrOB:i.createEl("div"),asld:i.createEl("div"),asldB:i.createEl("div"),asldM:i.createEl("div"),asldGrad:i.createASliderGradient(),asldPtrS:i.createEl("div"),asldPtrIB:i.createEl("div"),asldPtrMB:i.createEl("div"),asldPtrOB:i.createEl("div"),pal:i.createEl("div"),btn:i.createEl("div"),btnT:i.createEl("span")},i.picker.pad.appendChild(i.picker.padCanvas.elm),i.picker.padB.appendChild(i.picker.pad),i.picker.cross.appendChild(i.picker.crossBY),i.picker.cross.appendChild(i.picker.crossBX),i.picker.cross.appendChild(i.picker.crossLY),i.picker.cross.appendChild(i.picker.crossLX),i.picker.padB.appendChild(i.picker.cross),i.picker.box.appendChild(i.picker.padB),i.picker.box.appendChild(i.picker.padM),i.picker.sld.appendChild(i.picker.sldGrad.elm),i.picker.sldB.appendChild(i.picker.sld),i.picker.sldB.appendChild(i.picker.sldPtrOB),i.picker.sldPtrOB.appendChild(i.picker.sldPtrMB),i.picker.sldPtrMB.appendChild(i.picker.sldPtrIB),i.picker.sldPtrIB.appendChild(i.picker.sldPtrS),i.picker.box.appendChild(i.picker.sldB),i.picker.box.appendChild(i.picker.sldM),i.picker.asld.appendChild(i.picker.asldGrad.elm),i.picker.asldB.appendChild(i.picker.asld),i.picker.asldB.appendChild(i.picker.asldPtrOB),i.picker.asldPtrOB.appendChild(i.picker.asldPtrMB),i.picker.asldPtrMB.appendChild(i.picker.asldPtrIB),i.picker.asldPtrIB.appendChild(i.picker.asldPtrS),i.picker.box.appendChild(i.picker.asldB),i.picker.box.appendChild(i.picker.asldM),i.picker.box.appendChild(i.picker.pal),i.picker.btn.appendChild(i.picker.btnT),i.picker.box.appendChild(i.picker.btn),i.picker.boxB.appendChild(i.picker.box),i.picker.wrap.appendChild(i.picker.boxS),i.picker.wrap.appendChild(i.picker.boxB),i.picker.wrap.addEventListener("touchstart",i.onPickerTouchStart,!!i.isPassiveEventSupported&&{passive:!1}));var t,o,a=i.picker,r=!!i.getSliderChannel(n),l=n.hasAlphaChannel(),p=i.getPickerDims(n),u=2*n.pointerBorderWidth+n.pointerThickness+2*n.crossSize,m=i.getControlPadding(n),h=Math.min(n.borderRadius,Math.round(n.padding*Math.PI));a.wrap.className="jscolor-picker-wrap",a.wrap.style.clear="both",a.wrap.style.width=p.borderW+"px",a.wrap.style.height=p.borderH+"px",a.wrap.style.zIndex=n.zIndex,a.box.className="jscolor-picker",a.box.style.width=p.paddedW+"px",a.box.style.height=p.paddedH+"px",a.box.style.position="relative",a.boxS.className="jscolor-picker-shadow",a.boxS.style.position="absolute",a.boxS.style.left="0",a.boxS.style.top="0",a.boxS.style.width="100%",a.boxS.style.height="100%",i.setBorderRadius(a.boxS,h+"px"),a.boxB.className="jscolor-picker-border",a.boxB.style.position="relative",a.boxB.style.border=n.borderWidth+"px solid",a.boxB.style.borderColor=n.borderColor,a.boxB.style.background=n.backgroundColor,i.setBorderRadius(a.boxB,h+"px"),a.padM.style.background="rgba(255,0,0,.2)",a.sldM.style.background="rgba(0,255,0,.2)",a.asldM.style.background="rgba(0,0,255,.2)",a.padM.style.opacity=a.sldM.style.opacity=a.asldM.style.opacity="0",a.pad.style.position="relative",a.pad.style.width=n.width+"px",a.pad.style.height=n.height+"px",a.padCanvas.draw(n.width,n.height,i.getPadYChannel(n)),a.padB.style.position="absolute",a.padB.style.left=n.padding+"px",a.padB.style.top=n.padding+"px",a.padB.style.border=n.controlBorderWidth+"px solid",a.padB.style.borderColor=n.controlBorderColor,a.padM.style.position="absolute",a.padM.style.left="0px",a.padM.style.top="0px",a.padM.style.width=n.padding+2*n.controlBorderWidth+n.width+m+"px",a.padM.style.height=2*n.controlBorderWidth+2*n.padding+n.height+"px",a.padM.style.cursor="crosshair",i.setData(a.padM,{instance:n,control:"pad"}),a.cross.style.position="absolute",a.cross.style.left=a.cross.style.top="0",a.cross.style.width=a.cross.style.height=u+"px",a.crossBY.style.position=a.crossBX.style.position="absolute",a.crossBY.style.background=a.crossBX.style.background=n.pointerBorderColor,a.crossBY.style.width=a.crossBX.style.height=2*n.pointerBorderWidth+n.pointerThickness+"px",a.crossBY.style.height=a.crossBX.style.width=u+"px",a.crossBY.style.left=a.crossBX.style.top=Math.floor(u/2)-Math.floor(n.pointerThickness/2)-n.pointerBorderWidth+"px",a.crossBY.style.top=a.crossBX.style.left="0",a.crossLY.style.position=a.crossLX.style.position="absolute",a.crossLY.style.background=a.crossLX.style.background=n.pointerColor,a.crossLY.style.height=a.crossLX.style.width=u-2*n.pointerBorderWidth+"px",a.crossLY.style.width=a.crossLX.style.height=n.pointerThickness+"px",a.crossLY.style.left=a.crossLX.style.top=Math.floor(u/2)-Math.floor(n.pointerThickness/2)+"px",a.crossLY.style.top=a.crossLX.style.left=n.pointerBorderWidth+"px",a.sld.style.overflow="hidden",a.sld.style.width=n.sliderSize+"px",a.sld.style.height=n.height+"px",a.sldGrad.draw(n.sliderSize,n.height,"#000","#000"),a.sldB.style.display=r?"block":"none",a.sldB.style.position="absolute",a.sldB.style.left=n.padding+n.width+2*n.controlBorderWidth+2*m+"px",a.sldB.style.top=n.padding+"px",a.sldB.style.border=n.controlBorderWidth+"px solid",a.sldB.style.borderColor=n.controlBorderColor,a.sldM.style.display=r?"block":"none",a.sldM.style.position="absolute",a.sldM.style.left=n.padding+n.width+2*n.controlBorderWidth+m+"px",a.sldM.style.top="0px",a.sldM.style.width=n.sliderSize+2*m+2*n.controlBorderWidth+(l?0:Math.max(0,n.padding-m))+"px",a.sldM.style.height=2*n.controlBorderWidth+2*n.padding+n.height+"px",a.sldM.style.cursor="default",i.setData(a.sldM,{instance:n,control:"sld"}),a.sldPtrIB.style.border=a.sldPtrOB.style.border=n.pointerBorderWidth+"px solid "+n.pointerBorderColor,a.sldPtrOB.style.position="absolute",a.sldPtrOB.style.left=-(2*n.pointerBorderWidth+n.pointerThickness)+"px",a.sldPtrOB.style.top="0",a.sldPtrMB.style.border=n.pointerThickness+"px solid "+n.pointerColor,a.sldPtrS.style.width=n.sliderSize+"px",a.sldPtrS.style.height=i.pub.sliderInnerSpace+"px",a.asld.style.overflow="hidden",a.asld.style.width=n.sliderSize+"px",a.asld.style.height=n.height+"px",a.asldGrad.draw(n.sliderSize,n.height,"#000"),a.asldB.style.display=l?"block":"none",a.asldB.style.position="absolute",a.asldB.style.left=n.padding+n.width+2*n.controlBorderWidth+m+(r?n.sliderSize+3*m+2*n.controlBorderWidth:0)+"px",a.asldB.style.top=n.padding+"px",a.asldB.style.border=n.controlBorderWidth+"px solid",a.asldB.style.borderColor=n.controlBorderColor,a.asldM.style.display=l?"block":"none",a.asldM.style.position="absolute",a.asldM.style.left=n.padding+n.width+2*n.controlBorderWidth+m+(r?n.sliderSize+2*m+2*n.controlBorderWidth:0)+"px",a.asldM.style.top="0px",a.asldM.style.width=n.sliderSize+2*m+2*n.controlBorderWidth+Math.max(0,n.padding-m)+"px",a.asldM.style.height=2*n.controlBorderWidth+2*n.padding+n.height+"px",a.asldM.style.cursor="default",i.setData(a.asldM,{instance:n,control:"asld"}),a.asldPtrIB.style.border=a.asldPtrOB.style.border=n.pointerBorderWidth+"px solid "+n.pointerBorderColor,a.asldPtrOB.style.position="absolute",a.asldPtrOB.style.left=-(2*n.pointerBorderWidth+n.pointerThickness)+"px",a.asldPtrOB.style.top="0",a.asldPtrMB.style.border=n.pointerThickness+"px solid "+n.pointerColor,a.asldPtrS.style.width=n.sliderSize+"px",a.asldPtrS.style.height=i.pub.sliderInnerSpace+"px",a.pal.className="jscolor-palette",a.pal.style.display=p.palette.rows?"block":"none",a.pal.style.position="absolute",a.pal.style.left=n.padding+"px",a.pal.style.top=2*n.controlBorderWidth+2*n.padding+n.height+"px",a.pal.innerHTML="";for(var g=i.genColorPreviewCanvas("rgba(0,0,0,0)"),f=0,C=0;C<p.palette.rows;C++)for(var y=0;y<p.palette.cols&&f<n._palette.length;y++,f++){var b=n._palette[f],k=i.rgbaColor.apply(null,b.rgba),G=i.createEl("div");G.style.width=p.palette.cellW-2*n.controlBorderWidth+"px",G.style.height=p.palette.cellH-2*n.controlBorderWidth+"px",G.style.backgroundColor=k;var v=i.createEl("div");v.className="jscolor-palette-sample",v.style.display="block",v.style.position="absolute",v.style.left=(p.palette.cols<=1?0:Math.round(y*((p.contentW-p.palette.cellW)/(p.palette.cols-1))*10)/10)+"px",v.style.top=C*(p.palette.cellH+n.paletteSpacing)+"px",v.style.border=n.controlBorderWidth+"px solid",v.style.borderColor=n.controlBorderColor,v.style.cursor="pointer",null!==b.rgba[3]&&b.rgba[3]<1&&(v.style.backgroundImage="url('"+g.canvas.toDataURL()+"')",v.style.backgroundRepeat="repeat",v.style.backgroundPosition="center center"),i.setData(v,{instance:n,control:"palette-sample",color:b}),v.addEventListener("click",i.onPaletteSampleClick,!1),v.appendChild(G),a.pal.appendChild(v)}a.btn.className="jscolor-btn-close",a.btn.style.display=n.closeButton?"block":"none",a.btn.style.position="absolute",a.btn.style.left=n.padding+"px",a.btn.style.bottom=n.padding+"px",a.btn.style.padding="0 15px",a.btn.style.maxWidth=p.contentW-2*n.controlBorderWidth-30+"px",a.btn.style.overflow="hidden",a.btn.style.height=n.buttonHeight+"px",a.btn.style.whiteSpace="nowrap",a.btn.style.border=n.controlBorderWidth+"px solid",o=(t=n.controlBorderColor.split(/\s+/)).length<2?t[0]:t[1]+" "+t[0]+" "+t[0]+" "+t[1],a.btn.style.borderColor=o,a.btn.style.color=n.buttonColor,a.btn.style.font="12px sans-serif",a.btn.style.textAlign="center",a.btn.style.cursor="pointer",a.btn.onmousedown=function(){n.hide()},a.btnT.style.lineHeight=n.buttonHeight+"px",a.btnT.innerHTML="",a.btnT.appendChild(e.document.createTextNode(n.closeText)),s(),c(),d(),i.picker.owner&&i.picker.owner!==n&&i.removeClass(i.picker.owner.targetElement,i.pub.activeClassName),i.picker.owner=n,n.container===e.document.body?i.redrawPosition():i._drawPosition(n,0,0,"relative",!1),a.wrap.parentNode!==n.container&&n.container.appendChild(a.wrap),i.addClass(n.targetElement,i.pub.activeClassName)}function s(){var e=i.getPadYChannel(n),t=Math.round(n.channels.h/360*(n.width-1)),o=Math.round((1-n.channels[e]/100)*(n.height-1)),a=2*n.pointerBorderWidth+n.pointerThickness+2*n.crossSize,r=-Math.floor(a/2);switch(i.picker.cross.style.left=t+r+"px",i.picker.cross.style.top=o+r+"px",i.getSliderChannel(n)){case"s":var l=i.HSV_RGB(n.channels.h,100,n.channels.v),s=i.HSV_RGB(n.channels.h,0,n.channels.v),c="rgb("+Math.round(l[0])+","+Math.round(l[1])+","+Math.round(l[2])+")",d="rgb("+Math.round(s[0])+","+Math.round(s[1])+","+Math.round(s[2])+")";i.picker.sldGrad.draw(n.sliderSize,n.height,c,d);break;case"v":var p=i.HSV_RGB(n.channels.h,n.channels.s,100);c="rgb("+Math.round(p[0])+","+Math.round(p[1])+","+Math.round(p[2])+")",d="#000",i.picker.sldGrad.draw(n.sliderSize,n.height,c,d)}i.picker.asldGrad.draw(n.sliderSize,n.height,n.toHEXString())}function c(){var e=i.getSliderChannel(n);if(e){var t=Math.round((1-n.channels[e]/100)*(n.height-1));i.picker.sldPtrOB.style.top=t-(2*n.pointerBorderWidth+n.pointerThickness)-Math.floor(i.pub.sliderInnerSpace/2)+"px"}i.picker.asldGrad.draw(n.sliderSize,n.height,n.toHEXString())}function d(){var e=Math.round((1-n.channels.a)*(n.height-1));i.picker.asldPtrOB.style.top=e-(2*n.pointerBorderWidth+n.pointerThickness)-Math.floor(i.pub.sliderInnerSpace/2)+"px"}function p(){return i.picker&&i.picker.owner===n}if(o||(o={}),this.channels={r:255,g:255,b:255,h:0,s:0,v:100,a:1},this.format="auto",this.value=void 0,this.alpha=void 0,this.onChange=void 0,this.onInput=void 0,this.valueElement=void 0,this.alphaElement=void 0,this.previewElement=void 0,this.previewPosition="left",this.previewSize=32,this.previewPadding=8,this.required=!0,this.hash=!0,this.uppercase=!0,this.forceStyle=!0,this.width=181,this.height=101,this.mode="HSV",this.alphaChannel="auto",this.position="bottom",this.smartPosition=!0,this.showOnClick=!0,this.hideOnLeave=!0,this.palette=[],this.paletteCols=10,this.paletteSetsAlpha="auto",this.paletteHeight=16,this.paletteSpacing=4,this.hideOnPaletteClick=!1,this.sliderSize=16,this.crossSize=8,this.closeButton=!1,this.closeText="Close",this.buttonColor="rgba(0,0,0,1)",this.buttonHeight=18,this.padding=12,this.backgroundColor="rgba(255,255,255,1)",this.borderWidth=1,this.borderColor="rgba(187,187,187,1)",this.borderRadius=8,this.controlBorderWidth=1,this.controlBorderColor="rgba(187,187,187,1)",this.shadow=!0,this.shadowBlur=15,this.shadowColor="rgba(0,0,0,0.2)",this.pointerColor="rgba(76,76,76,1)",this.pointerBorderWidth=1,this.pointerBorderColor="rgba(255,255,255,1)",this.pointerThickness=2,this.zIndex=5e3,this.container=void 0,this.minS=0,this.maxS=100,this.minV=0,this.maxV=100,this.minA=0,this.maxA=1,this.option=function(){if(!arguments.length)throw new Error("No option specified");if(1===arguments.length&&"string"==typeof arguments[0]){try{return r(arguments[0])}catch(e){console.warn(e)}return!1}if(arguments.length>=2&&"string"==typeof arguments[0]){try{if(!a(arguments[0],arguments[1]))return!1}catch(e){return console.warn(e),!1}return this.redraw(),this.exposeColor(),!0}if(1===arguments.length&&"object"==typeof arguments[0]){var e=arguments[0],t=!0;for(var o in e)if(e.hasOwnProperty(o))try{a(o,e[o])||(t=!1)}catch(e){console.warn(e),t=!1}return this.redraw(),this.exposeColor(),t}throw new Error("Invalid arguments")},this.channel=function(e,t){if("string"!=typeof e)throw new Error("Invalid value for channel name: "+e);if(void 0===t)return this.channels.hasOwnProperty(e.toLowerCase())?this.channels[e.toLowerCase()]:(console.warn("Getting unknown channel: "+e),!1);var o=!1;switch(e.toLowerCase()){case"r":o=this.fromRGBA(t,null,null,null);break;case"g":o=this.fromRGBA(null,t,null,null);break;case"b":o=this.fromRGBA(null,null,t,null);break;case"h":o=this.fromHSVA(t,null,null,null);break;case"s":o=this.fromHSVA(null,t,null,null);break;case"v":o=this.fromHSVA(null,null,t,null);break;case"a":o=this.fromHSVA(null,null,null,t);break;default:return console.warn("Setting unknown channel: "+e),!1}return!!o&&(this.redraw(),!0)},this.trigger=function(e){for(var t=i.strList(e),o=0;o<t.length;o+=1){var n=t[o].toLowerCase(),a=null;switch(n){case"input":a="onInput";break;case"change":a="onChange"}a&&i.triggerCallback(this,a),i.triggerInputEvent(this.valueElement,n,!0,!0)}},this.fromHSVA=function(e,t,o,n,a){if(void 0===e&&(e=null),void 0===t&&(t=null),void 0===o&&(o=null),void 0===n&&(n=null),null!==e){if(isNaN(e))return!1;this.channels.h=Math.max(0,Math.min(360,e))}if(null!==t){if(isNaN(t))return!1;this.channels.s=Math.max(0,Math.min(100,this.maxS,t),this.minS)}if(null!==o){if(isNaN(o))return!1;this.channels.v=Math.max(0,Math.min(100,this.maxV,o),this.minV)}if(null!==n){if(isNaN(n))return!1;this.channels.a=this.hasAlphaChannel()?Math.max(0,Math.min(1,this.maxA,n),this.minA):1}var r=i.HSV_RGB(this.channels.h,this.channels.s,this.channels.v);return this.channels.r=r[0],this.channels.g=r[1],this.channels.b=r[2],this.exposeColor(a),!0},this.fromRGBA=function(e,t,o,n,a){if(void 0===e&&(e=null),void 0===t&&(t=null),void 0===o&&(o=null),void 0===n&&(n=null),null!==e){if(isNaN(e))return!1;e=Math.max(0,Math.min(255,e))}if(null!==t){if(isNaN(t))return!1;t=Math.max(0,Math.min(255,t))}if(null!==o){if(isNaN(o))return!1;o=Math.max(0,Math.min(255,o))}if(null!==n){if(isNaN(n))return!1;this.channels.a=this.hasAlphaChannel()?Math.max(0,Math.min(1,this.maxA,n),this.minA):1}var r=i.RGB_HSV(null===e?this.channels.r:e,null===t?this.channels.g:t,null===o?this.channels.b:o);null!==r[0]&&(this.channels.h=Math.max(0,Math.min(360,r[0]))),0!==r[2]&&(this.channels.s=Math.max(0,this.minS,Math.min(100,this.maxS,r[1]))),this.channels.v=Math.max(0,this.minV,Math.min(100,this.maxV,r[2]));var l=i.HSV_RGB(this.channels.h,this.channels.s,this.channels.v);return this.channels.r=l[0],this.channels.g=l[1],this.channels.b=l[2],this.exposeColor(a),!0},this.fromHSV=function(e,t,o,n){return console.warn("fromHSV() method is DEPRECATED. Using fromHSVA() instead."+i.docsRef),this.fromHSVA(e,t,o,null,n)},this.fromRGB=function(e,t,o,n){return console.warn("fromRGB() method is DEPRECATED. Using fromRGBA() instead."+i.docsRef),this.fromRGBA(e,t,o,null,n)},this.fromString=function(e,t){if(!this.required&&""===e.trim())return this.setPreviewElementBg(null),this.setValueElementValue(""),!0;var o=i.parseColorString(e);return!!o&&("any"===this.format.toLowerCase()&&(this._setFormat(o.format),i.isAlphaFormat(this.getFormat())||(o.rgba[3]=1)),this.fromRGBA(o.rgba[0],o.rgba[1],o.rgba[2],o.rgba[3],t),!0)},this.toString=function(e){switch(void 0===e&&(e=this.getFormat()),e.toLowerCase()){case"hex":return this.toHEXString();case"hexa":return this.toHEXAString();case"rgb":return this.toRGBString();case"rgba":return this.toRGBAString()}return!1},this.toHEXString=function(){return i.hexColor(this.channels.r,this.channels.g,this.channels.b)},this.toHEXAString=function(){return i.hexaColor(this.channels.r,this.channels.g,this.channels.b,this.channels.a)},this.toRGBString=function(){return i.rgbColor(this.channels.r,this.channels.g,this.channels.b)},this.toRGBAString=function(){return i.rgbaColor(this.channels.r,this.channels.g,this.channels.b,this.channels.a)},this.toGrayscale=function(){return.213*this.channels.r+.715*this.channels.g+.072*this.channels.b},this.toCanvas=function(){return i.genColorPreviewCanvas(this.toRGBAString()).canvas},this.toDataURL=function(){return this.toCanvas().toDataURL()},this.toBackground=function(){return i.pub.background(this.toRGBAString())},this.isLight=function(){return this.toGrayscale()>127.5},this.hide=function(){p()&&(i.removeClass(n.targetElement,i.pub.activeClassName),i.picker.wrap.parentNode.removeChild(i.picker.wrap),delete i.picker.owner)},this.show=function(){l()},this.redraw=function(){p()&&l()},this.getFormat=function(){return this._currentFormat},this._setFormat=function(e){this._currentFormat=e.toLowerCase()},this.hasAlphaChannel=function(){return"auto"===this.alphaChannel?"any"===this.format.toLowerCase()||i.isAlphaFormat(this.getFormat())||void 0!==this.alpha||void 0!==this.alphaElement:this.alphaChannel},this.processValueInput=function(e){this.fromString(e)||this.exposeColor()},this.processAlphaInput=function(e){this.fromHSVA(null,null,null,parseFloat(e))||this.exposeColor()},this.exposeColor=function(e){var t=this.toString(),o=this.getFormat();if(i.setDataAttr(this.targetElement,"current-color",t),e&i.flags.leaveValue||!this.valueElement||("hex"!==o&&"hexa"!==o||(this.uppercase||(t=t.toLowerCase()),this.hash||(t=t.replace(/^#/,""))),this.setValueElementValue(t)),!(e&i.flags.leaveAlpha)&&this.alphaElement){var n=Math.round(100*this.channels.a)/100;this.setAlphaElementValue(n)}e&i.flags.leavePreview||!this.previewElement||((i.isTextInput(this.previewElement)||i.isButton(this.previewElement)&&!i.isButtonEmpty(this.previewElement))&&this.previewPosition,this.setPreviewElementBg(this.toRGBAString())),p()&&(s(),c(),d())},this.setPreviewElementBg=function(e){if(this.previewElement){var t=null,o=null;(i.isTextInput(this.previewElement)||i.isButton(this.previewElement)&&!i.isButtonEmpty(this.previewElement))&&(t=this.previewPosition,o=this.previewSize);var n=[];if(e){n.push({image:i.genColorPreviewGradient(e,t,o?o-i.pub.previewSeparator.length:null),position:"left top",size:"auto",repeat:t?"repeat-y":"repeat",origin:"padding-box"});var a=i.genColorPreviewCanvas("rgba(0,0,0,0)",t?{left:"right",right:"left"}[t]:null,o,!0);n.push({image:"url('"+a.canvas.toDataURL()+"')",position:(t||"left")+" top",size:a.width+"px "+a.height+"px",repeat:t?"repeat-y":"repeat",origin:"padding-box"})}else n.push({image:"none",position:"left top",size:"auto",repeat:"no-repeat",origin:"padding-box"});for(var r={image:[],position:[],size:[],repeat:[],origin:[]},l=0;l<n.length;l+=1)r.image.push(n[l].image),r.position.push(n[l].position),r.size.push(n[l].size),r.repeat.push(n[l].repeat),r.origin.push(n[l].origin);var s={"background-image":r.image.join(", "),"background-position":r.position.join(", "),"background-size":r.size.join(", "),"background-repeat":r.repeat.join(", "),"background-origin":r.origin.join(", ")};i.setStyle(this.previewElement,s,this.forceStyle);var c={left:null,right:null};t&&(c[t]=this.previewSize+this.previewPadding+"px"),s={"padding-left":c.left,"padding-right":c.right},i.setStyle(this.previewElement,s,this.forceStyle,!0)}},this.setValueElementValue=function(e){this.valueElement&&("input"===i.nodeName(this.valueElement)?this.valueElement.value=e:this.valueElement.innerHTML=e)},this.setAlphaElementValue=function(e){this.alphaElement&&("input"===i.nodeName(this.alphaElement)?this.alphaElement.value=e:this.alphaElement.innerHTML=e)},this._processParentElementsInDOM=function(){if(!this._parentElementsProcessed){this._parentElementsProcessed=!0;var e=this.targetElement;do{var t=i.getCompStyle(e);t.position&&"fixed"===t.position.toLowerCase()&&(this.fixed=!0),e!==this.targetElement&&(i.getData(e,"hasScrollListener")||(e.addEventListener("scroll",i.onParentScroll,!1),i.setData(e,"hasScrollListener",!0)))}while((e=e.parentNode)&&"body"!==i.nodeName(e))}},this.tryHide=function(){this.hideOnLeave&&this.hide()},this.set__palette=function(e){this.palette=e,this._palette=i.parsePaletteValue(e),this._paletteHasTransparency=i.containsTranparentColor(this._palette)},i.pub.options)for(var u in i.pub.options)if(i.pub.options.hasOwnProperty(u))try{a(u,i.pub.options[u])}catch(e){console.warn(e)}var m=[];o.preset&&("string"==typeof o.preset?m=o.preset.split(/\s+/):Array.isArray(o.preset)?m=o.preset.slice():console.warn("Unrecognized preset value")),-1===m.indexOf("default")&&m.push("default");for(var h=m.length-1;h>=0;h-=1){var g=m[h];if(g)if(i.pub.presets.hasOwnProperty(g)){for(var u in i.pub.presets[g])if(i.pub.presets[g].hasOwnProperty(u))try{a(u,i.pub.presets[g][u])}catch(e){console.warn(e)}}else console.warn("Unknown preset: %s",g)}var f=["preset"];for(var u in o)if(o.hasOwnProperty(u)&&-1===f.indexOf(u))try{a(u,o[u])}catch(e){console.warn(e)}if(void 0===this.container?this.container=e.document.body:this.container=i.node(this.container),!this.container)throw new Error("Cannot instantiate color picker without a container element");if(this.targetElement=i.node(t),!this.targetElement){if("string"==typeof t&&/^[a-zA-Z][\w:.-]*$/.test(t))throw new Error("If '"+t+"' is supposed to be an ID, please use '#"+t+"' or any valid CSS selector.");throw new Error("Cannot instantiate color picker without a target element")}if(this.targetElement.jscolor&&this.targetElement.jscolor instanceof i.pub)throw new Error("Color picker already installed on this element");if(this.targetElement.jscolor=this,i.addClass(this.targetElement,i.pub.className),i.instances.push(this),i.isButton(this.targetElement)&&("button"!==this.targetElement.type.toLowerCase()&&(this.targetElement.type="button"),i.isButtonEmpty(this.targetElement))){i.removeChildren(this.targetElement),this.targetElement.appendChild(e.document.createTextNode(" "));var C=i.getCompStyle(this.targetElement);(parseFloat(C["min-width"])||0)<this.previewSize&&i.setStyle(this.targetElement,{"min-width":this.previewSize+"px"},this.forceStyle)}if(void 0===this.valueElement?i.isTextInput(this.targetElement)&&(this.valueElement=this.targetElement):null===this.valueElement||(this.valueElement=i.node(this.valueElement)),this.alphaElement&&(this.alphaElement=i.node(this.alphaElement)),void 0===this.previewElement?this.previewElement=this.targetElement:null===this.previewElement||(this.previewElement=i.node(this.previewElement)),this.valueElement&&i.isTextInput(this.valueElement)){var y={onInput:this.valueElement.oninput};this.valueElement.oninput=null,this.valueElement.addEventListener("keydown",(function(e){"Enter"===i.eventKey(e)&&(n.valueElement&&n.processValueInput(n.valueElement.value),n.tryHide())}),!1),this.valueElement.addEventListener("change",(function(e){if(!i.getData(e,"internal")){var t=n.valueElement.value;n.processValueInput(n.valueElement.value),i.triggerCallback(n,"onChange"),n.valueElement.value!==t&&i.triggerInputEvent(n.valueElement,"change",!0,!0)}}),!1),this.valueElement.addEventListener("input",(function(e){i.getData(e,"internal")||(n.valueElement&&n.fromString(n.valueElement.value,i.flags.leaveValue),i.triggerCallback(n,"onInput"))}),!1),y.onInput&&this.valueElement.addEventListener("input",y.onInput,!1),this.valueElement.setAttribute("autocomplete","off"),this.valueElement.setAttribute("autocorrect","off"),this.valueElement.setAttribute("autocapitalize","off"),this.valueElement.setAttribute("spellcheck",!1)}this.alphaElement&&i.isTextInput(this.alphaElement)&&(this.alphaElement.addEventListener("keydown",(function(e){"Enter"===i.eventKey(e)&&(n.alphaElement&&n.processAlphaInput(n.alphaElement.value),n.tryHide())}),!1),this.alphaElement.addEventListener("change",(function(e){if(!i.getData(e,"internal")){var t=n.alphaElement.value;n.processAlphaInput(n.alphaElement.value),i.triggerCallback(n,"onChange"),i.triggerInputEvent(n.valueElement,"change",!0,!0),n.alphaElement.value!==t&&i.triggerInputEvent(n.alphaElement,"change",!0,!0)}}),!1),this.alphaElement.addEventListener("input",(function(e){i.getData(e,"internal")||(n.alphaElement&&n.fromHSVA(null,null,null,parseFloat(n.alphaElement.value),i.flags.leaveAlpha),i.triggerCallback(n,"onInput"),i.triggerInputEvent(n.valueElement,"input",!0,!0))}),!1),this.alphaElement.setAttribute("autocomplete","off"),this.alphaElement.setAttribute("autocorrect","off"),this.alphaElement.setAttribute("autocapitalize","off"),this.alphaElement.setAttribute("spellcheck",!1));var b="FFFFFF";void 0!==this.value?b=this.value:this.valueElement&&void 0!==this.valueElement.value&&(b=this.valueElement.value);var k=void 0;if(void 0!==this.alpha?k=""+this.alpha:this.alphaElement&&void 0!==this.alphaElement.value&&(k=this.alphaElement.value),this._currentFormat=null,["auto","any"].indexOf(this.format.toLowerCase())>-1){var G=i.parseColorString(b);this._currentFormat=G?G.format:"hex"}else this._currentFormat=this.format.toLowerCase();this.processValueInput(b),void 0!==k&&this.processAlphaInput(k)}}).pub.className="jscolor",i.pub.activeClassName="jscolor-active",i.pub.looseJSON=!0,i.pub.presets={},i.pub.presets.default={},i.pub.presets.light={backgroundColor:"rgba(255,255,255,1)",controlBorderColor:"rgba(187,187,187,1)",buttonColor:"rgba(0,0,0,1)"},i.pub.presets.dark={backgroundColor:"rgba(51,51,51,1)",controlBorderColor:"rgba(153,153,153,1)",buttonColor:"rgba(240,240,240,1)"},i.pub.presets.small={width:101,height:101,padding:10,sliderSize:14,paletteCols:8},i.pub.presets.medium={width:181,height:101,padding:12,sliderSize:16,paletteCols:10},i.pub.presets.large={width:271,height:151,padding:12,sliderSize:24,paletteCols:15},i.pub.presets.thin={borderWidth:1,controlBorderWidth:1,pointerBorderWidth:1},i.pub.presets.thick={borderWidth:2,controlBorderWidth:2,pointerBorderWidth:2},i.pub.sliderInnerSpace=3,i.pub.chessboardSize=8,i.pub.chessboardColor1="#666666",i.pub.chessboardColor2="#999999",i.pub.previewSeparator=["rgba(255,255,255,.65)","rgba(128,128,128,.65)"],i.pub.init=function(){if(!i.initialized)for(e.document.addEventListener("mousedown",i.onDocumentMouseDown,!1),e.document.addEventListener("keyup",i.onDocumentKeyUp,!1),e.addEventListener("resize",i.onWindowResize,!1),e.addEventListener("scroll",i.onWindowScroll,!1),i.pub.install(),i.initialized=!0;i.readyQueue.length;)i.readyQueue.shift()()},i.pub.install=function(e){var t=!0;try{i.installBySelector("[data-jscolor]",e)}catch(e){t=!1,console.warn(e)}if(i.pub.lookupClass)try{i.installBySelector("input."+i.pub.lookupClass+", button."+i.pub.lookupClass,e)}catch(e){}return t},i.pub.ready=function(e){return"function"!=typeof e?(console.warn("Passed value is not a function"),!1):(i.initialized?e():i.readyQueue.push(e),!0)},i.pub.trigger=function(e){var t=function(){i.triggerGlobal(e)};i.initialized?t():i.pub.ready(t)},i.pub.hide=function(){i.picker&&i.picker.owner&&i.picker.owner.hide()},i.pub.chessboard=function(e){return e||(e="rgba(0,0,0,0)"),i.genColorPreviewCanvas(e).canvas.toDataURL()},i.pub.background=function(e){var t=[];t.push(i.genColorPreviewGradient(e));var o=i.genColorPreviewCanvas();return t.push(["url('"+o.canvas.toDataURL()+"')","left top","repeat"].join(" ")),t.join(", ")},i.pub.options={},i.pub.lookupClass="jscolor",i.pub.installByClassName=function(){return console.error('jscolor.installByClassName() is DEPRECATED. Use data-jscolor="" attribute instead of a class name.'+i.docsRef),!1},i.register(),i.pub);return void 0===e.jscolor&&(e.jscolor=e.JSColor=a),a}))}},t={};function o(n){var i=t[n];if(void 0!==i)return i.exports;var a=t[n]={exports:{}};return e[n].call(a.exports,a,a.exports,o),a.exports}o.n=e=>{var t=e&&e.__esModule?()=>e.default:()=>e;return o.d(t,{a:t}),t},o.d=(e,t)=>{for(var n in t)o.o(t,n)&&!o.o(e,n)&&Object.defineProperty(e,n,{enumerable:!0,get:t[n]})},o.o=(e,t)=>Object.prototype.hasOwnProperty.call(e,t),(()=>{"use strict";let e={};function t(){e.WrinklerButtons&&Game.elderWrath?(l("PopAllNormalWrinklerButton").style.display="",l("PopFattestWrinklerButton").style.display=""):(l("PopAllNormalWrinklerButton").style.display="none",l("PopFattestWrinklerButton").style.display="none")}const n=["","","M","G","T","P","E","Z","Y"],i=["","","M","B","Tr","Quadr","Quint","Sext","Sept","Oct","Non","Dec","Undec","Duodec","Tredec","Quattuordec","Quindec","Sexdec","Septendec","Octodec","Novemdec","Vigint","Unvigint","Duovigint","Trevigint","Quattuorvigint"],a=["","K","M","B","T","Qa","Qi","Sx","Sp","Oc","No","De","UDe","DDe","TDe","QaDe","QiDe","SxDe","SpDe","ODe","NDe","Vi","UVi","DVi","TVi","QaVi","QiVi","SxVi","SpVi","OVi","NVi","Tr","UTr","DTr","TTr","QaTr","QiTr","SxTr","SpTr","OTr","NTr","Qaa","UQa","DQa","TQa","QaQa","QiQa","SxQa","SpQa","OQa","NQa","Qia","UQi","DQi","TQi","QaQi","QiQi","SxQi","SpQi","OQi","NQi","Sxa","USx","DSx","TSx","QaSx","QiSx","SxSx","SpSx","OSx","NSx","Spa","USp","DSp","TSp","QaSp","QiSp","SxSp","SpSp","OSp","NSp","Oco","UOc","DOc","TOc","QaOc","QiOc","SxOc","SpOc","OOc","NOc","Noa","UNo","DNo","TNo","QaNo","QiNo","SxNo","SpNo","ONo","NNo","Ct","UCt"];let r,s,c,d,p,u,m,h,g,f,C,y,b,k,G,v,w=[],x=[],T=[],M={},S=Date.now(),P=Date.now();function B(t,o,r){const l=e.ScaleDecimals+1;if(t===1/0)return"Infinity";if(void 0===t)return"0";if(0===e.Scale)return M.Beautify(t,o);if(Number.isFinite(t)){if(t<0)return`-${B(Math.abs(t))}`;let s="";if(0===t)return t.toString();if(t>.001&&t<e.ScaleCutoff)return s=e.ScaleSeparator?t.toLocaleString("nl"):t.toLocaleString("en"),s;if(4===e.Scale&&!r||4===r)s=t.toExponential(l).toString().replace("e","E");else{const o=t.toExponential().toString(),c=Math.floor(o.slice(o.indexOf("e")+1)/3);s=(t/Number("1e"+3*c)).toFixed(l),1===e.Scale&&!r||1===r?t>=.01&&t<Number("1e"+3*n.length)?s+=` ${n[c]}`:s=B(t,0,4):2===e.Scale&&!r||2===r?t>=.01&&t<Number("1e"+3*i.length)?s+=` ${i[c]}`:s=B(t,0,4):3===e.Scale&&!r||3===r?t>=.01&&t<Number("1e"+3*a.length)?s+=` ${a[c]}`:s=B(t,0,4):(5===e.Scale&&!r||5===r)&&(s+="E"+3*c)}return""===s&&(console.log(`Could not beautify number with Cookie Monster Beautify: ${t}`),s=M.Beautify(t,o)),e.ScaleSeparator&&(s=s.replace(".",",")),s}return console.log(`Could not beautify number with Cookie Monster Beautify: ${t}`),M.Beautify(t,o)}let E,N,O,F,L,W,A,j,D,H,U,$,R,I,z,V,_,Q,X,q,Y,K,J,Z,ee,te,oe,ne,ie,ae,re,le,se,ce,de,pe,ue,me=0,he=0,ge=0,fe=0,Ce=0,ye=0,be=0,ke=0,Ge=0,ve=0,we=0,xe=0,Te=0,Me=0,Se=0,Pe=1,Be=1,Ee=1,Ne=0,Oe=0,Fe=0,Le=0,We={},Ae={},je={},De={},He=0,Ue=0,$e=[0,null],Re=0,Ie=0,ze={},Ve={},_e={},Qe={},Xe={},qe=0,Ye=0,Ke=[],Je=0,Ze=0,et=[],tt=0,ot=0,nt=[],it=0,at=0,rt=[],lt=0,st=0,ct={},dt={},pt=0,ut=0,mt=[],ht={},gt=0,ft={0:[0,0,0],1:[0,0,0],2:[0,0,0],3:[0,0,0],4:[0,0,0],5:[0,0,0],6:[0,0,0],7:[0,0,0],8:[0,0,0],9:[0,0,0],10:[0,0,0]},Ct=[],yt=[],bt=[];function kt(e){let t=0;return Game.dragonAuras[ne].name!==e&&Game.dragonAuras[ie].name!==e||(t=1),"Reality Bending"!==Game.dragonAuras[ne].name&&"Reality Bending"!==Game.dragonAuras[ie].name||(t+=.1),t}function Gt(){let e=.25;return e*=1+kt("Earth Shatterer"),e}function vt(e,t){return void 0===ce[e]?void 0===t?1:t:ce[e]}function wt(e){const t=yt[e];return(1!==Game.ascensionMode||"prestige"!==t.pool&&"fortune"!==t.tier)&&t?t.bought:0}function xt(e){if(Game.hasGod){void 0===Ct.Temple.minigame&&(Ct.Temple.minigame=Game.Objects.Temple.minigame);const t=Ct.Temple.minigame.gods[e];if(ae===t.id)return 1;if(re===t.id)return 2;if(le===t.id)return 3}return!1}function Tt(e,t){let o=t;if(wt("Season savings")&&(o*=.99),wt("Santa's dominion")&&(o*=.99),wt("Faberge egg")&&(o*=.99),wt("Divine discount")&&(o*=.99),wt("Fortune #100")&&(o*=.99),o*=1-.02*kt("Fierce Hoarder"),Game.hasBuff("Everything must go")&&(o*=.95),Game.hasBuff("Crafty pixies")&&(o*=.98),Game.hasBuff("Nasty goblins")&&(o*=1.02),e.fortune&&wt(e.fortune.name)&&(o*=.93),o*=vt("buildingCost"),Ct.Temple.minigameLoaded){const e=xt("creation");1===e?o*=.93:2===e?o*=.95:3===e&&(o*=.98)}return o}function Mt(e,t,o,n,i,a){const r=void 0===a?0:a;let l=i,s=o,c=0;-1===i&&(l=s),i||(l=Game.buyBulk);for(let o=0;o<l;o++){let o=t*Game.priceIncrease**Math.max(0,s-n);o=r?Game.modifyBuildingPrice(e,o):Tt(e,o),o=Math.ceil(o);const i=r?e.getSellMultiplier():Gt();o=Math.floor(o*i),s>0&&(c+=o,s-=1)}return c}const St="CMText",Pt="CMBack",Bt="Blue",Et="Green",Nt="Yellow",Ot="Orange",Ft="Red",Lt="Purple",Wt="Gray",At="Pink",jt="Brown",Dt=[Wt,Bt,Et,Nt,Ot,Ft,Lt,At,jt],Ht=[["GoldCookTooltipPlaceholder","Calculated with Golden Switch off","200px"],["GoldCookDragonsFortuneTooltipPlaceholder","Calculated with Golden Switch off and at least one golden cookie on-screen","240px"],["PrestMaxTooltipPlaceholder","The MAX prestige is calculated with the cookies gained from popping all wrinklers with Skruuia god in Diamond slot, selling all stock market goods, selling all buildings with Earth Shatterer and Reality Bending auras, and buying Chocolate egg","320px"],["NextPrestTooltipPlaceholder","Calculated with cookies gained from wrinklers and Chocolate egg","200px"],["HeavenChipMaxTooltipPlaceholder","The MAX heavenly chips is calculated with the cookies gained from popping all wrinklers with Skruuia god in Diamond slot, selling all stock market goods, selling all buildings with Earth Shatterer and Reality Bending auras, and buying Chocolate egg","330px"],["ResetTooltipPlaceholder","The bonus income you would get from new prestige levels unlocked at 100% of its potential and from ascension achievements if you have the same buildings/upgrades after reset","370px"],["ChoEggTooltipPlaceholder","The amount of cookies you would get from popping all wrinklers with Skruuia god in Diamond slot, selling all stock market goods, selling all buildings with Earth Shatterer and Reality Bending auras, and then buying Chocolate egg","300px"],["ChainNextLevelPlaceholder","Cheated cookies might break this formula","250px"],["FavouriteSettingPlaceholder","Click to set this setting as favourite and show it in 'favourite' settings at the top of the Cookie Monster Settings","250px"]],Ut={};let $t,Rt,It=0,zt=-1,Vt=[],_t=1,Qt=1,Xt=1;const qt=[10,15,30,60,300,600,900,1800],Yt=[1,5,10,15,30],Kt={Frenzy:Nt,"Dragon Harvest":jt,"Elder frenzy":Et,Clot:Ft,"Click frenzy":Bt,Dragonflight:At};let Jt,Zt,eo,to,oo,no,io,ao={},ro="",lo=[];function so(){let t=Game.buyBulk;if(1===Game.buyMode?Qt=t:t=Qt,1===t?t=We:10===t?t=Ae:100===t&&(t=je),1===Game.buyMode?1===e.BuildColour?Object.keys(t).forEach((o=>{l(`productPrice${Game.Objects[o].id}`).style.color=e[`Colour${t[o].color}`]})):Object.keys(Game.Objects).forEach((e=>{l(`productPrice${Game.Objects[e].id}`).style.removeProperty("color")})):-1===Game.buyMode&&Object.keys(We).forEach((e=>{const t=Game.Objects[e];l(`productPrice${t.id}`).style.color="",l(`productPrice${t.id}`).innerHTML=B(Mt(t,t.basePrice,t.amount,t.free,Game.buyBulk,1))})),1===Game.buyMode&&e.SortBuildings){let o;1===e.SortBuildings?(o=Object.keys(We).map((e=>{const t=We[e];return t.name=e,t.id=Game.Objects[e].id,t})),o.sort((function(e,t){return Dt.indexOf(e.color)>Dt.indexOf(t.color)?1:Dt.indexOf(e.color)<Dt.indexOf(t.color)||e.pp<t.pp?-1:0}))):2===e.SortBuildings&&(o=Object.keys(t).map((e=>{const o=t[e];return o.name=e,o.id=Game.Objects[e].id,o})),o.sort((function(e,t){return Dt.indexOf(e.color)>Dt.indexOf(t.color)?1:Dt.indexOf(e.color)<Dt.indexOf(t.color)||e.pp<t.pp?-1:0})));for(let e=0;e<o.length;e++)Game.Objects[o[e].name].l.style.gridRow=`${e+2}/${e+2}`}else{const e=Object.keys(We).map((e=>{const t=We[e];return t.name=e,t.id=Game.Objects[e].id,t}));e.sort(((e,t)=>e.id-t.id));for(let t=0;t<e.length;t++)Game.Objects[e[t].name].l.style.gridRow=`${t+2}/${t+2}`}}function co(){if(e.UpBarColour>0){let e=0,t=0,o=0,n=0,i=0,a=0,r=0;Object.keys(Game.UpgradesInStore).forEach((s=>{const c=Game.UpgradesInStore[s];let d=!1;for(let e=0;e<l(`upgrade${s}`).childNodes.length;e+=1)if(-1!==l(`upgrade${s}`).childNodes[e].className.indexOf(Pt)){l(`upgrade${s}`).childNodes[e].className=Pt+ze[c.name].color,d=!0;break}if(!d){const e=document.createElement("div");e.style.width="10px",e.style.height="10px",e.className=Pt+ze[c.name].color,l(`upgrade${s}`).appendChild(e)}ze[c.name].color===Bt?e+=1:ze[c.name].color===Et?t+=1:ze[c.name].color===Nt?o+=1:ze[c.name].color===Ot?n+=1:ze[c.name].color===Ft?i+=1:ze[c.name].color===Lt?a+=1:ze[c.name].color===Wt&&(r+=1)})),l("CMUpgradeBarBlue").textContent=e,l("CMUpgradeBarGreen").textContent=t,l("CMUpgradeBarYellow").textContent=o,l("CMUpgradeBarOrange").textContent=n,l("CMUpgradeBarRed").textContent=i,l("CMUpgradeBarPurple").textContent=a,l("CMUpgradeBarGray").textContent=r}const t=[];for(let e=0;e<Game.UpgradesInStore.length;e+=1){const o={};o.name=Game.UpgradesInStore[e].name,o.price=Game.UpgradesInStore[e].basePrice,o.pp=ze[o.name].pp,t.push(o)}e.SortUpgrades?t.sort((function(e,t){return Dt.indexOf(e.color)>Dt.indexOf(t.color)?1:Dt.indexOf(e.color)<Dt.indexOf(t.color)||e.pp<t.pp?-1:0})):t.sort(((e,t)=>e.price-t.price));const o=function(e,t){return e.findIndex((e=>e.name===t.name))};for(let e=0;e<Game.UpgradesInStore.length;e+=1)l(`upgrade${e}`).style.order=o(t,Game.UpgradesInStore[e])+1}function po(t,o){let n=t;if(t===1/0)return t;if(t<0)return"Negative time period";n=Math.ceil(t);const i=Math.floor(n/31557600),a=Math.floor(n%31557600/86400),r=Math.floor(n%86400/3600),l=Math.floor(n%3600/60),s=Math.floor(n%60);let c="";if(e.TimeFormat){if(n>315576e4)return"XX:XX:XX:XX:XX";c+=(i<10?"0":"")+i+":",c+=(a<10?"0":"")+a+":",c+=(r<10?"0":"")+r+":",c+=(l<10?"0":"")+l+":",c+=(s<10?"0":"")+s}else{if(n>7776e5)return o?"Over 9000 days!":">9000d";c+=i>0?i+(o?1===i?" year":" years":"y")+", ":"",c+=a>0?a+(o?1===a?" day":" days":"d")+", ":"",(c.length>0||r>0)&&(c+=r+(o?1===r?" hour":" hours":"h")+", "),(c.length>0||l>0)&&(c+=l+(o?1===l?" minute":" minutes":"m")+", "),c+=s+(o?1===s?" second":" seconds":"s")}return c}function uo(t){let o,n;return t<=0?(n=e.TimeFormat?"00:00:00:00:00":"Done!",o=Et):(n=po(t),o=t>300?Ft:t>60?Ot:Nt),{text:n,color:o}}function mo(){return e.CPSMode?_:0===e.CalcWrink?Game.cookiesPs*(1-Game.cpsSucked):1===e.CalcWrink?Game.cookiesPs*(Re+(1-.05*Ie)):null!==$e[1]&&2===e.CalcWrink&&1===Game.wrinklers[$e[1]].type?Game.cookiesPs*(3*Re/Ie+(1-.05*Ie)):Game.cookiesPs*(Re/Ie+(1-.05*Ie))}function ho(){return 1===e.CalcWrink?He:2===e.CalcWrink?$e[0]:0}function go(e,t,o){const n=document.createElement("div");n.id=e,n.style.height="12px",n.style.margin="0px 10px",n.style.position="relative";const i=document.createElement("div");i.style.width="100%",i.style.height="10px",i.style.margin="auto",i.style.position="absolute",i.style.left="0px",i.style.top="0px",i.style.right="0px",i.style.bottom="0px";const a=document.createElement("span");a.style.display="inline-block",a.style.textAlign="right",a.style.fontSize="10px",a.style.width="108px",a.style.marginRight="5px",a.style.verticalAlign="text-top",a.textContent=t,i.appendChild(a);for(let e=0;e<o.length;e++){const t=document.createElement("span");t.id=o[e].id,t.style.display="inline-block",t.style.height="10px",t.style.verticalAlign="text-top",t.style.textAlign="center",o.length-1===e&&(t.style.borderTopRightRadius="10px",t.style.borderBottomRightRadius="10px"),void 0!==o[e].color&&(t.className=Pt+o[e].color),i.appendChild(t)}const r=document.createElement("span");return r.id=`${e}Time`,r.style.marginLeft="5px",r.style.verticalAlign="text-top",i.appendChild(r),n.appendChild(i),n}function fo(e){if(null!==l("CMBotBar")){const t=l("CMBotBar").firstChild.firstChild.childNodes[0],o=l("CMBotBar").firstChild.firstChild.childNodes[1],n=l("CMBotBar").firstChild.firstChild.childNodes[2],i=l("CMBotBar").firstChild.firstChild.childNodes[3],a=e,r=t.appendChild(document.createElement("td"));r.appendChild(document.createTextNode(`${-1!==a.indexOf(" ")?a.substring(0,a.indexOf(" ")):a} (`)),r.appendChild(document.createElement("span")).className="CMTextBlue",r.appendChild(document.createTextNode(")")),t.lastChild.style.paddingLeft="8px",o.appendChild(document.createElement("td")),o.lastChild.style.paddingLeft="8px",n.appendChild(document.createElement("td")),n.lastChild.style.paddingLeft="8px",i.appendChild(document.createElement("td")),i.lastChild.style.paddingLeft="2px"}}function Co(){if(1===e.BotBar&&We&&1===Game.buyMode){let t=0;Object.keys(We).forEach((o=>{let n,i=Game.buyBulk;1===Game.buyMode?_t=i:i=_t,1===i&&(i=We),10===i&&(i=Ae),100===i&&(i=je),t+=1,l("CMBotBar").firstChild.firstChild.childNodes[0].childNodes[t].childNodes[1].textContent=Game.Objects[o].amount,l("CMBotBar").firstChild.firstChild.childNodes[1].childNodes[t].textContent=B(i[o].bonus,2),l("CMBotBar").firstChild.firstChild.childNodes[2].childNodes[t].className=St+i[o].color,n=e.PPDisplayTime?po(Math.round(i[o].pp)):B(Math.round(i[o].pp),2),l("CMBotBar").firstChild.firstChild.childNodes[2].childNodes[t].textContent=n;const a=uo((Game.Objects[o].bulkPrice-(Game.cookies+ho()))/mo());l("CMBotBar").firstChild.firstChild.childNodes[3].childNodes[t].className=St+a.color,"Done!"===a.text&&Game.cookies<Game.Objects[o].bulkPrice?l("CMBotBar").firstChild.firstChild.childNodes[3].childNodes[t].textContent=`${a.text} (with Wrink)`:l("CMBotBar").firstChild.firstChild.childNodes[3].childNodes[t].textContent=a.text}))}}function yo(){Game.Background.canvas.width=Game.Background.canvas.parentNode.offsetWidth,Game.Background.canvas.height=Game.Background.canvas.parentNode.offsetHeight,Game.LeftBackground.canvas.width=Game.LeftBackground.canvas.parentNode.offsetWidth,Game.LeftBackground.canvas.height=Game.LeftBackground.canvas.parentNode.offsetHeight,Game.DrawBackground()}function bo(){1===e.BotBar&&1===e.TimerBar&&1===e.TimerBarPos?(l("CMBotBar").style.bottom=l("CMTimerBar").style.height,l("game").style.bottom=`${Number(l("CMTimerBar").style.height.replace("px",""))+70}px`):1===e.BotBar?(l("CMBotBar").style.bottom="0px",l("game").style.bottom="70px"):1===e.TimerBar&&1===e.TimerBarPos?l("game").style.bottom=l("CMTimerBar").style.height:l("game").style.bottom="0px",1===e.TimerBar&&0===e.TimerBarPos?l("sectionLeft").style.top=l("CMTimerBar").style.height:l("sectionLeft").style.top="",yo()}function ko(){1===e.TimerBar?l("CMTimerBar").style.display="":l("CMTimerBar").style.display="none",bo()}function Go(){me=Game.dragonAura,he=Game.dragonAura2}function vo(e){const t=Game.Achievements[e],o={};return o.name=t.name,o}function wo(e){let t=1;if(Object.keys(e.tieredUpgrades).forEach((o=>{!Game.Tiers[e.tieredUpgrades[o].tier].special&&wt(e.tieredUpgrades[o].name)&&(t*=2)})),Object.keys(e.synergies).forEach((o=>{if(wt(e.synergies[o].name)){const n=e.synergies[o];n.buildingTie1.name===e.name?t*=1+.05*n.buildingTie2.amount:n.buildingTie2.name===e.name&&(t*=1+.001*n.buildingTie1.amount)}})),e.fortune&&wt(e.fortune.name)&&(t*=1.07),e.grandma&&wt(e.grandma.name)&&(t*=1+.01*Ct.Grandma.amount*(1/(e.id-1))),"object"==typeof e.tieredUpgrades.misfortune&&1===e.vanilla&&wt(e.tieredUpgrades.misfortune.name))switch(Game.elderWrath){default:t*=1;break;case 1:t*=1.02;break;case 2:t*=1.04;break;case 3:t*=1.06}return t}function xo(e){const t=Game.Objects[e],o={};return"Cursor"===t.name?o.cps=function(e){let t=0;wt("Thousand fingers")&&(t+=.1),wt("Million fingers")&&(t*=5),wt("Billion fingers")&&(t*=10),wt("Trillion fingers")&&(t*=20),wt("Quadrillion fingers")&&(t*=20),wt("Quintillion fingers")&&(t*=20),wt("Sextillion fingers")&&(t*=20),wt("Septillion fingers")&&(t*=20),wt("Octillion fingers")&&(t*=20),wt("Nonillion fingers")&&(t*=20);let o=1,n=0;return Object.keys(Ct).forEach((e=>{"Cursor"!==Ct[e].name&&(n+=Ct[e].amount)})),t*=n,o*=wo(e),o*=Game.magicCpS("Cursor"),o*=vt("cursorCps"),Game.ComputeCps(.1,wt("Reinforced index finger")+wt("Carpal tunnel prevention cream")+wt("Ambidextrous"),t)*o}:"Grandma"===t.name?o.cps=function(e){let o=1;if(Object.keys(Game.GrandmaSynergies).forEach((e=>{wt(Game.GrandmaSynergies[e])&&(o*=2)})),wt("Bingo center/Research facility")&&(o*=4),wt("Ritual rolling pins")&&(o*=2),wt("Naughty list")&&(o*=2),wt("Elderwort biscuits")&&(o*=1.02),o*=vt("grandmaCps"),wt("Cat ladies"))for(let e=0;e<Game.UpgradesByPool.kitten.length;e++)wt(Game.UpgradesByPool.kitten[e].name)&&(o*=1.29);o*=wo(e);let n=0;wt("One mind")&&(n+=.02*Ct.Grandma.amount),wt("Communal brainsweep")&&(n+=.02*Ct.Grandma.amount),wt("Elder Pact")&&(n+=.05*Ct.Portal.amount);let i=0;return Object.keys(Ct).forEach((e=>{"Grandma"!==Ct[e].name&&(i+=Ct[e].amount)})),o*=1+.01*kt("Elder Battalion")*i,o*=Game.magicCpS(t.name),(t.baseCps+n)*o}:o.cps=function(e){let t=1;return t*=wo(e),t*=Game.magicCpS(e.name),e.baseCPS*t},o.baseCps=t.baseCps,o.name=t.name,o.tieredUpgrades=t.tieredUpgrades,o.synergies=t.synergies,o.fortune=t.fortune,o.grandma=t.grandma,o.baseCPS=t.baseCps,o.id=t.id,o.vanilla=t.vanilla,o}function To(e){const t=Game.Upgrades[e],o={};return o.power=t.power,"function"==typeof o.power&&("Sugar crystal cookies"===t.name?o.power=function(){let e=5;return Object.keys(Ct).forEach((t=>{Ct[t].level>=10&&(e+=1)})),e}:o.power=function(){let e=2;if(wt("Starlove")&&(e=3),Game.hasGod){const t=xt("seasons");1===t?e*=1.3:2===t?e*=1.2:3===t&&(e*=1.1)}return e}),o.pool=t.pool,o.name=t.name,o}function Mo(){J=Game.UpgradesOwned,Z=Game.pledges,ee=Game.AchievementsOwned,te=Game.heavenlyPower,oe=Game.prestige,Object.keys(Game.Objects).forEach((e=>{const t=Game.Objects[e];let o=Ct[e];void 0===o&&(Ct[e]=xo(e),o=Ct[e],fo(e)),o.amount=t.amount,o.level=t.level,o.totalCookies=t.totalCookies,o.basePrice=t.basePrice,o.free=t.free,t.minigameLoaded&&("Temple"===t.name&&(ae=t.minigame.slot[0],re=t.minigame.slot[1],le=t.minigame.slot[2]),o.minigameLoaded=t.minigameLoaded,o.minigame=t.minigame),Ct[e]=o})),Object.keys(Game.Upgrades).forEach((e=>{const t=Game.Upgrades[e];let o=yt[e];void 0===o&&(yt[e]=To(e),o=yt[e]),o.bought=t.bought,yt[e]=o})),Object.keys(Game.Achievements).forEach((e=>{const t=Game.Achievements[e];let o=bt[e];void 0===o&&(bt[e]=vo(e),o=bt[e]),o.won=t.won,bt[e]=o})),Go(),ne=me,ie=he}function So(e){const t=document.createElement("div");return t.style.fontWeight="bold",t.id=`${e}Title`,t.className="CMTextBlue",t.textContent=e,t}function Po(e){e.appendChild(So("Bonus Income"));const t=document.createElement("div");t.style.marginBottom="4px",t.style.color="white",t.id="CMTooltipIncome",e.appendChild(t),e.appendChild(So("Bonus Cookies per Click")),e.lastChild.style.display="none";const o=document.createElement("div");o.style.marginBottom="4px",o.style.color="white",o.style.display="none",o.id="CMTooltipCookiePerClick",e.appendChild(o),e.appendChild(So("Payback Period"));const n=document.createElement("div");n.style.marginBottom="4px",n.id="CMTooltipPP",e.appendChild(n),e.appendChild(So("Time Left"));const i=document.createElement("div");if(i.id="CMTooltipTime",e.appendChild(i),"b"===eo){e.appendChild(So("Production left till next achievement")),e.lastChild.id="CMTooltipProductionLeftHeader";const t=document.createElement("div");t.id="CMTooltipProductionLeft",e.appendChild(t)}if("b"===eo){e.appendChild(So("Buildings (price / PP) left till next achievement")),e.lastChild.id="CMTooltipNextAchievementHeader";const t=document.createElement("div");t.id="CMTooltipNextAchievement",e.appendChild(t)}}function Bo(t,o){let n="";return n=t.pp<=0||t.pp===1/0?Wt:t.pp<ut?Bt:t.pp===ut?Et:t.pp<mt[10][0]?Nt:t.pp<mt[20][0]?Ot:t.pp<mt[30][0]?Ft:Lt,0!==Number(e.PPSecondsLowerLimit)&&o/mo()<Number(e.PPSecondsLowerLimit)&&(n=Bt),e.PPOnlyConsiderBuyable&&o-Game.cookies>0&&(n=Ft),n}function Eo(){let e=1;return Object.keys(Game.buffs).forEach((t=>{void 0!==Game.buffs[t].multCpS&&(e*=Game.buffs[t].multCpS)})),e}function No(e){bt[e]&&0===bt[e].won&&(bt[e].won=1,"shadow"!==Game.Achievements[e].pool&&(ee+=1))}function Oo(){de=0;let e=1;const t={};Object.keys(Game.Objects).forEach((e=>{if(Game.Objects[e].minigameLoaded&&Game.Objects[e].minigame.effs){const o=Game.Objects[e].minigame.effs;Object.keys(o).forEach((e=>{t[e]?t[e]*=o[e]:t[e]=o[e]}))}})),ce=t,1!==Game.ascensionMode&&(e+=.01*parseFloat(oe)*te*function(){let e=0;if(wt("Heavenly chip secret")&&(e+=.05),wt("Heavenly cookie stand")&&(e+=.2),wt("Heavenly bakery")&&(e+=.25),wt("Heavenly confectionery")&&(e+=.25),wt("Heavenly key")&&(e+=.25),e*=1+.05*kt("Dragon God"),wt("Lucky digit")&&(e*=1.01),wt("Lucky number")&&(e*=1.01),wt("Lucky payout")&&(e*=1.01),Game.hasGod){const t=xt("creation");1===t?e*=.7:2===t?e*=.8:3===t&&(e*=.9)}return e}()),e*=vt("cps"),wt("Heralds")&&1!==Game.ascensionMode&&(e*=1+.01*Game.heralds),Object.keys(Game.cookieUpgrades).forEach((t=>{const o=Game.cookieUpgrades[t];wt(o.name)&&("function"==typeof o.power?e*=1+.01*yt[o.name].power(yt[o.name]):e*=1+.01*o.power)})),wt("Specialized chocolate chips")&&(e*=1.01),wt("Designer cocoa beans")&&(e*=1.02),wt("Underworld ovens")&&(e*=1.03),wt("Exotic nuts")&&(e*=1.04),wt("Arcane sugar")&&(e*=1.05),wt("Increased merriness")&&(e*=1.15),wt("Improved jolliness")&&(e*=1.15),wt("A lump of coal")&&(e*=1.01),wt("An itchy sweater")&&(e*=1.01),wt("Santa's dominion")&&(e*=1.2),wt("Fortune #100")&&(e*=1.01),wt("Fortune #101")&&(e*=1.07),wt("Dragon scale")&&(e*=1.03);let o=1;if(xt){let t=xt("asceticism");1===t?e*=1.15:2===t?e*=1.1:3===t&&(e*=1.05),t=xt("ages"),1===t?e*=1+.15*Math.sin(S/1e3/10800*Math.PI*2):2===t?e*=1+.15*Math.sin(S/1e3/43200*Math.PI*2):3===t&&(e*=1+.15*Math.sin(S/1e3/86400*Math.PI*2)),t=xt("decadence"),1===t?o*=.93:2===t?o*=.95:3===t&&(o*=.98),t=xt("industry"),1===t?o*=1.1:2===t?o*=1.06:3===t&&(o*=1.03),t=xt("labor"),1===t?o*=.97:2===t?o*=.98:3===t&&(o*=.99)}wt("Santa's legacy")&&(e*=1+.03*(Game.santaLevel+1));const n=ee/25;let i=1;if(wt("Santa's milk and cookies")&&(i*=1.05),i*=1+.05*kt("Breath of Milk"),xt){const e=xt("mother");1===e?i*=1.1:2===e?i*=1.05:3===e&&(i*=1.03)}i*=vt("milk");let a=1;wt("Kitten helpers")&&(a*=1+.1*n*i),wt("Kitten workers")&&(a*=1+.125*n*i),wt("Kitten engineers")&&(a*=1+.15*n*i),wt("Kitten overseers")&&(a*=1+.175*n*i),wt("Kitten managers")&&(a*=1+.2*n*i),wt("Kitten accountants")&&(a*=1+.2*n*i),wt("Kitten specialists")&&(a*=1+.2*n*i),wt("Kitten experts")&&(a*=1+.2*n*i),wt("Kitten consultants")&&(a*=1+.2*n*i),wt("Kitten assistants to the regional manager")&&(a*=1+.175*n*i),wt("Kitten marketeers")&&(a*=1+.15*n*i),wt("Kitten analysts")&&(a*=1+.125*n*i),wt("Kitten executives")&&(a*=1+.115*n*i),wt("Kitten angels")&&(a*=1+.1*n*i),wt("Fortune #103")&&(a*=1+.05*n*i),Object.keys(Ct).forEach((e=>{const t=Ct[e];let a=t.cps(t);1!==Game.ascensionMode&&(a*=(1+.01*t.level)*o),"Grandma"===t.name&&wt("Milkhelp&reg; lactose intolerance relief tablets")&&(a*=1+.05*n*i),de+=t.amount*a})),wt('"egg"')&&(de+=9),e*=a;let r=1;if(wt("Chicken egg")&&(r*=1.01),wt("Duck egg")&&(r*=1.01),wt("Turkey egg")&&(r*=1.01),wt("Quail egg")&&(r*=1.01),wt("Robin egg")&&(r*=1.01),wt("Ostrich egg")&&(r*=1.01),wt("Cassowary egg")&&(r*=1.01),wt("Salmon roe")&&(r*=1.01),wt("Frogspawn")&&(r*=1.01),wt("Shark egg")&&(r*=1.01),wt("Turtle egg")&&(r*=1.01),wt("Ant larva")&&(r*=1.01),wt("Century egg")){let e=10*Math.floor((P-Game.startDate)/1e3/10)/60/60/24;e=Math.min(e,100),Oe=1+.1*(1-(1-e/100)**3),r*=Oe}e*=r,wt("Sugar baking")&&(e*=1+.01*Math.min(100,Game.lumps)),e*=1+kt("Radiant Appetite");const l=de*e;Object.keys(Game.CpsAchievements).forEach((e=>{l>=Game.CpsAchievements[e].threshold&&No(Game.CpsAchievements[e].name)})),pe=l;const{n:s}=Game.shimmerTypes.golden,c=kt("Dragon's Fortune");for(let t=0;t<s;t++)e*=1+1.23*c;const d=Game.bakeryName.toLowerCase();if("orteil"===d?e*=.99:"ortiel"===d&&(e*=.98),wt("Elder Covenant")&&(e*=.95),wt("Golden switch [off]")){let t=1.5;if(wt("Residual luck")){const e=Game.goldenCookieUpgrades;Object.keys(e).forEach((o=>{wt(e[o])&&(t+=.1)}))}e*=t}if(wt("Shimmering veil [off]")){let t=.5;wt("Reinforced membrane")&&(t+=.1),e*=1+t}wt("Magic shenanigans")&&(e*=1e3),wt("Occult obstruction")&&(e*=0),de=Game.runModHookOnValue("cps",de),e*=Eo(),de*=e}const Fo=["Fortune #001","Fortune #002","Fortune #003","Fortune #004","Fortune #005","Fortune #006","Fortune #007","Fortune #008","Fortune #009","Fortune #010","Fortune #011","Fortune #012","Fortune #013","Fortune #014","Fortune #015","Fortune #016","Fortune #017","Fortune #018","Fortune #100","Fortune #101","Fortune #102","Fortune #103","Fortune #104"],Lo=["Skull cookies","Ghost cookies","Bat cookies","Slime cookies","Pumpkin cookies","Eyeball cookies","Spider cookies"],Wo=["Christmas tree biscuits","Snowflake biscuits","Snowman biscuits","Holly biscuits","Candy cane biscuits","Bell biscuits","Present biscuits"],Ao=["Pure heart biscuits","Ardent heart biscuits","Sour heart biscuits","Weeping heart biscuits","Golden heart biscuits","Eternal heart biscuits","Prism heart biscuits"],jo=["Elderwort biscuits","Bakeberry cookies","Duketater cookies","Green yeast digestives","Wheat slims","Fern tea","Ichor syrup"];function Do(e){return bt[e]?bt[e].won:0}function Ho(){let e=0;Object.keys(Game.GrandmaSynergies).forEach((t=>{wt(Game.GrandmaSynergies[t])&&(e+=1)})),!Do("Elder")&&e>=7&&No("Elder"),!Do("Veteran")&&e>=14&&No("Veteran");let t=0,o=1,n=1,i=1e5;Object.keys(Ct).forEach((e=>{t+=Ct[e].amount,i=Math.min(Ct[e].amount,i),Do("Mathematician")||Ct[e].amount<Math.min(128,2**(Game.ObjectsById.length-Game.Objects[e].id-1))&&(o=0),Do("Base 10")||Ct[e].amount<10*(Game.ObjectsById.length-Game.Objects[e].id)&&(n=0)})),i>=1&&No("One with everything"),1===o&&No("Mathematician"),1===n&&No("Base 10"),i>=100&&No("Centennial"),i>=150&&No("Centennial and a half"),i>=200&&No("Bicentennial"),i>=250&&No("Bicentennial and a half"),i>=300&&No("Tricentennial"),i>=350&&No("Tricentennial and a half"),i>=400&&No("Quadricentennial"),i>=450&&No("Quadricentennial and a half"),i>=500&&No("Quincentennial"),i>=550&&No("Quincentennial and a half"),i>=600&&No("Sexcentennial"),t>=100&&No("Builder"),t>=500&&No("Architect"),t>=1e3&&No("Engineer"),t>=2e3&&No("Lord of Constructs"),t>=4e3&&No("Grand design"),t>=8e3&&No("Ecumenopolis"),J>=20&&No("Enhancer"),J>=50&&No("Augmenter"),J>=100&&No("Upgrader"),J>=200&&No("Lord of Progress"),J>=300&&No("The full picture"),J>=400&&No("When there's nothing left to add"),t>=4e3&&J>=300&&No("Polymath"),t>=8e3&&J>=400&&No("Renaissance baker"),Ct.Cursor.amount+Ct.Grandma.amount>=777&&No("The elder scrolls");let a=!0;Object.keys(Lo).forEach((e=>{wt(Lo[e])||(a=!1)})),a&&No("Spooky cookies");let r=!0;if(Object.keys(Wo).forEach((e=>{wt(Wo[e])||(r=!1)})),r&&No("Let it snow"),wt("Fortune cookies")){const e=Game.Tiers.fortune.upgrades;let t=0;Object.keys(e).forEach((o=>{wt(e[o].name)&&(t+=1)})),t>=e.length&&No("O Fortuna")}}function Uo(e,t){Mo(),Ct[e].amount+=t;const o=Ct[e];"Cursor"===e?(o.amount>=1&&No("Click"),o.amount>=2&&No("Double-click"),o.amount>=50&&No("Mouse wheel"),o.amount>=100&&No("Of Mice and Men"),o.amount>=200&&No("The Digital"),o.amount>=300&&No("Extreme polydactyly"),o.amount>=400&&No("Dr. T"),o.amount>=500&&No("Thumbs, phalanges, metacarpals"),o.amount>=600&&No("With her finger and her thumb"),o.amount>=700&&No("Gotta hand it to you"),o.amount>=800&&No("The devil's workshop")):Object.keys(Game.Objects[o.name].tieredAchievs).forEach((e=>{o.amount>=Game.Tiers[Game.Objects[o.name].tieredAchievs[e].tier].achievUnlock&&No(Game.Objects[o.name].tieredAchievs[e].name)}));const n=ee;return Oo(),Ho(),n!==ee&&Oo(),de-Game.cookiesPs}function $o(e,t,o){let n=e,i=0;for(;n<o;)n+=.002*Math.max(.002,(n/Math.max(t,100))**.5),i+=1;return i/Game.fps}function Ro(){null!==l("CMDispTooltipWarningParent")&&(0===e.ToolWarnPos?(l("CMDispTooltipWarningParent").style.top="auto",l("CMDispTooltipWarningParent").style.margin="4px -4px",l("CMDispTooltipWarningParent").style.padding="3px 4px"):(l("CMDispTooltipWarningParent").style.right="auto",l("CMDispTooltipWarningParent").style.margin="4px",l("CMDispTooltipWarningParent").style.padding="4px 3px"))}function Io(){if(Mo(),"none"!==l("tooltipAnchor").style.display&&l("CMTooltipArea")){l("CMTooltipArea").innerHTML="";const t=function(){l("tooltip").firstChild.style.paddingBottom="4px";const e=document.createElement("div");return e.style.border="1px solid",e.style.padding="4px",e.style.margin="0px -4px",e.id="CMTooltipBorder",e.className="CMTextGray",e}();l("CMTooltipArea").appendChild(t),"b"===eo?function(){if(1===e.TooltipBuildUpgrade&&1===Game.buyMode){let t;if(Po(l("CMTooltipBorder")),1===Game.buyMode?Xt=t:t=Xt,1===Game.buyBulk?t=We:10===Game.buyBulk?t=Ae:100===Game.buyBulk&&(t=je),Jt=Game.Objects[to].bulkPrice,Zt=t[to].bonus,1===e.TooltipBuildUpgrade&&1===Game.buyMode){l("CMTooltipIncome").textContent=B(Zt,2);const o=Math.round(Zt/Game.cookiesPs*1e4);Number.isFinite(o)&&0!==o?l("CMTooltipIncome").textContent+=` (${o/100}% of income)`:l("CMTooltipIncome").textContent+=` (<0${e.ScaleSeparator?",":"."}01% of income)`,l("CMTooltipBorder").className=St+t[to].color,e.PPDisplayTime?l("CMTooltipPP").textContent=po(t[to].pp):l("CMTooltipPP").textContent=B(t[to].pp,2),l("CMTooltipPP").className=St+t[to].color;const n=uo((Jt-(Game.cookies+ho()))/mo());l("CMTooltipTime").textContent=n.text,"Done!"===n.text&&Game.cookies<t[to].price?l("CMTooltipTime").textContent=`${n.text} (with Wrink)`:l("CMTooltipTime").textContent=n.text,l("CMTooltipTime").className=St+n.color}l("CMTooltipProductionLeftHeader").style.display="none",l("CMTooltipTime").style.marginBottom="0px";for(const e of Object.keys(Game.Objects[to].productionAchievs))if(!Game.HasAchiev(Game.Objects[to].productionAchievs[e].achiev.name)){const t=Game.Objects[to].productionAchievs[e];l("CMTooltipTime").style.marginBottom="4px",l("CMTooltipProductionLeftHeader").style.display="",l("CMTooltipProductionLeft").className=`ProdAchievement${to}`,l("CMTooltipProductionLeft").textContent=B(t.pow-Ct[to].totalCookies,15),l("CMTooltipProductionLeft").style.color="white";break}const o=De[to];if(o.AmountNeeded<101){let t;l("CMTooltipProductionLeft").style.marginBottom="4px",l("CMTooltipNextAchievementHeader").style.display="",t=Game.cookiesPs?Math.max(o.price-(Game.cookies+ho()),0)/Game.cookiesPs+o.price/Uo(to,o.AmountNeeded):o.price/Uo(to,o.AmountNeeded),l("CMTooltipNextAchievement").textContent=`${B(o.AmountNeeded)} / ${B(o.price)} / `,l("CMTooltipNextAchievement").style.color="white";const n=document.createElement("span");e.PPDisplayTime?n.textContent=po(t):n.textContent=B(t),n.className=St+Bo({pp:t},o.price),l("CMTooltipNextAchievement").appendChild(n)}else l("CMTooltipNextAchievementHeader").style.display="none",l("CMTooltipProductionLeft").style.marginBottom="0px"}else l("CMTooltipArea").style.display="none"}():"u"===eo?function(){if(Po(l("CMTooltipBorder")),Zt=ze[Game.UpgradesInStore[to].name].bonus,Jt=Game.Upgrades[Game.UpgradesInStore[to].name].getPrice(),oo=ze[Game.UpgradesInStore[to].name].bonusMouse,1===e.TooltipBuildUpgrade){l("CMTooltipIncome").textContent=B(Zt,2);const t=Math.round(Zt/Game.cookiesPs*1e4);"0"!==l("CMTooltipIncome").textContent||"b"!==eo&&"u"!==eo?(Number.isFinite(t)&&0!==t?l("CMTooltipIncome").textContent+=` (${t/100}% of income)`:l("CMTooltipIncome").textContent+=` (<0${e.ScaleSeparator?",":"."}01% of income)`,l("CMTooltipBorder").className=St+ze[Game.UpgradesInStore[to].name].color,oo&&(l("CMTooltipCookiePerClick").textContent=B(oo),l("CMTooltipCookiePerClick").style.display="block",l("CMTooltipCookiePerClick").previousSibling.style.display="block"),0===Zt&&oo?(l("CMTooltipPP").textContent=`${B(Jt/oo)} Clicks`,l("CMTooltipPP").style.color="white"):(e.PPDisplayTime?l("CMTooltipPP").textContent=po(ze[Game.UpgradesInStore[to].name].pp):l("CMTooltipPP").textContent=B(ze[Game.UpgradesInStore[to].name].pp,2),l("CMTooltipPP").className=St+ze[Game.UpgradesInStore[to].name].color)):(l("Bonus IncomeTitle").style.display="none",l("CMTooltipIncome").style.display="none",l("Payback PeriodTitle").style.display="none",l("CMTooltipPP").style.display="none");const o=uo((Jt-(Game.cookies+ho()))/mo());if(l("CMTooltipTime").textContent=o.text,"Done!"===o.text&&Game.cookies<Game.UpgradesInStore[to].getPrice()?l("CMTooltipTime").textContent=`${o.text} (with Wrink)`:l("CMTooltipTime").textContent=o.text,l("CMTooltipTime").className=St+o.color,"Chocolate egg"===Game.UpgradesInStore[to].name){l("CMTooltipBorder").lastChild.style.marginBottom="4px",l("CMTooltipBorder").appendChild(So("Cookies to be gained (Currently/Max)"));const e=document.createElement("div");e.style.color="white",e.textContent=`${B(.05*Game.cookies)} / ${B(Le)}`,l("CMTooltipBorder").appendChild(e)}}else l("CMTooltipArea").style.display="none"}():"s"===eo?function(){if(1===e.TooltipLump){const e=l("CMTooltipBorder");e.appendChild(So("Current Sugar Lump"));const o=document.createElement("div");o.id="CMTooltipTime",e.appendChild(o);const n=0===(t=Game.lumpCurrentType)?{text:"Normal",color:Wt}:1===t?{text:"Bifurcated",color:Et}:2===t?{text:"Golden",color:Nt}:3===t?{text:"Meaty",color:Ot}:4===t?{text:"Caramelized",color:Lt}:{text:"Unknown Sugar Lump",color:Ft};o.textContent=n.text,o.className=St+n.color}else l("CMTooltipArea").style.display="none";var t}():"g"===eo?function(){const{minigame:t}=Game.Objects["Wizard tower"],o=t.getSpellCost(t.spellsById[to]);if(1===e.TooltipGrim&&o<=t.magicM){const e=l("CMTooltipBorder");e.appendChild(So("Time Left"));const n=document.createElement("div");n.id="CMTooltipTime",e.appendChild(n);const i=uo($o(t.magic,t.magicM,o));if(n.textContent=i.text,n.className=St+i.color,o<=t.magic){e.appendChild(So("Recover Time"));const n=document.createElement("div");n.id="CMTooltipRecover",e.appendChild(n);const i=uo($o(Math.max(0,t.magic-o),t.magicM,t.magic));n.textContent=i.text,n.className=St+i.color}if("0"===to){e.appendChild(So("Cookies to be gained/lost"));const t=document.createElement("div");t.id="x",e.appendChild(t);const o=document.createElement("span");o.style.color="#33FF00",o.textContent=B(Math.min(.15*(Game.cookies+ho()),60*Se*30),2),t.appendChild(o);const n=document.createElement("span");n.textContent=" / ",t.appendChild(n);const i=document.createElement("span");i.style.color="red",i.textContent=B(60*Se*15,2),t.appendChild(i)}l("CMTooltipArea").appendChild(e)}else l("CMTooltipArea").style.display="none"}():"p"===eo?function(){const{minigame:t}=Game.Objects.Farm;if(e.TooltipPlots&&0!==t.plot[to[1]][to[0]][0]){const e=t.plot[to[1]][to[0]][1]>t.plantsById[t.plot[to[1]][to[0]][0]-1].mature,o=t.plantsById[t.plot[to[1]][to[0]][0]-1].name;l("CMTooltipBorder").appendChild(So("Reward (Current / Maximum)"));const n=document.createElement("div");n.id="CMTooltipPlantReward",l("CMTooltipBorder").appendChild(n),"Bakeberry"===o?l("CMTooltipPlantReward").textContent=`${e?B(Math.min(.03*Game.cookies,60*Game.cookiesPs*30)):"0"} / ${B(60*Game.cookiesPs*30)}`:"Chocoroot"===o||"White chocoroot"===o?l("CMTooltipPlantReward").textContent=`${e?B(Math.min(.03*Game.cookies,60*Game.cookiesPs*3)):"0"} / ${B(60*Game.cookiesPs*3)}`:"Queenbeet"===o?l("CMTooltipPlantReward").textContent=`${e?B(Math.min(.04*Game.cookies,60*Game.cookiesPs*60)):"0"} / ${B(60*Game.cookiesPs*60)}`:"Duketater"===o?l("CMTooltipPlantReward").textContent=`${e?B(Math.min(.08*Game.cookies,60*Game.cookiesPs*120)):"0"} / ${B(60*Game.cookiesPs*120)}`:l("CMTooltipArea").style.display="none"}else l("CMTooltipArea").style.display="none"}():"ha"===eo?function(){const{minigame:t}=Game.Objects.Farm;if(e.TooltipLump){l("CMTooltipBorder").appendChild(So("Cookies gained from harvesting:"));let e=0,o=0;Game.keys[16]&&Game.keys[17]&&(o=1);for(let n=0;n<6;n++)for(let i=0;i<6;i++)if(t.plot[n][i][0]>=1){const a=t.plot[n][i],r=t.plantsById[a[0]-1],l=r.name;let s=!0;o&&r.immortal&&(s=!1),a[1]<r.matureBase&&(s=!1),s&&"Bakeberry"===l?e+=Math.min(.03*Game.cookies,60*Game.cookiesPs*30):s&&"Chocoroot"===l||"White chocoroot"===l?e+=Math.min(.03*Game.cookies,60*Game.cookiesPs*3):s&&"Queenbeet"===l?e+=Math.min(.04*Game.cookies,60*Game.cookiesPs*60):s&&"Duketater"===l&&(e+=Math.min(.08*Game.cookies,60*Game.cookiesPs*120))}l("CMTooltipBorder").appendChild(document.createTextNode(B(e)))}else l("CMTooltipArea").style.display="none"}():"wb"===eo?function(){l("tooltip").innerHTML="",l("tooltip").appendChild(So("Reward:"));const e=document.createElement("div");e.id="CMWrinklerReward","PopAll"===to?e.textContent=B(He):"PopFattest"===to&&(e.textContent=B($e[0])),l("tooltip").appendChild(e)}():("pag"===eo||"pas"===eo&&-1!==to[1])&&function(){if(1===e.TooltipPantheon){const t=l("CMTooltipBorder");let o;o="pas"===eo?to[1]:to,t.appendChild(So("Effect in position 1:"));const n=document.createElement("div");if(n.id="CMPantheonTooltipPosition1",0!==ft[o][0]){n.textContent=B(ft[o][0]);const t=Math.round(ft[o][0]/Game.cookiesPs*1e4);Number.isFinite(t)&&0!==t?n.textContent+=` (${t/100}% of income)`:n.textContent+=` (<0${e.ScaleSeparator?",":"."}01% of income)`}else n.textContent="No effect to CPS";t.appendChild(n),t.appendChild(So("Effect in position 2:"));const i=document.createElement("div");if(i.id="CMPantheonTooltipPosition2",0!==ft[o][1]){i.textContent=B(ft[o][1]);const t=Math.round(ft[o][1]/Game.cookiesPs*1e4);Number.isFinite(t)&&0!==t?i.textContent+=` (${t/100}% of income)`:i.textContent+=` (<0${e.ScaleSeparator?",":"."}01% of income)`}else i.textContent="No effect to CPS";t.appendChild(i),t.appendChild(So("Effect in position 3:"));const a=document.createElement("div");if(a.id="CMPantheonTooltipPosition2",0!==ft[o][2]){a.textContent=B(ft[o][2]);const t=Math.round(ft[o][2]/Game.cookiesPs*1e4);Number.isFinite(t)&&0!==t?a.textContent+=` (${t/100}% of income)`:a.textContent+=` (<0${e.ScaleSeparator?",":"."}01% of income)`}else a.textContent="No effect to CPS";t.appendChild(a),l("CMTooltipArea").appendChild(t)}else l("CMTooltipArea").style.display="none"}(),function(){if("b"===eo||"u"===eo){null===document.getElementById("CMDispTooltipWarningParent")&&(l("tooltipAnchor").appendChild(function(){const t=document.createElement("div");t.style.position="absolute",t.style.display="block",t.style.left="auto",t.style.bottom="auto",t.id="CMDispTooltipWarningParent";const o=function(e,t,o,n,i){const a=document.createElement("div");a.id=e,a.style.display="none",a.style.transition="opacity 0.1s ease-out",a.className="CMBorder"+t,a.style.padding="2px",a.style.background="#000 url(img/darkNoise.png)";const r=document.createElement("div");a.appendChild(r);const l=document.createElement("span");l.className=St+t,l.style.fontWeight="bold",l.textContent=o,r.appendChild(l),r.appendChild(document.createTextNode(n));const s=document.createElement("div");a.appendChild(s);const c=document.createElement("span");return c.id=i,s.appendChild(document.createTextNode("Deficit: ")),s.appendChild(c),a};return t.appendChild(o("CMDispTooltipWarnLucky",Ft,"Warning: ",'Purchase of this item will put you under the number of Cookies required for "Lucky!"',"CMDispTooltipWarnLuckyText")),t.firstChild.style.marginBottom="4px",t.appendChild(o("CMDispTooltipWarnLuckyFrenzy",Nt,"Warning: ",'Purchase of this item will put you under the number of Cookies required for "Lucky!" (Frenzy)',"CMDispTooltipWarnLuckyFrenzyText")),t.lastChild.style.marginBottom="4px",t.appendChild(o("CMDispTooltipWarnConjure",Lt,"Warning: ",'Purchase of this item will put you under the number of Cookies required for "Conjure Baked Goods"',"CMDispTooltipWarnConjureText")),t.lastChild.style.marginBottom="4px",t.appendChild(o("CMDispTooltipWarnConjureFrenzy",Lt,"Warning: ",'Purchase of this item will put you under the number of Cookies required for "Conjure Baked Goods" (Frenzy)',"CMDispTooltipWarnConjureFrenzyText")),t.lastChild.style.marginBottom="4px",t.appendChild(o("CMDispTooltipWarnEdifice",Lt,"Warning: ",'Purchase of this item will put you under the number of Cookies needed for "Spontaneous Edifice" to possibly give you your most expensive building"',"CMDispTooltipWarnEdificeText")),t.lastChild.style.marginBottom="4px",t.appendChild(o("CMDispTooltipWarnUser",Ft,"Warning: ",`Purchase of this item will put you under the number of Cookies equal to ${e.ToolWarnUser} seconds of CPS`,"CMDispTooltipWarnUserText")),t}()),Ro()),0===e.ToolWarnPos?l("CMDispTooltipWarningParent").style.right="0px":l("CMDispTooltipWarningParent").style.top=`${l("tooltip").offsetHeight}px`,l("CMDispTooltipWarningParent").style.width=l("tooltip").offsetWidth-6+"px";const t=Game.cookies+ho()-Jt,o=e.ToolWarnBon?Zt:0;let n=Ce;if(1===e.ToolWarnBon){let e=Zt;e/=Eo(),n+=60*e*15/.15}if(1===e.ToolWarnLucky&&t<n&&("b"!==eo||1===Game.buyMode)?(l("CMDispTooltipWarnLucky").style.display="",l("CMDispTooltipWarnLuckyText").textContent=`${B(n-t)} (${po((n-t)/(mo()+o))})`):l("CMDispTooltipWarnLucky").style.display="none",1===e.ToolWarnLuckyFrenzy){const e=7*n;t<e&&("b"!==eo||1===Game.buyMode)?(l("CMDispTooltipWarnLuckyFrenzy").style.display="",l("CMDispTooltipWarnLuckyFrenzyText").textContent=`${B(e-t)} (${po((e-t)/(mo()+o))})`):l("CMDispTooltipWarnLuckyFrenzy").style.display="none"}else l("CMDispTooltipWarnLuckyFrenzy").style.display="none";if(1===e.ToolWarnConjure){const e=2*n;t<e&&("b"!==eo||1===Game.buyMode)?(l("CMDispTooltipWarnConjure").style.display="",l("CMDispTooltipWarnConjureText").textContent=`${B(e-t)} (${po((e-t)/(mo()+o))})`):l("CMDispTooltipWarnConjure").style.display="none"}else l("CMDispTooltipWarnConjure").style.display="none";if(1===e.ToolWarnConjureFrenzy){const e=2*n*7;t<e&&("b"!==eo||1===Game.buyMode)?(l("CMDispTooltipWarnConjureFrenzy").style.display="",l("CMDispTooltipWarnConjureFrenzyText").textContent=`${B(e-t)} (${po((e-t)/(mo()+o))})`):l("CMDispTooltipWarnConjureFrenzy").style.display="none"}else l("CMDispTooltipWarnConjureFrenzy").style.display="none";1===e.ToolWarnEdifice&&Game.Objects["Wizard tower"].minigameLoaded&&Te&&t<Te&&("b"!==eo||1===Game.buyMode)?(l("CMDispTooltipWarnEdifice").style.display="",l("CMDispTooltipWarnEdificeText").textContent=`${B(Te-t)} (${po((Te-t)/(mo()+o))})`):l("CMDispTooltipWarnEdifice").style.display="none",e.ToolWarnUser>0&&t<e.ToolWarnUser*mo()&&("b"!==eo||1===Game.buyMode)?(l("CMDispTooltipWarnUser").style.display="",l("CMDispTooltipWarnUser").children[0].textContent=`Purchase of this item will put you under the number of Cookies equal to ${e.ToolWarnUser} seconds of CPS`,l("CMDispTooltipWarnUserText").textContent=`${B(e.ToolWarnUser*mo()-t)} (${po((e.ToolWarnUser*mo()-t)/(mo()+o))})`):l("CMDispTooltipWarnUser").style.display="none"}else null!==l("CMDispTooltipWarningParent")&&l("CMDispTooltipWarningParent").remove()}()}else null===l("CMTooltipArea")&&null!==l("CMDispTooltipWarningParent")&&l("CMDispTooltipWarningParent").remove()}function zo(){if(Game.prefs.autosave&&Game.drawT%10==0&&"stats"===Game.onMenu&&e.Stats){const e=document.getElementById("CMStatsAutosaveTimer");e&&(e.innerText=Game.sayTime(60*Game.fps-Game.T%(60*Game.fps),4))}if(so(),co(),function(){if(1===e.TimerBar){const t=l("CMTimerBar").offsetWidth-163,o=l("CMTimerBar").offsetWidth-133;let n=0;0!==Game.shimmerTypes.golden.spawned||Game.Has("Golden switch [off]")?l("CMTimerBarGC").style.display="none":(l("CMTimerBarGC").style.display="",l("CMTimerBarGCMinBar").style.width=`${Math.round(Math.max(0,Game.shimmerTypes.golden.minTime-Game.shimmerTypes.golden.time)*t/Game.shimmerTypes.golden.maxTime)}px`,e.TimerBarOverlay>=1?l("CMTimerBarGCMinBar").textContent=Math.ceil((Game.shimmerTypes.golden.minTime-Game.shimmerTypes.golden.time)/Game.fps):l("CMTimerBarGCMinBar").textContent="",Game.shimmerTypes.golden.minTime===Game.shimmerTypes.golden.maxTime?(l("CMTimerBarGCMinBar").style.borderTopRightRadius="10px",l("CMTimerBarGCMinBar").style.borderBottomRightRadius="10px"):(l("CMTimerBarGCMinBar").style.borderTopRightRadius="",l("CMTimerBarGCMinBar").style.borderBottomRightRadius=""),l("CMTimerBarGCBar").style.width=`${Math.round(Math.min(Game.shimmerTypes.golden.maxTime-Game.shimmerTypes.golden.minTime,Game.shimmerTypes.golden.maxTime-Game.shimmerTypes.golden.time)*t/Game.shimmerTypes.golden.maxTime)}px`,e.TimerBarOverlay>=1?l("CMTimerBarGCBar").textContent=Math.ceil(Math.min(Game.shimmerTypes.golden.maxTime-Game.shimmerTypes.golden.minTime,Game.shimmerTypes.golden.maxTime-Game.shimmerTypes.golden.time)/Game.fps):l("CMTimerBarGCBar").textContent="",l("CMTimerBarGCTime").textContent=Math.ceil((Game.shimmerTypes.golden.maxTime-Game.shimmerTypes.golden.time)/Game.fps),n+=1),"christmas"===Game.season&&0===Game.shimmerTypes.reindeer.spawned?(l("CMTimerBarRen").style.display="",l("CMTimerBarRenMinBar").style.width=`${Math.round(Math.max(0,Game.shimmerTypes.reindeer.minTime-Game.shimmerTypes.reindeer.time)*t/Game.shimmerTypes.reindeer.maxTime)}px`,e.TimerBarOverlay>=1?l("CMTimerBarRenMinBar").textContent=Math.ceil((Game.shimmerTypes.reindeer.minTime-Game.shimmerTypes.reindeer.time)/Game.fps):l("CMTimerBarRenMinBar").textContent="",l("CMTimerBarRenBar").style.width=`${Math.round(Math.min(Game.shimmerTypes.reindeer.maxTime-Game.shimmerTypes.reindeer.minTime,Game.shimmerTypes.reindeer.maxTime-Game.shimmerTypes.reindeer.time)*t/Game.shimmerTypes.reindeer.maxTime)}px`,e.TimerBarOverlay>=1?l("CMTimerBarRenBar").textContent=Math.ceil(Math.min(Game.shimmerTypes.reindeer.maxTime-Game.shimmerTypes.reindeer.minTime,Game.shimmerTypes.reindeer.maxTime-Game.shimmerTypes.reindeer.time)/Game.fps):l("CMTimerBarRenBar").textContent="",l("CMTimerBarRenTime").textContent=Math.ceil((Game.shimmerTypes.reindeer.maxTime-Game.shimmerTypes.reindeer.time)/Game.fps),n+=1):l("CMTimerBarRen").style.display="none";const i={};l("CMTimerBarBuffTimers").innerHTML="",Object.keys(Game.buffs).forEach((t=>{if(Game.buffs[t]){const a=go(Game.buffs[t].name,Game.buffs[t].name,[{id:`${Game.buffs[t].name}Bar`}]);a.style.display="";let r="";r=void 0!==Kt[Game.buffs[t].name]?Kt[Game.buffs[t].name]:Lt,a.lastChild.children[1].className=Pt+r,a.lastChild.children[1].style.color="black",2===e.TimerBarOverlay?a.lastChild.children[1].textContent=`${Math.round(Game.buffs[t].time/Game.buffs[t].maxTime*100)}%`:a.lastChild.children[1].textContent="",a.lastChild.children[1].style.width=`${Math.round(Game.buffs[t].time*(o-8*Math.ceil(Game.buffs[t].time/Game.fps).toString().length)/Game.buffs[t].maxTime)}px`,a.lastChild.children[2].textContent=Math.ceil(Game.buffs[t].time/Game.fps),n+=1,i[Game.buffs[t].name]=a}})),Object.keys(i).forEach((e=>{l("CMTimerBarBuffTimers").appendChild(i[e])})),0!==n&&(l("CMTimerBar").style.height=12*n+2+"px"),io!==n&&(io=n,bo())}}(),Co(),Io(),function(){if(1===e.TooltipWrink&&1===It){let e=!1;Object.keys(Game.wrinklers).forEach((t=>{const o=Game.wrinklers[t];if(o.phase>0&&o.selected){if(e=!0,0===Vt[t]||void 0===Vt[t]){const e=document.createElement("div"),o=document.createElement("div");o.style.minWidth="120px",o.style.marginBottom="4px";const n=document.createElement("div");n.style.textAlign="center",n.id="CMTooltipWrinkler",o.appendChild(n),e.appendChild(o),Game.tooltip.draw(this,escape(e.innerHTML)),zt=t,Vt[t]=1}}else Vt[t]=0})),e||Game.tooltip.hide()}}(),function(){if(1===e.TooltipWrink&&null!==l("CMTooltipWrinkler")){let{sucked:e}=Game.wrinklers[zt],t=1.1;if(Game.Has("Sacrilegious corruption")&&(t*=1.05),1===Game.wrinklers[zt].type&&(t*=3),e*=t,Game.Has("Wrinklerspawn")&&(e*=1.05),Ct.Temple.minigameLoaded){const t=Game.hasGod("scorn");1===t?e*=1.15:2===t?e*=1.1:3===t&&(e*=1.05)}l("CMTooltipWrinkler").textContent=B(e)}}(),e.UpStats&&"stats"===Game.onMenu&&(Game.drawT-1)%(5*Game.fps)!=0&&(Game.drawT-1)%Game.fps==0&&Game.UpdateMenu(),t(),e.Scale){let e=l("cookies").innerHTML.replace(/.*(?=<br>)/i,B(Game.cookies));Game.prefs.monospace&&(e=`<span class="monospace">${e}</span>`),l("cookies").innerHTML=e}}function Vo(){E.addLatest(Game.computedMouseCps)}class _o{constructor(e){this.maxLength=e,this.queue=[]}addLatest(e){this.queue.push(e)>this.maxLength&&this.queue.shift()}calcAverage(e){let t=e;t>this.maxLength&&(t=this.maxLength),t>this.queue.length&&(t=this.queue.length);let o=0;for(let e=this.queue.length-1;e>=0&&e>this.queue.length-1-t;e--)o+=this.queue[e];return 0===o?0:o/t}calcSum(e){let t=e;return t>this.maxLength&&(t=this.maxLength),t>this.queue.length&&(t=this.queue.length),0===t?0:this.queue.slice(-t).reduce(((e,t)=>e+t),0)}}function Qo(){const t=Math.floor(Date.now()/1e3);if(Game.T/Game.fps%1==0){let o=Game.cookies+gt;Game.cpsSucked>0&&(o+=He),lt=Math.max(Game.cookiesEarned,o),o*=.05;const n=t-j,i=Math.max(0,Game.cookies-D)/n,a=Math.max(0,He-H)/n,r=Math.max(0,$e[0]-U)/n,l=Math.max(0,o-Le)/n,s=(Game.cookieClicks-$)/n;for(let e=0;e<n;e++)N.addLatest(i),O.addLatest(a),F.addLatest(r),L.addLatest(l),W.addLatest(s);j=t,D=Game.cookies,H=He,U=$e[0],Le=o,$=Game.cookieClicks;const c=qt[e.AvgCPSHist];R=N.calcAverage(c),I=O.calcAverage(c),z=F.calcAverage(c),V=L.calcAverage(c),_=R,1===e.CalcWrink&&(_+=I),2===e.CalcWrink&&(_+=z);const d=Game.HasUnlocked("Chocolate egg")&&!Game.Has("Chocolate egg");st=d||0===e.CalcWrink?R+I+(d?V:0):_,Ve=W.calcAverage(Yt[e.AvgClicksHist])}}function Xo(t,o){Object.keys(t).forEach((n=>{if(e.PPRigidelMode&&1===o)t[n].color=Wt;else{t[n].color=Bo(t[n],Game.Objects[n].getSumPrice(o));for(let o=0;o<e.PPExcludeTop;o++)t[n].pp===mt[o][0]&&(t[n].color=Wt)}}))}function qo(t,o){Object.keys(t).forEach((n=>{const i=Game.Objects[n].getSumPrice(o);Game.cookiesPs?t[n].pp=Math.max(i-(Game.cookies+ho()),0)/Game.cookiesPs+i/t[n].bonus:t[n].pp=i/t[n].bonus,e.PPRigidelMode&&1===o||mt.push([t[n].pp,o])}))}function Yo(){!function(){ut=1/0,mt=[],void 0===e.PPExcludeTop&&(e.PPExcludeTop=0),qo(We,1),qo(Ae,10),qo(je,100),mt.sort(((e,t)=>e[0]-t[0]));let t=0;if(e.PPOnlyConsiderBuyable)for(;mt[t][1]>Game.cookies&&(t+=1,mt.length!==t+1););ut=mt[e.PPExcludeTop][t],Xo(We,1),Xo(Ae,10),Xo(je,100)}(),Object.keys(ze).forEach((e=>{Game.cookiesPs?ze[e].pp=Math.max(Game.Upgrades[e].getPrice()-(Game.cookies+ho()),0)/Game.cookiesPs+Game.Upgrades[e].getPrice()/ze[e].bonus:ze[e].pp=Game.Upgrades[e].getPrice()/ze[e].bonus,Number.isNaN(ze[e].pp)&&(ze[e].pp=1/0),ze[e].color=Bo(ze[e],Game.Upgrades[e].getPrice())}))}function Ko(e,t,o,n,i){let a=o,r=0;for(let o=0;o<i;o+=1){let o=t*Game.priceIncrease**Math.max(0,a-n);o=Game.modifyBuildingPrice(e,o),o=Math.ceil(o),r+=o,a+=1}return r}function Jo(e){const t={};return Object.keys(Game.Objects).forEach((o=>{t[o]={},t[o].bonus=Uo(o,e),1!==e&&(q=1)})),t}function Zo(){Object.keys(Game.Objects).forEach((e=>{We[e].price=Ko(Game.Objects[e],Game.Objects[e].basePrice,Game.Objects[e].amount,Game.Objects[e].free,1),Ae[e].price=Ko(Game.Objects[e],Game.Objects[e].basePrice,Game.Objects[e].amount,Game.Objects[e].free,10),je[e].price=Ko(Game.Objects[e],Game.Objects[e].basePrice,Game.Objects[e].amount,Game.Objects[e].free,100),De[e].price=Ko(Game.Objects[e],Game.Objects[e].basePrice,Game.Objects[e].amount,Game.Objects[e].free,De[e].AmountNeeded)}))}function en(){We=Jo(1),Ae=Jo(10),je=Jo(100),ze=[],Object.keys(Game.Upgrades).forEach((t=>{const o=function(e){if("toggle"===Game.Upgrades[e].pool||0===Game.Upgrades[e].bought&&Game.Upgrades[e].unlocked&&"prestige"!==Game.Upgrades[e].pool){Mo(),"Shimmering veil [on]"===yt[e].name?yt["Shimmering veil [off]"].bought=0:"Golden switch [on]"===yt[e].name?yt["Golden switch [off]"].bought=0:yt[e].bought=(yt[e].bought+1)%2,Game.CountsAsUpgradeOwned(Game.Upgrades[e].pool)&&(J+=1),"Elder Pledge"===e?(Z+=1,Z>0&&No("Elder nap"),Z>=5&&No("Elder slumber")):"Elder Covenant"===e?No("Elder calm"):"Prism heart biscuits"===e?No("Lovely cookies"):"Heavenly key"===e&&No("Wholesome");const t=ee;Oo(),Ho(),t!==ee&&Oo();const o=function(){let e=0;wt("Thousand fingers")&&(e+=.1),wt("Million fingers")&&(e*=5),wt("Billion fingers")&&(e*=10),wt("Trillion fingers")&&(e*=20),wt("Quadrillion fingers")&&(e*=20),wt("Quintillion fingers")&&(e*=20),wt("Sextillion fingers")&&(e*=20),wt("Septillion fingers")&&(e*=20),wt("Octillion fingers")&&(e*=20),wt("Nonillion fingers")&&(e*=20);let t=0;Object.keys(Ct).forEach((e=>{t+=Ct[e].amount})),t-=Ct.Cursor.amount,e*=t,wt("Plastic mouse")&&(e+=.01*de),wt("Iron mouse")&&(e+=.01*de),wt("Titanium mouse")&&(e+=.01*de),wt("Adamantium mouse")&&(e+=.01*de),wt("Unobtainium mouse")&&(e+=.01*de),wt("Eludium mouse")&&(e+=.01*de),wt("Wishalloy mouse")&&(e+=.01*de),wt("Fantasteel mouse")&&(e+=.01*de),wt("Nevercrack mouse")&&(e+=.01*de),wt("Armythril mouse")&&(e+=.01*de),wt("Technobsidian mouse")&&(e+=.01*de),wt("Plasmarble mouse")&&(e+=.01*de),wt("Miraculite mouse")&&(e+=.01*de),wt("Fortune #104")&&(e+=.01*de);let o=1;if(wt("Santa's helpers")&&(o*=1.1),wt("Cookie egg")&&(o*=1.1),wt("Halo gloves")&&(o*=1.1),wt("Dragon claw")&&(o*=1.03),wt("Aura gloves")&&(o*=1+.05*Math.min(Game.Objects.Cursor.level,wt("Luminous gloves")?20:10)),o*=vt("click"),Ct.Temple.minigameLoaded&&xt){const e=xt("labor");1===e?o*=1.15:2===e?o*=1.1:3===e&&(o*=1.05)}Object.keys(Game.buffs).forEach((e=>{void 0!==Game.buffs[e].multClick&&(o*=Game.buffs[e].multClick)})),o*=1+.05*kt("Dragon Cursor");let n=o*Game.ComputeCps(1,wt("Reinforced index finger")+wt("Carpal tunnel prevention cream")+wt("Ambidextrous"),e);return n=Game.runModHookOnValue("cookiesPerClick",n),Game.hasBuff("Cursed finger")&&(n=Game.buffs["Cursed finger"].power),n}()-Game.computedMouseCps;return o?[de-Game.cookiesPs,o]:[de-Game.cookiesPs]}return[]}(t);"Elder Pledge"===t?(ze[t]={bonus:Game.cookiesPs-R},1===e.CalcWrink?ze[t].bonus-=I:2===e.CalcWrink&&(ze[t].bonus-=z),Number.isFinite(ze[t].bonus)||(ze[t].bonus=0)):(ze[t]={},o[0]&&(ze[t].bonus=o[0]),o[1]&&(ze[t].bonusMouse=o[1]))}))}function tn(e,t,o){let n=0,i=0,a=0,r=0,l=1+Math.max(0,Math.ceil(Math.log(Game.cookies)/Math.LN10)-10);for(;a<t*o;)i=Math.max(e,Math.min(Math.floor(1/9*10**l*e*o),t*o)),a=Math.max(e,Math.min(Math.floor(1/9*10**(l+1)*e*o),t*o)),r=Math.floor(1/9*10**(l+1)*e*o),n+=i,l+=1;return[n,i,r]}function on(){let e=60*Se*60*6*Pe;const t=Eo();t>0?e/=t:e=0,Ke=tn(7,e,Be),qe=2*Ke[1]/Be,Ye=Ke[2]/60/60/6/Pe,et=tn(6,e,Ee),Je=2*et[1]/Ee,Ze=et[2]/60/60/6/Pe,nt=tn(7,7*e,Be),tt=2*nt[1]/Be,ot=nt[2]/60/60/6/Pe,rt=tn(6,7*e,Ee),it=2*rt[1]/Ee,at=rt[2]/60/60/6/Pe}function nn(){const e=Math.floor(Date.now()/1e3);if(Game.T/Game.fps%1==0){const t=Game.HowMuchPrestige(Game.cookiesReset),o=Math.floor(Game.HowMuchPrestige(Game.cookiesReset+Game.cookiesEarned))-Math.floor(t),n=e-Q,i=Math.max(0,o-X)/n;for(let e=0;e<n;e++)A.addLatest(i);Q=e,X=o,Ne=A.calcAverage(5)}}function an(){_e="",Xe="",Qe="";const e=[];Object.keys(Game.Upgrades).forEach((t=>{e.push(Game.Upgrades[t])})),e.sort((function(e,t){return e.order>t.order?1:e.order<t.order?-1:0})),Object.keys(e).forEach((t=>{const o=e[t];if(0===o.bought){let e="";e+=function(e){let t="crate upgrade missing";"prestige"===e.pool&&(t+=" heavenly");let o=0;Game.prefs.crates||(o=1),o&&(t+=" noFrame");let{icon:n}=e;e.iconFunction&&(n=e.iconFunction());const i=`function() {return Game.crateTooltip(Game.UpgradesById[${e.id}], 'stats');}`;return`<div class="${t}"\n\t${Game.getDynamicTooltip(i,"top",!0)}\n\tstyle = "${n[2]?`background-image: url(${n[2]});`:""}background-position:${48*-n[0]}px ${48*-n[1]}px;">\n\t</div>`}(o),"prestige"===o.pool?Qe+=e:"cookie"===o.pool?Xe+=e:"toggle"!==o.pool&&"unused"!==o.pool&&"debug"!==o.pool&&(_e+=e)}}))}function rn(){if("christmas"===Game.season){let e=60*Game.cookiesPs;Game.hasBuff("Elder frenzy")&&(e*=.5),Game.hasBuff("Frenzy")&&(e*=.75),Fe=Math.max(25,e),Game.Has("Ho ho ho-flavored frosting")&&(Fe*=2)}}function ln(){Ce=900*Se/.15,Ce*=Pe;const e=Eo();e>0?Ce/=e:Ce=0,ye=Be*(.15*Ce)+13,be=Ee*(.15*Ce)+13,ke=7*Ce,Ge=Be*(.15*ke)+13,ve=Ee*(.15*ke)+13,we=2*Ce,xe=.15*we,Te=0;let t=0,o=0;Object.keys(Game.Objects).forEach((e=>{Game.Objects[e].amount>t&&(t=Game.Objects[e].amount),Game.Objects[e].amount>0&&(o+=1)})),Object.keys(Game.Objects).forEach((e=>{(Game.Objects[e].amount<t||1===o)&&Game.Objects[e].amount<400&&2*Game.Objects[e].price>Te&&(Te=2*Game.Objects[e].price,Me=e)}))}function sn(){let e=1,t=1,o=1;wt("Green yeast digestives")&&(o*=1.01),wt("Dragon fang")&&(o*=1.03),e*=1+.1*Game.auraMult("Ancestral Metamorphosis"),e*=Game.eff("goldenCookieGain"),t*=1+.1*Game.auraMult("Unholy Dominion"),t*=Game.eff("wrathCookieGain"),Be=o*e,Ee=o*t,Pe=1,0===Game.shimmerTypes.golden.n&&(Pe*=1+1.23*Game.auraMult("Dragon's Fortune"))}function cn(){const e={};Object.keys(Game.Objects).forEach((t=>{if(0!==Object.keys(De).length&&De[t].TotalNeeded>Game.Objects[t].amount)e[t]={AmountNeeded:De[t].TotalNeeded-Game.Objects[t].amount,TotalNeeded:De[t].TotalNeeded};else{const o=function(e){const t=Game.AchievementsOwned;for(let o=0;o<101;o++)if(Uo(e,o),ee>t)return o;return 101}(t);e[t]={AmountNeeded:o,TotalNeeded:Game.Objects[t].amount+o}}})),De=e}function dn(){He=0,Ue=0,$e=[0,null];for(let e=0;e<Game.wrinklers.length;e++){let{sucked:t}=Game.wrinklers[e],o=1.1;if(Game.Has("Sacrilegious corruption")&&(o*=1.05),1===Game.wrinklers[e].type&&(o*=3),t*=o,Game.Has("Wrinklerspawn")&&(t*=1.05),Ct.Temple.minigameLoaded){const e=Game.hasGod("scorn");1===e?t*=1.15:2===e?t*=1.1:3===e&&(t*=1.05)}He+=t,0===Game.wrinklers[e].type&&(Ue+=t,t>$e[0]&&($e=[t,e]))}}function pn(){Go(),dn(),ln(),sn(),on(),an(),rn(),N=new _o(qt[qt.length-1]),O=new _o(qt[qt.length-1]),F=new _o(qt[qt.length-1]),L=new _o(qt[qt.length-1]),W=new _o(Yt[Yt.length-1]),A=new _o(5),E=new _o(20*Yt[Yt.length-1]),nn(),cn(),Qo(),en(),Zo(),Yo()}const un={CPSMode:1,AvgCPSHist:3,AvgClicksHist:0,CalcWrink:0,Scale:2,ScaleDecimals:2,ScaleSeparator:0,ScaleCutoff:999999,TimeFormat:0,DetailedTime:1,PPDisplayTime:0,BuildColour:1,PPOnlyConsiderBuyable:0,PPExcludeTop:0,PPRigidelMode:0,PPSecondsLowerLimit:0,ColourBlue:"#4bb8f0",ColourGreen:"#00ff00",ColourYellow:"#ffff00",ColourOrange:"#ff7f00",ColourRed:"#ff0000",ColourPurple:"#ff00ff",ColourGray:"#b3b3b3",ColourPink:"#ff1493",ColourBrown:"#8b4513",BotBar:1,TimerBar:1,TimerBarPos:0,TimerBarOverlay:2,UpBarColour:1,UpgradeBarFixedPos:1,SortBuildings:0,SortUpgrades:0,DragonAuraInfo:1,GrimoireBar:1,GCTimer:1,Favicon:1,WrinklerButtons:1,HideSectionsButtons:0,TooltipBuildUpgrade:1,TooltipAmor:0,ToolWarnLucky:1,ToolWarnLuckyFrenzy:1,ToolWarnConjure:1,ToolWarnConjureFrenzy:1,ToolWarnEdifice:1,ToolWarnUser:0,ToolWarnBon:1,ToolWarnPos:1,TooltipGrim:1,TooltipWrink:1,TooltipLump:1,TooltipPlots:1,TooltipPantheon:1,TooltipAscendButton:1,Stats:1,MissingUpgrades:1,MissingAchievements:0,UpStats:1,HeavenlyChipsTarget:1,ShowMissedGC:1,Title:1,GeneralSound:1,GCNotification:0,GCFlash:1,ColourGCFlash:"#ffffff",GCSound:1,GCVolume:100,GCSoundURL:"https://freesound.org/data/previews/66/66717_931655-lq.mp3",FortuneNotification:0,FortuneFlash:1,ColourFortuneFlash:"#ffffff",FortuneSound:1,FortuneVolume:100,FortuneSoundURL:"https://freesound.org/data/previews/174/174027_3242494-lq.mp3",SeaNotification:0,SeaFlash:1,ColourSeaFlash:"#ffffff",SeaSound:1,SeaVolume:100,SeaSoundURL:"https://www.freesound.org/data/previews/121/121099_2193266-lq.mp3",GardFlash:1,ColourGardFlash:"#ffffff",GardSound:1,GardVolume:100,GardSoundURL:"https://freesound.org/data/previews/103/103046_861714-lq.mp3",MagicNotification:0,MagicFlash:1,ColourMagicFlash:"#ffffff",MagicSound:1,MagicVolume:100,MagicSoundURL:"https://freesound.org/data/previews/221/221683_1015240-lq.mp3",WrinklerNotification:0,WrinklerFlash:1,ColourWrinklerFlash:"#ffffff",WrinklerSound:1,WrinklerVolume:100,WrinklerSoundURL:"https://freesound.org/data/previews/124/124186_8043-lq.mp3",WrinklerMaxNotification:0,WrinklerMaxFlash:1,ColourWrinklerMaxFlash:"#ffffff",WrinklerMaxSound:1,WrinklerMaxVolume:100,WrinklerMaxSoundURL:"https://freesound.org/data/previews/152/152743_15663-lq.mp3",BulkBuyBlock:0,FavouriteSettings:1,Header:{Favourite:1,Calculation:1,Notation:1,Colours:1,BarsDisplay:1,Tooltip:1,Statistics:1,Notification:1,NotificationGeneral:1,NotificationGC:1,NotificationFC:1,NotificationSea:1,NotificationGard:1,NotificationMagi:1,NotificationWrink:1,NotificationWrinkMax:1,Miscellaneous:1,Lucky:1,Chain:1,Spells:1,Garden:1,Prestige:1,Wrink:1,Sea:1,Misc:1,InfoTab:1}},mn=function(e){if(1===e){const e=function(){try{Notification.requestPermission().then()}catch(e){return!1}return!0};"Notification"in window?e()?Notification.requestPermission().then():Notification.requestPermission():console.log("This browser does not support notifications.")}};function hn(){1===e.BotBar?(l("CMBotBar").style.display="",Co()):l("CMBotBar").style.display="none",bo()}function gn(){BeautifyAll(),Game.RefreshStore(),Game.RebuildUpgrades(),Co(),so(),co()}function fn(){1===e.Favicon&&y>0?ct.wrath?l("CMFavicon").href="https://aktanusa.github.io/CookieMonster/favicon/wrathCookie.ico":l("CMFavicon").href="https://aktanusa.github.io/CookieMonster/favicon/goldenCookie.ico":l("CMFavicon").href="https://orteil.dashnet.org/cookieclicker/favicon.ico"}class Cn{constructor(e,t){this.type=e,this.group=t}}class yn extends Cn{constructor(e,t,o){super(e,t),this.desc=o}}class bn extends Cn{constructor(e,t,o,n,i,a){super(e,t),this.label=o,this.desc=n,this.min=i,this.max=a}}class kn extends Cn{constructor(e,t,o,n,i,a){super(e,t),this.label=o,this.desc=n,this.toggle=i,void 0!==a&&(this.func=a)}}class Gn extends Cn{constructor(e,t,o,n){super(e,t),this.label=o,this.desc=n;for(let e=0;e<101;e++)this.label[e]=`${e}%`}}const vn={CPSMode:new kn("bool","Calculation",["Current cookies per second","Average cookies per second"],"Calculate times using current cookies per second or average cookies per second",!1),AvgCPSHist:new kn("bool","Calculation",["Average CPS in past 10s","Average CPS in past 15s","Average CPS in past 30s","Average CPS in past 1m","Average CPS in past 5m","Average CPS in past 10m","Average CPS in past 15m","Average CPS in past 30m"],"How much time average Cookies Per Second should consider",!1),AvgClicksHist:new kn("bool","Calculation",["Average clicks in past 1s","Average clicks in past 5s","Average clicks in past 10s","Average clicks in past 15s","Average clicks in past 30s"],"How much time average Cookie Clicks should consider",!1),CalcWrink:new kn("bool","Calculation",["Calculate with wrinklers OFF","Calculate with wrinklers ON","Calculate with single fattest wrinkler ON"],"Calculate times and average Cookies Per Second with (only the single non-shiny fattest) wrinklers",!0,(function(){se=!0})),Scale:new kn("bool","Notation",["Game's setting scale","Metric","Short scale","Short scale (Abbreviated)","Scientific notation","Engineering notation"],"Change how long numbers are formatted",!1,(function(){gn()})),ScaleDecimals:new kn("bool","Notation",["1 decimals","2 decimals","3 decimals"],'Set the number of decimals used when applicable. This only works with Cookie Monster scales and not with "Game\'s Setting Scale"',!1,(function(){gn()})),ScaleSeparator:new kn("bool","Notation",[". for decimals (standard)",". for thousands"],"Set the separator used for decimals and thousands",!1,(function(){gn()})),ScaleCutoff:new bn("numscale","Notation","Notation cut-off point: ","The number from which Cookie Monster will start formatting numbers based on chosen scale. Standard is 999,999. Setting this above 999,999,999 might break certain notations",1,999999999),TimeFormat:new kn("bool","Notation",["Time XXd, XXh, XXm, XXs","Time XX:XX:XX:XX:XX"],"Change the time format",!1),DetailedTime:new kn("bool","Notation",["Detailed time OFF","Detailed time ON"],"Change how time is displayed in certain statistics and tooltips",!0,(function(){1===e.DetailedTime?Game.sayTime=Rt:Game.sayTime=M.sayTime})),PPDisplayTime:new kn("bool","Notation",["PP as value (standard)","PP as time unit"],"Display PP as calculated value or as approximate time unit. Note that PP does not translate directly into a time unit and this is therefore only an approximation.",!1),BuildColour:new kn("bool","Colours",["Building colours OFF","Building colours ON"],"Colour code buildings",!0,(function(){so()})),PPOnlyConsiderBuyable:new kn("bool","Colours",["Don't ignore non-buyable","Ignore non-buyable"],"Makes Cookie Monster label buildings and upgrades you can't buy right now red, useful in those situations where you just want to spend your full bank 'most optimally'",!0),PPExcludeTop:new kn("bool","Colours",["Don't ignore any","Ignore 1st best","Ignore 1st and 2nd best","Ignore 1st, 2nd and 3rd best"],"Makes Cookie Monster ignore the 1st, 2nd or 3rd best buildings in labeling and colouring PP values",!0),PPRigidelMode:new kn("bool","Colours",["Rigidel mode OFF","Rigidel mode ON"],'Makes Cookie Monster ignore all "buy 1" options when colouring PP in order to stay at a total building count ending in 10 for pantheon god Rigidel',!0),PPSecondsLowerLimit:new bn("numscale","Colours","Lower limit for PP (in seconds): ",'If a building or upgrade costs less than the specified seconds of CPS it will also be considered optimal and label it as such ("PP is less than xx seconds of CPS"); setting to 0 ignores this option',0,1/0),ColourBlue:new yn("colour","Colours","Standard colour is blue. Used to show upgrades better than best PP building, for Click Frenzy bar, and for various labels"),ColourGreen:new yn("colour","Colours","Standard colour is green. Used to show best PP building, for Blood Frenzy bar, and for various labels"),ColourYellow:new yn("colour","Colours","Standard colour is yellow. Used to show buildings within the top 10 of PP, for Frenzy bar, and for various labels"),ColourOrange:new yn("colour","Colours","Standard colour is orange. Used to show buildings within the top 20 of PP, for Next Reindeer bar, and for various labels"),ColourRed:new yn("colour","Colours","Standard colour is Red. Used to show buildings within the top 30 of PP, for Clot bar, and for various labels"),ColourPurple:new yn("colour","Colours","Standard colour is purple. Used to show buildings outside of the top 30 of PP, for Next Cookie bar, and for various labels"),ColourGray:new yn("colour","Colours","Standard colour is gray. Used to show negative or infinity PP, and for Next Cookie/Next Reindeer bar"),ColourPink:new yn("colour","Colours","Standard colour is pink. Used for Dragonflight bar"),ColourBrown:new yn("colour","Colours","Standard colour is brown. Used for Dragon Harvest bar"),BotBar:new kn("bool","BarsDisplay",["Bottom bar OFF","Bottom bar ON"],"Building information",!0,(function(){hn()})),TimerBar:new kn("bool","BarsDisplay",["Timer bar OFF","Timer bar ON"],"Bar with timers for golden cookie, season popup, Frenzy (Normal, Clot, Elder), Click Frenzy",!0,(function(){ko()})),TimerBarPos:new kn("bool","BarsDisplay",["Timer bar position (top left)","Timer bar position (bottom)"],"Placement of the timer bar",!1,(function(){0===e.TimerBarPos?(l("CMTimerBar").style.width="30%",l("CMTimerBar").style.bottom="",l("game").insertBefore(l("CMTimerBar"),l("sectionLeft"))):(l("CMTimerBar").style.width="100%",l("CMTimerBar").style.bottom="0px",l("wrapper").appendChild(l("CMTimerBar"))),bo()})),TimerBarOverlay:new kn("bool","BarsDisplay",["Timer bar overlay OFF","Timer bar overlay only seconds","Timer bar overlay full"],"Overlay on timers displaying seconds and/or percentage left",!0),UpBarColour:new kn("bool","BarsDisplay",["Upgrade colours/bar OFF","Upgrade colours with bar ON","Upgrade colours without bar ON"],"Colour code upgrades and optionally add a counter bar",!1,(function(){1===e.UpBarColor?(l("CMUpgradeBar").style.display="",co()):2===e.UpBarColor?(l("CMUpgradeBar").style.display="none",co()):(l("CMUpgradeBar").style.display="none",Game.RebuildUpgrades())})),UpgradeBarFixedPos:new kn("bool","BarsDisplay",["Upgrade bar fixed position OFF","Upgrade bar fixed position ON"],"Lock the upgrade bar at top of the screen to prevent it from moving ofscreen when scrolling",!0,(function(){1===e.UpgradeBarFixedPos?(l("CMUpgradeBar").style.position="sticky",l("CMUpgradeBar").style.top="0px"):l("CMUpgradeBar").style.position=""})),SortBuildings:new kn("bool","BarsDisplay",["Sort buildings: default","Sort buildings: PP of x1 purchase","Sort buildings: PP of selected bulk mode"],"Sort the display of buildings in either default order or by PP",!1,(function(){so()})),SortUpgrades:new kn("bool","BarsDisplay",["Sort upgrades: default","Sort upgrades: PP"],"Sort the display of upgrades in either default order or by PP",!1,(function(){co()})),DragonAuraInfo:new kn("bool","BarsDisplay",["Extra dragon aura info OFF","Extra dragon aura info ON"],"Shows information about changes in CPS and costs in the dragon aura interface.",!0),GrimoireBar:new kn("bool","BarsDisplay",["Grimoire magic meter timer OFF","Grimoire magic meter timer ON"],"A timer on how long before the Grimoire magic meter is full",!0),GCTimer:new kn("bool","BarsDisplay",["Golden cookie timer OFF","Golden cookie timer ON"],"A timer on the golden cookie when it has been spawned",!0,(function(){1===e.GCTimer?Object.keys(ao).forEach((e=>{ao[e].style.display="block",ao[e].style.left=ht[e].l.style.left,ao[e].style.top=ht[e].l.style.top})):Object.keys(ao).forEach((e=>ao[e].style.display="none"))})),Favicon:new kn("bool","BarsDisplay",["Favicon OFF","Favicon ON"],"Update favicon with golden/wrath cookie",!0,(function(){fn()})),WrinklerButtons:new kn("bool","BarsDisplay",["Extra wrinkler buttons OFF","Extra wrinkler buttons ON"],"Show buttons for popping wrinklers at bottom of cookie section",!0,(function(){t()})),HideSectionsButtons:new kn("bool","BarsDisplay",["Hide buildings/upgrades button OFF","Hide buildings/upgrades button ON"],"Show buttons for hiding and showing the buildings and upgrades sections in the right column",!0,(function(){e.HideSectionsButtons?l("CMSectionHidButtons").style.display="":l("CMSectionHidButtons").style.display="none"})),TooltipBuildUpgrade:new kn("bool","Tooltip",["Building/upgrade tooltip information OFF","Building/upgrade tooltip information ON"],"Extra information in building/upgrade tooltips",!0),TooltipAmor:new kn("bool","Tooltip",["Buildings tooltip amortization information OFF","Buildings tooltip amortization information ON"],"Add amortization information to buildings tooltip",!0),ToolWarnLucky:new kn("bool","Tooltip",["Tooltip lucky warning OFF","Tooltip lucky warning ON"],'A warning when buying if it will put the bank under the amount needed for max "Lucky!" rewards',!0),ToolWarnLuckyFrenzy:new kn("bool","Tooltip",["Tooltip lucky frenzy warning OFF","Tooltip lucky frenzy warning ON"],'A warning when buying if it will put the bank under the amount needed for max "Lucky!" (Frenzy) rewards',!0),ToolWarnConjure:new kn("bool","Tooltip",["Tooltip conjure warning OFF","Tooltip conjure warning ON"],'A warning when buying if it will put the bank under the amount needed for max "Conjure Baked Goods" rewards',!0),ToolWarnConjureFrenzy:new kn("bool","Tooltip",["Tooltip conjure frenzy warning OFF","Tooltip conjure frenzy warning ON"],'A warning when buying if it will put the bank under the amount needed for max "Conjure Baked Goods" rewards with Frenzy active',!0),ToolWarnEdifice:new kn("bool","Tooltip",["Tooltip edifice warning OFF","Tooltip edifice warning ON"],'A warning when buying if it will put the bank under the amount needed for "Spontaneous Edifice" to possibly give you your most expensive building',!0),ToolWarnUser:new bn("numscale","Tooltip","Tooltip warning at x times CPS: ","Use this to show a customized warning if buying it will put the bank under the amount equal to value times cps; setting to 0 disables the function altogether",0,1/0),ToolWarnBon:new kn("bool","Tooltip",["Calculate tooltip warning with bonus CPS OFF","Calculate tooltip warning with bonus CPS ON"],"Calculate the warning with or without the bonus CPS you get from buying",!0),ToolWarnPos:new kn("bool","Tooltip",["Tooltip warning position (left)","Tooltip warning position (bottom)"],"Placement of the warning boxes",!1,(function(){Ro()})),TooltipGrim:new kn("bool","Tooltip",["Grimoire tooltip information OFF","Grimoire tooltip information ON"],"Extra information in tooltip for grimoire",!0),TooltipWrink:new kn("bool","Tooltip",["Wrinkler tooltip OFF","Wrinkler tooltip ON"],"Shows the amount of cookies a wrinkler will give when popping it",!0),TooltipLump:new kn("bool","Tooltip",["Sugar lump tooltip OFF","Sugar lump tooltip ON"],"Shows the current Sugar Lump type in Sugar lump tooltip.",!0),TooltipPlots:new kn("bool","Tooltip",["Garden plots tooltip OFF","Garden plots tooltip ON"],"Shows a tooltip for plants that have a cookie reward.",!0),TooltipPantheon:new kn("bool","Tooltip",["Pantheon tooltip OFF","Pantheon tooltip ON"],"Shows additional info in the pantheon tooltip",!0),TooltipAscendButton:new kn("bool","Tooltip",["Show Extra Info Ascend Tooltip OFF","Show Extra Info Ascend Tooltip ON"],"Shows additional info in the ascend tooltip",!0),Stats:new kn("bool","Statistics",["Statistics OFF","Statistics ON"],"Extra Cookie Monster statistics!",!0),MissingUpgrades:new kn("bool","Statistics",["Missing upgrades OFF","Missing upgrades ON"],"Shows missing upgrades in statistics menu",!0),MissingAchievements:new kn("bool","Statistics",["Missing Achievements OFF","Missing Normal Achievements ON"],"Shows missing normal achievements in statistics menu.",!0),UpStats:new kn("bool","Statistics",["Statistics update rate (default)","Statistics update rate (1s)"],"Default rate is once every 5 seconds",!1),HeavenlyChipsTarget:new bn("numscale","Statistics","Heavenly chips target: ",'Use this to set a heavenly chips target that will be counted towards in the "prestige" statsistics sections',1,1/0),ShowMissedGC:new kn("bool","Statistics",["Missed GC OFF","Missed GC ON"],"Show a stat in the statistics screen that counts how many golden cookies you have missed",!0),Title:new kn("bool","NotificationGeneral",["Title OFF","Title ON","Title pinned tab highlight"],'Update title with colden cookie/season popup timers; pinned tab highlight only changes the title when a golden cookie/season popup spawns; "!" means that golden cookie/reindeer can spawn',!0),GeneralSound:new kn("bool","NotificationGeneral",["Consider game volume setting OFF","Consider game volume setting ON"],'Turning this toggle to "off" makes Cookie Monster no longer consider the volume setting of the base game, allowing mod notifications to play with base game volume turned down',!0),GCNotification:new kn("bool","NotificationGC",["Notification OFF","Notification ON"],"Create a notification when golden cookie spawns",!0,(function(){mn(e.GCNotification)})),GCFlash:new kn("bool","NotificationGC",["Flash OFF","Flash ON"],"Flash screen on golden cookie",!0),ColourGCFlash:new yn("colour","NotificationGC","The colour of the GC flash, standard colour is white"),GCSound:new kn("bool","NotificationGC",["Sound OFF","Sound ON"],"Play a sound on golden cookie",!0),GCVolume:new Gn("vol","NotificationGC",[],"Volume"),GCSoundURL:new kn("url","NotificationGC","Sound URL:","URL of the sound to be played when a golden cookie spawns"),FortuneNotification:new kn("bool","NotificationFC",["Notification OFF","Notification ON"],"Create a notification when fortune cookie is on the ticker",!0,(function(){mn(e.FortuneNotification)})),FortuneFlash:new kn("bool","NotificationFC",["Flash OFF","Flash ON"],"Flash screen on fortune cookie spawn",!0),ColourFortuneFlash:new yn("colour","NotificationFC","The colour of the fortune flash, standard colour is white"),FortuneSound:new kn("bool","NotificationFC",["Sound OFF","Sound ON"],"Play a sound on fortune cookie spawn",!0),FortuneVolume:new Gn("vol","NotificationFC",[],"Volume"),FortuneSoundURL:new kn("url","NotificationFC","Sound URL:","URL of the sound to be played when the ticker has a fortune cookie"),SeaNotification:new kn("bool","NotificationSea",["Notification OFF","Notification ON"],"Create a notification on season popup",!0,(function(){mn(e.SeaNotification)})),SeaFlash:new kn("bool","NotificationSea",["Flash OFF","Flash ON"],"Flash screen on season popup",!0),ColourSeaFlash:new yn("colour","NotificationSea","The colour of the season popup flash, standard colour is white"),SeaSound:new kn("bool","NotificationSea",["Sound OFF","Sound ON"],"Play a sound on season popup",!0),SeaVolume:new Gn("vol","NotificationSea",[],"Volume"),SeaSoundURL:new kn("url","NotificationSea","Sound URL:","URL of the sound to be played when on season popup spawns"),GardFlash:new kn("bool","NotificationGard",["Garden Tick Flash OFF","Flash ON"],"Flash screen on garden tick",!0),ColourGardFlash:new yn("colour","NotificationGard","The colour of the garden flash, standard colour is white"),GardSound:new kn("bool","NotificationGard",["Sound OFF","Sound ON"],"Play a sound on garden tick",!0),GardVolume:new Gn("vol","NotificationGard",[],"Volume"),GardSoundURL:new kn("url","NotificationGard","Garden Tick Sound URL:","URL of the sound to be played when the garden ticks"),MagicNotification:new kn("bool","NotificationMagi",["Notification OFF","Notification ON"],"Create a notification when magic reaches maximum",!0,(function(){mn(e.MagicNotification)})),MagicFlash:new kn("bool","NotificationMagi",["Flash OFF","Flash ON"],"Flash screen when magic reaches maximum",!0),ColourMagicFlash:new yn("colour","NotificationMagi","The colour of the magic flash, standard colour is white"),MagicSound:new kn("bool","NotificationMagi",["Sound OFF","Sound ON"],"Play a sound when magic reaches maximum",!0),MagicVolume:new Gn("vol","NotificationMagi",[],"Volume"),MagicSoundURL:new kn("url","NotificationMagi","Sound URL:","URL of the sound to be played when magic reaches maxium"),WrinklerNotification:new kn("bool","NotificationWrink",["Notification OFF","Notification ON"],"Create a notification when a wrinkler appears",!0,(function(){mn(e.WrinklerNotification)})),WrinklerFlash:new kn("bool","NotificationWrink",["Flash OFF","Flash ON"],"Flash screen when a wrinkler appears",!0),ColourWrinklerFlash:new yn("colour","NotificationWrink","The colour of the wrinkler flash, standard colour is white"),WrinklerSound:new kn("bool","NotificationWrink",["Sound OFF","Sound ON"],"Play a sound when a wrinkler appears",!0),WrinklerVolume:new Gn("vol","NotificationWrink",[],"Volume"),WrinklerSoundURL:new kn("url","NotificationWrink","Sound URL:","URL of the sound to be played when a wrinkler appears"),WrinklerMaxNotification:new kn("bool","NotificationWrinkMax",["Notification OFF","Notification ON"],"Create a notification when the maximum amount of wrinklers has appeared",!0,(function(){mn(e.WrinklerMaxNotification)})),WrinklerMaxFlash:new kn("bool","NotificationWrinkMax",["Flash OFF","Flash ON"],"Flash screen when the maximum amount of Wrinklers has appeared",!0),ColourWrinklerMaxFlash:new yn("colour","NotificationWrinkMax","The colour of the maximum wrinkler flash, standard colour is white"),WrinklerMaxSound:new kn("bool","NotificationWrinkMax",["Sound OFF","Sound ON"],"Play a sound when the maximum amount of wrinklers has appeared",!0),WrinklerMaxVolume:new Gn("vol","NotificationWrinkMax",[],"Volume"),WrinklerMaxSoundURL:new kn("url","NotificationWrinkMax","Sound URL:","URL of the sound to be played when the maximum amount of wrinklers has appeared"),BulkBuyBlock:new kn("bool","Miscellaneous",["Block bulk buying OFF","Block bulk buying ON"],"Block clicking bulk buying when you can't buy all. This prevents buying 7 of a building when you are in buy-10 or buy-100 mode.",!0),FavouriteSettings:new kn("bool","Miscellaneous",["Favourite settings section OFF","Favourite settings section ON","Favourite settings section ON (Locked)"],"Show stars before each setting which allows selecting it for a 'favourites' section at the top of the Cookie Monster settings. Setting this to Locked removes the stars but shows the 'favourites' section",!0,(function(){Game.UpdateMenu()}))};function wn(){return JSON.stringify({favouriteSettings:lo,settings:e,version:"2.031.7"})}function xn(){if(ge!==Game.dragonLevel||se){if(Game.dragonLevel<25&&Game.dragonLevels[Game.dragonLevel].buy.toString().includes("sacrifice")){let e=Game.dragonLevels[Game.dragonLevel].buy.toString().match(/Objects\[(.*)\]/)[1];const t=Game.dragonLevels[Game.dragonLevel].buy.toString().match(/sacrifice\((.*?)\)/)[1];if("i"!==e)if(e=e.replaceAll("'",""),Game.Objects[e].amount<t)fe="Not enough buildings to sell";else{let o=0;Mo();for(let n=0;n<t;n++){let t=Ct[e].basePrice*Game.priceIncrease**Math.max(0,Ct[e].amount-1-Ct[e].free);t=Game.modifyBuildingPrice(Ct[e],t),t=Math.ceil(t),o+=t,Ct[e].amount-=1}fe=`Cost to rebuy: ${B(o)}`}else{let o=0;Mo(),Object.keys(Game.Objects).forEach((n=>{if(e=n,Game.Objects[e].amount<t)fe="Not enough buildings to sell";else{for(let n=0;n<t;n++){let t=Ct[e].basePrice*Game.priceIncrease**Math.max(0,Ct[e].amount-1-Ct[e].free);t=Game.modifyBuildingPrice(Ct[e],t),t=Math.ceil(t),o+=t,Ct[e].amount-=1}fe=`Cost to rebuy: ${B(o)}`}}))}}ge=Game.dragonLevel}}function Tn(e,t){if(!Game.Objects.Temple.minigameLoaded)return 0;Mo();const{minigame:o}=Game.Objects.Temple,n=o.godsById[e].slot;"0"===n?ae=o.slot[t]:"1"===n?re=o.slot[t]:"2"===n&&(le=o.slot[t]),0===t?ae=e:1===t?re=e:2===t&&(le=e);const i=ee;return Oo(),Ho(),i!==ee&&Oo(),de-Game.cookiesPs}function Mn(){Ct=[],Object.keys(Game.Objects).forEach((e=>{Ct[e]=xo(e)})),yt=[],Object.keys(Game.Upgrades).forEach((e=>{yt[e]=To(e)})),bt=[],Object.keys(Game.Achievements).forEach((e=>{bt[e]=vo(e)})),Mo()}let Sn=!1;function Pn(t,o,n){(1===e[o]||n)&&3===t&&!1===Sn||1===t?(l("CMFlashScreen").style.backgroundColor=e[`Colour${o}`],l("CMFlashScreen").style.opacity="0.5",3===t?(l("CMFlashScreen").style.display="inline",setTimeout((function(){Pn(2,o,!0)}),1e3/Game.fps)):setTimeout((function(){Pn(0,o,!0)}),1e3/Game.fps)):2===t?(l("CMFlashScreen").style.opacity="1",setTimeout((function(){Pn(1,o,!0)}),1e3/Game.fps)):0===t&&(l("CMFlashScreen").style.display="none")}function Bn(t,o,n,i){if((1===e[o]||i)&&!1===Sn){const o=new Audio(t);e.GeneralSound?o.volume=e[n]/100*(Game.volume/100):o.volume=e[n]/100,o.play()}}function En(t,o,n){1===e[t]&&"hidden"===document.visibilityState&&!1===Sn&&new Notification(o,{body:n,badge:"https://orteil.dashnet.org/cookieclicker/favicon.ico"})}function Nn(){if(no!==Game.OnAscend&&(no=Game.OnAscend,Game.OnAscend?(l("game").style.bottom="0px",1===e.BotBar&&(l("CMBotBar").style.display="none"),1===e.TimerBar&&(l("CMTimerBar").style.display="none")):(hn(),ko()),yo()),!Game.OnAscend&&0===Game.AscendTimer){r!==Object.keys(Game.mods).length&&(Mn(),pn(),r=Object.keys(Game.mods).length),se&&(cn(),en(),Game.Has("Golden switch [off]")?(Mo(),yt["Golden switch [off]"].bought=0,Oo(),Se=de):Se=Game.cookiesPs,sn(),ln(),an(),on(),xn(),function(){for(let e=0;e<11;e+=1)for(let t=0;t<3;t+=1)ft[e][t]=Tn(e,t)}(),rn(),function(){let e=0;if(Game.Objects.Bank.minigameLoaded){const t=Game.Objects.Bank.minigame.goods;let o=0;Object.keys(t).forEach((e=>{const n=t[e];o+=n.stock*n.val})),e+=o*Game.cookiesPsRawHighest}e+=function(){let e=0;Mo();let t=2;5!==ne&&18!==ne||(t-=1),5!==ie&&18!==ie||(t-=1),ne=5,ie=18;for(let e=0;e<t;++e){let e="Cursor";Object.keys(Ct).forEach((t=>{Ct[t].amount>0&&(e=t)})),Ct[e].amount-=1,K-=1}return Object.keys(Ct).forEach((t=>{const o=Ct[t];e+=Mt(Game.Objects[o.name],Game.Objects[t].basePrice,o.amount,Game.Objects[t].free,o.amount)})),e}(),gt=e}(),se=0);const t=Game.auraMult("Fierce Hoarder")>0;!Y&&t?(Y=!0,q=1):Y&&!t&&(Y=!1,q=1),q&&(Zo(),q=0),function(){dn(),Yo(),function(){Re=0;let e=0;Object.keys(Game.wrinklers).forEach((t=>{2===Game.wrinklers[t].phase&&(e+=1)}));let t=1;if(Ct.Temple.minigameLoaded){const e=Game.hasGod("scorn");1===e?t*=1.15:2===e?t*=1.1:3===e&&(t*=1.05)}Ie=e,Re=e*(.05*e*1.1)*(.05*Game.Has("Sacrilegious corruption")+1)*(.05*Game.Has("Wrinklerspawn")+1)*t}(),Qo(),nn();const e=Game.HowManyCookiesReset(Math.floor(Game.HowMuchPrestige(Game.cookiesReset+Game.cookiesEarned))+1)-(Game.cookiesEarned+Game.cookiesReset);pt=po(e/mo())}(),G=0,ht={},Object.keys(Game.shimmers).forEach((e=>{ht[Game.shimmers[e].id]=Game.shimmers[e],Game.shimmers[e].spawnLead&&"golden"===Game.shimmers[e].type&&(ct=Game.shimmers[e],G+=1)})),Object.keys(ao).forEach((e=>{void 0===ht[e]&&(ao[e].parentNode.removeChild(ao[e]),delete ao[e])})),y!==Game.shimmerTypes.golden.n?(y=Game.shimmerTypes.golden.n,y&&(b<G&&(Pn(3,"GCFlash",!1),Bn(e.GCSoundURL,"GCSound","GCVolume",!1),En("GCNotification","Golden Cookie Spawned","A Golden Cookie has spawned. Click it now!")),Object.keys(Game.shimmers).forEach((t=>{void 0===ao[Game.shimmers[t].id]&&function(t){const o=document.createElement("div");o.id=`GCTimer${t.id}`,o.style.width="96px",o.style.height="96px",o.style.position="absolute",o.style.zIndex="10000000001",o.style.textAlign="center",o.style.lineHeight="96px",o.style.fontFamily='"Kavoon", Georgia, serif',o.style.fontSize="35px",o.style.cursor="pointer",o.style.display="block",0===e.GCTimer&&(o.style.display="none"),o.style.left=t.l.style.left,o.style.top=t.l.style.top,o.onclick=function(){t.pop()},o.onmouseover=function(){t.l.style.filter="brightness(125%) drop-shadow(0px 0px 3px rgba(255,255,255,1))",t.l.style.webkitFilter="brightness(125%) drop-shadow(0px 0px 3px rgba(255,255,255,1))"},o.onmouseout=function(){t.l.style.filter="",t.l.style.webkitFilter=""},ao[t.id]=o,l("shimmers").appendChild(o)}(Game.shimmers[t])}))),fn(),b=G,0===G&&(ct=0)):1===e.GCTimer&&y&&Object.keys(ao).forEach((e=>{ao[e].style.opacity=ht[e].l.style.opacity,ao[e].style.transform=ht[e].l.style.transform,ao[e].textContent=Math.ceil(ht[e].life/Game.fps)})),f!==(Game.TickerEffect&&"fortune"===Game.TickerEffect.type)&&(f=Game.TickerEffect&&"fortune"===Game.TickerEffect.type,f&&(Pn(3,"FortuneFlash",!1),Bn(e.FortuneSoundURL,"FortuneSound","FortuneVolume",!1),En("FortuneNotification","Fortune Cookie found","A Fortune Cookie has appeared on the Ticker."))),g!==Game.shimmerTypes.reindeer.spawned&&(g=Game.shimmerTypes.reindeer.spawned,Object.keys(Game.shimmers).forEach((e=>{Game.shimmers[e].spawnLead&&"reindeer"===Game.shimmers[e].type&&(dt=Game.shimmers[e])})),Pn(3,"SeaFlash",!1),Bn(e.SeaSoundURL,"SeaSound","SeaVolume",!1),En("SeaNotification","Reindeer sighted!","A Reindeer has spawned. Click it now!")),Game.Objects.Farm.minigameLoaded&&C!==Game.Objects.Farm.minigame.nextStep&&(0!==C&&C<Date.now()&&(Pn(3,"GardFlash",!1),Bn(e.GardSoundURL,"GardSound","GardVolume",!1)),C=Game.Objects.Farm.minigame.nextStep),function(){if(Game.Objects["Wizard tower"].minigameLoaded&&1===e.GrimoireBar){const{minigame:t}=Game.Objects["Wizard tower"];t.magic<t.magicM?k=!1:k||(k=!0,Pn(3,"MagicFlash",!1),Bn(e.MagicSoundURL,"MagicSound","MagicVolume",!1),En("MagicNotification","Magic Meter full","Your Magic Meter is full. Cast a spell!"))}}(),function(){if(Game.elderWrath>0){let t=0;Object.keys(Game.wrinklers).forEach((e=>{2===Game.wrinklers[e].phase&&(t+=1)})),t>v?(v=t,t===Game.getWrinklersMax()&&e.WrinklerMaxFlash?Pn(3,"WrinklerMaxFlash",!1):Pn(3,"WrinklerFlash",!1),t===Game.getWrinklersMax()&&e.WrinklerMaxSound?Bn(e.WrinklerMaxSoundURL,"WrinklerMaxSound","WrinklerMaxVolume",!1):Bn(e.WrinklerSoundURL,"WrinklerSound","WrinklerVolume",!1),t===Game.getWrinklersMax()&&e.WrinklerMaxNotification?En("WrinklerMaxNotification","Maximum Wrinklers Reached","You have reached your maximum ammount of wrinklers"):En("WrinklerNotification","A Wrinkler appeared","A new wrinkler has appeared")):v=t}}()}}function On(){let t="";for(let o=0;o<Dt.length;o++)t+=`.CMText${Dt[o]} { color: ${e[`Colour${Dt[o]}`]}; }\n`;for(let o=0;o<Dt.length;o++)t+=`.CMBack${Dt[o]} { background-color: ${e[`Colour${Dt[o]}`]}; }\n`;for(let o=0;o<Dt.length;o++)t+=`.CMBorder${Dt[o]} { border: 1px solid ${e[`Colour${Dt[o]}`]}; }\n`;l("CMCSS").textContent=t,so()}function Fn(){const e=b64_to_utf8(unescape(localStorage.getItem("CookieClickerGame")).split("!END!")[0]),t=e.match(/CookieMonster.*(;|$)/);if(null!==t){const o=e.replace(t[0],`CookieMonster:${wn()}`);localStorage.setItem("CookieClickerGame",escape(`${utf8_to_b64(o)}!END!`))}}function Ln(t){if(void 0!==localStorage.CMConfig&&delete localStorage.CMConfig,void 0!==t){e=t,void 0!==e.Colors&&delete e.Colors,void 0!==e.Colours&&delete e.Colours;let o=!1;Object.keys(un).forEach((t=>{void 0===e[t]?(o=!0,e[t]=un[t]):"Header"===t&&Object.keys(un.Header).forEach((n=>{void 0!==e[t][n]&&e[t][n]>-1&&e[t][n]<2||(o=!0,e[t][n]=un[t][n])}))})),o&&Fn(),Nn(),Object.keys(un).forEach((e=>{"Header"!==e&&void 0!==vn[e].func&&vn[e].func()}))}else Ln(un);Game.UpdateMenu(),On()}function Wn(){Object.keys(Game.wrinklers).forEach((e=>{Game.wrinklers[e].sucked>0&&0===Game.wrinklers[e].type&&(Game.wrinklers[e].hp=0)}))}function An(t,o){if("b"===t){if(l("tooltip").innerHTML=Game.Objects[o].tooltip(),1===e.TooltipAmor){const e=Ko(Game.Objects[o],Game.Objects[o].basePrice,0,Game.Objects[o].free,Game.Objects[o].amount),t=e-Game.Objects[o].totalCookies;t>0&&(l("tooltip").innerHTML=l("tooltip").innerHTML.split("so far</div>").join(`so far<br/>&bull; <b>${B(t)}</b> ${1===Math.floor(t)?"cookie":"cookies"} left to amortize (${uo((e-Game.Objects[o].totalCookies)/(Game.Objects[o].storedTotalCps*Game.globalCpsMult)).text})</div>`))}-1===Game.buyMode&&(l("tooltip").innerHTML=l("tooltip").innerHTML.split(B(Game.Objects[o].bulkPrice)).join(B((Game.Objects[o],Game.Objects[o].basePrice,Game.Objects[o].amount,Game.Objects[o].free,Game.buyBulk,1))))}else if("u"===t){if(!Game.UpgradesInStore[o])return"";l("tooltip").innerHTML=Game.crateTooltip(Game.UpgradesInStore[o],"store")}else"s"===t?l("tooltip").innerHTML=Game.lumpTooltip():"g"===t?l("tooltip").innerHTML=Game.Objects["Wizard tower"].minigame.spellTooltip(o)():"p"===t?l("tooltip").innerHTML=Game.ObjectsById[2].minigame.tileTooltip(o[0],o[1])():"ha"===t?l("tooltip").innerHTML=Game.ObjectsById[2].minigame.toolTooltip(1)():"wb"===t?l("tooltip").innerHTML="":"pag"===t?l("tooltip").innerHTML=Game.Objects.Temple.minigame.godTooltip(o)():"pas"===t&&(l("tooltip").innerHTML=Game.Objects.Temple.minigame.slotTooltip(o[0])());if("b"===t&&1===Game.buyMode||"u"===t||"s"===t||"g"===t||"p"===t&&!Game.keys[16]||"ha"===t||"wb"===t||"pag"===t||"pas"===t&&-1!==o[1]){const e=document.createElement("div");e.id="CMTooltipArea",l("tooltip").appendChild(e)}return eo=t,to=o,Io(),l("tooltip").innerHTML}function jn(){Game.Objects["Wizard tower"].minigameLoaded&&Object.keys(Game.Objects["Wizard tower"].minigame.spellsById).forEach((e=>{null!==l(`grimoireSpell${e}`).onmouseover&&(x[e]=l(`grimoireSpell${e}`).onmouseover,l(`grimoireSpell${e}`).onmouseover=function(){Game.tooltip.dynamic=1,Game.tooltip.draw(this,(function(){return An("g",`${e}`)}),"this"),Game.tooltip.wobble()})}))}function Dn(){if(!m&&Game.Objects["Wizard tower"].minigameLoaded){const{minigame:t}=Game.Objects["Wizard tower"];p=t.draw,Game.Objects["Wizard tower"].minigame.draw=function(){p(),1===e.GrimoireBar&&t.magic<t.magicM&&(t.magicBarTextL.innerHTML+=` (${po($o(t.magic,t.magicM,t.magicM))})`)},m=!0}}function Hn(){!function(){if(!u&&Game.Objects["Wizard tower"].minigameLoaded){const{minigame:e}=Game.Objects["Wizard tower"];c=e.launch,d=new Function(`return ${e.launch.toString().split("=this").join("= Game.Objects['Wizard tower'].minigame")}`),Game.Objects["Wizard tower"].minigame.launch=function(){d(),jn(),m=!1,Dn(),u=!0}}}(),Dn()}var Un=o(877),$n=o.n(Un);const Rn="CMConfig";function In(t){null!==l(`slider${t}`)&&(l(`slider${t}right`).innerHTML=`${l(`slider${t}`).value}%`,e[t]=Math.round(l(`slider${t}`).value)),Fn()}function zn(t){e.Header[t]+=1,e.Header[t]>1&&(e.Header[t]=0),Fn()}function Vn(t,o){const n=document.createElement("div");n.className="title",n.style.padding="0px 16px",n.style.opacity="0.7",n.style.fontSize="17px",n.style.fontFamily='"Kavoon", Georgia, serif',n.appendChild(document.createTextNode(`${t} `));const i=document.createElement("span");return i.style.cursor="pointer",i.style.display="inline-block",i.style.height="14px",i.style.width="14px",i.style.borderRadius="7px",i.style.textAlign="center",i.style.backgroundColor="#C0C0C0",i.style.color="black",i.style.fontSize="13px",i.style.verticalAlign="middle",i.textContent=e.Header[o]?"-":"+",i.onclick=function(){zn(o),Game.UpdateMenu()},n.appendChild(i),n}function _n(e,t,o,n){const i=document.createElement("div");i.className="listing";const a=document.createElement("b");if(a.textContent=t,i.appendChild(a),"withTooltip"===e){i.className="listing",i.appendChild(document.createTextNode(" "));const e=document.createElement("span");e.onmouseout=function(){Game.tooltip.hide()},e.onmouseover=function(){Game.tooltip.draw(this,escape(Ut[n].innerHTML))},e.style.cursor="default",e.style.display="inline-block",e.style.height="10px",e.style.width="10px",e.style.borderRadius="5px",e.style.textAlign="center",e.style.backgroundColor="#C0C0C0",e.style.color="black",e.style.fontSize="9px",e.style.verticalAlign="bottom",e.textContent="?",i.appendChild(e)}return i.appendChild(document.createTextNode(": ")),i.appendChild(o),i}function Qn(e){const t=document.createDocumentFragment();t.appendChild(document.createTextNode(`${e.length} `));const o=document.createElement("span");o.onmouseout=function(){Game.tooltip.hide()};const n=document.createElement("div"),i=document.createElement("div");i.style.minWidth="140px",i.style.marginBottom="4px";const a=document.createElement("div");return a.className="name",a.style.marginBottom="4px",a.style.textAlign="center",a.textContent="Missing",i.appendChild(a),Object.keys(e).forEach((t=>{const o=document.createElement("div");o.style.textAlign="center",o.appendChild(document.createTextNode(e[t])),i.appendChild(o)})),n.appendChild(i),o.onmouseover=function(){Game.tooltip.draw(this,escape(n.innerHTML))},o.style.cursor="default",o.style.display="inline-block",o.style.height="10px",o.style.width="10px",o.style.borderRadius="5px",o.style.textAlign="center",o.style.backgroundColor="#C0C0C0",o.style.color="black",o.style.fontSize="9px",o.style.verticalAlign="bottom",o.textContent="?",t.appendChild(o),t}function Xn(t){const o=document.createElement("div");if(o.className="subsection",o.appendChild(t),o.appendChild(Vn("Lucky Cookies","Lucky")),e.Header.Lucky&&o.appendChild(function(){const e=Game.auraMult("Dragon's Fortune")?"GoldCookDragonsFortuneTooltipPlaceholder":"GoldCookTooltipPlaceholder",t=document.createElement("div");t.className="CMStatsLuckySection";const o=Game.cookies+ho()<Ce?Ft:Et,n=Game.cookies+ho()<Ce?po((Ce-(Game.cookies+ho()))/mo()):"",i=document.createDocumentFragment(),a=document.createElement("span");if(a.style.fontWeight="bold",a.className=St+o,a.textContent=B(Ce),i.appendChild(a),""!==n){const e=document.createElement("small");e.textContent=` (${n})`,i.appendChild(e)}t.appendChild(_n("withTooltip",'"Lucky!" cookies required',i,e));const r=Game.cookies+ho()<ke?Ft:Et,l=Game.cookies+ho()<ke?po((ke-(Game.cookies+ho()))/mo()):"",s=document.createDocumentFragment(),c=document.createElement("span");if(c.style.fontWeight="bold",c.className=St+r,c.textContent=B(ke),s.appendChild(c),""!==l){const e=document.createElement("small");e.textContent=` (${l})`,s.appendChild(e)}t.appendChild(_n("withTooltip",'"Lucky!" cookies required (frenzy)',s,e));const d=ye!==be,p=document.createElement("span");p.style.fontWeight="bold",p.className=St+ye,p.textContent=B(ye)+(d?` / ${B(be)}`:""),t.appendChild(_n("withTooltip",'"Lucky!" reward (max)'+(d?" (golden / wrath)":""),p,e));const u=document.createElement("span");u.style.fontWeight="bold",u.className=St+u,u.textContent=B(Ge)+(d?` / ${B(ve)}`:""),t.appendChild(_n("withTooltip",'"Lucky!" reward (max) (frenzy)'+(d?" (golden / wrath)":""),u,e));const m=Math.min(.15*(Game.cookies+ho()),Se*Pe*60*15)+13,h=document.createElement("span");return h.style.fontWeight="bold",h.className=St+h,h.textContent=B(Be*m)+(d?` / ${B(Ee*m)}`:""),t.appendChild(_n("withTooltip",'"Lucky!" reward (cur)'+(d?" (golden / wrath)":""),h,e)),t}()),o.appendChild(Vn("Chain Cookies","Chain")),e.Header.Chain&&o.appendChild(function(){const e=Game.auraMult("Dragon's Fortune")?"GoldCookDragonsFortuneTooltipPlaceholder":"GoldCookTooltipPlaceholder",t=document.createElement("div");t.className="CMStatsChainSection";const o=Game.cookies+ho()<qe?Ft:Et,n=Game.cookies+ho()<qe?po((qe-(Game.cookies+ho()))/mo()):"",i=document.createDocumentFragment(),a=document.createElement("span");if(a.style.fontWeight="bold",a.className=St+o,a.textContent=B(qe),i.appendChild(a),""!==n){const e=document.createElement("small");e.textContent=` (${n})`,i.appendChild(e)}t.appendChild(_n("withTooltip",'"Chain" cookies required',i,e));const r=Game.cookies+ho()<Je?Ft:Et,l=Game.cookies+ho()<Je?po((Je-(Game.cookies+ho()))/mo()):"",s=document.createDocumentFragment(),c=document.createElement("span");if(c.style.fontWeight="bold",c.className=St+r,c.textContent=B(Je),s.appendChild(c),""!==l){const e=document.createElement("small");e.textContent=` (${l})`,s.appendChild(e)}t.appendChild(_n("withTooltip",'"Chain" cookies required (Wrath)',s,e));const d=Game.cookies+ho()<tt?Ft:Et,p=Game.cookies+ho()<tt?po((tt-(Game.cookies+ho()))/mo()):"",u=document.createDocumentFragment(),m=document.createElement("span");if(m.style.fontWeight="bold",m.className=St+d,m.textContent=B(tt),u.appendChild(m),""!==p){const e=document.createElement("small");e.textContent=` (${p})`,u.appendChild(e)}t.appendChild(_n("withTooltip",'"Chain" cookies required (Frenzy)',u,e));const h=Game.cookies+ho()<it?Ft:Et,g=Game.cookies+ho()<it?po((it-(Game.cookies+ho()))/mo()):"",f=document.createDocumentFragment(),C=document.createElement("span");if(C.style.fontWeight="bold",C.className=St+h,C.textContent=B(it),f.appendChild(C),""!==g){const e=document.createElement("small");e.textContent=` (${g})`,f.appendChild(e)}t.appendChild(_n("withTooltip",'"Chain" cookies required (frenzy) (Wrath)',f,e)),t.appendChild(_n("withTooltip",'"Chain" reward (max) (golden / wrath)',document.createTextNode(`${B(Ke[0])} / ${B(et[0])}`),e)),t.appendChild(_n("withTooltip",'"Chain" reward (max) (frenzy) (golden / wrath)',document.createTextNode(`${B(nt[0])} / ${B(nt[0])}`),e));const y=Math.min(60*Game.cookiesPs*60*6*Pe,.5*Game.cookies),b=tn(7,y,Be)[0],k=tn(6,y,Ee)[0];return t.appendChild(_n("withTooltip",'"Chain" reward (cur) (golden / wrath)',document.createTextNode(`${B(b)} / ${B(k)}`),e)),t.appendChild(_n("withTooltip","CPS needed for next level (g / w)",document.createTextNode(`${B(Ye)} / ${B(Ze)}`),"ChainNextLevelPlaceholder")),t.appendChild(_n("withTooltip","CPS needed for next level (frenzy) (g / w)",document.createTextNode(`${B(ot)} / ${B(at)}`),"ChainNextLevelPlaceholder")),t}()),Game.Objects["Wizard tower"].minigameLoaded&&(o.appendChild(Vn("Spells","Spells")),e.Header.Spells&&o.appendChild(function(){const e=document.createElement("div");e.className="CMStatsSpellsSection";const t=Game.cookies+ho()<we?Ft:Et,o=Game.cookies+ho()<we?po((we-(Game.cookies+ho()))/mo()):"",n=document.createDocumentFragment(),i=document.createElement("span");if(i.style.fontWeight="bold",i.className=St+t,i.textContent=B(we),n.appendChild(i),""!==o){const e=document.createElement("small");e.textContent=` (${o})`,n.appendChild(e)}e.appendChild(_n("withTooltip",'"Conjure Baked Goods" cookies required',n,"GoldCookTooltipPlaceholder")),e.appendChild(_n("withTooltip",'"Conjure Baked Goods" reward (max)',document.createTextNode(B(xe)),"GoldCookTooltipPlaceholder"));const a=Game.cookies+ho()<7*we?Ft:Et,r=Math.min(.15*(Game.cookies+ho()),60*Se*30),l=Game.cookies+ho()<7*we?po((7*we-(Game.cookies+ho()))/mo()):"",s=document.createDocumentFragment(),c=document.createElement("span");if(c.style.fontWeight="bold",c.className=St+a,c.textContent=B(7*we),s.appendChild(c),""!==l){const e=document.createElement("small");e.textContent=` (${l})`,s.appendChild(e)}return e.appendChild(_n("withTooltip",'"Conjure Baked Goods" cookies required (frenzy)',s,"GoldCookTooltipPlaceholder")),e.appendChild(_n("withTooltip",'"Conjure Baked Goods" reward (max) (frenzy)',document.createTextNode(B(7*xe)),"GoldCookTooltipPlaceholder")),e.appendChild(_n("withTooltip",'"Conjure Baked Goods" reward (cur)',document.createTextNode(B(r)),"GoldCookTooltipPlaceholder")),Te&&e.appendChild(_n("withTooltip",'"Spontaneous Edifice" cookies required (most expensive building)',document.createTextNode(`${B(Te)} (${Me})`),"GoldCookTooltipPlaceholder")),e}())),Game.Objects.Farm.minigameLoaded&&(o.appendChild(Vn("Garden","Garden")),e.Header.Garden&&o.appendChild(function(){const e=document.createElement("div");e.className="CMStatsGardenSection";const t=Game.cookies<60*Game.cookiesPs*10*100?Ft:Et,o=document.createElement("span");o.style.fontWeight="bold",o.className=St+t,o.textContent=B(60*Game.cookiesPs*10*100),e.appendChild(_n("basic","Cookies required for max reward of Bakeberry: ",o));const n=Game.cookies<60*Game.cookiesPs*100?Ft:Et,i=document.createElement("span");i.style.fontWeight="bold",i.className=St+n,i.textContent=B(60*Game.cookiesPs*100),e.appendChild(_n("basic","Cookies required for max reward of Chocoroot: ",i));const a=Game.cookies<60*Game.cookiesPs*60*25?Ft:Et,r=document.createElement("span");r.style.fontWeight="bold",r.className=St+a,r.textContent=B(60*Game.cookiesPs*60*25),e.appendChild(_n("basic","Cookies required for max reward of Queenbeet: ",r));const l=Game.cookies<60*Game.cookiesPs*15*100?Ft:Et,s=document.createElement("span");s.style.fontWeight="bold",s.className=St+l,s.textContent=B(60*Game.cookiesPs*15*100),e.appendChild(_n("basic","Cookies required for max reward of Duketater: ",s));const c=[];return Object.keys(jo).forEach((e=>{Game.HasUnlocked(jo[e])||c.push(jo[e])})),0!==c.length&&e.appendChild(_n("basic","Rare plant drops left to unlock",Qn(c))),e}())),o.appendChild(Vn("Prestige","Prestige")),e.Header.Prestige&&o.appendChild(function(){const t=document.createElement("div");t.className="CMStatsPrestigeSection";const o=Math.floor(Game.HowMuchPrestige(lt+Game.cookiesReset+He+(Game.HasUnlocked("Chocolate egg")&&!Game.Has("Chocolate egg")?Le:0)));t.appendChild(_n("withTooltip","Prestige level (cur / max)",document.createTextNode(`${B(Game.prestige)} / ${B(o)}`),"PrestMaxTooltipPlaceholder"));const n=Math.max(0,Game.HowManyCookiesReset(o+1)-(lt+Game.cookiesReset+He+(Game.HasUnlocked("Chocolate egg")&&!Game.Has("Chocolate egg")&&Le?Le:0))),i=document.createDocumentFragment();i.appendChild(document.createTextNode(B(n)));const a=document.createElement("small");a.textContent=` (${po(n/st,1)})`,i.appendChild(a),t.appendChild(_n("withTooltip","Cookies to next level",i,"NextPrestTooltipPlaceholder")),t.appendChild(_n("withTooltip","Heavenly chips (cur / max)",document.createTextNode(`${B(Game.heavenlyChips)} / ${B(o-Game.prestige+Game.heavenlyChips)}`),"HeavenChipMaxTooltipPlaceholder")),t.appendChild(_n("basic","Heavenly chips per second (last 5 seconds)",document.createTextNode(B(Ne,2))));const r=Number(e.HeavenlyChipsTarget);if(!Number.isNaN(r)){const e=r-Math.floor(Game.HowMuchPrestige(Game.cookiesReset+Game.cookiesEarned));e>0&&(t.appendChild(_n("basic","Heavenly chips to target set in settings (cur)",document.createTextNode(B(e)))),t.appendChild(_n("basic","Time till target (cur, current 5 second average)",document.createTextNode(po(e/Ne)))))}const l=function(e){let t=Game.cookiesPs;Mo(),0===yt["Heavenly key"].bought&&(yt["Heavenly chip secret"].bought=1,yt["Heavenly cookie stand"].bought=1,yt["Heavenly bakery"].bought=1,yt["Heavenly confectionery"].bought=1,yt["Heavenly key"].bought=1,Oo(),t=de,Mo()),lt>=1e6&&No("Sacrifice"),lt>=1e9&&No("Oblivion"),lt>=1e12&&No("From scratch"),lt>=1e15&&No("Nihilism"),lt>=1e18&&No("Dematerialize"),lt>=1e21&&No("Nil zero zilch"),lt>=1e24&&No("Transcendence"),lt>=1e27&&No("Obliterate"),lt>=1e30&&No("Negative void"),lt>=1e33&&No("To crumbs, you say?"),lt>=1e36&&No("You get nothing"),lt>=1e39&&No("Humble rebeginnings"),lt>=1e42&&No("The end of the world"),lt>=1e45&&No("Oh, you're back"),lt>=1e48&&No("Lazarus"),lt>=1e51&&No("Smurf account"),lt>=1e54&&No("If at first you don't succeed"),yt["Heavenly chip secret"].bought=1,yt["Heavenly cookie stand"].bought=1,yt["Heavenly bakery"].bought=1,yt["Heavenly confectionery"].bought=1,yt["Heavenly key"].bought=1,oe=e;const o=ee;Oo(),Ho(),o!==ee&&Oo();const n=de-t;return oe=Game.prestige,n}(o),s=document.createDocumentFragment();s.appendChild(document.createTextNode(B(l)));const c=Math.round(l/Game.cookiesPs*1e4);if(Number.isFinite(c)&&0!==c){const e=document.createElement("small");e.textContent=` (${c/100}% of income)`,s.appendChild(e)}t.appendChild(_n("withTooltip","Reset bonus income",s,"ResetTooltipPlaceholder"));const d=Math.floor(Game.HowMuchPrestige(Game.cookiesReset)),p=Math.floor(Game.HowMuchPrestige(Game.cookiesReset+Game.cookiesEarned)),u=p-d;if(!Game.Has("Lucky digit")){let e=7-p%10;e<0&&(e+=10);const o=u+e,n=p+e,i=document.createDocumentFragment();i.appendChild(document.createTextNode(`${n.toLocaleString()} / ${o.toLocaleString()} (+${e})`)),t.appendChild(_n("basic",'Next "Lucky Digit" (total / reset)',i))}if(!Game.Has("Lucky number")){let e=777-p%1e3;e<0&&(e+=1e3);const o=u+e,n=p+e,i=document.createDocumentFragment();i.appendChild(document.createTextNode(`${n.toLocaleString()} / ${o.toLocaleString()} (+${e})`)),t.appendChild(_n("basic",'Next "Lucky Number" (total / reset)',i))}if(!Game.Has("Lucky payout")){let e=777777-p%1e6;e<0&&(e+=1e6);const o=u+e,n=p+e,i=document.createDocumentFragment();i.appendChild(document.createTextNode(`${n.toLocaleString()} / ${o.toLocaleString()} (+${e})`)),t.appendChild(_n("basic",'Next "Lucky Payout" (total / reset)',i))}return t}()),Game.cpsSucked>0&&(o.appendChild(Vn("Wrinklers","Wrink")),e.Header.Wrink)){const e=document.createDocumentFragment();e.appendChild(document.createTextNode(`${B(He)} / ${B(Ue)} `));const t=document.createElement("a");t.textContent="Pop All Normal",t.className="option",t.onclick=function(){Wn()},e.appendChild(t),o.appendChild(_n("basic","Rewards of Popping (All/Normal)",e));const n=document.createDocumentFragment();n.appendChild(document.createTextNode(`${B($e[0])} `));const i=document.createElement("a");i.textContent="Pop Single Fattest",i.className="option",i.onclick=function(){null!==$e[1]&&(Game.wrinklers[$e[1]].hp=0)},n.appendChild(i),o.appendChild(_n("basic",`Rewards of Popping Single Fattest Non-Shiny Wrinkler (id: ${null!==$e[1]?$e[1]:"None"})`,n))}if(o.appendChild(function(){const t=document.createElement("div");t.className="CMStatsSeasonSection";let o=!1;const n=[];Object.keys(Lo).forEach((e=>{Game.Has(Lo[e])||(n.push(Lo[e]),o=!0)}));const i=[];Object.keys(Wo).forEach((e=>{Game.Has(Wo[e])||(i.push(Wo[e]),o=!0)}));const a=[];Object.keys(Ao).forEach((e=>{Game.Has(Ao[e])||(a.push(Ao[e]),o=!0)}));const r=[];Object.keys(Game.eggDrops).forEach((e=>{Game.HasUnlocked(Game.eggDrops[e])||(r.push(Game.eggDrops[e]),o=!0)}));const l=[];Object.keys(Game.rareEggDrops).forEach((e=>{Game.HasUnlocked(Game.rareEggDrops[e])||(l.push(Game.rareEggDrops[e]),o=!0)}));const s=Game.HasUnlocked("Chocolate egg")&&!Game.Has("Chocolate egg"),c=Game.Has("Century egg");if(("christmas"===Game.season||o||s||c)&&(t.appendChild(Vn("Season Specials","Sea")),e.Header.Sea)){if(0!==n.length){t.appendChild(_n("basic","Halloween cookies left to buy",Qn(n)));let e=.95;if(Game.HasAchiev("Spooky cookies")&&(e=.8),Game.Has("Starterror")&&(e*=.9),e*=1/Game.dropRateMult(),Game.hasGod){const t=Game.hasGod("seasons");1===t?e*=.9:2===t?e*=.95:3===t&&(e*=.97)}const o=n.length/7;t.appendChild(_n("basic","Chance of receiving a cookie from wrinkler/shiny wrinkler",document.createTextNode(`${B((1-e)*o*100)}% / ${B((1-.9*e)*o*100)}%`)))}if(0!==i.length){t.appendChild(_n("basic","Christmas cookies left to buy",Qn(i)));let e=.8;if(Game.HasAchiev("Let it snow")&&(e=.6),e*=1/Game.dropRateMult(),Game.Has("Starsnow")&&(e*=.95),Game.hasGod){const t=Game.hasGod("seasons");1===t?e*=.9:2===t?e*=.95:3===t&&(e*=.97)}const o=i.length/7;t.appendChild(_n("basic","Chance of receiving a cookie from reindeer",document.createTextNode(`${B((1-e)*o*100)}%`)))}0!==a.length&&t.appendChild(_n("basic","Valentine cookies left to buy",Qn(a)));const e=function(e){let t=e*(1/Game.dropRateMult());if(Game.HasAchiev("Hide & seek champion")&&(t*=.7),Game.Has("Omelette")&&(t*=.9),Game.Has("Starspawn")&&(t*=.9),Game.hasGod){const e=Game.hasGod("seasons");1===e?t*=.9:2===e?t*=.95:3===e&&(t*=.97)}const o=1-t,n=Game.eggDrops.length-r.length,i=Game.rareEggDrops.length-l.length,a=.9*o*(1-n/Game.eggDrops.length),s=.1*o*(1-i/Game.rareEggDrops.length),c=.9*o*(n/Game.eggDrops.length)+.1*o*(i/Game.rareEggDrops.length);return[a+.9*c*(1-n/Game.eggDrops.length),s+.1*c*(1-i/Game.rareEggDrops.length)]};0!==r.length&&(t.appendChild(_n("basic","Normal easter eggs left to unlock",Qn(r))),t.appendChild(_n("basic","Chance of receiving an egg from wrinkler/golden cookie",document.createTextNode(`${B(100*e(.98)[0])}% / ${B(100*e(.9)[0])}%`)))),0!==l.length&&(t.appendChild(_n("basic","Rare easter eggs left to unlock",Qn(l))),t.appendChild(_n("basic","Chance of receiving a rare egg from wrinkler/golden cookie",document.createTextNode(`${B(100*e(.98)[1])}% / ${B(100*e(.9)[1])}%`)))),"christmas"===Game.season&&t.appendChild(_n("basic","Reindeer reward",document.createTextNode(B(Fe)))),s&&t.appendChild(_n("withTooltip","Chocolate egg cookies",document.createTextNode(B(Le)),"ChoEggTooltipPlaceholder")),c&&t.appendChild(_n("basic","Century egg multiplier",document.createTextNode(Math.round(1e4*(Oe-1))/100+"%")))}return t}()),o.appendChild(Vn("Miscellaneous","Misc")),e.Header.Misc){if(o.appendChild(_n("basic",`Average cookies per second (past ${qt[e.AvgCPSHist]<60?`${qt[e.AvgCPSHist]} seconds`:qt[e.AvgCPSHist]/60+(3===e.AvgCPSHist?" minute":" minutes")})`,document.createTextNode(B(mo(),3)))),o.appendChild(_n("basic",`Average cookie clicks per second (past ${Yt[e.AvgClicksHist]}${0===e.AvgClicksHist?" second":" seconds"})`,document.createTextNode(B(Ve,1)))),o.appendChild(_n("basic",`Cookies from clicking (past ${Yt[e.AvgClicksHist]}${0===e.AvgClicksHist?" second":" seconds"})`,document.createTextNode(B(E.calcSum(Ve*Yt[e.AvgClicksHist]))))),Game.Has("Fortune cookies")){const e=[];Object.keys(Fo).forEach((t=>{Game.Has(Fo[t])||e.push(Fo[t])})),0!==e.length&&o.appendChild(_n("basic","Fortune Upgrades Left to Buy",Qn(e)))}if(e.ShowMissedGC&&o.appendChild(_n("basic","Missed golden cookies",document.createTextNode(B(Game.missedGoldenClicks)))),Game.prefs.autosave){const e=document.createElement("span");e.id="CMStatsAutosaveTimer",e.innerText=Game.sayTime(60*Game.fps-(Game.OnAscend?0:Game.T%(60*Game.fps)),4),o.appendChild(_n("basic","Time till autosave",e))}}l("menu").insertBefore(o,l("menu").childNodes[2]),e.MissingUpgrades&&l("menu").childNodes.forEach((e=>{if(e.children[0])if("Prestige"===e.children[0].innerHTML&&Qe){const t=Qe.match(new RegExp("div","g")||0).length/2,o=document.createElement("div");o.id="CMMissingUpgradesPrestigeTitle",o.className="listing";const n=document.createElement("div");n.innerHTML=`<b>Missing Prestige upgrades:</b> ${t}/${Game.PrestigeUpgrades.length} (${Math.floor(t/Game.PrestigeUpgrades.length*100)}%)`,o.appendChild(n),e.appendChild(o);const i=document.createElement("div");i.className="listing crateBox",i.innerHTML=Qe,e.appendChild(i)}else if("Upgrades"===e.children[0].innerHTML){if(_e){const t=_e.match(new RegExp("div","g")||0).length/2,o=document.createElement("div");o.id="CMMissingUpgradesTitle",o.className="listing";const n=document.createElement("div");n.innerHTML=`<b>Missing normal upgrades:</b> ${t}/${Game.UpgradesByPool[""].length+Game.UpgradesByPool.tech.length} (${Math.floor(t/(Game.UpgradesByPool[""].length+Game.UpgradesByPool.tech.length)*100)}%)`,o.appendChild(n),e.insertBefore(o,e.childNodes[3]);const i=document.createElement("div");i.className="listing crateBox",i.innerHTML=_e,e.insertBefore(i,document.getElementById("CMMissingUpgradesTitle").nextSibling)}if(Xe){const t=Xe.match(new RegExp("div","g")||0).length/2,o=document.createElement("div");o.id="CMMissingUpgradesCookiesTitle",o.className="listing";const n=document.createElement("div");n.innerHTML=`<b>Missing Cookie upgrades:</b> ${t}/${Game.UpgradesByPool.cookie.length} (${Math.floor(t/Game.UpgradesByPool.cookie.length*100)}%)`,o.appendChild(n),e.appendChild(o);const i=document.createElement("div");i.className="listing crateBox",i.innerHTML=Xe,e.appendChild(i)}}})),e.MissingAchievements&&function(){let t;Object.values(document.querySelectorAll("div.title")).forEach((e=>{e.textContent.includes("Achievements")&&(t=e.parentElement.querySelectorAll("div.listing.crateBox")[0])})),e.MissingAchievements&&Object.values(t.children).forEach((e=>{if(!e.className.includes("enabled")){const t=e.onclick.toString().split(/\[(.*)\]/gi)[1],{icon:o}=Game.AchievementsById[t];e.style.backgroundPosition=`${48*-o[0]}px ${48*-o[1]}px`,e.onmouseover=function(){Game.mouseDown||(Game.setOnCrate(this),Game.tooltip.dynamic=1,Game.tooltip.draw(this,(function(){return function(e){const t=[];"shadow"===e.pool?t.push("Shadow Achievement","#9700cf"):t.push("Achievement",0),t.push("Locked",0);let o=0;(Game.Has("Neuromancy")||Game.sesame&&"debug"===e.pool)&&(o=1),o&&0===e.won?t.push("Click to win!","#00c462"):o&&e.won>0&&t.push("Click to lose!","#00c462");let{icon:n}=e;e.iconFunction&&(n=e.iconFunction());let{desc:i}=e;e.descFunc&&(i=e.descFunc("stats"));let a="";for(let e=0;e<t.length;e+=2)e%2==0&&(a+=` <div class="tag" style="color:${0===t[e+1]?"#fff":t[e+1]};">[${t[e]}]</div>`);return a=a.substring(1),`<div style="padding:8px 4px;min-width:350px;opacity:0.5">\n  <div class="icon" style="float:left;margin-left:-8px;margin-top:-8px;background-position:${48*-n[0]}px ${48*-n[1]}px;"></div>\n  <div class="name">${e.name}</div>\n  ${a}<div class="line"></div><div class="description">${i}</div></div>\n  ${Game.sesame?`<div style="font-size:9px;">Id : ${e.id} | Order : ${Math.floor(e.order)}${e.tier?` | Tier : ${e.tier}`:""}</div>`:""}`}(Game.AchievementsById[t])}),"top"),Game.tooltip.wobble())}}}))}()}const qn={Favourite:"Favourite Settings",Calculation:"Calculation",Notation:"Notation",Colours:"Colours and colour coding",BarsDisplay:"Infobars and visual settings",Tooltip:"Tooltips",Statistics:"Statistics",Notification:"Notifications",Miscellaneous:"Miscellaneous"},Yn={NotificationGeneral:"General Notifications",NotificationGC:"Golden Cookie",NotificationFC:"Fortune Cookie",NotificationSea:"Season Special",NotificationGard:"Garden Tick",NotificationMagi:"Full Magic Bar",NotificationWrink:"Wrinkler",NotificationWrinkMax:"Maximum Wrinklers"};function Kn(t,o){const n=document.createElement("div");n.className="title",n.style.opacity="0.7",n.style.fontSize="17px",n.appendChild(document.createTextNode(`${o} `));const i=document.createElement("span");return i.style.cursor="pointer",i.style.display="inline-block",i.style.height="14px",i.style.width="14px",i.style.borderRadius="7px",i.style.textAlign="center",i.style.backgroundColor="#C0C0C0",i.style.color="black",i.style.fontSize="13px",i.style.verticalAlign="middle",i.textContent=e.Header[t]?"-":"+",i.onclick=function(){zn(t),Game.UpdateMenu()},n.appendChild(i),n}function Jn(t){const o=document.createElement("div");if(o.className="listing",1===e.FavouriteSettings&&o.appendChild(function(e){const t=document.createElement("a");return lo.includes(e)?(t.innerText="★",t.style.color="yellow"):t.innerText="☆",t.className="option",t.onclick=function(){!function(e){lo.includes(e)?lo=lo.filter((function(t){return t!==e})):lo.push(e)}(e),Fn(),Game.UpdateMenu()},t.onmouseover=function(){Game.tooltip.draw(this,escape(Ut.FavouriteSettingPlaceholder.innerHTML))},t.onmouseout=function(){Game.tooltip.hide()},t.appendChild(document.createTextNode(" ")),t}(t)),"bool"===vn[t].type){const n=document.createElement("a");vn[t].toggle&&0===e[t]?n.className="option off":n.className="option",n.id=Rn+t,n.onclick=function(){!function(t){e[t]+=1,e[t]===vn[t].label.length?(e[t]=0,vn[t].toggle&&(l(Rn+t).className="option off")):l(Rn+t).className="option",void 0!==vn[t].func&&vn[t].func(),Fn()}(t),Game.UpdateMenu()},n.textContent=vn[t].label[e[t]],o.appendChild(n);const i=document.createElement("label");return i.textContent=vn[t].desc,i.style.lineHeight="1.6",o.appendChild(i),o}if("vol"===vn[t].type){const n=document.createElement("div");n.className="sliderBox";const i=document.createElement("div");i.style.float="left",i.innerHTML=vn[t].desc,n.appendChild(i);const a=document.createElement("div");a.id=`slider${t}right`,a.style.float="right",a.innerHTML=`${e[t]}%`,n.appendChild(a);const r=document.createElement("input");r.className="slider",r.id=`slider${t}`,r.style.clear="both",r.type="range",r.min="0",r.max="100",r.step="1",r.value=e[t],r.oninput=function(){In(t),Game.UpdateMenu()},r.onchange=function(){In(t),Game.UpdateMenu()},n.appendChild(r),o.appendChild(n);const l=document.createElement("a");return l.className="option",l.onclick=function(){Bn(e[t.replace("Volume","SoundURL")],t.replace("Volume","Sound"),t,!0)},l.textContent="Test sound",o.appendChild(l),o}if("url"===vn[t].type){const n=document.createElement("span");n.className="option",n.textContent=`${vn[t].label} `,n.style.lineHeight="1.6",o.appendChild(n);const i=document.createElement("input");i.id=Rn+t,i.className="option",i.type="text",i.readOnly=!0,i.setAttribute("value",e[t]),i.style.width="300px",o.appendChild(i),o.appendChild(document.createTextNode(" "));const a=document.createElement("input");a.id=`${Rn+t}Prompt`,a.className="option",a.type="text",a.setAttribute("value",e[t]);const r=document.createElement("a");r.className="option",r.onclick=function(){!function(e,t){Game.promptWrapL.className="framed";const o=e;Game.promptL.innerHTML=`${o}<div class="optionBox"></div>`,Object.keys(t).forEach((e=>{const o=document.createElement("a");o.id=`promptOption${e}`,o.className="option",o.onclick=function(){PlaySound("snd/tick.mp3"),t[e][1]()},o.textContent=t[e][0],Game.promptL.children[1].appendChild(o)})),Game.promptAnchorL.style.display="block",Game.darkenL.style.display="block",Game.promptL.focus(),Game.promptOn=1,Game.UpdatePrompt()}(a.outerHTML,[["Save",function(){e[t]=l(`CMConfig${t}Prompt`).value,Fn(),Game.ClosePrompt(),Game.UpdateMenu()}],["Cancel",function(){Game.ClosePrompt()}]])},r.textContent="Edit",o.appendChild(r);const s=document.createElement("label");return s.textContent=vn[t].desc,s.style.lineHeight="1.6",o.appendChild(s),o}if("colour"===vn[t].type){const n=document.createElement("span");n.className="option";const i=document.createElement("input");i.id=t,i.style.width="65px",i.setAttribute("value",e[t]),n.appendChild(i),new Un(i,{hash:!0,position:"right",onInput:function(){e[this.targetElement.id]=this.toHEXString(),On(),Fn(),Game.UpdateMenu()}});const a=document.createElement("label");if(a.textContent=vn[t].desc,a.style.lineHeight="1.6",n.appendChild(a),t.includes("Flash")){const e=document.createElement("a");e.className="option",e.onclick=function(){Pn(3,t.replace("Colour",""),!0)},e.textContent="Test flash",n.appendChild(e)}return o.appendChild(n),$n().init(),o}if("numscale"===vn[t].type){const n=document.createElement("span");n.className="option",n.textContent=`${vn[t].label} `,n.style.lineHeight="1.6",o.appendChild(n);const i=document.createElement("input");i.id=Rn+t,i.className="option",i.type="number",i.value=e[t],i.min=vn[t].min,i.max=vn[t].max,i.oninput=function(){e[t]=this.value,Fn(),gn(),Game.UpdateMenu()},o.appendChild(i),o.appendChild(document.createTextNode(" "));const a=document.createElement("label");return a.textContent=vn[t].desc,a.style.lineHeight="1.6",o.appendChild(a),o}return o}function Zn(t){if(1===e.TimerBar&&0===e.TimerBarPos){const e=parseInt(l("CMTimerBar").style.height,10);Game.mouseY-=e,t(),Game.mouseY+=e}else t()}function ei(){M.Beautify=B,B=CMBeautify,M.CalculateGains=Game.CalculateGains,Game.CalculateGains=function(){M.CalculateGains(),se=1,S=Date.now(),P=Date.now()},M.tooltip={},M.tooltip.draw=Game.tooltip.draw,M.tooltip.drawMod=new Function(`return ${Game.tooltip.draw.toString().split("this").join("Game.tooltip")}`)(),Game.tooltip.draw=function(e,t,o){M.tooltip.drawMod(e,t,o)},M.tooltip.update=Game.tooltip.update,M.tooltip.updateMod=new Function(`return ${Game.tooltip.update.toString().split("this.").join("Game.tooltip.")}`)(),Game.tooltip.update=function(){M.tooltip.updateMod(),function(){if("store"===Game.tooltip.origin){let t=0;1===e.ToolWarnLucky&&1===e.ToolWarnPos&&null!==l("CMDispTooltipWarningParent")&&(t=l("CMDispTooltipWarningParent").clientHeight-4),Game.tooltip.tta.style.top=`${Math.min(parseInt(Game.tooltip.tta.style.top,10),l("game").clientHeight+l("topBar").clientHeight-Game.tooltip.tt.clientHeight-t-46)}px`}}()},M.UpdateWrinklers=Game.UpdateWrinklers,Game.UpdateWrinklers=function(){Zn(M.UpdateWrinklers)},M.UpdateSpecial=Game.UpdateSpecial,Game.UpdateSpecial=function(){Zn(M.UpdateSpecial)},l("bigCookie").removeEventListener("click",Game.ClickCookie,!1),l("bigCookie").addEventListener("click",(function(){Zn(Game.ClickCookie)}),!1),M.RebuildUpgrades=Game.RebuildUpgrades,Game.RebuildUpgrades=function(){M.RebuildUpgrades(),T=[],Object.keys(Game.UpgradesInStore).forEach((e=>{null!==l(`upgrade${e}`).onmouseover&&(T[e]=l(`upgrade${e}`).onmouseover,l(`upgrade${e}`).onmouseover=function(){Game.mouseDown||(Game.setOnCrate(this),Game.tooltip.dynamic=1,Game.tooltip.draw(this,(function(){return An("u",`${e}`)}),"store"),Game.tooltip.wobble())})})),Game.CalculateGains()},M.ClickProduct=Game.ClickProduct,Game.ClickProduct=function(t){(!e.BulkBuyBlock||Game.ObjectsById[t].bulkPrice<Game.cookies||-1===Game.buyMode)&&M.ClickProduct(t)},M.DescribeDragonAura=Game.DescribeDragonAura,Game.DescribeDragonAura=function(t){M.DescribeDragonAura(t),function(t){if(1===e.DragonAuraInfo){const[e,o]=function(e){Mo(),l("promptContent").children[0].innerHTML.includes("secondary")?ie=e:ne=e;let t=0;if(ne!==me||ie!==he)for(let e=Game.ObjectsById.length-1;e>-1;--e)if(Game.ObjectsById[e].amount>0){const o=Ct[Game.ObjectsById[e].name].name;Ct[o].amount-=1,K-=1,t=Ct[o].basePrice*Game.priceIncrease**Math.max(0,Ct[o].amount-1-Ct[o].free),t=Game.modifyBuildingPrice(Ct[o],t),t=Math.ceil(t);break}const o=ee;return Oo(),Ho(),o!==ee&&Oo(),[de-Game.cookiesPs,t]}(t),n=po(o/(e+Game.cookiesPs));let i;i=0===Game.cookiesPs?B(1/0):B(e/Game.cookiesPs*100),l("dragonAuraInfo").style.minHeight="60px",l("dragonAuraInfo").style.margin="8px",l("dragonAuraInfo").appendChild(document.createElement("div")).className="line";const a=document.createElement("div");a.style.minWidth="200px",a.style.textAlign="center",a.textContent=`Picking this aura will change CPS by ${B(e)} (${i}% of current CPS).`,l("dragonAuraInfo").appendChild(a);const r=document.createElement("div");r.style.minWidth="200px",r.style.textAlign="center",r.textContent=`It will take ${n} to recover the cost.`,l("dragonAuraInfo").appendChild(r)}}(t)},M.ToggleSpecialMenu=Game.ToggleSpecialMenu,Game.ToggleSpecialMenu=function(e){M.ToggleSpecialMenu(e),function(){if(null!==(l("specialPopup").className.match(/onScreen/)&&l("specialPopup").children[0].style.background.match(/dragon/)))for(let e=0;e<l("specialPopup").childNodes.length;e++)"optionBox"===l("specialPopup").childNodes[e].className&&(l("specialPopup").children[e].onmouseover=function(){xn(),Game.tooltip.dynamic=1,Game.tooltip.draw(l("specialPopup"),`<div style="min-width:200px;text-align:center;">${fe}</div>`,"this"),Game.tooltip.wobble()},l("specialPopup").children[e].onmouseout=function(){Game.tooltip.shouldHide=1})}()},M.UpdateMenu=Game.UpdateMenu,Game.UpdateMenu=function(){void 0!==$n().picker&&void 0!==$n().picker.owner||(M.UpdateMenu(),function(){const t=document.createElement("div");t.className="title","prefs"===Game.onMenu?(t.textContent="Cookie Monster Settings",function(t){const o=document.createDocumentFragment();o.appendChild(t),Object.keys(qn).forEach((t=>{if("Favourite"===t){if(0!==lo.length&&e.FavouriteSettings>0&&(o.appendChild(Kn(t,qn[t])),e.Header[t]))for(let e=0;e<lo.length;e++)o.appendChild(Jn(lo[e]))}else o.appendChild(Kn(t,qn[t])),e.Header[t]&&("Notification"===t?Object.keys(Yn).forEach((t=>{const n=Kn(t,Yn[t]);n.style.fontSize="15px",n.style.opacity="0.5",o.appendChild(n),e.Header[t]&&Object.keys(vn).forEach((e=>{vn[e].group===t&&o.appendChild(Jn(e))}))})):Object.keys(vn).forEach((e=>{vn[e].group===t&&o.appendChild(Jn(e))})))}));const n=document.createElement("div");n.className="listing";const i=document.createElement("a");i.className="option",i.onclick=function(){Ln(un)},i.textContent="Restore Default",n.appendChild(i),o.appendChild(n),l("menu").childNodes[2].insertBefore(o,l("menu").childNodes[2].childNodes[l("menu").childNodes[2].childNodes.length-1])}(t)):"stats"===Game.onMenu?e.Stats&&(t.textContent="Cookie Monster Statistics",Xn(t)):"log"===Game.onMenu&&(t.textContent="Cookie Monster ",function(t){const o=document.createElement("div");o.className="subsection";const n=document.createElement("span");if(n.style.cursor="pointer",n.style.display="inline-block",n.style.height="14px",n.style.width="14px",n.style.borderRadius="7px",n.style.textAlign="center",n.style.backgroundColor="#C0C0C0",n.style.color="black",n.style.fontSize="13px",n.style.verticalAlign="middle",n.textContent=e.Header.InfoTab?"-":"+",n.onclick=function(){zn("InfoTab"),Game.UpdateMenu()},t.appendChild(n),o.appendChild(t),e.Header.InfoTab){const e=document.createElement("div");e.innerHTML='<div class="listing">\n <a href="https://github.com/Aktanusa/CookieMonster" target="blank">Cookie Monster</a>\n offers a wide range of tools and statistics to enhance your game experience.\n It is not a cheat interface – although it does offer helpers for golden cookies and such, everything can be toggled off at will to only leave how much information you want.</br>\n Progess on new updates and all previous release notes can be found on the GitHub page linked above!</br>\n Please also report any bugs you may find over there!</br>\n </div>\n ',o.appendChild(e);const t=document.createElement("div");t.innerHTML="<div class=\"listing\">\nThese are the release notes for the latest update (v 2.031.7).</br>\n</br>\nThis update implements the following functions:</br>\n- Updated the ordering and description of all settings of Cookie Monster</br>\n- Colours of Flash notifications can now be set by the user</br>\n- You can now select a number of settings to be your 'favourite'. They will be displayed at the top of all other settings in a new section. This functionality is toggleable and you can find it in the miscellaneous section</br>\n- The statistics page now shows cookies earned by clicking for a selected time period</br>\n- The statistics page can now show missing normal achievements, this can be toggled in the settings</br>\n</br>\nThis update fixes the following bugs:</br>\n- Fixed some typo's</br>\n- No longer show the wrinkler pop buttons when wrinklers can't spawn</br>\n- Fixed incorrect calculation of plant maturity for relevant tooltips</br>\n- Fixed edit prompts not working in settings</br>\n- Fixed red and purple being switched around for PP colour coding</br>\n- Fixed incorrect notation of numbers in Dragon tooltip and interface</br>\n- Fixed incorrect calculation of changing Dragon aura cost</br>\n- Fixed incorrect calculation of changing Pantheon gods effect</br>\n- Fixed fluctuating PP values when god Cyclius was selected</br>\n- Fixed black flickering when Golden Cookies or buffs start</br>\n</div>\n",o.appendChild(t)}const i=l("menu").children[1];i.insertBefore(o,i.children[1])}(t))}())},M.sayTime=Game.sayTime,Rt=function(e,t){return Number.isNaN(e)||e<=0?M.sayTime(e,t):po(e/Game.fps,1)},M.Logic=Game.Logic,Game.Logic=function(){M.Logic();let t="Cookie Clicker";"fools"===Game.season&&(t="Cookie Baker"),ro=`${Game.OnAscend?"Ascending! ":""}${B(Game.cookies)} ${1===Game.cookies?"cookie":"cookies"} - ${t}`,function(){if(Game.OnAscend||0===e.Title)document.title=ro;else if(1===e.Title){let e,t,o,n=!1,i=!1;e=ct?ct.wrath?`[W${Math.ceil(ct.life/Game.fps)}]`:`[G${Math.ceil(ct.life/Game.fps)}]`:Game.Has("Golden switch [off]")?"[GS]":`[${Number(l("CMTimerBarGCMinBar").textContent)<0?"!":""}${Math.ceil((Game.shimmerTypes.golden.maxTime-Game.shimmerTypes.golden.time)/Game.fps)}]`,f&&(n=!0,t="[F]"),"christmas"===Game.season&&(i=!0,o=g?`[R${Math.ceil(dt.life/Game.fps)}]`:`[${Number(l("CMTimerBarRenMinBar").textContent)<0?"!":""}${Math.ceil((Game.shimmerTypes.reindeer.maxTime-Game.shimmerTypes.reindeer.time)/Game.fps)}]`);let a=ro;"["===a.charAt(0)&&(a=a.substring(a.lastIndexOf("]")+1)),document.title=`${e+(n?t:"")+(i?o:"")} ${a}`}else if(2===e.Title){let e="",t=!1;ct&&(t=!0,ct.wrath?e+=`[W${Math.ceil(ct.life/Game.fps)}]`:e+=`[G${Math.ceil(ct.life/Game.fps)}]`),f&&(t=!0,e+="[F]"),"christmas"===Game.season&&g&&(e+=`[R${Math.ceil(dt.life/Game.fps)}]`,t=!0),t&&(e+=" - ");let o="Cookie Clicker";"fools"===Game.season&&(o="Cookie Baker"),e+=o,document.title=e}}(),function(){const t=Math.max(0,Game.HowManyCookiesReset(Math.floor(Game.HowMuchPrestige(Game.cookiesReset+Game.cookiesEarned))+1)-(Game.cookiesEarned+Game.cookiesReset)),o=Game.sayTime((Date.now()-Game.startDate)/1e3*Game.fps,-1);let n="";n+=`You've been on this run for <b>${""===o?"not very long":o}</b>.<br>`,n+='<div class="line"></div>',Game.prestige>0&&(n+=`Your prestige level is currently <b>${B(Game.prestige)}</b>.<br>(CpS +${B(Game.prestige)}%)`,n+='<div class="line"></div>'),n+=X<1?"Ascending now would grant you no prestige.":X<2?"Ascending now would grant you<br><b>1 prestige level</b> (+1% CpS)<br>and <b>1 heavenly chip</b> to spend.":`Ascending now would grant you<br><b>${B(X)} prestige levels</b> (+${B(X)}% CpS)<br>and <b>${B(X)} heavenly chips</b> to spend.`,n+='<div class="line"></div>',n+=`You need <b>${B(t)} more cookies</b> for the next level.<br>`,n+=e.TooltipAscendButton?`<div class='line'></div>It takes ${pt} to reach the next level and you were making ${B(Ne,2)} chips on average in the last 5 seconds.<br>`:"",l("ascendTooltip").innerHTML=n}()}}function ti(){Mn(),ln(),pn(),r=Object.keys(Game.mods).length,ue=document.createElement("style"),ue.type="text/css",ue.id="CMCSS",document.head.appendChild(ue),function(){const e=document.createElement("div");e.id="CMBotBar",e.style.height="69px",e.style.width="100%",e.style.position="absolute",e.style.display="none",e.style.backgroundColor="#262224",e.style.backgroundImage="linear-gradient(to bottom, #4d4548, #000000)",e.style.borderTop="1px solid black",e.style.overflow="auto",e.style.textShadow="-1px 0 black, 0 1px black, 1px 0 black, 0 -1px black";const t=e.appendChild(document.createElement("table"));t.style.width="100%",t.style.textAlign="center",t.style.whiteSpace="nowrap";const o=t.appendChild(document.createElement("tbody")),n=function(e,t){const o=document.createElement("td");return o.style.textAlign="right",o.className=St+t,o.textContent=e,o},i=o.appendChild(document.createElement("tr"));i.style.fontWeight="bold",i.appendChild(n("CM 2.031.7",Nt)),o.appendChild(document.createElement("tr")).appendChild(n("Bonus Income",Bt)),o.appendChild(document.createElement("tr")).appendChild(n("Payback Period",Bt)),o.appendChild(document.createElement("tr")).appendChild(n("Time Left",Bt)),l("wrapper").appendChild(e),Object.keys(Game.Objects).forEach((e=>{fo(e)}))}(),function(){const e=document.createElement("div");e.id="CMTimerBar",e.style.position="absolute",e.style.display="none",e.style.height="0px",e.style.fontSize="10px",e.style.fontWeight="bold",e.style.backgroundColor="black";const t=go("CMTimerBarGC","Next Cookie",[{id:"CMTimerBarGCMinBar",color:Wt},{id:"CMTimerBarGCBar",color:Lt}]);e.appendChild(t);const o=go("CMTimerBarRen","Next Reindeer",[{id:"CMTimerBarRenMinBar",color:Wt},{id:"CMTimerBarRenBar",color:Ot}]);e.appendChild(o);const n=document.createElement("div");n.id="CMTimerBarBuffTimers",e.appendChild(n),l("wrapper").appendChild(e)}(),function(){const e=document.createElement("div");e.id="CMUpgradeBar",e.style.width="100%",e.style.backgroundColor="black",e.style.textAlign="center",e.style.fontWeight="bold",e.style.display="none",e.style.zIndex="21",e.onmouseout=function(){Game.tooltip.hide()};const t=document.createElement("div");t.appendChild(function(){const e=document.createElement("div");e.style.minWidth="330px",e.style.marginBottom="4px";const t=document.createElement("div");t.className="name",t.style.marginBottom="4px",t.textContent="Legend",e.appendChild(t);const o=function(e,t){const o=document.createElement("div");o.style.verticalAlign="middle";const n=document.createElement("span");return n.className=Pt+e,n.style.display="inline-block",n.style.height="10px",n.style.width="10px",n.style.marginRight="4px",o.appendChild(n),o.appendChild(document.createTextNode(t)),o};return e.appendChild(o(Bt,"Better than the best PP of a building option")),e.appendChild(o(Et,"Same as the best PP building option")),e.appendChild(o(Nt,"Within the top 10 of PP for buildings")),e.appendChild(o(Ot,"Within the top 20 of PP for buildings")),e.appendChild(o(Ft,"Within the top 30 of PP for buildings")),e.appendChild(o(Lt,"Outside of the top 30 of PP for buildings")),e.appendChild(o(Wt,"Negative or infinity PP")),e}()),e.onmouseover=function(){Game.tooltip.draw(this,escape(t.innerHTML),"store")};const o=function(e,t){const o=document.createElement("span");return o.id=e,o.className=St+t,o.style.width="14.28571428571429%",o.style.display="inline-block",o.textContent="0",o};e.appendChild(o("CMUpgradeBarBlue",Bt)),e.appendChild(o("CMUpgradeBarGreen",Et)),e.appendChild(o("CMUpgradeBarYellow",Nt)),e.appendChild(o("CMUpgradeBarOrange",Ot)),e.appendChild(o("CMUpgradeBarRed",Ft)),e.appendChild(o("CMUpgradeBarPurple",Lt)),e.appendChild(o("CMUpgradeBarGray",Wt)),l("upgrades").parentNode.insertBefore(e,l("upgrades").parentNode.childNodes[3])}(),function(){const e=document.createElement("div");e.id="CMFlashScreen",e.style.width="100%",e.style.height="100%",e.style.backgroundColor="white",e.style.display="none",e.style.zIndex="9999999999",e.style.position="absolute",l("wrapper").appendChild(e)}(),function(){const e=document.createElement("div");e.id="CMSectionHidButtons",e.style.textAlign="center";const t=document.createElement("a");t.className="option",t.onclick=function(){"flex"===l("upgrades").style.display?(l("upgrades").style.display="none",l("toggleUpgrades").style.display="none",l("techUpgrades").style.display="none",l("vaultUpgrades").style.display="none"):(l("upgrades").style.display="flex",0!==l("toggleUpgrades").children.length&&(l("toggleUpgrades").style.display="block"),0!==l("techUpgrades").children.length&&(l("techUpgrades").style.display="block"),0!==l("vaultUpgrades").children.length&&(l("vaultUpgrades").style.display="block"))},t.textContent="Hide/Show Upgrades",e.appendChild(t);const o=document.createElement("a");o.className="option",o.onclick=function(){"grid"===l("products").style.display?l("products").style.display="none":l("products").style.display="grid"},o.textContent="Hide/Show Buildings",e.appendChild(o),l("store").insertBefore(e,l("store").childNodes[2])}(),function(){const e=document.createElement("link");e.id="CMFavicon",e.rel="shortcut icon",e.href="https://orteil.dashnet.org/cookieclicker/favicon.ico",document.getElementsByTagName("head")[0].appendChild(e)}(),Object.keys(Ht).forEach((e=>{!function(e,t,o){const n=document.createElement("div");n.id=e;const i=document.createElement("div");i.style.minWidth=o,i.style.marginBottom="4px";const a=document.createElement("div");a.style.textAlign="left",a.textContent=t,i.appendChild(a),n.appendChild(i),Ut[e]=n}(Ht[e][0],Ht[e][1],Ht[e][2])})),function(){const e=document.createElement("a");e.id="PopAllNormalWrinklerButton",e.textContent="Pop All Normal",e.className="option",e.onclick=function(){Wn()},e.onmouseout=function(){Game.tooltip.shouldHide=1},e.onmouseover=function(){Game.tooltip.dynamic=1,Game.tooltip.draw(this,(function(){return An("wb","PopAll")}),"this"),Game.tooltip.wobble()},l("sectionLeftExtra").children[0].append(e);const t=document.createElement("a");t.id="PopFattestWrinklerButton",t.textContent="Pop Single Fattest",t.className="option",t.onclick=function(){null!==$e[1]&&(Game.wrinklers[$e[1]].hp=0)},t.onmouseout=function(){Game.tooltip.shouldHide=1},t.onmouseover=function(){Game.tooltip.dynamic=1,Game.tooltip.draw(this,(function(){return An("wb","PopFattest")}),"this"),Game.tooltip.wobble()},l("sectionLeftExtra").children[0].append(t)}(),l("products").style.display="grid",l("storeBulk").style.gridRow="1/1",l("upgrades").style.display="flex",l("upgrades").style["flex-wrap"]="wrap",Object.keys(Game.Objects).forEach((e=>{const t=Game.Objects[e];null!==l(`product${t.id}`).onmouseover&&(w[e]=l(`product${t.id}`).onmouseover,l(`product${t.id}`).onmouseover=function(){Game.tooltip.dynamic=1,Game.tooltip.draw(this,(function(){return An("b",`${e}`)}),"store"),Game.tooltip.wobble()})})),Game.canLumps()&&(s=l("lumps").onmouseover,l("lumps").onmouseover=function(){Game.tooltip.dynamic=1,Game.tooltip.draw(this,(function(){return An("s","Lump")}),"this"),Game.tooltip.wobble()}),h=Game.LoadMinigames,Game.LoadMinigames=function(){h(),Game.Objects.Farm.minigameLoaded&&(l("gardenTool-1").onmouseover=function(){Game.tooltip.dynamic=1,Game.tooltip.draw(this,(function(){return An("ha","HarvestAllButton")}),"this"),Game.tooltip.wobble()},Array.from(l("gardenPlot").children).forEach((e=>{const t=e.id.slice(-3);e.onmouseover=function(){Game.tooltip.dynamic=1,Game.tooltip.draw(this,(function(){return An("p",[`${t[0]}`,`${t[2]}`])}),"this"),Game.tooltip.wobble()}}))),jn(),function(){if(Game.Objects.Temple.minigameLoaded){for(let e=0;e<11;e+=1)l(`templeGod${e}`).onmouseover=function(){Game.tooltip.dynamic=1,Game.tooltip.draw(this,(function(){return An("pag",e)}),"this"),Game.tooltip.wobble()};for(let e=0;e<3;e+=1)l(`templeSlot${e}`).onmouseover=function(){Game.tooltip.dynamic=1,Game.tooltip.draw(this,(function(){return An("pas",[e,Game.Objects.Temple.minigame.slot[e]])}),"this"),Game.tooltip.wobble()}}}(),Hn()},Game.LoadMinigames(),l("backgroundLeftCanvas").onmouseover=function(){It=1},l("backgroundLeftCanvas").onmouseout=function(){It=0,Game.tooltip.hide(),Object.keys(Game.wrinklers).forEach((e=>{Vt[e]=0}))},ei(),Hn(),Game.CalculateGains(),Ln(),$t=Game.OnAscend,Game.prefs.popups?Game.Popup("Cookie Monster version 2.031.7 loaded!"):Game.Notify("Cookie Monster version 2.031.7 loaded!","","",1,1),Game.Win("Third-party")}const oi={init:function(){Sn=!0;let e=!0;Game.version!==Number("2.031")&&(e=confirm("Cookie Monster version 2.031.7 is meant for Game version 2.031. Loading a different version may cause errors. Do you still want to load Cookie Monster?")),e&&(ti(),Game.registerHook("draw",zo),Game.registerHook("logic",Nn),Game.registerHook("click",Vo),Sn=!1)},load:function(e){const t=JSON.parse(e);Mn(),void 0!==t.favouriteSettings&&(lo=t.favouriteSettings),Ln(t.settings),"2.031.7"!==t.version&&(Game.prefs.popups?Game.Popup("A new version of Cookie Monster has been loaded, check out the release notes in the info tab!"):Game.Notify("A new version of Cookie Monster has been loaded, check out the release notes in the info tab!","","",0,1))},save:wn};Game.registerMod("CookieMonster",oi)})()})();
