+++ src/Disp/Tooltips/TypesOfTooltips/Upgrade.js
@@ -3,11 +3,9 @@ import {
   CacheUpgrades,
 } from '../../../Cache/VariablesAndData';
 import { CMOptions } from '../../../Config/VariablesAndData';
-import {
-  Beautify,
-  FormatTime,
-  GetTimeColour,
-} from '../../BeautifyAndFormatting/BeautifyFormatting';
+import Beautify from '../../BeautifyAndFormatting/Beautify';
+import FormatTime from '../../BeautifyAndFormatting/FormatTime';
+import GetTimeColour from '../../BeautifyAndFormatting/GetTimeColour';
 import GetCPS from '../../HelperFunctions/GetCPS';
 import GetWrinkConfigBank from '../../HelperFunctions/GetWrinkConfigBank';
 import {
