+++ src/Disp/Tooltips/TypesOfTooltips/WrinklerButton.js
@@ -2,7 +2,7 @@ import {
   CacheWrinklersFattest,
   CacheWrinklersTotal,
 } from '../../../Cache/VariablesAndData';
-import { Beautify } from '../../BeautifyAndFormatting/BeautifyFormatting';
+import Beautify from '../../BeautifyAndFormatting/Beautify';
 import { TooltipName } from '../../VariablesAndData';
 import * as Create from '../CreateTooltip';
