+++ src/Main/ReplaceGameFunctions/ReplaceNative.js
@@ -1,9 +1,7 @@
 import jscolor from '@eastdesire/jscolor';
 import { CMOptions } from '../../Config/VariablesAndData';
-import {
-  Beautify as CMBeautify,
-  FormatTime,
-} from '../../Disp/BeautifyAndFormatting/BeautifyFormatting';
+import Beautify from '../../Disp/BeautifyAndFormatting/Beautify';
+import FormatTime from '../../Disp/BeautifyAndFormatting/FormatTime';
 import { AddAuraInfo, AddDragonLevelUpTooltip } from '../../Disp/Dragon/Dragon';
 import AddMenu from '../../Disp/MenuSections/AddMenus';
 import UpdateTitle from '../../Disp/TabTitle/TabTitle';
@@ -146,7 +144,7 @@ export default function ReplaceNative() {
     let title = 'Cookie Clicker';
     if (Game.season === 'fools') title = 'Cookie Baker';
     // eslint-disable-next-line no-unused-vars
-    Title = `${Game.OnAscend ? 'Ascending! ' : ''}${CMBeautify(Game.cookies)} ${
+    Title = `${Game.OnAscend ? 'Ascending! ' : ''}${Beautify(Game.cookies)} ${
       Game.cookies === 1 ? 'cookie' : 'cookies'
     } - ${title}`;
     UpdateTitle();
